% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_convert_mT_to_MHz.R
\name{a_convert_mT_to_MHz}
\alias{a_convert_mT_to_MHz}
\title{TODO}
\usage{
a_convert_mT_to_MHz(a, g)
}
\arguments{
\item{a}{TODO}

\item{g}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
