% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_MHz_convert_to_mT.R
\name{A_MHz_convert_to_mT}
\alias{A_MHz_convert_to_mT}
\title{TODO}
\usage{
A_MHz_convert_to_mT(A, g)
}
\arguments{
\item{A}{TODO}

\item{g}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
