% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aAiso_rearrng_fromQCHcomp.R
\name{aAiso_rearrng_fromQCHcomp}
\alias{aAiso_rearrng_fromQCHcomp}
\title{TODO}
\usage{
aAiso_rearrng_fromQCHcomp(path_to_ASC, data.col.names, nuclei.list.slct)
}
\arguments{
\item{path_to_ASC}{TODO}

\item{data.col.names}{TODO}

\item{nuclei.list.slct}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
