% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aAiso_rearrng_QCHcomp.R
\name{aAiso_rearrng_QCHcomp}
\alias{aAiso_rearrng_QCHcomp}
\title{Rearrangement of \eqn{A_{iso}}/\eqn{a_{iso}} from Quantum Chemical Computations
According to Proposed Molecular Structure/Symmetry}
\usage{
aAiso_rearrng_QCHcomp(path_to_ASC, data.col.names, nuclei.list.slct)
}
\arguments{
\item{path_to_ASC}{String/Character, pointing to path of ASCII file (\code{txt},\code{csv}...etc) with characteristic
\eqn{A_{iso}}/\eqn{a_{iso}} values (either in \code{cm-1} or other units) presented for each atom/nucleus
(with its corresponding \code{atomic number within the structure} as well as with characteristic
\code{isotopic number/value}), such an entire table can be copied from \strong{Gaussian} output
(after \code{'Isotropic Fermi Contact Couplings'} line) or can be constructed from \strong{ORCA} output,
example for such file (from \strong{Gaussian}):
\tabular{ccccc}{
 \strong{No_atom} \tab \strong{Atom_Nucleus} \tab \strong{MegaHertz} \tab \strong{Gauss} \tab \strong{1e-4_cm-1} \cr
 1 \tab N(14) \tab 0.00643 \tab 0.00229 \tab 0.00214 \cr
 17 \tab N(14) \tab 13.99707 \tab 4.9945 \tab 4.66892 \cr
 28 \tab H(1) \tab 16.34971 \tab 5.83398 \tab 5.45368 \cr
}}

\item{data.col.names}{String/Character vector containing names of all columns from QCH Computational output,
for the names see the example in \code{path_to_ASC}, they must contain atomic/structure number, isotopic value
with element label (nucleus characterization) and \eqn{A} in MHz as well as \eqn{a} in Gauss}

\item{nuclei.list.slct}{Values/Numeric list for rearrangement of selected atoms/nuclei according to symmetry,
e.g. like: \code{nuclei.list.slct <- list(3,c(21,22),c(20,23),c(24,25),c(27,26))}}
}
\value{
Data frame/Table of \eqn{A_{iso}}/\eqn{a_{iso}} mean values corresponding to nuclei group
  structure/symmetry
}
\description{
tbc
}
\examples{
\dontrun{
aAiso_rearrng_QCHcomp(
"./iso_values_Gaussian.txt",
c("No","Nucleus","au","Megahertz","Gauss","10^n4_cm^n1"),
list(3,c(21,22),c(20,23),c(24,25),c(27,26))
)
}


}
