% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bextrem_fromSpectr.R
\name{Bextrem_fromSpectr}
\alias{Bextrem_fromSpectr}
\title{TODO}
\usage{
Bextrem_fromSpectr(spectrum.data, B.reg.start, B.reg.end, extrem)
}
\arguments{
\item{spectrum.data}{TODO}

\item{B.reg.start}{TODO}

\item{B.reg.end}{TODO}

\item{extrem}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
