% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bextrem_fromSpectrum.R
\name{Bextrem_fromSpectrum}
\alias{Bextrem_fromSpectrum}
\title{TODO}
\usage{
Bextrem_fromSpectrum(
  spectrum.data,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  B.reg.start,
  B.reg.end,
  extrem
)
}
\arguments{
\item{spectrum.data}{TODO}

\item{B}{TODO}

\item{Intensity}{TODO}

\item{B.reg.start}{TODO}

\item{B.reg.end}{TODO}

\item{extrem}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
