% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bextreme_fromSpectrum.R
\name{Bextreme_fromSpectrum}
\alias{Bextreme_fromSpectrum}
\title{Find Extreme (\strong{Min}imum or \strong{Max}imum) within the EPR Spectrum (or its Integrated Form)}
\usage{
Bextreme_fromSpectrum(
  spectrum.data,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  B.reg.start,
  B.reg.end,
  extreme
)
}
\arguments{
\item{spectrum.data}{TODO}

\item{B}{TODO}

\item{Intensity}{TODO}

\item{B.reg.start}{TODO}

\item{B.reg.end}{TODO}

\item{extreme}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
