% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bextreme_Spec.R
\name{Bextreme_Spec}
\alias{Bextreme_Spec}
\title{Find Intensity Extremes within the EPR Spectrum (or its Integrated Form).}
\usage{
Bextreme_Spec(
  data.spectrum,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  Blim,
  extreme
)
}
\arguments{
\item{data.spectrum}{EPR spectrum data frame/table with magnetic flux density (in \code{mT} or \code{G}) column
and that of the derivative \code{dIepr_over_dB} or integrated e.g. \code{Intensity}.
\code{Index} column may be included as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{data.spectrum} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (\strong{default})
or \code{B = "B_G"} or \code{B = "B_G_Sim"} to include simulated EPR spectra as well}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(3495.4,3595.4)`}

\item{extreme}{Character/String with only a two values allowed: \code{"min"} or \code{"max"}}
}
\value{
Numeric value of magnetic flux density \eqn{B} corresponding to \code{minimal} or \code{maximal}
  \code{intensity} within the EPR spectra (or their integrated form)
}
\description{
Finding positions (\eqn{B} in \code{mT} or \code{G}) of intensity minimum and/or maximum
within selected region of the EPR spectrum
}
\examples{
\dontrun{
Bextreme_Spec(data.spectrum,Blim = c(349.54,359.54),extreme = 'maximum')
Bextreme_Spec(data.frame,"B_G",Intensity = "Integrated_Intensity",c(3495.4,3595.4),'minimum')
}


}
