% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_a_mT_2A.R
\name{convert_a_mT_2A}
\alias{convert_a_mT_2A}
\title{Convert Splitting Constants into Coupling Ones.}
\usage{
convert_a_mT_2A(a.mT, g.val = 2.0023193)
}
\arguments{
\item{a.mT}{Numeric value/vector of HFSCs in \code{mT} ('line distances' from EPR spectrum)}

\item{g.val}{Numeric value/vector, corresponding to actual \eqn{g}-factor (\code{unitless}).
\strong{Default:} \code{g.val = 2.00231930} (corresponding to free electron).}
}
\value{
Numeric value/vector corresponding to HFCCs (\eqn{A}) in \code{MHz}.
}
\description{
Converting hyperfine splitting constants (HFSCs, \eqn{a} values in \code{mT})
  into hyperfine coupling ones (HFCCs, \eqn{A} values in \code{MHz}).
}
\details{
Conversion performed according to the following relation:
  \deqn{A = (a\,g\,\mu_{\text{B}}\,10^{-3}) / h\,10^6}
  where \eqn{h} corresponds to Planck's constant and \eqn{\mu_{\text{B}}} to Bohr's magneton.
  Both constants are obtained by the \code{constans::syms$h} and \code{constants::syms$mub}, respectively,
  using the \href{https://r-quantities.github.io/constants/}{constants} package (see \code{\link[constants]{syms}}).
  The \eqn{10^{-3}\,/\,10^6} factor is introduced due to specific \eqn{[a] = \text{mT}} \eqn{\rightarrow}
  \eqn{[A] = \text{MHz}} conversion. The latter is suitable for the EPR simulations and/or ENDOR.
}
\examples{
# all A (HFCCs) in MHz
convert_a_mT_2A(a.mT = 0.5) # 5 G
#
convert_a_mT_2A(0.6,2.0059) # 6 G
#
convert_a_mT_2A(0.15,g.val = 2.00036) # 1.5 G


}
\seealso{
Other Conversions and Corrections:
\code{\link[=convert_A_MHz_2a]{convert_A_MHz_2a()}},
\code{\link[=convert_B]{convert_B()}},
\code{\link[=convert_time2var]{convert_time2var()}},
\code{\link[=correct_time_Exp_Specs]{correct_time_Exp_Specs()}}
}
\concept{Conversions and Corrections}
