% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_B.R
\name{convert_B}
\alias{convert_B}
\title{Conversion of Magnetic Flux Density According to Input-Output Units}
\usage{
convert_B(B.val, B.unit, B.2unit)
}
\arguments{
\item{B.val}{Numeric value/vector corresponding to input value(s) of magnetic flux density.}

\item{B.unit}{Character string referring to input magnetic flux density units.
Usually `T`, `mT` or `G` are used.}

\item{B.2unit}{Character string referring to input magnetic flux density units.
Usually `T`, `mT` or `G` are used.}
}
\value{
Numeric value or vector as a result of the B conversion.
}
\description{
Conversion of magnetic flux density/field (\emph{B}) values depending on the input and the required
  output units.
}
\examples{
## simple conversion
convert_B(B.val = 3500,B.unit = "G",B.2unit = "T")
#
## conversion of B.seq vector with Sweep Width = 100 G
## and central filed 3496 G
B.seq <- seq(3496-100/2,3496+100/2,length.out = 1024)
Bnew <- convert_B(B.seq,B.unit = "G",B.2unit = "mT")
head(as.matrix(Bnew),n = 20)


}
\seealso{
Other Conversions and Corrections: 
\code{\link{convert_A_MHz_2a}()},
\code{\link{convert_a_mT_2A}()},
\code{\link{convert_time2var}()},
\code{\link{correct_time_Exp_Specs}()}
}
\concept{Conversions and Corrections}
