% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time2var.R
\name{convert_time2var}
\alias{convert_time2var}
\title{Convert Time \eqn{t} into Variable Linearly Depending on \eqn{t}.}
\usage{
convert_time2var(
  time.vals,
  time.unit = "s",
  var0,
  var.switch = NULL,
  var.rate,
  var.rate.unit = "s^{-1}"
)
}
\arguments{
\item{time.vals}{Numeric value or vector corresponding to time (points) where the variable \code{var}
is changed.}

\item{time.unit}{Character string time unit defined by \code{s},\code{min} or \code{h}.
\strong{Default}: \code{time.unit = "s"}.}

\item{var0}{Numeric, the initial value (INCL. ALSO NEGATIVE SIGN, if required, e.g.
negative electrochemical potential) of the variable (\code{var}).}

\item{var.switch}{Numeric, the switching point \code{var} value, in case when a linear CYCLIC CHANGE
(or 'triangular ramp') of \code{var} upon time is applied (e.g. in cyclic voltammetry).
\strong{Default}: \code{var.switch = NULL} (in case there is no such cyclic change).}

\item{var.rate}{Numeric, corresponding to rate of linear \code{var} change (INCL. ALSO NEGATIVE SIGN,
if required, e.g. in the case of electrochemical reduction or sample cooling).}

\item{var.rate.unit}{Character string corresponding to \code{var.rate} unit defined
by following strings \code{"s^{-1}"} \eqn{\equiv \text{s}^{-1}},
\code{"min^{-1}"} \eqn{\equiv \text{min}^{-1}} or \code{"h^{-1}"} \eqn{\equiv \text{h}^{-1}}.
\strong{Default}: \code{var.rate.unit = "s^{-1}"}.}
}
\value{
Numeric value or vector of the variable such as electrochemical potential or temperature,
  linearly changing upon time.
}
\description{
Conversion of time (\eqn{t}) into variable (\eqn{var}) which is linearly changed upon time.
}
\details{
The linear time change of \eqn{var} can be expressed like
 \deqn{var = var0 + rate~ t}
 This is especially suitable for time conversion of EPR time series experiments
 (see e.g. \code{\link{readEPR_Exp_Specs_kin}}) simultaneously performed either during
 electrochemical/voltammetric or variable temperature experiment. When cyclic series experiment
 is performed (e.g. cyclic voltammetry), that \eqn{var} value depends on the switching one,
 like =>
 \deqn{var = var0 + rate~ t ~~ \text{for} ~~ t \leq t_{\text{switch}}}
 \deqn{var = var_{\text{switch}} - rate\, (t - t_{\text{switch}}) ~~ \text{for} ~~ t \geq t_{\text{switch}}}
 where the \eqn{t_{\text{switch}}} corresponding to \eqn{var_{\text{switch}}} are the quantities
 at the turning point( see also \code{var.switch} argument).
}
\examples{
## calculate potential after 30 s, starting from 200 mV
## into cathodic direction (reduction) by 5 mV s^{-1}
convert_time2var(30,var0 = 0.2,var.rate = - 0.005)
#
## heating sample after 5 min starting from 293 K
## by the temperature rate of 4 K min^{-1}
convert_time2var(5,
                 time.unit = "min",
                 var0 = 293,
                 var.rate = 4,
                 var.rate.unit = "min^{-1}")
#
## create/evaluate vector containing the applied
## cell potential (in V) from the simultaneously
## performed electrochemical oxidation experiment
## (e.g. cyclic voltammetry from -0.1V to 0.45V and back
## to -0.1V). Time series vector is labeled as "time_s".
time_s <- seq(0,360,by = 18)
E_V <- convert_time2var(time.vals = time_s,
                        var0 = -0.1,
                        var.switch = 0.45,
                        var.rate = 0.003)
## preview
as.matrix(E_V)



}
\seealso{
Other Conversions and Corrections: 
\code{\link{convert_A_MHz_2a}()},
\code{\link{convert_B}()},
\code{\link{convert_a_mT_2A}()},
\code{\link{correct_time_Exp_Specs}()}
}
\concept{Conversions and Corrections}
