% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_integ_EPR_Specs.R
\name{correct_integ_EPR_Specs}
\alias{correct_integ_EPR_Specs}
\title{Polynomial Baseline Correction for Single-Integrated EPR Spectra}
\usage{
correct_integ_EPR_Specs(
  data.spec.integ,
  B = "B_G",
  integ.single = "single_integ",
  Blim = NULL,
  BpeaKlim,
  poly.degree,
  double.integ = FALSE,
  output.vector = FALSE
)
}
\arguments{
\item{data.spec.integ}{Spectrum data frame/table with magnetic flux density (in \code{mT} or \code{G}) column
can be labeled as \code{B_mT} (or \code{B_G}) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column may be included as well, \code{sinlge integrated EPR} \strong{spectrum must be included}
as column.}

\item{B}{Character/String pointing to magnetic flux density \code{column} (in the original \code{data.spec.integ})
of EPR spectrum data frame \code{spectrum.data} either in \code{millitesla} or in \code{Gauss},
that is \code{B = "B_mT"} or \code{B = "B_G"} (\strong{default}).}

\item{integ.single}{Character/String pointing to single integral \code{column} of EPR spectrum data frame
\code{spectrum.spe.integ} (\strong{default}: \code{integ.single = "single_integ"}).}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(3495.4,3595.4)`. \strong{Default}: \code{Blim = NULL}
(corresponding to entire `B` range).}

\item{BpeaKlim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `BpeaKlim = c(3535.4,3555.4)`.}

\item{poly.degree}{Numeric, degree of polynomial function used to fit the baseline under the single integrated
curve of the original EPR spectrum.}

\item{double.integ}{Logical, whether to present (column in data frame) the double integral of \emph{dIepr_over_dB},
which is required for quantitative analysis, \strong{default}: \code{double.integ = FALSE}}

\item{output.vector}{Logical, whether to "vectorize" the result in order to use it (in a form of additional
data frame column) for quantitative analysis of a EPR spectral series (for each individual spectrum).
\strong{Default}: \code{output.vector = FALSE}.}
}
\value{
Data frame/table  (or vector/column) including the EPR spectral data (\emph{dIepr_over_dB} vs \eqn{B}) as well as its
  corresponding \code{single} (column \code{sIntegCorr}) and/or \code{double} (column \code{dIntegCorr})
  integral corrected against the baseline fit.
}
\description{
Evaluates/Plots integrated EPR spectrum incl. the plynomial baseline correction.
 Single-integrated spectrum (from original derivative form) can be obtained
 by \code{\link[pracma:trapz]{pracma::cumtrapz}} function
 taken into account \emph{B} and \emph{dIepr_over_dB}. Double Integral Calculation/Presentation
 may be also provided by this function, if \code{double.integ = T}
}
\examples{
\dontrun{
correct_integ_EPR_Specs(EPR_spectral_data_table,
                        B = "B_mT",
                        Blim = c(348.2,351.1),
                        BpeaKlim = c(349,350),
                        poly.degree = 3,
                        output.vector = TRUE)
}


}
