% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaBpp_fromSpectr.R
\name{DeltaBpp_fromSpectr}
\alias{DeltaBpp_fromSpectr}
\title{Calculation of Spectral Linewidth (\eqn{\Delta B_pp}) of the EPR Spectrum}
\usage{
DeltaBpp_fromSpectr(spectrum.data, B.reg.start, B.reg.end)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column can be included as well}

\item{B.reg.start}{Numeric, magnetic flux density in \code{mT} corresponding to \code{starting} border
of the \code{selected B region} (therefore abbreviation \code{.reg.})}

\item{B.reg.end}{Numeric, magnetic flux density in \code{mT} corresponding to \code{ending} border
of the \code{selected B region} (therefore abbreviation \code{.reg.})}
}
\value{
Difference (the absolute value) of the magnetic flux densities (\code{B}) corresponding to \code{minimum}
  and \code{maximum} of the derivative intensity (\code{dIepr_over_dB}) => \eqn{\Delta B_pp}
}
\description{
The function calculates the \eqn{\Delta B_pp} (peak-to-peak) linewidth
  of EPR spectrum. The difference corresponds to magnetic flux densities corresponding
  to minimum and maximum of the \code{dIepr_over_dB} in the selected \code{B} region.
}
\examples{
\dontrun{
DeltaBpp_fromSpectr(spectrum.data,320.221,328.331)
DeltaBpp_fromSpectr(spectrum.data,B.reg.start = 320.221,B.reg.end = 328.331)
}

}
