% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaBpp_fromSpectrum.R
\name{DeltaBpp_fromSpectrum}
\alias{DeltaBpp_fromSpectrum}
\title{Calculation of Spectral Linewidth (\eqn{\Delta B_{pp}}) of the EPR Spectrum}
\usage{
DeltaBpp_fromSpectrum(
  spectrum.data,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  B.reg.start,
  B.reg.end
)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table with magnetic flux density (in \code{mT} or \code{G}) column
must be labeled as \code{B_mT} (or \code{B_G}) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column may be included as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (default)
or \code{B = "B_G"}}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used, default: \code{Intesity = "dIepr_over_dB"}}

\item{B.reg.start}{Numeric, magnetic flux density in \code{mT} (\code{G}) corresponding to \code{starting} border
of the \code{selected \emph{B} region} (therefore abbreviation \code{.reg.})}

\item{B.reg.end}{Numeric, magnetic flux density in \code{mT} (\code{G}) corresponding to \code{ending} border
of the \code{selected \emph{B} region} (therefore abbreviation \code{.reg.})}
}
\value{
Difference (the absolute value) of the magnetic flux densities (\code{B}) corresponding to \code{minimum}
  and \code{maximum} of the derivative intensity (\code{dIepr_over_dB}) => \eqn{ \Delta B_{pp}}
}
\description{
The function calculates the \eqn{\Delta B_{pp}} (peak-to-peak) linewidth
  of EPR spectrum. The difference corresponds to magnetic flux densities corresponding
  to minimum and maximum of the \code{dIepr_over_dB} in the selected \code{\emph{B}} region.
}
\examples{
\dontrun{
DeltaBpp_fromSpectr(spectrum.data,320.221,328.331)
DeltaBpp_fromSpectr(spectrum.data,B = "B_G",Intensity = "dIepr_over_dB",3202.11,3283.31)
DeltaBpp_fromSpectr(spectrum.data,"B_mT",B.reg.start = 320.221,B.reg.end = 328.331)
}

}
