% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaBpp_Spec.R
\name{DeltaBpp_Spec}
\alias{DeltaBpp_Spec}
\title{Calculation of Spectral Linewidth (\eqn{\Delta B_{pp}}) of the EPR Spectrum}
\usage{
DeltaBpp_Spec(data.spectrum, B = "B_mT", Intensity = "dIepr_over_dB", Blim)
}
\arguments{
\item{data.spectrum}{Spectrum data frame/table with magnetic flux density (in \code{mT} or \code{G}) column
must be labeled as \code{B_mT} (or \code{B_G}) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column may be included as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{data.spectrum} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (\strong{default})
or \code{B = "B_G"} or \code{B = "B_G_Sim"} to include simulated EPR spectra as well}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(349.54,359.54)`}
}
\value{
NUmeric Value of \eqn{B} difference (the absolute value) corresponding to \code{minimum}
  and \code{maximum} of the derivative intensity (\code{dIepr_over_dB}) => \eqn{ \Delta B_{pp}}
}
\description{
The function calculates the \eqn{\Delta B_{pp}} (peak-to-peak) linewidth
of EPR spectrum. The difference corresponds to magnetic flux densities corresponding
to minimum and maximum of the \code{dIepr_over_dB} in the selected \code{\emph{B}} region.
}
\examples{
\dontrun{
DeltaBpp_Spec(data.spectrum,c(320.221,328.331))
DeltaBpp_Spec(data.spectrum,B = "B_G",Intensity = "dIepr_over_dB",c(3202.11,3283.31))
DeltaBpp_Spec(data.spectrum,"B_mT",Blim = c(320.221,328.331))
DeltaBpp_Spec(data.spectrum,"B_mT_Sim",c(320.221,328.331))
}

}
