% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaXpp_Spec.R
\name{DeltaXpp_Spec}
\alias{DeltaXpp_Spec}
\title{Calculation of Spectral Linewidth ( e.g. \eqn{\Delta B_{pp}}) of the EPR/ENDOR Spectrum}
\usage{
DeltaXpp_Spec(data.spectrum, x = "B_mT", Intensity = "dIepr_over_dB", xlim)
}
\arguments{
\item{data.spectrum}{EPR/ENDOR spectrum data frame/table with magnetic flux density \eqn{B} (in \code{mT} or \code{G})
or \eqn{g}-Value or \eqn{RF} (in \code{MHz}) column and that of the derivative \code{dIepr_over_dB}
or integrated \code{Intensity}. \code{Index} column may be included as well.}

\item{x}{Character/String pointing to \code{x}-axis/column quantity like magnetic flux density \eqn{B}, \eqn{g}-Value
or \eqn{RF} (radio frequency), \strong{default}: \code{x = "B_mT"}}

\item{Intensity}{CCharacter/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{xlim}{Numeric vector corresponding to border limits of the selected \eqn{x} region,
e.g. like `xlim = c(3495.4,3595.4)` (\eqn{B} in \code{G}) or `xlim = c(12.5,21.2)` (\eqn{RF} in \code{MHz})
or `xlim = c(2.004,2.001)` (\eqn{g} dimensionless)}
}
\value{
Numeric value of difference of \code{x}-axis quantity like \eqn{B},\eqn{g},\eqn{RF} (the absolute value)
  corresponding to \code{minimum} and \code{maximum} of the derivative intensity (\code{dIepr_over_dB})
  in EPR/ENDOR spectrum
}
\description{
The function calculates the \code{peak-to-peak} (distance between of the \eqn{x}-axis
projection of "min" and "max" derivative intensities) linewidth
of an EPR/ENDOR spectrum.
}
\examples{
\dontrun{
DeltaBpp_Spec(data.spectrum,c(320.221,328.331))
DeltaBpp_Spec(data.spectrum,B = "B_G",Intensity = "dIepr_over_dB",c(3202.11,3283.31))
DeltaBpp_Spec(data.spectrum,"RF_MHz",xlim = c(10,42))
DeltaBpp_Spec(data.spectrum,"B_mT_Sim",c(320.221,328.331))
}

}
