% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_molecule_by_rcdk.R
\name{draw_molecule_by_rcdk}
\alias{draw_molecule_by_rcdk}
\title{Draw Molecule by `rcdk` Defined by SMILES or SDF}
\usage{
draw_molecule_by_rcdk(
  molecule,
  type = "smiles",
  mol.label = NULL,
  mol.label.color = "black",
  mol.label.xy.posit = c(8.2, 1.2),
  sma = NULL,
  annotate = "off",
  style = "cow",
  abbr = "off",
  suppressh = TRUE,
  ...
)
}
\arguments{
\item{molecule}{Character string => either SMILES strings (in case
of \code{type = "smiles"}/\code{type = "SMILES"}) or file path strings
pointing to \code{".sdf"} files.}

\item{type}{Character string referred to original input molecular type =>
either \code{type = "smiles"}/\code{type = "SMILES"} or \code{type = "sdf"}/\code{type = "SDF"}.
\strong{Default}: \code{type = "smiles"}.}

\item{mol.label}{Character string pointing to name label of the molecule/compound, e.g. like
\code{mol.label = c("acetone")} or \code{mol.label = c("PBN")}. If \code{mol.label = NULL}
(\strong{default}) a character string "mol. structure viewer" with gray color
is shown.}

\item{mol.label.color}{Character string pointing to displayed font color of the chemical structure
label. \strong{Default}: \code{mol.label.color = "black"}.}

\item{mol.label.xy.posit}{Numeric vector matching the x,y-position of the \code{mol.label}
label within the image having the size of "(1,1,10,10)" =>
see \code{\link[graphics]{rasterImage}}. \strong{Default}: \code{mol.label.xy.posit = c(8.2,1,2)}
=> the label is positioned in the bottom right corner.}

\item{sma}{Character string allowing to highlight sub-structures using `SMARTS`
(\strong{SM}ILES \strong{AR}bitrary \strong{T}arget \strong{S}pecification) to highlight
the common substructures in a set of molecules, e.g. like \code{sma = "C=O"}.}

\item{annotate}{Character string, whether to display (\code{annotate = "number"}) or not display
(\code{annotate = "off"}) atomic numbers/indexes. \strong{Default}: \code{annotate = "off"}.}

\item{style}{Character string denoting the plotting style like =>
\tabular{rcl}{
----------------  \tab ----------------------------- \cr
\strong{Plotting Style} \tab \strong{Style Text String} \cr
----------------  \tab ----------------------------- \cr
`color on black`  \tab \code{style = "cob"} \cr
`color on white`  \tab \code{style = "cow"} \cr
`black on white`  \tab \code{style = "bow"} \cr
`white on black` \tab \code{style = "wob"} \cr
`neon on black`  \tab \code{style = "nob"} \cr
---------------- \tab ----------------------------- \cr
}
\strong{Default}: \code{style = "cow"}.}

\item{abbr}{Character string which controls how the structure is displayed. Following options
can be set => \code{abbr = "off"} (\strong{default}) pointing to present structure as is;
\code{abbr = "groups"} creating an abbreviation for `groups`; \code{abbr = "reagents"} creating
an abbreviation for `reagents` or \code{abbr = "on"} to abbreviate the both latter.
The \code{abbr = "groups"} WORKS ONLY IF \code{annotate = "of"}!}

\item{suppressh}{Logical, denoting whether to suppress displaying the hydrogen atoms.
The `smiles` or `sdf` STRUCTURE MUST CONTAIN \strong{H} ATOMS!
\strong{Default}: \code{supressh = TRUE}.}

\item{...}{additional options/arguments for \code{\link[rcdk]{get.depictor}}.}
}
\value{
Graphics/Figure object with molecular structure.
}
\description{
Drawing a molecular structure based on \pkg{rcdk} package and using `SMILES` or `SDF` input data.
}
\examples{
## draw N,N,N',N'-tetramethyl-p-phenylenediamine based
## on `smiles` code character with highlighting
## the "C(aromatic)--N" bond
draw_molecule_by_rcdk("CN(C)C1=C([H])C([H])=C(N(C)C)C([H])=C1[H]",
                      type = "smiles",
                      sma = "cN")
#
## draw N,N,N',N'-tetramethyl-p-phenylenediamine (TMPD) radical
## cation based on `smiles` code character, with hydrogen atoms
## and molecule name label = "TMPD^(+.)"
draw_molecule_by_rcdk("CN(C)[C+]1C([H])=C([H])[C.]([N](C)C)C([H])=C1[H]",
                      type = "smiles",
                      mol.label = expression(TMPD^+.),
                      mol.label.color = "blue",
                      suppressh = FALSE)
#
## draw N,N,N',N'-tetramethyl-p-phenylenediamine based
## on `sdf` file path ("TMPD.sdf") with "color on black"
## style + atom numbering
draw_molecule_by_rcdk(molecule = load_data_example("TMPD.sdf"),
                      type = "sdf",
                      annotate = "number",
                      style = "cob")


}
\seealso{
Other Visualizations and Graphics: 
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{present_EPR_Sim_Spec}()}
}
\concept{Visualizations and Graphics}
