% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_molecules_rcdk.R
\name{draw_molecules_rcdk}
\alias{draw_molecules_rcdk}
\title{Draw Molecule by `rcdk` Defined by SMILES}
\usage{
draw_molecules_rcdk(molecules, name = NULL, sma = NULL, ...)
}
\arguments{
\item{molecules}{tbc}

\item{name}{tbc}

\item{sma}{tbc}

\item{...}{tbc}
}
\value{
tbc
}
\description{
A short description...
}
\examples{
\dontrun{
tbc
tbc
}

}
