% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawMolecule_rcdk.R
\name{drawMolecule_rcdk}
\alias{drawMolecule_rcdk}
\title{Draw Molecule by `rcdk` Defined by SMILES}
\usage{
drawMolecule_rcdk(molecule, name = NULL, sma = NULL, ...)
}
\arguments{
\item{molecule}{tbc}

\item{name}{tbc}

\item{sma}{tbc}

\item{...}{tbc}
}
\value{
tbc
}
\description{
Draw Molecule by `rcdk` Defined by SMILES
}
\examples{
tbc

}
