% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawMolecules_rcdk.R
\name{drawMolecules_rcdk}
\alias{drawMolecules_rcdk}
\title{Draw Molecule by `rcdk` Defined by SMILES}
\usage{
drawMolecules_rcdk(molecules, name = NULL, sma = NULL, ...)
}
\arguments{
\item{molecules}{tbc}

\item{name}{tbc}

\item{sma}{tbc}

\item{...}{tbc}
}
\value{
tbc
}
\description{
Draw Molecule by `rcdk` Defined by SMILES
}
\examples{
\dontrun{
tbc
tbc
}

}
