% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_DeltaXpp_Spec.R
\name{eval_DeltaXpp_Spec}
\alias{eval_DeltaXpp_Spec}
\title{Calculation of Spectral Linewidth ( e.g. \eqn{\Delta B_{pp}}) of the EPR/ENDOR Spectrum}
\usage{
eval_DeltaXpp_Spec(
  data.spectr,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  xlim = NULL
)
}
\arguments{
\item{data.spectr}{EPR/ENDOR spectrum data frame object with magnetic flux density \emph{B} (in `mT` or `G`)
or \emph{g}-Value or \emph{RF} (in `MHz`) column/variable and that of the derivative \code{dIepr_over_dB}
\code{Intensity}. \code{Index} column may be included as well.}

\item{x}{Character string pointing to name of the \code{x}-axis/column/variable (in the original \code{data.spectr})
like magnetic flux density \eqn{B}, \eqn{g}-Value or \eqn{RF} (radio frequency), \strong{default}: \code{x = "B_mT"}.}

\item{Intensity}{Character string pointing to name of the \code{intensity column/variable} ((in the original \code{data.spectr}))
if other than \code{dIepr_over_dB} name/label is used (e.g. for simulated spectra).
\strong{Default}: \code{Intesity = "dIepr_over_dB"}.}

\item{xlim}{Numeric vector corresponding to border limits of the selected \emph{x} region,
e.g. like \code{xlim = c(3495.4,3595.4)} (\emph{B} in `G`) or \code{xlim = c(12.5,21.2)} (\emph{RF} in `MHz`)
or \code{xlim = c(2.004,2.001)} (\emph{g} dimensionless). \strong{Default}: \code{xlim = NULL} (corresponding
to entire `x` range).}
}
\value{
Numeric value of difference of \code{x}-axis quantity like \emph{B},\emph{g},\emph{RF} (the absolute value)
  corresponding to \code{minimum} and \code{maximum} of the derivative intensity (\code{dIepr_over_dB})
  in EPR/ENDOR spectrum.
}
\description{
Calculating the `peak-to-peak` (distance between of the \emph{x}-axis
  projection of "min" and "max" derivative intensities) linewidth of an EPR/ENDOR spectrum.
}
\examples{
\dontrun{
eval_DeltaXpp_Spec(data.spectr,
                   c(320.221,328.331))
#
eval_DeltaXpp_Spec(data.spectr,
                   B = "B_G",
                   Intensity = "dIepr_over_dB",
                   c(3202.11,3283.31))
#
eval_DeltaXpp_Spec(data.spectr,
                   "RF_MHz",
                   xlim = c(10,42))
#
eval_DeltaXpp_Spec(data.spectr,
                   "B_mT_Sim",
                   c(320.221,328.331))
}

}
\seealso{
Other Evaluations: 
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()}
}
\concept{Evaluations}
