% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_ECh_QNe_chronoamp.R
\name{eval_ECh_QNe_chronoamp}
\alias{eval_ECh_QNe_chronoamp}
\title{Transferred Charge and Number of Electrons from Chronoamperogram}
\usage{
eval_ECh_QNe_chronoamp(
  data.at,
  time = "time_s",
  time.unit = "s",
  tlim = NULL,
  Current = "I_A",
  Current.unit = "A",
  E = NULL,
  E.unit = NULL,
  ref.electrode = NULL,
  Ne.output = TRUE,
  separate.plots = FALSE
)
}
\arguments{
\item{data.at}{Data frame (table) object including required columns like \code{Time} (\eqn{t}),
\code{Current} (\eqn{I}). Even though an arbitrary column label can be used, the best option
is to use labels e.g. \code{time_s}, \code{I_uA} or \code{I_mA}. Optionally, column related
to potential (\eqn{E}) may be present as well for transferred charge (\eqn{Q}) or number
of electrons (\eqn{N_{\text{e}}}) \emph{vs} \eqn{E} visualization (see also arguments
\code{E}, \code{E.unit} and \code{ref.electrode}).}

\item{time}{Character string pointing to \code{time}-axis/column header in the original \code{data.at}.
\strong{Default}: \code{time = "time_s"} (time in seconds).}

\item{time.unit}{Character string pointing to \code{time}-quantity unit. There are following units available:
\code{time.unit = "s"} (\strong{default}), \code{time.unit = "ms"}, \code{time.unit = "us"} (microseconds),
\code{time.unit = "ns"} or \code{time.unit = "min"}.}

\item{tlim}{Numeric vector of the \code{time}-quantity lower and upper limit, e.g. \code{xlim = c(5,400)}
(time in seconds. \strong{Default}: \code{tlim = NULL} actually setting the entire \code{time} span
from the original dataset.}

\item{Current}{Character string indicating the \code{Current}(\eqn{I})-axis/column quantity name
in the original \code{data.at} object. \strong{Default}: \code{Current = "I_A"} (current in \eqn{\text{A}}).}

\item{Current.unit}{Character string pointing to \code{Current} quantity unit like \code{Current.unit = "uA"}
(microamps) \code{Current.unit = "A"} (\strong{default}), \code{Current.unit = "mA"}
and \code{Current.unit = "nA"}.}

\item{E}{Character string referring to \eqn{E}(potential) column name within the input \code{data.at}
dataset. \strong{Default}: \code{E = NULL} corresponding to situation, when one doesn't want to
visualize transferred charge (or number of electrons) \emph{vs} \eqn{E}.}

\item{E.unit}{Character string, setting the potential unit (see \code{E} argument), usually
\code{E.unit = "mV"} or \code{E.unit = "V"}. \strong{Default}: \code{E.unit = NULL} corresponding
to situation, when one doesn't want to visualize transferred charge (or number of electrons)
\emph{vs} \eqn{E}.}

\item{ref.electrode}{Character string corresponding to reference electrode label, e.g.
\code{ref.electrode = "Ag-quasiref"} or \code{ref.electrode = "Fc/Fc+"}. \strong{Default}: \code{ref.electrode = NULL}
(displayed potential is not related to any \code{ref.electrode}).}

\item{Ne.output}{Logical. Should be the number of transferred electrons (\code{Ne}) presented within
the plot ? \strong{Default}: \code{Ne.output = TRUE}.}

\item{separate.plots}{Logical. By \strong{default}, both relations: \eqn{Q(N_{\text{e}})} \emph{vs}
\eqn{t,E} (time or potential) are shown in one plot (\code{separate.plots = FALSE}). One can separate
\eqn{N_{\text{e}}} \emph{vs} \eqn{t,E} and \eqn{Q} \emph{vs} \eqn{t,E} into individual plots setting
up the \code{separate.plots = TRUE}.}
}
\value{
List containing the following elements, depending on \code{separate.plots}:
  \enumerate{
  \item If \code{separate.plots = FALSE}
    \describe{
    \item{df}{Original \code{data.at} data frame object with the following additional
    columns/variables: \code{Q_C} (charge in coulombs), \code{Q_mC} (charge in millicoulombs,
    if the maximal charge \eqn{\leq 0.099\,\text{C}}) and \code{Ne} (number of transferred
    electrons, if \code{Ne.output = TRUE}).}
    \item{plot}{Side-by-side plot object (list) of \eqn{N_{\text{e}}} \emph{vs} \eqn{t,E}
    as well as \eqn{Q} \emph{vs} \eqn{t,E}.}
    }

  \item If \code{separate.plots = TRUE}
    \describe{
    \item{df}{Original \code{data.at} data frame object with the following additional
    columns/variables: \code{Q_C} (charge in coulombs), \code{Q_mC} (charge in millicoulombs,
    if the maximal charge \eqn{\leq 0.099\,\text{C}}) and \code{Ne} (number of transferred
    electrons, if \code{Ne.output = TRUE}).}
    \item{plot.Ne}{Plot object (list) of \eqn{N_{\text{e}}} \emph{vs} \eqn{t,E}.}
    \item{plot.Q}{Plot object (list) of \eqn{Q} \emph{vs} \eqn{t,E}.}
    }
  }
}
\description{
Evaluating transferred charge and the corresponding number of electrons
  from \href{https://doi.org/10.1515/pac-2018-0109}{chronoamperogram} related to electrochemical
  experiment, performed simultaneously with the EPR time series measurement or independently of the latter.
  To acquire charge, the input \eqn{I} \emph{vs} \eqn{time} relation (coming from
  \code{data.at}) is integrated by the \code{\link[pracma:trapz]{pracma::cumtrapz}} function.
  \strong{Prior to integration a capacitive current correction must be done}, especially if it is relatively
  high in comparison to Faradaic one. Afterwards, the number of electrons is calculated
  by Faraday's law (see details). The output plot can be visualized either as \eqn{Q(N_{\text{e}})} \emph{vs}
  \eqn{t} (time) or as \eqn{Q(N_{\text{e}})} \emph{vs} \eqn{E} (potential, if available
  in the input \code{data.at}).
}
\details{
When quantitative EPR is carried out along with electrochemistry simultaneously,
  by this function one can easily compare the number of radicals with the number of transferred electrons.
  Number of radicals (\eqn{N_{\text{R}}}) are evaluated from quantitative measurements (see also
  \code{\link{quantify_EPR_Abs}}), whereas number of transferred electrons (\eqn{N_{\text{e}}}) is related
  to charge (\eqn{Q}), according to:
  \deqn{N_{\text{e}} = (Q\,N_{\text{A}})/F}
  where \eqn{N_{\text{A}}} stands for Avogadro's number and \eqn{F} for Faraday's constants.
  Both obtained by \code{constans::syms$na} and \code{constants::syms$f}, respectively,
  using the \href{https://r-quantities.github.io/constants/}{constants} package
  (see \code{\link[constants]{syms}}). If both numbers match (\eqn{N_{\text{R}} = N_{\text{e}}}),
  it reveals the presence of one-electron oxidation/reduction, while different numbers may point
  to a more complex mechanism (such as comproportionation, follow-up reactions, multiple electron transfer).
}
\examples{
## loading package built-in example file =>
## `.txt` file generated by the IVIUM potentiostat software
triarylamine.path.cv <-
load_data_example(file = "Triarylamine_ECh_CV_ivium.txt")
## the data frame contains following variables:
## time, desired potential, current and actual/applied
## potential
triarylamine.data.cv <-
  data.table::fread(file = triarylamine.path.cv,
    skip = 2,
    col.names = c("time_s",
                  "E_V_des", # desired potential
                  "I_A",
                  "E_V_app") # applied potential
  )
#
## simple chronoamperogram plot
plot_ECh_VoC_amperogram(data.vat = triarylamine.data.cv,
  x = "time_s",
  x.unit = "s",
  Current = "I_A",
  Current.unit = "A",
  ticks = "in"
 )
#
## transferred charge and number of electrons
## with default parameters
triarylamine.data.QNe <-
  eval_ECh_QNe_chronoamp(data.at = triarylamine.data.cv)
#
## data frame preview
triarylamine.data.QNe$df
#
## graphical representation
triarylamine.data.QNe$plot



}
\references{
\insertRef{Bard2022elcheM}{eprscope}

  \insertRef{BarekIUPAC2019}{eprscope}

  \insertRef{Neudeck1999}{eprscope}

  \insertRef{Rpracma2023}{eprscope}
}
\seealso{
Other EPR Spectroelectrochemistry: 
\code{\link{plot_ECh_VoC_amperogram}()}
}
\concept{EPR Spectroelectrochemistry}
