% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gFactors.R
\name{eval_gFactor_QCHcomp}
\alias{eval_gFactor_QCHcomp}
\title{Calculation of \eqn{g}-factor from Quantum Chemical Computational Output}
\usage{
eval_gFactor_QCHcomp(path_to_QCHoutput, mean = TRUE, origin = "gaussian")
}
\arguments{
\item{path_to_QCHoutput}{Character string corresponding to path of `Gaussian` or `ORCA` output text files
incl. all \eqn{g}-values. \code{\link[base]{file.path}} can be applied to get the full/relative path.}

\item{mean}{Logical, whether to calculate the \code{mean value/iso} from principal components,
\strong{default}: \code{mean = TRUE}, or save the entire vector with the all components.}

\item{origin}{Character string pointing to origin of DFT EPR calculation parameters <=> which
software package was used. Only two values are available => \code{"gaussian"} (\strong{default})
or \code{"orca"}.}
}
\value{
Numeric mean \eqn{g}-factor value from principal difference (from \eqn{g_e}) components
  calculated by QCH methods (e.g. by DFT) or numeric vector with principal \eqn{g}-components
  if \code{mean = FALSE}.
}
\description{
\eqn{g}-Values (3 principal ones) are presented in a form of differences from
  the \eqn{g_e} (\eqn{g} of the free electron). Therefore the function takes these values
  and calculates the entire \eqn{g}-values or parses the corresponding mean value
  from `Gaussian` or `ORCA` output.
}
\examples{
## built-in file and path
gauss.file.path <- load_data_example(file = "TMPDAradCatEPRa.inp.log.zip")
gauss.file <- unzip(gauss.file.path)
## g_iso-value calculation from Gaussian output file
eval_gFactor_QCHcomp(gauss.file)


}
\seealso{
Other Evaluations and Quantum Chemistry: 
\code{\link{rearrange_aAiso_QCHcomp}()},
\code{\link{rearrange_aAiso_QCHorgau}()}
}
\concept{Evaluations and Quantum Chemistry}
