% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gFactor_Spec.R
\name{eval_gFactor_Spec}
\alias{eval_gFactor_Spec}
\title{Calculation of \eqn{g}-factor ("Position") from the EPR Spectrum/Data}
\usage{
eval_gFactor_Spec(
  data.spectrum,
  nu.GHz,
  B.unit = "G",
  B = "B_G",
  Intensity = "dIepr_over_dB",
  Blim = NULL,
  iso = TRUE
)
}
\arguments{
\item{data.spectrum}{Spectrum data frame/table where the magnetic flux density (in \code{mT} or \code{G}) column
can be labeled as \code{Field} or \code{B_G} and that of the derivative intensity as \code{dIepr_over_dB}
or single integrated intensity like \code{Integrated_Intensity}, \code{index} column can be included as well.}

\item{nu.GHz}{Numeric, microwave frequency in \code{GHz}}

\item{B.unit}{description tbc}

\item{B}{Character string pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{data.spectrum} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (\strong{default})
or \code{B = "B_G"} or \code{B = "B_G_Sim"} to include simulated EPR spectra as well}

\item{Intensity}{Character string pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(3495.4,3595.4)`. \strong{Default}: \code{Blim = NULL} (corresponding
to entire `B` range).}

\item{iso}{Logical, whether to calculate the \eqn{g}-factor from the \eqn{B} value corresponding to
that between the \code{min.} and \code{max.} derivative intensities (\code{dIepr_over_dB}, that is \eqn{g_{iso}}
(this is the \strong{default}: \code{iso = TRUE}), or by finding the the \eqn{B} value corresponding
to \code{dIepr_over_dB = 0} (close/near zero, which is \code{iso = FALSE})}
}
\value{
Numeric \eqn{g_{iso}}-value ('iso' = 'isotropic') according to \eqn{(\nu h)/(\mu_{B} B)}
}
\description{
Calculation of g-value according to fundamental formula (\code{\link{eval_gFactor}}).
  \eqn{g}-related magnetic flux density (like \eqn{B_{iso}} or \eqn{B_{center}}) is directly taken
  from the EPR spectrum. If positive and negative derivative intensities of the spectral line are similar
  and their distance from the middle of the spectrum equals, the \eqn{B_{iso}} should be be considered,
  otherwise the \eqn{B_{center}} must be taken into account. In case of integrated EPR spectrum/data
  the \eqn{B_{max}} is used for the \eqn{g}-value evaluation.
}
\examples{
\dontrun{
eval_gFactor_Spec(data.spectrum,
                  9.82451,
                  B.unit = "mT",
                  "Field_mT",
                  Intensity = "dIepr_over_dB_Sim",
                  c(349.8841,351.112))
#
eval_gFactor_Spec(data.spectrum,
                  nu.GHz = 9.82451,
                  B.unit = "G",
                  B = "B_G",
                  Blim = c(3498.841,3511.12),
                  iso = FALSE)
#
eval_gFactor_Spec(data.spectrum,
                  9.91024,
                  B.unit = "G",
                  B = "B_G_Sim",
                  Intensity = "Integral_Intensity",
                  c(3499,3501))
}


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor_QCHcomp}()},
\code{\link{eval_interval_cnfd_tVec}()}
}
\concept{Evaluations}
