% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_integ_EPR_Spec.R
\name{eval_integ_EPR_Spec}
\alias{eval_integ_EPR_Spec}
\title{Integration of EPR Spectral Data for Quantitative Analysis}
\usage{
eval_integ_EPR_Spec(
  data.spectrum,
  B = "B_G",
  B.unit = "G",
  Intensity = "dIepr_over_dB",
  Blim = NULL,
  correct.integ = FALSE,
  BpeaKlim = NULL,
  poly.degree = NULL,
  double.integ = FALSE,
  output.vecs = FALSE
)
}
\arguments{
\item{data.spectrum}{Spectrum data frame/table with magnetic flux density (in \code{mT}
or \code{G})) and that of the derivative or already single integrated intensity.
\code{Index} column may be already present as well.}

\item{B}{Character/String pointing to magnetic flux density \code{column} (in the original
\code{data.spectrum}) either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"}
or \code{B = "BField"}...etc or \code{B = "B_G"} (\strong{default}).}

\item{B.unit}{Character/String pointing to unit of magnetic flux density (coming from original data) which
is to be presented on \eqn{B} abscissa of the EPR spectrum,
like \code{"G"} (`Gauss`) or \code{"mT"} (`millitesla`), \strong{default}: \code{B.unit = "mT"}.}

\item{Intensity}{Character/String pointing to \code{column} of either derivative
(e.g. \code{Intensity = "dIepr_over_dB"}, \strong{default}) or single integrated EPR
spectrum (e.g. \code{Intensity = "single_Integrated"}) within the actual data frame \code{data.spectrum}.}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(3495.4,3595.4)`. \strong{Default}: \code{Blim = NULL}
(corresponding to entire `B` range).}

\item{correct.integ}{Logical, whether to correct the integral by baseline model fit.
\strong{Default}: \code{correct.integ = FALSE}.}

\item{BpeaKlim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `BpeaKlim = c(3535.4,3555.4)`.}

\item{poly.degree}{Numeric, degree of polynomial function used to fit the baseline under the single integrated
curve of the original EPR spectrum.}

\item{double.integ}{Logical, whether to present (column in data frame) the double integral of \emph{dIepr_over_dB},
which is required for quantitative analysis, \strong{default}: \code{double.integ = FALSE}.}

\item{output.vecs}{Logical, whether the `integral` \code{columns} are presented within the entire
data frame (\code{output.vecs = FALSE}, \strong{default}) or called as a vectors or list for
additional processing by \pkg{dplyr}.}
}
\value{
The integration results may be divided into following types depending on the above-described
  parameters/arguments. Generally, they are either data frames incl. the original data and the integrals
  (\code{output.vecs = FALSE}) or vectors corresponding to individual baseline corrected/uncorrected
  integrals (\code{output.vecs = TRUE}). This is especially useful for spectral (time) series EPR data,
  which can be handily processed \code{\link[dplyr]{group_by}} using `pipe` operators (\code{\link[magrittr]{\%>\%}}).
  \itemize{
  \item Data frame/table including the EPR spectral data (general \emph{Intensity}
  (integrated or derivative) \emph{vs} \eqn{B}) as well as its corresponding \code{single}
  (\strong{column} \code{single_Integ}) integral. This is the case if only a single uncorrected integral
  is required.
  \item Data frame/table with single integral/intensity already corrected by a certain degree
  of polynomial baseline (fitted to experimental baseline without peak). Single integrals are referred either
  to derivative or already single integrated EPR spectra where corrected integral \code{column}
  id denoted as \code{single_Integ_correct}. This is the case for \code{correct.integ = TRUE}
  and \code{double.integ = FALSE} + \code{output.vecs = FALSE}.
  \item The above-described data frame/table with \code{single} and \code{double} integral \strong{column}
  (\code{double_Integ}) essential for quantitative analysis. This is the case for \code{output.vecs = FALSE}
  and \code{correct.integ = FALSE}.
  \item Data frame/table in case of \code{correct.integ = TRUE}, \code{double.integ = TRUE}
  and \code{output.vecs = FALSE}. It contains the original data (as in all previous cases) + corrected
  single integral (\code{single_Integ_correct}) and double integral (\code{double_Integ}) which
  is evaluated from the baseline corrected single one. Therefor such double integral is suitable for
  the accurate determination of radical (paramagnetic centers) amount.
  \item Numeric vector corresponding to baseline uncorrected (or corrected) single integral
  in case of \code{double.integ = FALSE} + \code{output.vecs = TRUE}.
  \item List of numeric vectors corresponding to  `single` (corrected or uncorrected denpending
  on \code{correct.integ} parameter/argument) and `double` integrals (\code{double.integ = TRUE}).
  In such case the integrals are called by \code{list} names like \code{...[["single"]]}
  (or \code{...$single}), corresponding to single integral (or integrated form of EPR spectrum),
  or by \code{...[["double"]]} (or \code{...$double}) corresponding to doubly integrated
  (or single integrated EPR spectra originally presented already in single integrals) EPR spectral data.
  }
}
\description{
Evaluates integrals of EPR spectra depending on input data (corresponding to either derivative or single integrated
 EPR signal form) with option to correct the single integral baseline by the polynomial of \code{poly.degree}
 level. Integration is done by \code{\link[pracma:trapz]{pracma::cumtrapz}} function. For the purpose
 of quantitative analysis the integrals are evaluated using the \code{B.units = "G"} (see below).
 Therefore, depending on \eqn{B} unit (either \code{"G"} or \code{"mT"}) the resulting integral data
 have to be optionally (in case of \code{"mT"}) multiplied by factor of \code{10} because
 \eqn{1 \text{mT}\equiv 10 \text{G}}. Such correction is already included in the `R` function/script.
}
\examples{
\dontrun{
## Evaluation of single corrected (by polynomial degree = 3) as well as double
## integrals corresponding to original derivative EPR spectrum.
## Integrals are presented as a columns within the en tire data frame
eval_integ_EPR_Spec(EPR_spectral_data_table,
                    B = "FieldB",
                    B.unit = "mT"
                    Blim = c(348.2,351.1),
                    correct.integ = TRUE,
                    BpeaKlim = c(349,350),
                    poly.degree = 3,
                    double.integ = TRUE)

## Double integral evaluation corresponding to original non-corrected
## integral/intensity (non-derivative EPR signal). Double Integral
## (which is actually single integral of the original form, however
## possesses the sigmoidal curve form) is exported into vectors list
eval_integ_EPR_Spec(EPR_spectral_data_table,
                    B = "B_G",
                    B.unit = "G",
                    double.integ = TRUE,
                    output.vecs = TRUE)

}


}
