% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_ODE_model.R
\name{eval_kinR_ODE_model}
\alias{eval_kinR_ODE_model}
\title{Quantitative EPR Kinetic Model Profiles by Numeric Solution of the ODE.}
\usage{
eval_kinR_ODE_model(
  model.react = "(r=1)R --> [k1] B",
  model.expr.diff = FALSE,
  elementary.react = TRUE,
  kin.params = c(k1 = 0.001, qvar0R = 0.02),
  time.unit = "s",
  time.interval.model = c(0, 1800),
  time.frame.model = 2,
  solve.ode.method = "lsoda",
  data.qt.expr = NULL,
  time.expr = NULL,
  qvar.expr = NULL,
  ...
)
}
\arguments{
\item{model.react}{Character string, denoting a specific radical (\code{"R"}) reaction related to
changes in integral intensities (or any other \strong{q}uantitative \strong{var}iable) in EPR spectral
time series. Arrow shows direction of the reaction (\code{"-->", forward} or \code{"<==>", forward + reverse}).
Rate constants are indicated by square brackets after the arrows. Following examples of the reaction schemes
are predefined and commonly used to describe the integral intensity and/or radical concentration/amount
changes during the EPR time series experiment (the \code{r,a,b} stoichiometric coefficients
may vary, see below).
\tabular{ll}{
\strong{Reaction Scheme} \tab \strong{model.react} \cr
\eqn{(r=1)\text{R} \xrightarrow{k_1} \text{B}} \tab \code{"(r=1)R --> [k1] B"} \cr
 \eqn{(a=2)\text{A} \xrightarrow{k_1} (r=2)\text{R}} \tab \code{"(a=2)A --> [k1] (r=2)R"} \cr
 \eqn{(a=2)\text{A} \xrightleftharpoons[k_4]{k_1} (r=2)\text{R}
 \xrightleftharpoons[k_3]{k_2} (b=1)\text{B}} \tab
 \code{"(a=2)A <==> [k1] [k4] (r=2)R <==> [k2] [k3] (b=1)B"} \cr
 \eqn{(r=1)\text{R} \xrightleftharpoons[k_2]{k_1} (b=1)\text{B}} \tab
 \code{"(r=1)R <==> [k1] [k2] (b=1)B"} \cr
 \eqn{(a=2)\text{A} \xrightleftharpoons[k_2]{k_1} (r=2)\text{R}} \tab
 \code{"(a=2)A <==> [k1] [k2] (r=2)R"} \cr
 \eqn{(a=1)\text{A} + (b=1)\text{B} \xrightarrow{k_1} (r=1)\text{R}} \tab
 \code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \cr
 \eqn{(a=1)\text{A} + (r=1)\text{R} \xrightarrow{k_1} \text{B}} \tab
 \code{"(a=1)A + (r=1)R --> [k1] B"} \cr
}
Couple of examples are also given in the \code{Details}. The function is relatively flexible and enables
later addition of any other reaction schemes describing the EPR time series experiments
(YOU MAY ASK DEVELOPER(S) via forum/help-channels). The stoichiometric coefficient (such as \code{(r=1)}
or \code{(a=1)}) can be varied within the \code{model.react} character string.
Defined/Allowed values are integers, e.g. 1,2,3...etc. The space character within the \code{model.react}
string is not fixed and can be skipped for the sake of simplicity.
If \code{elementary.react = FALSE} (the model reaction is not considered as an elementary one),
a possible non-integer partial coefficients (e.g. \code{alpha},\code{beta} or \code{gamma})
must be included in the \code{kin.params} (see also \code{kin.params} description). For the consecutive model reaction
presented above, it applies only to one part/step of the mechanism.}

\item{model.expr.diff}{Logical, difference between the integral intensities/areas under the EPR spectra calculated
using the experimental data and those generated by the model. By \strong{default} the argument
is \strong{FALSE} and it is ONLY ACTIVATED (\code{model.expr.diff = TRUE}) IN THE CASE WHEN THE KINETIC MODEL
FITTING PROCEDURE (see also \code{\link{eval_kinR_EPR_modelFit}} or examples below) IS PERFORMED.}

\item{elementary.react}{Logical, if the model reaction should be considered as an elementary one,
i.e. the stoichiometric coefficients equal to the partial reaction orders. Such reaction proceeds without
identifiable intermediate species forming. \strong{Default}: \code{elementary.react = TRUE}.
If \code{elementary.react = FALSE}, i.e. the \code{model.react} cannot be considered like an elementary one,
the user must include the parameterized reaction orders \eqn{\alpha}, \eqn{\beta} or \eqn{\gamma} in
the \code{kin.params}, e.g \code{kin.params = c(k1 = 0.01, qvar0A = 0.05, alpha = 1.5)}. For the consecutive
model reaction presented above, it applies only to one part/step of the mechanism.}

\item{kin.params}{Named numeric vector, containing rate constants as well as initial radical
or other reactant/product concentration/integral intensities/areas...etc. Therefore, a general
\code{qvar} (\strong{q}uantitative \strong{var}iable) is defined which may actually reflect
all the above-mentioned quantities. \strong{Default}: \code{kin.params = c(k1 = 0.001,qvar0R = 0.02)}.
The initial values are denoted as \code{qvar0X} (e.g. qvar0R for radical or qvar0A for the reactant \code{A}).
The components of \code{kin.params} depend on \code{model.react} as well as on the \code{elementary.react}.
If \code{elementary.react = FALSE} additional parameters like partial reaction orders (\code{alpha}
and/or \code{beta} and/or \code{gamma}) must be defined within the \code{kin.params} (see the examples below
in the following table):
\tabular{ll}{
\strong{model.react} \tab \strong{Essential kin.params components} \cr
\code{"(r=1)R --> [k1] B"} \tab \code{k1}, \code{qvar0R}, (\code{alpha}) \cr
\code{"(a=1)A --> [k1] (r=1)R"} \tab \code{k1}, \code{qvar0A}, \code{qvar0R}, (\code{alpha}) \cr
\code{"(a=1)A <==> [k1] [k4] (r=1)R <==> [k2] [k3] (b=1)B"} \tab \code{k1}, \code{k2},
\code{k3}, \code{k4}, \code{qvar0A}, \code{qvar0R}, \code{qvar0B}, (\code{alpha},
\code{beta}, \code{gamma}) \cr
\code{"(r=1)R <==> [k1] [k2] (b=1)B"} \tab \code{k1}, \code{k2}, \code{qvar0R},
\code{qvar0B}, (\code{alpha}, \code{beta}) \cr
\code{"(a=1)A <==> [k1] [k2] (r=1)R"} \tab \code{k1}, \code{k2}, \code{qvar0A},
\code{qvar0R}, (\code{alpha}, \code{beta}) \cr
\code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \tab \code{k1}, \code{qvar0A}, \code{qvar0B},
\code{qvar0R}, (\code{alpha},\code{beta}) \cr
\code{"(a=1)A + (r=1)R --> [k1] B"} \tab \code{k1}, \code{qvar0A},
\code{qvar0R}, (\code{alpha},\code{beta}) \cr
}
The \code{k1}-unit is eventually expressed in terms of \eqn{s^{-1}} as well as in units of the applied \code{qvar}
(e.g. \code{c}, concentration) and depends on the partial reaction order(s) (PRO, see above),
which power(s) the \code{qvar(s)}. For example, the \code{k1}-unit of elementary radical recombination,
evaluated by double integrals, like \code{model.react = "(r=2)R --> [k1] B"},
reads: \eqn{\text{s}^{-1}\,(\text{p.d.u.})^{-1}},
where \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6803776/}{p.d.u.} stands for the procedure defined units.}

\item{time.unit}{Character string, corresponding to time unit like \code{"s"} (\strong{default}),
\code{"min"} (minutes), \code{"h"} (hours) or \code{"ms"} (milliseconds).}

\item{time.interval.model}{Numeric vector, including two values: starting and final time/termination
of the model reaction in \code{time.unit} (e.g. \code{c(0,1800)} in seconds, \strong{default}).}

\item{time.frame.model}{Numeric value, corresponding to interval time resolution, i.e. the smallest time difference
between two consecutive points in \code{time.unit}. Therefore, the number of points is defined
by the \code{time.interval.model} argument:
\deqn{((Interval[2] - Interval[1])\,/\,Frame) + 1}
where the \eqn{Interval[2]}, \eqn{Interval[1]} equal to the 2nd and the 1st element of the \code{time.interval.model},
respectively; and the \eqn{Frame} corresponds to \code{time.frame.model}. This argument is required to numerically
solve the kinetic differential equations by the \code{\link[deSolve]{ode}}. For the default interval mentioned
above, the \strong{default} value reads \code{time.frame.model = 2} (in \code{time.unit}).}

\item{solve.ode.method}{Character string, setting up the integrator (the \code{method} argument in \code{\link[deSolve]{ode}}),
applied to find the numeric solution of ODE. \strong{Default}: \code{solve.ode.method = "lsoda"}
(\code{\link[deSolve]{lsoda}}, additional methods (e.g. \code{"rk4"}, \code{"daspk"}, \code{"euler"}),
see link for the \code{ode} function above).}

\item{data.qt.expr}{A data frame object, containing the concentrations/integral intensities/areas under
the EPR spectra (calculated using the \strong{experimental data}) as well as time column. These two essential
column headers are described by the character strings like those below \code{time.expr} and \code{qvar.expr}.
The \code{data.qt.expr} MUST BE USED ONLY IN SUCH CASE WHEN THE EXPERIMENTAL TIME HAS TO BE INCLUDED
IN THE KINETIC MODEL (e.g. also for THE FITTING of EXPERIMENTAL DATA BY THE KINETIC MODEL).
\strong{Default}: \code{data.qt.expr = NULL}.}

\item{time.expr}{Character string, pointing to \code{time} column/variable name in the original
\code{data.qt.expr} data frame. \strong{Default}: \code{time.expr = NULL} (when the experimental
data aren't taken into account). If the time has to be corrected (e.g. in the case of double integrals),
please use \code{\link{correct_time_Exp_Specs}} function prior to kinetic evaluation.}

\item{qvar.expr}{Character string, pointing to \code{qvar} column/variable name in the original
\code{data.qt.expr} data frame. \strong{Default}: \code{qvar.expr = NULL} (when the experimental
data aren't taken into account).}

\item{...}{additional arguments passed to the ODE (see also \code{\link[deSolve]{ode}}).}
}
\value{
If the function \strong{is not used for fitting} of the experimental and processed data,
  the result is \code{list} consisting of:
  \describe{
  \item{df}{Data frame, containing \code{time} column and \code{qvar}, quantitative variable,
  columns corresponding to quantities of different relevant species
  denoted as \code{"R"}, \code{"A"}, \code{"B"} + if \code{data.qt.expr} is NOT NULL
  additional experimental quantitative variable is present.}
  \item{plot}{Plot object, containing \code{time} as \eqn{x}-axis and \code{qvar}
  (see \code{df} above) as \eqn{y}-axis + if \code{data.qt.expr} is NOT NULL the experimental
  quantitative variable is presented as well.}
  }
  Applying function \strong{for the fitting} procedure
  requires \code{model.expr.diff = TRUE} and therefore the result is represented by the difference between
  the integral intensities/areas, calculated using the experimental data
  and those generated by the kinetic model.
}
\description{
Theoretical quantitative kinetic profiles (such as concentration/amount/integral intensity)
  as well as comparison with the experimental data for various predefined model reactions involving
  radical(s) (labeled as "R"). Profiles are evaluated by numeric solution of the rate equations,
  using the \strong{O}rdinary \strong{D}ifferential \strong{E}quations (ODE from
  the \href{https://github.com/tpetzoldt/deSolve/}{desolve R package}).
  This function is inspired by the
  \href{https://www.r-bloggers.com/2013/06/learning-r-parameter-fitting-for-models-involving-differential-equations/}{R-bloggers
  article}.
}
\details{
According to IUPAC (2019), see also the \code{References}, the rate of a chemical reaction
  with the radicals (\eqn{\text{R}}) involved (refer to the example in \code{model.react} argument)
  \deqn{a\text{A} + r\text{R} \xrightarrow\, b\text{B}}
  is expressed via the time change of extent of the reaction (\eqn{\text{d}\xi/\text{d}t}):
  \deqn{-(1/r)\,(\text{d}n_{\text{R}}/\text{d}t) = -(1/a)\,(\text{d}n_{\text{A}}/\text{d}t) =
  (1/b)\,(\text{d}n_{\text{B}}/\text{d}t)}
  where \eqn{a,r,b} are the stoichiometric coefficients. At constant volume (\eqn{V}) conditions
  (or if volume changes are negligible) the amount (\eqn{n} in mole) of reactant/product
  can be replaced by its corresponding concentration (\eqn{c = n/V}). Such reaction rate
  (expressed in moles per unit volume and per second) is function of temperature (\eqn{T}),
  pressure (\eqn{p}) as well as that of concentration of reactants/products.
  For the reaction example shown above it applies (for radical \eqn{\text{R}}):
  \deqn{\text{d}c_{\text{R}}/\text{d}t = - r\,k(T,p)\,c_{\text{A}}^{\alpha}\,c_{\text{R}}^{\beta}}
  where \eqn{k} is the rate constant and its pressure dependence is usually
  small and therefore can be ignored, in the first approach. The coefficients \eqn{\alpha} and \eqn{\beta},
  in general, correspond to fitting parameters, coming from experimental relation of the reaction
  rate and the concentration of reactants/products. These coefficients are called \strong{partial reaction orders}
  or \code{PROs} and their \strong{sum} represents \strong{total order of the reaction}. If the kinetic equation,
  for the reaction, corresponds to its stoichiometry, the reaction is described as the elementary
  one. In EPR spectroscopy, the number of radicals is directly proportional to (double) integral
  of the radical EPR spectrum (see also \code{\link{quantify_EPR_Abs}}). Therefore, for a quick evaluation
  and/or comparison of different kinetic data, one can also obtain the rate constant
  from the area/integral \emph{vs} time fit onto the experimental EPR spectral time series outputs
  (see also the \code{\link{eval_kinR_EPR_modelFit}}).
  Accordingly, the "R" concentration (or number of radicals/V) can be replaced by the corresponding integral.
  For such a purpose a more general \strong{q}uantitative \strong{var}iable (\eqn{qvar}) is defined.
  However, in such case, the unit of \eqn{k} must be expressed accordingly (see the \code{kin.params} argument).
  Quantitative kinetic profiles (such as that \eqn{\text{d}c_{\text{R}}/\text{d}t}
  or \eqn{\text{d}(qvar)_{\text{R}}/\text{d}t} described above) are not evaluated
  by the common "analytical" integration of the kinetic equations/rate laws, however by numeric solution of the Ordinary
  Differential Equations, \href{https://desolve.r-forge.r-project.org/index.html}{ODE in \code{{desolve}} R package}.
  Therefore, higher number of models might be available than for analytical equations, because for complex mechanisms
  it's quite often highly demanding to obtain the analytical solution by common integration.
  \strong{Several kinetic models for radical reactions} in EPR spectroscopy
  \strong{are predefined and summarized} below (see also the \code{model.react} function argument).
  \tabular{ll}{
  \strong{model.react} \tab \strong{Short Description} \cr
  \code{"(r=1)R --> [k1] B"} \tab Basic forward reaction,
  e.g. irrev. dimerization (if \code{(r=2)}). \cr
  \code{"(a=1)A --> [k1] (r=1)R"} \tab Basic forward radical formation \cr
  \code{"(a=1)A <==> [k1] [k4] (r=1)R <==> [k2] [k3] (b=1)B"} \tab Consecutive reactions,
  e.g. considering comproportionation (for \code{(a=2)} and \code{(r=2)}) + follow-up reversible
  dimerization (\code{(b=1)}). \cr
  \code{"(r=1)R <==> [k1] [k2] (b=1)B"} \tab Basic reversible radical quenching,
  e.g. rev. \eqn{\pi-\pi} dimerization for \code{(r=2)} and \code{(b=1)}. \cr
  \code{"(a=1)A <==> [k1] [k2] (r=1)R"} \tab Basic reversible radical formation,
  e.g. from rev. comproportionation of conjugated thiophene oligomers
  (\eqn{\text{A}^{++} + \text{A}^0 \xrightleftharpoons ~ 2\text{R}^{.+}}, for \code{(a=2)} and \code{(r=2)}). \cr
  \code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \tab Radical formation by chemical reaction like oxidation,
  reduction or spin trapping (if the \code{A} refers to transient radical, which is not visible within
  the common EPR time scale). \cr
  \code{"(a=1)A + (r=1)R --> [k1] B"} \tab General radical quenching by chemical reaction. \cr
  }
}
\examples{
## irreversible dimerization quantitative kinetic profile,
## table (df) with first 10 observations/rows and application
## of the "euler" integrator (to solve ODE) method
##
kin.test.01 <-
  eval_kinR_ODE_model(
    model.react = "(r=2)R --> [k1] B",
    kin.params = c(k1 = 0.012,
                   qvar0R = 0.08),
   solve.ode.method = "euler"
  )
## preview of data frame
head(kin.test.01$df,n = 10)
#
## half-life (t1/2) of the second-order from data frame
## search for the value/point which is close to 0.08/2
t.half.01.a.idx <- which.min(abs(kin.test.01$df$R - 0.04))
t.half.01.a <- kin.test.01$df[t.half.01.a.idx,] \%>\%
  dplyr::pull(time)
#
## theoretical (analytical) expression of half-life
t.half.01.b <- 1 / (2 * 0.08 * 0.012)
#
## comparison of both values
t.half.01.a
t.half.01.b
#
## plot kinetic model profile with t1/2
kin.test.01$plot +
  ggplot2::geom_hline(
    yintercept = 0.08/2,
    linetype = "dashed",
    color = "blue"
  ) + ggplot2::geom_vline(
    xintercept = t.half.01.a,
    linetype = "dashed",
    color = "blue"
  ) + ggplot2::annotate(
    geom = "text",
    x = 640,
    y = 0.045,
    color = "blue",
    size = 5.4,
    label = "italic(t)[1/2]",
    parse = TRUE
  )
#
## consecutive reactions and the corresponding plot
## (`model.react` character string without spaces)
kin.test.02 <-
 eval_kinR_ODE_model(
   model.react = "(a=2)A<==>[k1][k4](r=2)R<==>[k2][k3](b=1)B",
   kin.params = c(k1 = 0.225,
                  k2 = 0.1,
                  k3 = 2e-4,
                  k4 = 2e-5,
                  qvar0A = 0.02,
                  qvar0R = 0.002,
                  qvar0B = 0)
 )
## plot preview
kin.test.02$plot
#
## data frame/table preview
head(kin.test.02$df)
#
## loading example data (incl. `Area` and `time` variables)
## from Xenon software: decay of a triarylamine radical cation
## after its generation by electrochemical oxidation
triaryl_radCat_path <-
  load_data_example(file = "Triarylamine_radCat_decay_a.txt")
## corresponding data (double integrated
## EPR spectrum = `Area` vs `time`)
triaryl_radCat_data <-
  readEPR_Exp_Specs(
    triaryl_radCat_path,
    header = TRUE,
    fill = TRUE,
    select = c(3,7),
    col.names = c("time_s","Area"),
    x.unit = "s",
    x.id = 1,
    Intensity.id = 2,
    qValue = 1700,
    data.structure = "others") \%>\%
  na.omit()
## data preview
head(triaryl_radCat_data)
#
## comparison of the kinetic model with the experimental
## data `triaryl_radCat_data`, kinetic parameters were estimated
## to be as close as possible to the latter.
compar_model_expr_data_01 <-
  eval_kinR_ODE_model(
    model.react = "(r=2)R --> [k1] B",
    kin.params = c(qvar0R = 0.019,
                   k1 = 0.04),
    time.interval.model = c(0,1500),
    data.qt.expr = triaryl_radCat_data,
    qvar.expr = "Area",
    time.expr = "time_s"
   )
## plot preview
compar_model_expr_data_01$plot
#
## previous kinetic model with partial reaction
## order ("alpha") corresponding to "R" (radical species).
## In such case REACTION is NOT CONSIDERED
## as an ELEMENTARY one !
compar_model_expr_data_02 <-
  eval_kinR_ODE_model(
    model.react = "(r=2)R --> [k1] B",
    elementary.react = FALSE,
    kin.params = c(qvar0R = 0.019,
                   k1 = 0.04,
                   alpha = 1.9
                   ),
    time.interval.model = c(0,1500),
    data.qt.expr = triaryl_radCat_data,
    qvar.expr = "Area",
    time.expr = "time_s"
   )
## plot preview
compar_model_expr_data_02$plot


}
\references{
International Union of Pure and Applied Chemistry (IUPAC) (2019). “Rate of Reaction”,
 \url{https://goldbook.iupac.org/terms/view/R05156}.

 Quisenberry KT, Tellinghuisen J (2006). “Textbook Deficiencies: Ambiguities in Chemical Kinetics Rates
 and Rate Constants.” \emph{J. Chem. Educ.}, \strong{83}(3), 510, \url{https://doi.org/10.1021/ed083p510}.

 Levine IN (2009). \emph{Physical Chemistry}, 6th edition. McGraw-Hill, ISBN 978-0-072-53862-5,
 \url{https://books.google.cz/books/about/Physical_Chemistry.html?id=L5juAAAAMAAJ&redir_esc=y}.

 rdabbler (2013). “Learning R: Parameter Fitting for Models Involving Differential Equations”,
 \url{https://www.r-bloggers.com/2013/06/learning-r-parameter-fitting-for-models-involving-differential-equations/}.

 Soetaert K, Petzoldt T, Setzer  RW (2010). "Solving Differential Equations in R: Package deSolve."
 \emph{J. Stat. Softw.}, \strong{33}(9), 1-25, \url{https://doi.org/10.18637/jss.v033.i09}.
}
\seealso{
Other Evaluations and Quantification:
\code{\link[=eval_integ_EPR_Spec]{eval_integ_EPR_Spec()}},
\code{\link[=eval_kinR_EPR_modelFit]{eval_kinR_EPR_modelFit()}},
\code{\link[=quantify_EPR_Abs]{quantify_EPR_Abs()}},
\code{\link[=quantify_EPR_Norm_const]{quantify_EPR_Norm_const()}}
}
\concept{Evaluations and Quantification}
