% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_nu_ENDOR.R
\name{eval_nu_ENDOR}
\alias{eval_nu_ENDOR}
\title{ENDOR/Larmor Freqency \eqn{\nu_{\text{ENDOR}}} for a Specific Nucleus and Magnetic Flux Density \eqn{B}}
\usage{
eval_nu_ENDOR(nucleus, B.unit, B.val)
}
\arguments{
\item{nucleus}{Character string in form of e.g. \code{"14N"}...tbc.}

\item{B.unit}{Character string denoting the magnetic flux density \eqn{B} unit.}

\item{B.val}{Numeric, magnetic flux density \eqn{B} \code{value}.}
}
\value{
Numeric value or vector...tbc...
}
\description{
A short description...tbc...
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_interval_cnfd_tVec}()}
}
\concept{Evaluations}
