% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue.R
\name{gValue}
\alias{gValue}
\title{Basic Calculation of g-factor}
\usage{
gValue(nu, unit_nu = "GHz", B, unit_B = "mT")
}
\arguments{
\item{nu}{Numeric, microwave Frequency}

\item{unit_nu}{String, frequency unit defined by \code{"GHz"} or \code{"Hz"}}

\item{B}{Numeric, magnetic flux density}

\item{unit_B}{String, magnetic flux density unit in \code{"G"} or \code{"mT"} or \code{"T"}}
}
\value{
g-value from \eqn{(\nu * h)/(\mu_B * B)}. For variables and constants =>
  see description above
}
\description{
Calculation of g-factor according to fundamental formula (see below).
  The magnetic flux density (\code{B}) and microwave frequency (\code{nu},\eqn{\nu})
  can be entered with common units like \code{G} (Gauss) \code{mT}
  (millitesla) or \code{T} (tesla) as well as \code{GHz} or \code{Hz}, respectively.
  The Planck constant (\eqn{h}) and Bohr magneton (\eqn{\mu_B}) are included
  in \code{\link[constants]{syms}} function and their values are taken by \code{syms$h}
  and \code{syms$muB} commands, respectively.
}
\examples{
gValue(9.8020458,unit_nu = "GHz",350.214,unit_B = "mT")
gValue(nu = 9.8020458e+9,unit_nu = "Hz",B = 3502.14,unit_B = "G")
gValue(9.5421,"GHz",0.333251,"T")

}
