% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue_fromSpectr.R
\name{gValue_fromSpectr}
\alias{gValue_fromSpectr}
\title{Calculation of g-factor ("position") of the EPR Spectrum}
\usage{
gValue_fromSpectr(spectrum.data, nu, B.reg.start, B.reg.end)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column can be included as well}

\item{nu}{Numeric, microwave frequency in \code{GHz}}

\item{B.reg.start}{Numeric, magnetic flux density in \code{mT} corresponding to \code{starting} border
of the \code{selected B region} (therefore abbreviation \code{.reg.})}

\item{B.reg.end}{Numeric, magnetic flux density in \code{mT} corresponding to \code{ending} border
of the \code{selected B region} (therefore abbreviation \code{.reg.})}
}
\value{
\eqn{g_iso}-value ('iso' = 'isotropic') according to \eqn{(\nu * h)/(\mu_B * B)},
  where the \code{B} comes directly from the EPR spectrum and is actually calculated between
  the \code{B(maximum)} and \code{B(minumum)} (these do not equal to 'B.reg.start' and 'B.reg.min'!)
  corresponding to maximum and minimum of the derivative intensity (\code{dIepr_over_dB})
}
\description{
Calculation of g-value according to fundamental formula. g-related magnetic flux density
  (like \code{B_iso} or \code{B_center}) is directly taken from the symmetric
  (if positive and negative derivative intensities of the spectral line are similar)
  EPR spectrum. Otherwise the g will be determined inaccurately!.
  The g-related B is computed as the mid-point between the magnetic flux densities
  corresponding to \code{min.} and \code{max.} derivative intensities (\code{dIepr_over_dB}).
  One can select the B region/span/interval from the spectrum to determine the g-value.
  The Planck constant (\eqn{h}) and the Bohr magneton (\eqn{\mu_B}) are included
  in \code{\link{constants}} package and their values are taken by the \code{syms$h}
  and \code{syms$muB} commands, respectively.
}
\examples{
\dontrun{
gValue_fromSpectr(spectrum.data,9.82451,349.8841,351.112)
gValue_fromSpectr(spectrum.data,nu = 9.82451,B.reg.start = 349.8841,B.reg.end = 351.112)
}


}
