% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue_QCHcomp.R
\name{gValue_QCHcomp}
\alias{gValue_QCHcomp}
\title{Calculation of \eqn{g}-factor from Quantum Chemical Computational Output (e.g. from Gaussian)}
\usage{
gValue_QCHcomp(delta_g_vec, mean = TRUE)
}
\arguments{
\item{delta_g_vec}{Numeric vector of principal \code{g-value differences in 'ppm'}
from \eqn{g_e} within e.g. \strong{Gaussian} or \strong{ORCA} output}

\item{mean}{Boolean, whether to calculated the \code{mean value} of \code{delta_g_vec} elements,
\strong{default}: \code{mean = TRUE}, or save the entire vector with all principal components}
}
\value{
Numeric mean \eqn{g}-factor value from principal difference (from \eqn{g_e}) components
  calculated by QCH methods (e.g. DFT) or numeric vector with principal \eqn{g}-components
  if \code{mean = FALSE}
}
\description{
TODO
}
\examples{
gValue_QCHcomp(c(-806.1,1099.9,1417.1))
gValue_QCHcomp(c(-543.1,832.8,1089.5),mean = FALSE)


}
