% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integ_correct_EPRspecs.R
\name{integ_correct_EPRspecs}
\alias{integ_correct_EPRspecs}
\title{Polynomial Baseline Correction for single-Integrated EPR Spectra + Double Integral Calculation/Presentation}
\usage{
integ_correct_EPRspecs(
  spec.integ.data,
  B = "B_G",
  B.reg.start,
  B.reg.end,
  B.peak.start,
  B.peak.end,
  poly.degree,
  double.integ = FALSE
)
}
\arguments{
\item{spec.integ.data}{Spectrum data frame/table with magnetic flux density (in \code{mT} or \code{G}) column
must be labeled as \code{B_mT} (or \code{B_G}) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column may be included as well, \code{sinlge integrated EPR} \strong{spectrum must be included}
in column named by \code{"single|sinteg|s_integ|single_|singleinteg|sintegral|sInteg_"} ("|" == "or" operator),
this can be obtained by \code{\link[pracma:trapz]{pracma::cumtrapz}} function from \emph{B} and \emph{dIepr_over_dB}}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"}
or \code{B = "B_G"} (\strong{default})}

\item{B.reg.start}{TODO}

\item{B.reg.end}{TODO}

\item{B.peak.start}{TODO}

\item{B.peak.end}{TODO}

\item{poly.degree}{TODO}

\item{double.integ}{TODO}
}
\value{
data frame TODO
}
\description{
TODO
}
\examples{


}
