% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_example.R
\name{load_data_example}
\alias{load_data_example}
\title{Loading Built-in Data Files for Package Examples}
\usage{
load_data_example(file = NULL)
}
\arguments{
\item{file}{Character string corresponding to file name + extension.}
}
\description{
Loads built-in package data files which are required to demonstrate the the package functionality
by the examples within documentation.
}
\examples{
## List of all files within the "extdata" directory =>
load_data_example()



}
