% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_EPR_by_nloptr.R
\name{optim_EPR_by_nloptr}
\alias{optim_EPR_by_nloptr}
\title{General Function to Non-Linear Optimization of the EPR Spectral Data}
\usage{
optim_EPR_by_nloptr(method = "slsqp", x.0, fn, lower, upper, data, ...)
}
\arguments{
\item{method}{Character string, method...}

\item{x.0}{Numeric vector ...}

\item{fn}{Objective function that is to be minimized ...+ examples}

\item{lower, upper}{lower and upper bound constraints.}

\item{data}{Data object ...}

\item{...}{additional arguments passed to the function.}
}
\value{
description
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{eval_sim_EPR_iso}()},
\code{\link{quantify_EPR_Sim_series}()}
}
\concept{Simulations and Optimization}
