% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ECh_VoC_amperogram.R
\name{plot_ECh_VoC_amperogram}
\alias{plot_ECh_VoC_amperogram}
\title{Plot Voltammogram or Chronoamperogram from (Spectro)Electrochemical Experiment}
\usage{
plot_ECh_VoC_amperogram(
  data.voltamm,
  x = "E_V",
  x.unit = "V",
  Current = "I_uA",
  Current.unit = "uA",
  xlim = NULL,
  Ilim = NULL,
  line.color = "darkviolet",
  line.width = 0.75,
  theme.basic = "theme_gray",
  axis.title.size = 15,
  axis.text.size = 14,
  border.line.color = "black",
  border.line.width = 0.5,
  ref.electrode = NULL,
  plot.interact = FALSE,
  ticks = "out",
  grid = TRUE
)
}
\arguments{
\item{data.voltamm}{Data frame (table) object including required columns like \code{Potential} (\eqn{E}) and/or \code{Time}
(\eqn{t}) as well as \code{Current} (\eqn{I}). Even though an arbitrary column label can be used, the best option
is to use labels e.g. like \code{E_V}, \code{E_mV}, \code{time_s}, \code{I_uA} or \code{I_mA}.}

\item{x}{Character string pointing to \code{x}-axis/column quantity, in the original \code{data.voltamm}, like \strong{potential}
(e.g. \code{x = "E_V"}, \code{x = "E_mV"}) or \strong{time} (e.g. \code{x = "time_s"}, \code{x = "time_min"}
or \code{x = "time_ms"}). \strong{Default}: \code{x = "E_V"} (potential in volts).}

\item{x.unit}{Character string pointing to \code{x}-quantity unit. There are following units available: \code{x.unit = "V"},
\code{x.unit = "mV"} (in case of voltammetry, \code{x} corresponds to potential) as well as \code{x.unit = "s"}, \code{x.unit = "ms"}
or \code{x.unit = "min"} (in case of chronoamperometry, \code{x} corresponds to time). \strong{Default}: \code{x.unit = "V"}.}

\item{Current}{Character string indicating the \code{Current}(\eqn{I})-axis/column quantity in the original \code{data.voltamm} object.
\strong{Default}: \code{Current = "I_ua"} (current in \eqn{\text{\mu A}}).}

\item{Current.unit}{Character string pointing to \code{Current} quantity unit like \code{Current.unit = "uA"}
(microamps, \strong{default}), \code{Current.unit = "A"}, \code{Current.unit = "mA"} and \code{Current.unit = "nA"}.}

\item{xlim}{Numeric vector of the \code{x}-quantity border limits, e.g. like \code{xlim = c(-100,400)} (potential in "mV",
in case of voltammetry) or \code{xlim = c(0,1000)} (time in seconds, in case of chronoamperometry).}

\item{Ilim}{Numeric vector characterizing the border limits of the current...TBC...}

\item{line.color}{Character string...TBC...}

\item{line.width}{Numeric, linewidth of the plot line in \code{pt}, \strong{default}: \code{line.width = 0.75}}

\item{theme.basic}{Character/String, which calls a ggplot theme base. The following ones are defined:
\describe{
  \item{\code{"theme_gray"}}{(\strong{default} one) => the gray background with white grid lines}
  \item{\code{"theme_bw"}}{ => the white background with thin gray grid lines, theme is proposed \strong{for publications}}
  \item{\code{"theme_light"}}{ => similar to \code{theme_bw()} but without the pronounced axis black lines}
  \item{\code{"theme_classic"}}{ => without grid, pronounced axis lines, however no opposite ones}
  \item{\code{"theme_linedraw"}}{ => pronounced axis lines (both for origin and opposite)
  as well as the grid-lines, theme is proposed \strong{for publications}
  (if the \code{grid} is set to \code{FALSE})}
}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{border.line.color}{tbc}

\item{border.line.width}{tbc}

\item{ref.electrode}{Character string ...TBC...}

\item{plot.interact}{Logical, ...TBC...}

\item{ticks}{Character string ...the axis ticks pointing either}

\item{grid}{Logical, whether to dislay the \code{grid} within the plot/graph, \strong{default}: \code{grid = TRUE}}
}
\value{
Graphical representation (list object) of voltammogram or chronoamperogram either in static form by \pkg{ggplot2}
  (\code{plot.interact = FALSE}) or in interactive form by \pkg{plotly} (\code{plot.interact = TRUE}).
}
\description{
A short description...
}
\examples{
\dontrun{
TBC
TBC
}


}
\concept{EPR Spectroelectrochemistry}
