% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_present_interact.R
\name{plot_EPR_present_interact}
\alias{plot_EPR_present_interact}
\title{Present/Save Interactive Plotly-Plot/-Graph for Publication (HTML,PDF,Word)}
\usage{
plot_EPR_present_interact(
  p,
  size.width = 700,
  size.height = 500,
  size.resolv.f = 2
)
}
\arguments{
\item{p}{tbc}

\item{size.width}{tbc}

\item{size.height}{tbc}

\item{size.resolv.f}{tbc}
}
\value{
Plot format depending on the output document one
}
\description{
Depending on Output Document Format
}
\examples{
\dontrun{
TODO
TODO
}


}
