% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Spec.R
\name{plot_EPR_Spec}
\alias{plot_EPR_Spec}
\title{EPR/ENDOR Spectrum Simple Plot}
\usage{
plot_EPR_Spec(
  data.spectrum,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  line.color = "steelblue",
  line.width = 0.75,
  theme.basic = "theme_gray",
  axis.title.size = 15,
  axis.text.size = 14,
  grid = TRUE,
  yTicks = TRUE
)
}
\arguments{
\item{data.spectrum}{SEPR/ENDOR spectrum data frame/table with magnetic flux density \eqn{B} (in \code{mT} or \code{G})
or \eqn{g}-Value or \eqn{RF} (in \code{MHz}) column and that of the derivative \code{dIepr_over_dB}
or integrated \code{Intensity}. \code{Index} column may be included as well.}

\item{x}{Character/String pointing to \code{x}-axis/column quantity like magnetic flux density \eqn{B}, \eqn{g}-Value
or \eqn{RF} (radio frequency), \strong{default}: \code{x = "B_mT"}}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated or integrated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{line.color}{String, line color to plot simple EPR spectrum. All \pkg{ggplot2} compatible
colors are allowed, \strong{default}: \code{line.color = "steelblue"}}

\item{line.width}{Numeric, linewidth of the plot line in \code{pt}, \strong{default}: \code{line.width = 0.75}}

\item{theme.basic}{Character/String, which calls a ggplot theme base. The following ones are defined:
\itemize{
  \item \code{"theme_gray"} (\strong{default} one) => the gray background with white grid lines
  \item \code{"theme_bw"} => the white background with thin gray grid lines
  \item \code{"theme_light"} => similar to \code{theme_bw()} but without the pronounced axis black lines
  \item \code{"theme_classic"} => without grid, pronounced axis lines, however no opposite ones
  \item \code{"theme_linedraw"} => pronounced axis lines (both for origin and opposite) as well as the grid-lines,
  theme is proposed \strong{for publications} (if the \code{grid} is set to \code{FALSE})
}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{grid}{Boolean, whether to dislay the \code{grid} within the plot/graph, \strong{default}: \code{grid = TRUE}}

\item{yTicks}{Boolean, whether to display the \code{y} (\code{dIepr_over_dB}) ticks and the corresponding text
(not the axis title!), which is usually skipped in the EPR community, \strong{default}: \code{yTicks = TRUE}}
}
\value{
EPR simple spectrum \pkg{ggplot2} graph/plot with key parameter (e.g. line-color and theme,grid...etc.) variation
}
\description{
Graph/Plot of an EPR/ENDOR spectrum based on \pkg{ggplot2}-functionality. Spectral data
  are in the form of data frame, which must contain the \code{dIepr_over_dB} (or its corresponding integrated
  form,\code{Iepr}) and the following \code{x}-axis quantities like \eqn{B} (in \code{mT} or \code{G})
  or \eqn{g}-Value (dimensionless) or \eqn{RF} (radio-frequency in \code{MHz})
  Theme of the graphic spectrum representation as well its line color can be varied like
  in \pkg{ggplot2} (see below). Within a theme \code{y} ticks can be displayed
  or skipped \code{y} (\code{dIepr_over_dB} in 'procedure defined unit',
  see \href{http://www.iupac.org/divisions/VII/VII.C.1/C-NPU_Uppsala_081023_25_minutes_confirmed.pdf}{p.d.u.}),
  this is common for presenting the EPR spectra. Function can be additionally combined by \code{+} sign
  with other functions like in \pkg{ggplot2}, e.g. present or skip \code{grid} within the code.
}
\examples{
\dontrun{
plot_EPR_Spec(data.spectrum)
plot_EPR_Spec(data.spectrum,"B_G",Intensity = "dIepr_over_dB_Sim")
plot_EPR_Spec(data.spectrum,x = "B_mT_Sim","Integral")
plot_EPR_Spec(data.spectrum,line.color = "blue",basic.theme = "theme_linedraw",yTicks = FALSE)
plot_EPR_Spec(data.spectrum,x = "g_Value",theme.basic = "theme_bw",grid = TRUE)
plot_EPR_Spec(data.spectrum,x = "RF_MHz",line.color = "darkred",line.width = 1.2)
}


}
