% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs2D_interact.R
\name{plot_EPR_Specs2D_interact}
\alias{plot_EPR_Specs2D_interact}
\title{Interactive (incl. Zooming, Data Reading...etc) EPR Spectrum Plot}
\usage{
plot_EPR_Specs2D_interact(
  data.spectra,
  x = "B_mT",
  x.unit = "mT",
  Intensity = "dIepr_over_dB",
  var2nd = NULL,
  line.color = "darkviolet",
  bg.color = "#e5ecf6",
  grid.color = "#ffff",
  line.width = 0.75,
  border.line.width = 1.2,
  border.line.color = "black",
  legend.title = NULL,
  legend.title.size = 13,
  axis.title.size = 15,
  axis.text.size = 14,
  var2nd.series = FALSE
)
}
\arguments{
\item{data.spectra}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} in mT (or \code{B_G} in gauss) and that of the derivative
intensity as \code{dIepr_over_dB}, \code{index} column can be included as well, integrated/simulated spectra
(incl. other \code{Intensity} and \code{B} columns) can be read as well}

\item{x}{Character/String pointing to \code{x}-axis/column quantity like magnetic flux density \eqn{B}, \eqn{g}-Value
or \eqn{RF} (radio frequency), \strong{default}: \code{x = "B_mT"}}

\item{x.unit}{Character/String pointing to unit of quantity (coming from original ASCII data, see also
\code{column.names} parameter) which is to be presented on \eqn{x} abscissa of the EPR spectrum,
like \code{"G"} (`Gauss`), \code{"mT"} (`millitesla`), \code{"MHz"} (`megahertz` in case of ENDOR spectra)
or \code{"Unitless"} in case of \eqn{g}-values, \strong{default}: \code{x.unit = "mT"}.}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for integrated or simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{var2nd}{String/Character referred to name of the variable/quantity column (e.g. like `time`,`Temperature`,
`Electrochemical Potential`,`Microwave Power`...etc) altered upon individual experiments as a second variable
(\code{var2nd}) and related to spectra/data (see also \code{var2nd.series} parameter). Data must be available
in \strong{long table} (or \strong{tidy}) \strong{format} (see also \code{\link{readEPR_Exp_Specs_multif}}).
\strong{Default}: \code{var2nd = NULL}. If \code{var2nd.series = FALSE}, otherwise \strong{usually}
\code{var2nd = "time_s"} if \code{var2nd.series = TRUE}}

\item{line.color}{Character/String corresponding to \strong{line color} in case \strong{of simple spectrum}
(not for \code{var2nd.series}), therefore \strong{default:} \code{line.color = "darkviolet"}}

\item{bg.color}{Character/String corresponding to \strong{background color}}

\item{grid.color}{Character/String corresponding to \strong{grid color}}

\item{line.width}{Numeric, linewidth of the plot line in \code{pt}, \strong{default}: \code{line.width = 0.75}}

\item{border.line.width}{tbc}

\item{border.line.color}{tbc}

\item{legend.title}{tbc}

\item{legend.title.size}{tbc}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{var2nd.series}{Boolean, whether the input ASCII spectrum data comes from the series of experiments
each corresponding to alteration of a second variable (usually like time series, see also parameter \code{var2nd})
where the ASCII data are in the long table/tidy format (e.g. for time series => 3 columns like "B_mT","time_s"
and "Intensity" must be supplied). \strong{Deffault}: \code{var2nd.series = FALSE}.}
}
\value{
Interactive EPR spectrum/spectra plot/graph based on \pkg{plotly}
}
\description{
tbc
}
\examples{
\dontrun{
TODO
TODO
}

}
