% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs3D_interact.R
\name{plot_EPR_Specs3D_interact}
\alias{plot_EPR_Specs3D_interact}
\title{Interactive (incl. Zooming, Data Reading...etc) 3D Plot for Time Series of EPR Spectra}
\usage{
plot_EPR_Specs3D_interact(
  data.time.spectra,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  time = "time_s",
  plot.type = "surface",
  scheme.color = "Viridis",
  contour.labels = FALSE,
  xlim,
  xlab = "<i>B</i> (mT)",
  ylab = "<i>Time</i> (s)",
  zlab = "d <i>I</i><sub>EPR</sub> / d <i>B</i> (p.d.u.)",
  bg.x.color = "rgb(220, 220,220)",
  grid.x.color = "rgb(255, 255, 255)",
  bg.y.color = "rgb(220, 220,220)",
  grid.y.color = "rgb(255, 255, 255)",
  bg.z.color = "rgb(220, 220,220)",
  grid.z.color = "rgb(255, 255, 255)",
  output.matrix.table = FALSE
)
}
\arguments{
\item{data.time.spectra}{tbc}

\item{x}{tbc}

\item{Intensity}{tbc}

\item{time}{tbc}

\item{plot.type}{tbc}

\item{scheme.color}{tbc}

\item{contour.labels}{tbc}

\item{xlim}{tbc}

\item{xlab}{tbc}

\item{ylab}{tbc}

\item{zlab}{tbc}

\item{bg.x.color}{tbc}

\item{grid.x.color}{tbc}

\item{bg.y.color}{tbc}

\item{grid.y.color}{tbc}

\item{bg.z.color}{tbc}

\item{grid.z.color}{tbc}

\item{output.matrix.table}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
