% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs_integ.R
\name{plot_EPR_Specs_integ}
\alias{plot_EPR_Specs_integ}
\title{Plotting Integrated Forms of EPR Spectra Acquired by \code{\link{eval_integ_EPR_Spec}}}
\usage{
plot_EPR_Specs_integ(
  data.spectra.integ,
  B = "B_G",
  B.unit = "G",
  Blim = NULL,
  ylim = NULL,
  slct.integs = c("single_Integ", "baseline_Integ_fit", "single_Integ_correct"),
  line.width = 0.75,
  axis.title.size = 15,
  axis.text.size = 14,
  legend.title.size = 13,
  legend.text.size = 11,
  separate.integs = FALSE,
  separate.integ.scales = NULL,
  output.df = FALSE
)
}
\arguments{
\item{data.spectra.integ}{tbc}

\item{B}{tbc}

\item{B.unit}{tbc}

\item{Blim}{tbc}

\item{ylim}{tbc...doesn't apply for separated integrals...it works only for overlay}

\item{slct.integs}{tbc}

\item{line.width}{tbc}

\item{axis.title.size}{tbc}

\item{axis.text.size}{tbc}

\item{legend.title.size}{tbc}

\item{legend.text.size}{tbc}

\item{separate.integs}{tbc}

\item{separate.integ.scales}{Character/String corresponding to how the axis scales should or shouldn't be fixed,
unless the \code{separate.integs = FALSE}. Inherited from \code{\link[ggplot2]{facet_wrap}}.
Following expressions are available => \code{"fixed"}, \code{"free"} or in one dimension \code{"free_x"}
or \code{"free_y"}. \strong{Default}: \code{separate.integ.scales = NULL}
in case of \code{separate.integs = FALSE}.}

\item{output.df}{Logical, description tbc...data frame/table in `tidy` (long) format...for additional
plotting by `plotly` (see \code{\link{plot_EPR_Specs2D_interact}})}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Visualization and Graphics: 
\code{\link{draw_molecule_by_rcdk}()},
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{presentEPR_Sim_Spec}()}
}
\concept{Visualization and Graphics}
