% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval_EPRtheo_mltiplet.R
\name{plot_eval_EPRtheo_mltiplet}
\alias{plot_eval_EPRtheo_mltiplet}
\title{EPR Intensity Multiplet Prediction for Interactions of Electron with Selected Nucleus/Nuclei}
\usage{
plot_eval_EPRtheo_mltiplet(nucle_us_i = "1H", I = NULL, N.nuclei = 1)
}
\arguments{
\item{nucle_us_i}{Character string, pointing to specific nucleus/nuclei in the form like \code{"14N"} or \code{"2H"}.
If \code{nucle_us_i = "1H"}, i.e. interaction with proton(s) is considered as \strong{default} one. Based on the string,
the characteristic nuclear spin quantum number \code{I} is taken from the \code{\link{isotopes_ds}} dataset. If one wants
to define \code{I} in a more general way, just put the \code{nucle_us_i = NULL} and define the desired \code{I}
by the corresponding argument (see below).}

\item{I}{Numeric value, pointing to nuclear spin quantum number of proposed nucleus/nuclei interacting with unpaired electron.
\code{I} must be specified in the form like \code{I = 0.5} (for spin 1/2), \code{I = 1.5} (for spin 3/2) or \code{I = 1}.
The \strong{default} value, \code{I = NULL}, is applied only in such case when the \code{I} is defined
via the \code{nucle_us_i} argument (see above).}

\item{N.nuclei}{Numeric value (integer), corresponding to number of interacting equivalent nuclei (within a group).
\strong{Default}: \code{N.nuclei = 1}.}
}
\value{
A list consisting of:
  \describe{
  \item{intensity.vec}{A named vector of binomial/multinomial coefficients related to theoretical intensities
  of isotropic EPR spectrum when the coupling between unpaired electron and the surrounding nuclei is present.
  Coefficient/Vector names correspond to total spin quantum numbers (e.g. for one \code{"14N"} nucleus, the names
  are represented by the \code{c("-1","0","1")} vector).}
  \item{plot}{GGplot2 object - graphical representation of the \code{intensity.vec} list component.}
  }
}
\description{
What is the expected EPR intensity pattern for a group of equivalent nuclei? One may use this function for a quick
  prediction/visualization of EPR spectrum multiplets (without a specific hyperfine splitting). Its central code
  (computation of binomial/multinomial coefficients) is implemented in the \code{\link{eval_sim_EPR_iso}} to provide
  simulations of isotropic EPR spectra. The theoretical intensities/coefficients are returned either as a vector
  or as a barplot using the \code{\link[ggplot2]{geom_bar}}.
}
\examples{
## intensity multiplet for methyl group
## (3 x 1H equivalent nuclei)
methyl.multipl <-
  plot_eval_EPRtheo_mltiplet(
    N.nuclei = 3
  )
#
## vector of predicted
## intensities/coefficients
methyl.multipl$intensity.vec
#
## graphical repsentation
methyl.multipl$plot
#
## intensity multiplet for 2 x 14N
## in TMPD (Wuster's blue) radical cation
tmpd.14N.multipl <-
  plot_eval_EPRtheo_mltiplet(
    I = 1,
    N.nuclei = 2
  )
#
tmpd.14N.multipl$plot


}
\seealso{
Other Simulations and Optimization:
\code{\link[=eval_ABIC_forFit]{eval_ABIC_forFit()}},
\code{\link[=eval_sim_EPR_iso]{eval_sim_EPR_iso()}},
\code{\link[=eval_sim_EPR_isoFit]{eval_sim_EPR_isoFit()}},
\code{\link[=eval_sim_EPR_isoFit_space]{eval_sim_EPR_isoFit_space()}},
\code{\link[=eval_sim_EPR_isoFitb]{eval_sim_EPR_isoFitb()}},
\code{\link[=eval_sim_EPR_iso_combo]{eval_sim_EPR_iso_combo()}},
\code{\link[=optim_for_EPR_fitness]{optim_for_EPR_fitness()}},
\code{\link[=plot_eval_RA_forFit]{plot_eval_RA_forFit()}},
\code{\link[=quantify_EPR_Sim_series]{quantify_EPR_Sim_series()}},
\code{\link[=smooth_EPR_Spec_by_npreg]{smooth_EPR_Spec_by_npreg()}}
}
\concept{Simulations and Optimization}
