% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout2D_interact.R
\name{plot_layout2D_interact}
\alias{plot_layout2D_interact}
\title{Additional Layout for Interactive Plots Similar to Layout from Plotly Package}
\usage{
plot_layout2D_interact(
  p,
  data = NULL,
  x.label = "<i>B</i> (mT)",
  y.label = "d <i>I</i><sub>EPR</sub> / d <i>B</i>  (p.d.u.)",
  x.axis.title.family = "Arial",
  x.axis.title.color = "black",
  x.axis.title.size = 15,
  x.axis.tick.family = "Arial",
  x.axis.tick.color = "black",
  x.axis.tick.size = 14,
  y.axis.title.family = "Arial",
  y.axis.title.color = "black",
  y.axis.title.size = 15,
  y.axis.tick.family = "Arial",
  y.axis.tick.color = "black",
  y.axis.tick.size = 14,
  legend.label = NULL,
  legend.label.family = "Arial",
  legend.label.color = "black",
  legend.label.size = 13,
  bg.color = "#e5ecf6",
  grid.color = "#ffff",
  border.line.width = 1.2,
  border.line.color = "black"
)
}
\arguments{
\item{p}{tbc}

\item{data}{tbc}

\item{x.label}{tbc}

\item{y.label}{tbc}

\item{x.axis.title.family}{tbc}

\item{x.axis.title.color}{tbc}

\item{x.axis.title.size}{tbc}

\item{x.axis.tick.family}{tbc}

\item{x.axis.tick.color}{tbc}

\item{x.axis.tick.size}{tbc}

\item{y.axis.title.family}{tbc}

\item{y.axis.title.color}{tbc}

\item{y.axis.title.size}{tbc}

\item{y.axis.tick.family}{tbc}

\item{y.axis.tick.color}{tbc}

\item{y.axis.tick.size}{tbc}

\item{legend.label}{tbc}

\item{legend.label.family}{tbc}

\item{legend.label.color}{tbc}

\item{legend.label.size}{tbc}

\item{bg.color}{tbc}

\item{grid.color}{tbc}

\item{border.line.width}{tbc}

\item{border.line.color}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}

}
