% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_interact.R
\name{plot_layout_interact}
\alias{plot_layout_interact}
\title{Additional Layout for Interactive \code{plotly} Plots}
\usage{
plot_layout_interact(
  x.label = "<i>B</i> (mT)",
  y.label = "d <i>I</i><sub>EPR</sub> / d <i>B</i>  (p.d.u.)",
  x.axis.title.family = "Courier New",
  x.axis.title.color = "black",
  x.axis.title.size = 15,
  x.axis.tick.family = "Courier New",
  x.axis.tick.color = "black",
  x.axis.tick.size = 14,
  y.axis.title.family = "Courier New",
  y.axis.title.color = "black",
  y.axis.title.size = 15,
  y.axis.tick.family = "Courier New",
  y.axis.tick.color = "black",
  y.axis.tick.size = 14,
  legend.label = NULL,
  legend.label.family = "Courier New",
  legend.label.color = "black",
  legend.label.size = 13,
  bg.color = "#e5ecf6",
  grid.color = "#ffff"
)
}
\arguments{
\item{x.label}{tbc}

\item{y.label}{tbc}

\item{x.axis.title.family}{tbc}

\item{x.axis.title.color}{tbc}

\item{x.axis.title.size}{tbc}

\item{x.axis.tick.family}{tbc}

\item{x.axis.tick.color}{tbc}

\item{x.axis.tick.size}{tbc}

\item{y.axis.title.family}{tbc}

\item{y.axis.title.color}{tbc}

\item{y.axis.title.size}{tbc}

\item{y.axis.tick.family}{tbc}

\item{y.axis.tick.color}{tbc}

\item{y.axis.tick.size}{tbc}

\item{legend.label}{tbc}

\item{legend.label.family}{tbc}

\item{legend.label.color}{tbc}

\item{legend.label.size}{tbc}

\item{bg.color}{tbc}

\item{grid.color}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
tbc
tbc

}
