% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_In_ticks.R
\name{plot_theme_In_ticks}
\alias{plot_theme_In_ticks}
\title{TODO}
\usage{
plot_theme_In_ticks(
  axis.text.size = 15,
  axis.title.size = 17,
  grid = TRUE,
  plot.bg.transparent = FALSE
)
}
\arguments{
\item{axis.text.size}{TODO}

\item{axis.title.size}{TODO}

\item{grid}{TODO}

\item{plot.bg.transparent}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
