% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_NoY_ticks.R
\name{plot_theme_NoY_ticks}
\alias{plot_theme_NoY_ticks}
\title{Custom \code{ggplot2} Theme with NO \code{Y} Axis Ticks (as Common for EPR Spectra Presentation)}
\usage{
plot_theme_NoY_ticks(
  axis.text.size = 14,
  axis.title.size = 15,
  grid = TRUE,
  border.line.color = "black",
  border.line.type = "solid",
  border.line.width = 0.5,
  bg.transparent = FALSE
)
}
\arguments{
\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{grid}{Boolean, whether to dislay the \code{grid} within the plot/graph, \strong{default}: \code{grid = TRUE}}

\item{border.line.color}{description...tbc...}

\item{border.line.type}{description ...tbc...}

\item{border.line.width}{description ...tbc...}

\item{bg.transparent}{Boolean, whether the \code{entire plot background} (NOT the \code{panel}=\code{own graph})
should be transparent, \strong{default}: \code{bg.transparent = FALSE}, i.e. no transparent background}
}
\value{
Custom \pkg{ggplot2} \code{theme} \code{without Y axis ticks}, to show opposite X axis ticks
  use: \code{scale_x_continuous(sec.axis = dup_axis(name = "",labels = NULL))}
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Visualization and Graphics: 
\code{\link{draw_molecule_by_rcdk}()},
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{presentEPR_Sim_Spec}()}
}
\concept{Visualization and Graphics}
