% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_Out_ticks.R
\name{plot_theme_Out_ticks}
\alias{plot_theme_Out_ticks}
\title{Custom \code{ggplot2} Theme with Axis Ticks Oriented Outside of the Panel}
\usage{
plot_theme_Out_ticks(
  axis.text.size = 14,
  axis.title.size = 15,
  grid = TRUE,
  bg.transparent = FALSE
)
}
\arguments{
\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{grid}{Boolean, whether to dislay the \code{grid} within the plot/graph, \strong{default}: \code{grid = TRUE}}

\item{bg.transparent}{Boolean, whether the \code{entire plot background} (NOT the \code{panel}=\code{own graph})
should be transparent, \strong{default}: \code{bg.transparent = FALSE}, i.e. no transparent background}
}
\value{
Custom \pkg{ggplot2} \code{theme} with axis \code{ticks pointing outside} the graph panel
}
\description{
TODO
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Visualization and Graphics: 
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{presentEPR_Sim_Spec}()}
}
\concept{Visualization and Graphics}
