% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRlabel.R
\name{plotEPRlabel}
\alias{plotEPRlabel}
\title{Labels for Various EPR Spectroscopy Plots}
\usage{
plotEPRlabel(axis)
}
\arguments{
\item{axis}{String, which corresponds to axis label, labels
are defined for the following quantities and their units:
\itemize{
\item \code{"B_mT"} => magnetic flux density in mT (mostly for x-axis) \eqn{B} (mT)
\item \code{"Deriv_Intensity"} => derivative intensity d\eqn{I_EPR}/d\eqn{B} (p.d.u.),
"_" corresponds to subscript and p.d.u. is abbreviation of "procedure defined unit" according to
\href{http://www.iupac.org/divisions/VII/VII.C.1/C-NPU_Uppsala_081023_25_minutes_confirmed.pdf}{IUPAC}
\item \code{"Time_s"} => time \eqn{Time} (s)
\item \code{"Integral"} => integral of the derivative intensity EPR spectrum \eqn{Integral} (p.d.u.)
\item \code{"Double_Integral"} => double integral of the derivative intensity EPR spectrum
\eqn{Double Integral (p.d.u.)}
\item \code{"Potential_V"} => potential (for spectroelectrochemical plots) in V \eqn{E} (V)
\item \code{"Concentration_mM"} => concentration (of the radicals or paramagnet. species in analyte)
in mM (common concentration order used in EPR) \eqn{c} (mmol dm^-3) "^" corresponds to superscript
\item \code{"DeltaBpp_mT"} => Peak-to-Peak spectral linewidth in mT \eqn{Delta B_pp} (mT)
}}
}
\value{
TODO
}
\description{
TODO
}
\examples{
plotEPRlabel(axis = "B_mT")
plotEPRlabel(axis = "Deriv_Intensity")
plotEPRlabel(axis = "DeltaBpp_mT")

}
