% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRspecs3D_interact.R
\name{plotEPRspecs3D_interact}
\alias{plotEPRspecs3D_interact}
\title{Interactive (incl. Zooming, Data Reading...etc) 3D Plot for Time Series of EPR Spectra}
\usage{
plotEPRspecs3D_interact(
  data.time.spectra,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  time = "time_s",
  plot.type = "surface",
  scheme.color = "Viridis",
  contour.labels = FALSE,
  Blim,
  x.label = "<i>B</i> (mT)",
  y.label = "<i>Time</i> (s)",
  z.label = "d <i>I</i><sub>EPR</sub> / d <i>B</i> (p.d.u.)",
  x.bg.color = "rgb(220, 220,220)",
  x.grid.color = "rgb(255, 255, 255)",
  y.bg.color = "rgb(220, 220,220)",
  y.grid.color = "rgb(255, 255, 255)",
  z.bg.color = "rgb(220, 220,220)",
  z.grid.color = "rgb(255, 255, 255)",
  output.matrix.table = FALSE
)
}
\arguments{
\item{data.time.spectra}{tbc}

\item{B}{tbc}

\item{Intensity}{tbc}

\item{time}{tbc}

\item{plot.type}{tbc}

\item{scheme.color}{tbc}

\item{contour.labels}{tbc}

\item{Blim}{tbc}

\item{x.label}{tbc}

\item{y.label}{tbc}

\item{z.label}{tbc}

\item{x.bg.color}{tbc}

\item{x.grid.color}{tbc}

\item{y.bg.color}{tbc}

\item{y.grid.color}{tbc}

\item{z.bg.color}{tbc}

\item{z.grid.color}{tbc}

\item{output.matrix.table}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
