% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRspectr.R
\name{plotEPRspectr}
\alias{plotEPRspectr}
\title{EPR Spectrum Simple Plot}
\usage{
plotEPRspectr(spectrum.data, line.color, plot.theme = "theme_grey", yTicks = T)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} (in mT) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column can be included as well}

\item{line.color}{String, line color to plot simple EPR spectrum. All \pkg{ggplot2} compatible
colors are allowed}

\item{plot.theme}{String, which calls a ggplot theme. The following ones are defined:
\itemize{
  \item \code{"theme_grey"} (default one) => the gray background with white grid lines, default theme
  \item \code{"theme_bw"} => the white background with thin gray grid lines
  \item \code{"theme_light"} => similar to \code{"theme_bw"} but without the pronounced axis black lines
  \item \code{"theme_classic"} => without grid, pronounced axis lines, however no opposite ones
  \item \code{"theme_pubready"} => pronounced axis lines (both for origin and opposite) without the grid,
  theme is proposed for publications.
}
Except the last one all above-described themes are standard in \pkg{ggplot2}.
The last one (\code{"theme_pubready"}) is modified \code{"theme_linedraw"} from the \pkg{ggplot2}.}

\item{yTicks}{Boolean, whether to display the \code{y} (\code{dIepr_over_dB}) ticks and the corresponding text
(not the axis title!), which is usually skipped in the EPR community, default TRUE}
}
\value{
EPR simple spectrum graph/plot with key parameter (e.g. line-color and theme) variation
}
\description{
Graph/Plot of an EPR spectrum based on \code{\link{ggplot2}}-functionality. Spectral data
  are in the form of data frame (must contain the \code{dIepr_over_dB} and \code{B_mT} columns,
  i.e. derivative EPR intensity vs. magnetic flux density, respectively). Theme of the graphic
  spectrum representation as well its line color can be varied like in \pkg{ggplot2} (see below).
  A theme for \code{publication ready} figures can be also applied based on the \code{theme_linedraw()}
  with displayed or skipped \code{y} (\code{dIepr_over_dB} in 'procedure defined unit',
  see \href{http://www.iupac.org/divisions/VII/VII.C.1/C-NPU_Uppsala_081023_25_minutes_confirmed.pdf}{p.d.u.})
  ticks (this is common for presenting the EPR spectra). Function can be additionally combined by \code{+} sign
  with other functions like in \pkg{ggplot2}.
}
\examples{
\dontrun{
plotEPRspectr(spectrum.data,"blue",plot.theme = "theme_pubready",yTicks = FALSE)
plotEPRspectr(spectrum.data,line.color = "steelblue","theme_bw")
plotEPRspectr(spectrum.data,"darkred")
}


}
