% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRspectr_interactiv.R
\name{plotEPRspectr_interactiv}
\alias{plotEPRspectr_interactiv}
\title{TODO}
\usage{
plotEPRspectr_interactiv(spectrum.data, B = "B_mT", line.color = "steelblue")
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} in mT (or \code{B_G} in gauss) and that of the derivative
intensity as \code{dIepr_over_dB}, \code{index} column can be included as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (default)
or \code{B = "B_G"}}

\item{line.color}{String, line color to plot simple EPR spectrum. All \pkg{ggplot2} compatible
colors are allowed, default: \code{line.color = "steelblue"}}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO

}
