% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRspectra_interactiv.R
\name{plotEPRspectra_interactiv}
\alias{plotEPRspectra_interactiv}
\title{TODO}
\usage{
plotEPRspectra_interactiv(
  spectrum.data,
  B = "B_mT",
  line.size = 0.75,
  time.series = FALSE
)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} in mT (or \code{B_G} in gauss) and that of the derivative
intensity as \code{dIepr_over_dB}, \code{index} column can be included as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (default)
or \code{B = "B_G"}}

\item{line.size}{Numeric, linewidth of the plot line in \code{pt}, default: \code{line.size = 0.75}}

\item{time.series}{Boolean, whether the input ASCII spectrum data comes from the time series experiment
with the additional \code{time} (labeled as \code{time_s}) column (ONLY IN CASE of "Xenon" software)}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
