% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentSimEPRspectrum.R
\name{presentSimEPRspectrum}
\alias{presentSimEPRspectrum}
\title{Show Comparison Between the Experimental and Simulated EPR Spectrum either by Table or Graphic Format}
\usage{
presentSimEPRspectrum(
  exp.spectrum.data,
  sim.spectrum.data,
  B = "B_mT",
  Intensity.shift.ratio = 1.2,
  B.shift = 0,
  line.color.exp = "red",
  line.color.sim = "blue",
  line.size = 0.75,
  output.table = FALSE
)
}
\arguments{
\item{exp.spectrum.data}{\strong{Experimental} Spectrum data frame/table where the magnetic flux
density (in \code{mT}) column is labeled as \code{B_mT} in mT (or \code{B_G} in gauss)
and that of the derivative intensity as \code{dIepr_over_dB}, \code{index} column can be included as well}

\item{sim.spectrum.data}{Data frame/table corresponding to \strong{simulated} spectrum where the magnetic
flux density (in \code{mT}) column must be labeled as \code{B_mT_Sim} in mT (or \code{B_G_Sim} in gauss)
and that of the derivative intensity as \code{dIepr_over_dB_Sim}}

\item{B}{TODO}

\item{Intensity.shift.ratio}{TODO}

\item{B.shift}{TODO}

\item{line.color.exp}{TODO}

\item{line.color.sim}{TODO}

\item{line.size}{TODO}

\item{output.table}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO
TODO


}
