% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_abs.R
\name{quantify_EPR_abs}
\alias{quantify_EPR_abs}
\title{Absolute Quantification of Radicals/Spins in the Most Common EPR Cavities}
\usage{
quantify_EPR_abs(
  integ.sigmoid.max,
  nu.GHz,
  P.mW,
  Bm.mT,
  qValue = NULL,
  tube.sample.id.mm,
  fill.sample.h.mm,
  Norm.constant = NULL,
  Temp.K = 298,
  S = 0.5,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{integ.sigmoid.max}{tbc}

\item{nu.GHz}{tbc}

\item{P.mW}{tbc}

\item{Bm.mT}{tbc}

\item{qValue}{tbc}

\item{tube.sample.id.mm}{tbc}

\item{fill.sample.h.mm}{tbc}

\item{Norm.constant}{tbc}

\item{Temp.K}{tbc}

\item{S}{tbc}

\item{microW.cavity}{tbc}
}
\value{
tbc
}
\description{
There are two approaches how to quantify the number of paramagnetic species/radicals/spins. The \strong{relative} one
  needs a standard sample with a known number and can be evaluated by the sigmoid integral ratio of the sample under study
  and that of the standard. Whereas the \strong{absolute} method do not need the reference sample however, it requires
  a precise cavity signal calibration as well as standardized cell geometry. Both are provided by the EPR instrument
  and lab-glass manufacturers. In case of absolute quantitative EPR analysis the sigmoid integral (its maximum value)
  can be expressed as follows =>
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Norm_const}()},
\code{\link{quantify_EPR_sim}()}
}
\concept{Evaluations and Quantification}
