% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativ_EPR_abs.R
\name{quantitativ_EPR_abs}
\alias{quantitativ_EPR_abs}
\title{TODO}
\usage{
quantitativ_EPR_abs(
  double.integ,
  tube.id.mm,
  tube.h.mm,
  nu.GHz,
  power.mW,
  modul.amp.mT,
  qValue,
  Nscans,
  rc.gain.dB,
  conv.time.ms,
  Temp.K = 298,
  S = 0.5,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{double.integ}{TODO}

\item{tube.id.mm}{TODO}

\item{tube.h.mm}{TODO}

\item{nu.GHz}{TODO}

\item{power.mW}{TODO}

\item{modul.amp.mT}{TODO}

\item{qValue}{TODO}

\item{Nscans}{TODO}

\item{rc.gain.dB}{TODO}

\item{conv.time.ms}{TODO}

\item{Temp.K}{TODO}

\item{S}{TODO}

\item{microW.cavity}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
