% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativ_EPR_Norm_const.R
\name{quantitativ_EPR_Norm_const}
\alias{quantitativ_EPR_Norm_const}
\title{Normalization Constant Calculation for Quantitative EPR Analysis}
\usage{
quantitativ_EPR_Norm_const(conv.time.ms, Nscans, rg.dB)
}
\arguments{
\item{conv.time.ms}{Numeric, \strong{Conversion time in milliseconds}, can be obtained
from the spectrometer parameters file using \code{\link{readEPR_param_slct}}.}

\item{Nscans}{Numeric, \strong{Number of scans}, can be obtained from the spectrometer
parameters file using \code{\link{readEPR_param_slct}}.}

\item{rg.dB}{Numeric, \strong{Receiver gain in dB}, can be obtained from the spectrometer
parameters file using \code{\link{readEPR_param_slct}}.}
}
\value{
Numeric value of normalization constant for quantitative EPR
}
\description{
Normalization Constant used by \code{\link{quantitativ_EPR_abs}}
}
\examples{
quantitativ_EPR_Norm_const(conv.time.ms = 6.4,
                          Nscans = 6,
                          rg.dB = 30)
quantitativ_EPR_Norm_const(conv.time.ms = 13.52,
                          Nscans = 100,
                          rg.dB = 24)

}
