% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_slct.R
\name{readEPR_params_slct_kin}
\alias{readEPR_params_slct_kin}
\title{Read the Selected Instrumental Parameters of EPR Time Series Experiment}
\usage{
readEPR_params_slct_kin(path_to_dsc_par, origin = "xenon")
}
\arguments{
\item{path_to_dsc_par}{String, path to \code{.DSC/.dsc} or \code{.par} file including all instrumental
parameters provided by the EPR machine.}

\item{origin}{Character string, corresponding to software used to acquire EPR spectra.
The files are slightly different depending on whether
they were recorded by the "WinEpr",\code{origin = "winepr"}, "Xenon"
(\strong{default}: \code{origin = "xenon"}) or by the "Magnettech" (ESR5000 [11-0422],
\code{origin = "magnettech"}).}
}
\value{
List containing:
  \describe{
  \item{Nscans}{Number of scans.}
  \item{swTime}{Sweep time in \code{s} required for time correction during the \code{2D_Field_Delay}
  (time series EPR experiment).}
  \item{Npoints}{Number of points (spectral resolution).}
  }
}
\description{
Function takes selected instrumental parameters relevant to \strong{time series ("kinetic")}
 experiment from the \code{.DSC/.dsc} or \code{.par} file of an EPR Spectrum, obtained from the "Xenon",
 "WinEpr" or "Magnettech" software. These parameters are required for time correction
 of the CW (continuous wave) EPR spectra, see the \code{\link{correct_time_Exp_Specs}}.
}
\examples{
## loading `.DSC` (`Xenon`) parameter file example
aminoxyl_dsc_path <-
  load_data_example(file = "Triarylamine_radCat_decay_series.DSC")
#
readEPR_params_slct_kin(aminoxyl_dsc_path)


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
