% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_slct.R
\name{readEPR_params_slct_quant}
\alias{readEPR_params_slct_quant}
\title{Read the Selected Instrumental Parameters Relevant to EPR Quantitative Analysis}
\usage{
readEPR_params_slct_quant(path_to_dsc_par, origin = "xenon")
}
\arguments{
\item{path_to_dsc_par}{Character string, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on OS, see the \code{origin} argument)
\code{text} files including all instrumental parameters and provided by the EPR machine.}

\item{origin}{Character string, corresponding to software used to acquire EPR spectra.
The files are slightly different depending on whether
they were recorded by the "WinEpr",\code{origin = "winepr"}, "Xenon"
(\strong{default}: \code{origin = "xenon"}) or by the "Magnettech" (ESR5000 [11-0422],
\code{origin = "magnettech"}).}
}
\value{
List consisting of:
  \describe{
  \item{BmmT}{Modulation amplitude value in \code{mT}.}
  \item{PmW}{Microwave source power in \code{mW}.}
  \item{TK}{Experimental temperature in \code{K}.}
  \item{mwGHz}{Microwave frequency value in \code{GHz}.}
  }
}
\description{
Reading the \code{.DSC/.dsc} or \code{.par} file to extract the important parameters like
  "modulation amplitude", "temperature", "microwave power" as well as "microwave frequency"
  which are are required for the absolute EPR quantitative analysis (\eqn{\equiv}
  radical or paramagnetic species number determination, see the \code{\link{quantify_EPR_Abs}} function).
}
\examples{
## loading `.DSC` (`Xenon`) parameter file example
aminoxyl_dsc_path <-
  load_data_example(file = "Aminoxyl_radical_a.DSC")
#
readEPR_params_slct_quant(aminoxyl_dsc_path)


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
