% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_tabs.R
\name{readEPR_params_tabs}
\alias{readEPR_params_tabs}
\title{Read the EPR Instrumental Parameters and Experiment Information}
\usage{
readEPR_params_tabs(path_to_DSC_or_par, origin = "xenon")
}
\arguments{
\item{path_to_DSC_or_par}{String, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} parameter)
\code{text} files including all instrumental parameters and provided by the EPR machine}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}
}
\value{
List of data frames/tables containing instrumental parameters (\code{params}) and information (\code{info}),
 i.e. numeric values & strings related to an EPR experiment
}
\description{
Function takes the instrumental parameters (from \code{.DSC} or \code{.par} file) applied
to record the EPR Spectra and transfers them into list of \code{Tables/Data Frames} incl.
parameter values as well as character/string information about the measurement see also
\code{\link{readEPR_param_slct}}
}
\examples{
\dontrun{
readEPR_params_tabs(path_to_DSC_or_par)
readEPR_params_tabs(file.path(".",
                              "dir_par",
                              "EPR_spectrum.par"),
                    origin = "winepr")
}


}
