% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Sim_Spec.R
\name{readEPR_Sim_Spec}
\alias{readEPR_Sim_Spec}
\title{Read the ASCII Data of a Simulated EPR Spectrum}
\usage{
readEPR_Sim_Spec(
  path_to_ASC,
  B.unit = "G",
  col.names.sim = c("Bsim_G", "dIeprSim_over_dB"),
  x.sim.id = 1,
  Intensity.sim.id = 2,
  origin.sim = "easyspin"
)
}
\arguments{
\item{path_to_ASC}{Character string, path to ASCII file/table
with simulated spectral data (\eqn{Intensity\,\,vs\,\,B}(Field) obtained from various sources.
Path can be alternatively defined by the \code{\link[base]{file.path}} function.}

\item{B.unit}{Character string, pointing to unit of magnetic flux density coming from the original data
which is to be presented on the \eqn{B}-axis of an EPR spectrum,
like \code{"G"} ("Gauss"), \code{"mT"} ("millitesla"). \strong{Default}: \code{B.unit = "G"}.}

\item{col.names.sim}{Character string vector, pointing to column names/headers of the original ASCII data
(refer also to the \code{path_to_ASC} argument and description).
\strong{Default}: \code{col.names.sim = c("Bsim_G","dIeprSim_over_dB")},
corresponding to the \code{origin.sim = "easyspin"}.}

\item{x.sim.id}{Numeric index related to the \code{col.names.sim} vector pointing to independent variable (like "B" or "Bsim"
- magnetic flux density), which corresponds to \eqn{x}-axis in the simulated spectra (refer also to the original ASCII data
\code{path_to_ASC} of the simulated spectrum). \strong{Default}: \code{x.sim.id = 1} (see also default of the \code{col.names.sim}).}

\item{Intensity.sim.id}{Numeric index related to the \code{col.names.sim} vector pointing to simulated EPR intensity
column name/header in the original ASCII data (refer also to the \code{path_to_ASC} argument and description).
If used together with the quantification of radicals, this argument must be equal to that
of the \code{\link{quantify_EPR_Sim_series}}. \strong{Default}: \code{x.sim.id = 2} (see also default of the \code{col.names.sim}).}

\item{origin.sim}{Character string, referring to the "origin" of a simulated spectrum ASCII data.
There are four possibilities \eqn{\Rightarrow} \code{sim.orimgin = "easyspin"} (\strong{default}),
\code{"xenon"}, \code{"simfonia"} as well as universal \code{"csv"}.}
}
\value{
Data frame, consisting of magnetic flux density and intensity variable/column
  (depending on the \code{col.names.sim} and \code{.id} arguments), corresponding to simulated EPR spectrum.
}
\description{
Loading the ASCII data like \code{.txt},\code{.asc} or \code{.csv}, related to simulated EPR spectrum from different
  sources like "EasySpin" (\emph{Matlab}), "Xenon" (EPR spectrometer), "SimFonia" (WinEPR system)
  or "csv" (comma separated values, universal format or MS Excel). Finally, they are automatically converted
  into data frames by the \code{\link[data.table]{fread}}.
}
\examples{
\dontrun{
readEPR_Sim_Spec(path_to_ASC =
                  "./Simulations/TEMPO_simulation.txt",
                 origin.sim = "xenon")
#
readEPR_Sim_Spec("Cu_complex_simulation.txt",
                 B.unit = "mT",
                 col.names.sim = c("Bsim_G","dIeprSim_over_dB"),
                 x.sim.id = 1,
                 Intensity.sim.id = 2,
                 origin.sim = "easyspin")
}



}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
