% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_solvent_props.R
\name{readEPR_solvent_props}
\alias{readEPR_solvent_props}
\title{Reading Solvent Properties from `solvent_db` Data Frame (Database)}
\usage{
readEPR_solvent_props(solvent, prop = NULL)
}
\arguments{
\item{solvent}{Character string pointing to solvent name (or any string from solvent name/abbreviation),
e.g. like \code{solvent = "DMSO"},\code{solvent = "acetone"}, \code{solvent = "xylene"}.
If more than one rows/observations are being returned (e.g. in case of \code{solvent = "xylene"}) => additional
solvent specification must be provided like e.g. \code{solvent = "o-xylene"}.}

\item{prop}{Character string...tbc...}
}
\value{
description...tbc...
}
\description{
A short description...tbc...
}
\examples{
## Properties of `DMSO`
print(tibble::as_tibble(readEPR_solvent_props("DMSO")),
      width = 80,n = 40)
#
## All `xylene` solvent specifications
print(tibble::as_tibble(readEPR_solvent_props(solvent = "xylene")),
      width = 80,n = 40)
#
## Boiling point of `o-xylene`
readEPR_solvent_props(solvent = "o-xylene",
                      prop = "boiling")



}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
