% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPRparam_slct_val.R
\name{readEPRparam_slct_val}
\alias{readEPRparam_slct_val}
\title{TODO}
\usage{
readEPRparam_slct_val(path_to_DSC, string)
}
\arguments{
\item{path_to_DSC}{TODO}

\item{string}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
