% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPRparams_slct_kin.R
\name{readEPRparams_slct_kin}
\alias{readEPRparams_slct_kin}
\title{Read the \strong{selected} Instrumental Parameters from \code{.DSC} or \code{.par} to Record
  the EPR Spectra (within the \code{Xenon} or \code{WinEpr} Software, respectively), \strong{
  required for time correction} (\code{\link{time_correct_EPRspecs}}) of the "kinetic" series}
\usage{
readEPRparams_slct_kin(path_to_DSC_or_par, origin = "xenon")
}
\arguments{
\item{path_to_DSC_or_par}{String, path to \code{.DSC} or \code{.par} file including all instrumental
parameters provided by the EPR machine}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}
}
\value{
List containing number of scans (\code{Nscans}) and sweep time (\code{sweepTime}) required
  for time correction during the \code{2D_Field_Delay} experiment ("kinetic" series)
}
\description{
tbc
}
\examples{
tbc
tbc

}
