% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPRparams_vals.R
\name{readEPRparams_vals}
\alias{readEPRparams_vals}
\title{TODO}
\usage{
readEPRparams_vals(path_to_DSC_or_par, origin = "xenon")
}
\arguments{
\item{path_to_DSC_or_par}{TODO}

\item{origin}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
