% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExpEPRspecs.R
\name{readExpEPRspecs}
\alias{readExpEPRspecs}
\title{Read the Experimental ASCII (\code{.txt}, \code{.csv} or \code{.asc}) EPR Spectrum and Transfer it into Data Frame}
\usage{
readExpEPRspecs(
  path_to_ASC,
  qValue = 1,
  Ns = 1,
  cM = 1,
  m = 1,
  time.series = FALSE,
  origin = "xenon"
)
}
\arguments{
\item{path_to_ASC}{String, path to ASCII file/table (e.g. \code{.txt}, \code{.csv} or \code{.asc})
with spectral data (\eqn{Intensity vs B}(Field) with additional 'index' and/or 'time' variables)}

\item{qValue}{Numeric, Q value (quality factor, number) displayed at specific \code{dB} by spectrometer.
In case of "Xenon" software the parameter is included in \code{.DSC} file, \strong{default = 1}}

\item{Ns}{Number of scans/sweeps per spectrum, in the case of "Xenon" software, the parameter
is already included for the intensity normalization, \strong{default = 1}}

\item{cM}{Numeric, Concentration of the analyte (e.g. radical) in solution (sample) in mol*dm^{-3},
\strong{default = 1}}

\item{m}{Numeric, weight of the powder sample in mg, \strong{default = 1}}

\item{time.series}{Boolean, whether the input ASCII spectrum comes from the time series experiment
with the additional \code{time} column (ONLY IN CASE of "Xenon" software), \strong{default = FALSE}}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}
}
\value{
Data frame/table consisting of the magnetic flux density column \code{B_mT}
  in millitesla (as well as \code{B_G} in gauss) and the derivative intensity column (\code{dIepr_over_dB})
  in \code{procedure defined unit}
  (see \href{http://www.iupac.org/divisions/VII/VII.C.1/C-NPU_Uppsala_081023_25_minutes_confirmed.pdf}{p.d.u.}),
  which is normalized by the above-described parameters and finally the \code{index} and/or a \code{time}
  (in the case of time series experiment) columns are displayed as well.
}
\description{
Function reads/loads the experimental EPR spectrum recorded by BRUKER spectrometers
  in ASCII format by the \code{\link[data.table]{fread}} function and transforms it into \code{data frame},
  which can be easily processed by other R \code{data science} packages (e.g. by \code{\pkg{tidyverse}} system),
   afterwards. For this purpose a \code{pipe} operator from \code{\pkg{magrittr}} is applied.
  Spectral data are normalized by the common experimental parameters in order to qualitatively compare
  the intensities of several spectra. ASCII files/tables depend on the software used to record the EPR spectra
  on BRUKER spectrometers and are slightly different. This is mirrored by \code{origin} parameter (with \code{"xenon"}
  or \code{"winepr"}).Time series spectra (time evolution of EPR spectra) can be read
  by the \code{time.series = TRUE/FALSE} parameter (ONLY IN CASE of \code{"xenon"} software).
}
\examples{
\dontrun{
readExpEPRspecs(path_to_ASCII_file,qValue = 3500) ## for the spectrum acquired by "xenon"
## for the spectrum acquired by "winepr" (and 20 scans) on a 10 mg powder sample:
readExpEPRspecs(path_to_ASCII_file,qValue = 2000,Ns = 20,m = 10,origin = "winepr")
## if no parameter intensity normalization is required and spectrum
## was recorded by "xenon" software:
readExpEPRspecs(path_to_ASCII_file)
## Example for time series experiment (evolution of EPR spectra in time, e.g. in case of
## EPR spectroelectrochemistry or photochemistry):
readExpEPRspecs(path_to_ASC_file,2800,time.series = T)
}


}
