% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAT_params_file.R
\name{readMAT_params_file}
\alias{readMAT_params_file}
\title{Reading EPR Simulation Parameters and Information from the \emph{MATLAB} \code{.mat} File}
\usage{
readMAT_params_file(path_to_MAT, str.var = NULL, field.var = NULL)
}
\arguments{
\item{path_to_MAT}{Character string, path to \code{.mat} MATLAB file with all variables saved in MATLAB workspace.
The file path can be also defined by the \code{\link[base]{file.path}}.}

\item{str.var}{Character string, denoting structure/variable, which may contain \code{fields}, such as \code{Sys}
and \code{g} \eqn{\Rightarrow} \strong{Sys}.g, respectively. \strong{Default}: \code{str.var = NULL}.}

\item{field.var}{Character string, corresponding to field variable after 'dot', which is available only for certain
structures/variables, see e.g. example above (Sys.\strong{g}). Therefore,
the \strong{default} value is \code{NULL} and the \code{string} \strong{is applied only for structures with fields}.}
}
\value{
Unless the \code{str.var} and/or \code{field.var} are not specified, the output is \code{list} consisting
  of all original parameters/structures from MATLAB file. Otherwise, the function returns either numeric/character
  \code{vector/value} or list depending on \code{class} of the original parameter/field variable.
}
\description{
Function is based on the \code{\link[R.matlab]{readMat}} and provides the reading of a \code{.mat} (MATLAB workspace)
  simulation (+ fitting) file content from \emph{EasySpin}, including structures/variables and fields.
  It can be also used to read and store simulated EPR spectrum in the form of R data frame (see \code{Examples}).
  In order to present the simulated/fitted spectra, represented by such a data frame, please refer
  to the \code{\link{present_EPR_Sim_Spec}} function.
}
\examples{
## loading the package built-in
## `Aminoxyl_radical_a.mat` file as an example
aminoxyl.mat.file <-
  load_data_example(file = "Aminoxyl_radical_a.mat")
#
## reading the entire `mat` file as list
## and assign variable
aminoxyl.mat.list <-
  readMAT_params_file(aminoxyl.mat.file)
#
## structure of the `fit1`
str(aminoxyl.mat.list$fit1)
#
## read the `Sim1` structure/variable content into list
aminoxyl.mat.sim1 <-
  readMAT_params_file(aminoxyl.mat.file,
                      str.var = "Sim1")
#
## list preview
aminoxyl.mat.sim1
#
## compare the previous simulation parameters with
## those obtained by the `eval_sim_EPR_isoFit()`
## function (look at the corresponding examples)
#
## alternatively the `Sim1` (its dimension > 2)
## can be also read by the following command,
## however the returned output has a complex
## array-list structure
aminoxyl.mat.list$Sim1[, , 1]
#
## read the `Sim1` structure/variable
## and the field `Nucs` corresponding to nuclei
## considered in the EPR simulation
aminoxyl.mat.sim1.nucs <-
  readMAT_params_file(aminoxyl.mat.file,
                      str.var = "Sim1",
                      field.var = "Nucs")
#
## preview
aminoxyl.mat.sim1.nucs
#
## reading the magnetic flux density
## `B` column/vector corresponding to simulated
## and experimental EPR spectrum
aminoxyl.B.G <-
  readMAT_params_file(aminoxyl.mat.file,
                      str.var = "B")
#
## preview of the first 6 values
aminoxyl.B.G[1:6]
#
## reading the intensity related to simulated/fitted
## EPR spectrum
aminoxyl.sim.fitSpec <-
  readMAT_params_file(aminoxyl.mat.file,
                      str.var = "fit1",
                      field.var = "fitSpec")
#
## for the newer EasySpin version (> 6.0.0),
## the "fitSpec" is replaced by the simple "fit"
## or "fitraw", corresponding to scaled and raw intensity
## of the simulated EPR spectrum, please refer also
## to the EasySpin documentantion:
## https://easyspin.org/easyspin/documentation/esfit.html,
## or to the example below
#
## preview of the first 6 values
aminoxyl.sim.fitSpec[1:6]
#
## In the following example load
## the simulated/fitted EPR spectrum
## by the EasySpin from `mat` workspace file =>
simulation.aminoxyl.spectr.df <-
  data.frame(Bsim_G = aminoxyl.B.G,
             dIeprSim_over_dB = aminoxyl.sim.fitSpec) \%>\%
             dplyr::mutate(Bsim_mT = Bsim_G * 0.1)
#
## preview
head(simulation.aminoxyl.spectr.df)
#
## load newer EasySpin (> 6.0.0) workspace file
## related to TMPD radical cation
tmpd.easyspin6.path <-
  load_data_example(file = "TMPD_specelchem_accu_b.mat")
#
## create a vector of scaled intensity
## of the simulated/fitted EPR spectrum
tmpd.easyspin6.fit <-
  readMAT_params_file(
    tmpd.easyspin6.path,
    str.var = "fit1",
    field.var = "fit"
  )
#
## preview of the first 6 values
tmpd.easyspin6.fit[1:6]
#
## load all simulation fitted arguments from
## the `.mat` workspace
tmpd.easyspin6.argsfit <-
  readMAT_params_file(
    tmpd.easyspin6.path,
    str.var = "fit1",
    field.var = "argsfit"
  )
#
## create lists (temporary variables)
list.argsfit.a <- tmpd.easyspin6.argsfit[[1]][[1]][ , , 1]
list.argsfit.b <- tmpd.easyspin6.argsfit[[2]][[1]][ , , 1]
#
## final list
list.argsfit.final <-
  do.call(c,list(list.argsfit.a,list.argsfit.b))
#
## convert it into nicely structured lists with
## all arguments/parameters
list.argsfit.final <-
  lapply(list.argsfit.final, c)
#
## preview
list.argsfit.final



}
\seealso{
Other Data Reading:
\code{\link[=readEPR_Exp_Specs]{readEPR_Exp_Specs()}},
\code{\link[=readEPR_Exp_Specs_kin]{readEPR_Exp_Specs_kin()}},
\code{\link[=readEPR_Exp_Specs_multif]{readEPR_Exp_Specs_multif()}},
\code{\link[=readEPR_Sim_Spec]{readEPR_Sim_Spec()}},
\code{\link[=readEPR_param_slct]{readEPR_param_slct()}},
\code{\link[=readEPR_params_slct_kin]{readEPR_params_slct_kin()}},
\code{\link[=readEPR_params_slct_quant]{readEPR_params_slct_quant()}},
\code{\link[=readEPR_params_slct_sim]{readEPR_params_slct_sim()}},
\code{\link[=readEPR_params_tabs]{readEPR_params_tabs()}},
\code{\link[=readEPR_solvent_props]{readEPR_solvent_props()}}
}
\concept{Data Reading}
