% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMATfile_params.R
\name{readMATfile_params}
\alias{readMATfile_params}
\title{TODO}
\usage{
readMATfile_params(path_to_MAT, str.var, var.field = NULL)
}
\arguments{
\item{path_to_MAT}{TODO}

\item{str.var}{TODO}

\item{var.field}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
