% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSimEPRspec.R
\name{readSimEPRspec}
\alias{readSimEPRspec}
\title{Read the Simulated ASCII (\code{.txt}) EPR Spectrum from \emph{MATLAB} and Transfer it into Data Frame}
\usage{
readSimEPRspec(path_to_ASC)
}
\arguments{
\item{path_to_ASC}{String, path to ASCII file/table (\code{.txt})
with spectral data (\eqn{Intensity vs B}(Field) obtained from \emph{MATLAB})}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO

}
