% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_correct_EPRspecs.R
\name{time_correct_EPRspecs}
\alias{time_correct_EPRspecs}
\title{Correction for EPR spectral Time Series Providing the Accurate Time for EPR Spectral Line Appearance}
\usage{
time_correct_EPRspecs(time, N_scans, sweep.time)
}
\arguments{
\item{time}{Vector/Column in \code{data frame} corresponding to \code{time} at which
the individual EPR spectra were recorded}

\item{N_scans}{Numeric, number of accumulation (number of scans \code{AVGS}) for each spectrum
in EPR time series}

\item{sweep.time}{Numeric, time for recording of one EPR spectrum \eqn{=} one accumulation}
}
\value{
Numeric value/vector corresponding to accurate time at which the EPR spectra were recorded during
  the kinetic measurements (e.g. like radical stability, electrochemical and/or photochemical measurements)
}
\description{
TODO

 The time is recorded according to the following scheme:

 \tabular{cccccc}{

           \tab     EPR Spectr.     \tab                   \tab     \tab             \tab   EPR Spectr. \cr

 \code{t[1]} \tab  ------^v------> \tab \code{t[2]}-delay \tab ---->  \tab \code{t[2]} \tab ------^v-------> ... \cr

        \tab  \code{N_scans}x\code{swt}  \tab          \tab \code{delay} \tab        \tab \code{N_scans}x\code{swt}
}
The recorded times are: \code{t[1]},\code{t[2]},\code{t[3]},...
and \code{N_scans} corresponds to \code{number of scans} and \code{swt} to \code{sweep time} for individual scan
}
\examples{
TODO
TODO


}
