% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_correct_EPRspectra.R
\name{time_correct_EPRspectra}
\alias{time_correct_EPRspectra}
\title{TODO}
\usage{
time_correct_EPRspectra(time, N_scans, sweep.time)
}
\arguments{
\item{time}{TODO}

\item{N_scans}{TODO}

\item{sweep.time}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
