% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dfs_2tidyDF.R
\name{transform_dfs_2tidyDF}
\alias{transform_dfs_2tidyDF}
\title{Transformation of Several Data Frame Objects into one Tidy (Long Table) Form}
\usage{
transform_dfs_2tidyDF(
  ...,
  df.names,
  which.coly.norm = "dIepr_over_dB",
  norm.vec = rep(1, times = length(df.names)),
  var2nd.series = "Spectrum"
)
}
\arguments{
\item{...}{Data frame object names (separated by comma \code{,}) to be transformed into long
table/tidy form in order to quickly compare plots/EPR spectra in one graph panel (see also \code{Examples}).
The number of data frame objects is not limited, however for the series of \eqn{\geq 10-12}
objects/spectra, the graph may look cluttered, depending on the complexity of EPR spectra.}

\item{df.names}{Character string vector of labels (in the form of e.g. \code{c("Spectr_01","Spectr_02")}
or \code{c("283 K","273 K","263 K")}), describing the individual data frames/plots in the returned
tidy/long table and corresponding to series of plots/spectra (be aware of the names order,
see also the \code{norm.vec} argument).}

\item{which.coly.norm}{Character string, pointing to name of the column (in all data frame
objects) to be normalized, by the \code{norm.vec} vector. This column actually corresponds to quantity/variable
presented on graph as \emph{y}-axis. \strong{Default}: \code{which.coly.norm = "dIepr_over_dB"} (that is the
derivative intensity in CW EPR spectra).}

\item{norm.vec}{Numeric vector, consisting of (division) normalization constants (see also
the \code{which.coly.norm} argument) for each of the data frame objects defined by the ellipsis
\code{...} argument. \strong{Default}: \code{norm.vec = rep(1, times = length(df.names))}, i.e. all normalization
constants equal to \code{1}. If needed, the vector may be redefined (please, be aware of the vector elements order,
see also the \code{df.names} argument).}

\item{var2nd.series}{Character string, pointing to name of the second variable which is "common denominator"
for the series of \code{df.names} vector elements like \code{var2nd.series = "Temperature"}
or \code{var2nd.series = "Spectrum"} (\strong{default}).}
}
\value{
Tidy (long form) data frame/table object, carrying all the individual inputs, defined by the \code{...}
  argument and ready for the overlay/offset plot to compare the data in the desired series.
}
\description{
Taking several data frame objects from the (local/Global) R environment and transforming them into
  one with the \href{https://r4ds.hadley.nz/data-tidy.html}{long table/tidy form}. This is especially handy
  if a quick comparison of similar plots/EPR spectra in one graph panel is required. Difference between
  the \code{\link{readEPR_Exp_Specs_multif}} and the actual function lies in the object input/origin.
  While the \code{\link{readEPR_Exp_Specs_multif}} takes the original files/data, coming from spectrometer,
  the \code{transform_dfs_2tidyDF} collects temporary processed/stored data frame objects/variables
  created meanwhile in the R environment.
}
\note{
In order to compare \eqn{g_{\text{iso}}}-values of different EPR spectra, prior to own transformation
  check that all data frame objects contain \code{g_Val(ue)} variable/column. If this is not case,
  the user may create those columns by the \code{\link{eval_gFactor}} function.
}
\examples{
## compare TMPD*+ EPR spectrum with that of aminoxyl,
## first, define the path and variable for TMPD*+ data
tmpd.data.path <-
  load_data_example(file = "TMPD_specelchem_accu_b.asc")
tmpd.data.df <-
  readEPR_Exp_Specs(
    tmpd.data.path,
    col.names = c("B_G","dIepr_over_dB"),
    qValue = 3500,
    origin = "winepr"
   )
#
## preview
head(tmpd.data.df)
#
## second, do the same for the  aminoxyl data
aminoxyl.data.path <-
  load_data_example(file = "Aminoxyl_radical_a.txt")
aminoxyl.data.df <-
  readEPR_Exp_Specs(
    aminoxyl.data.path,
    qValue = 2100
  )
#
## preview
aminoxyl.data.df \%>\% head()
#
## ...and the own transformation
spectra.data.tidy.df <-
  transform_dfs_2tidyDF(
    tmpd.data.df,
    aminoxyl.data.df,
    df.names = c("TMPD","Aminoxyl"),
    norm.vec = c(5e+3,1e-5),
    var2nd.series = "Radical"
  )
#
## preview
head(spectra.data.tidy.df)
#
## plotting both EPR spectra together
plot_EPR_Specs(
  data.spectra = spectra.data.tidy.df,
  x = "B_G",
  x.unit = "G",
  var2nd.series = "Radical",
  var2nd.series.slct.by = 1,
  xlim = c(3400,3600),
  line.colors = c("blue2","darkorange"),
  legend.title = "Radical"
)
#
## ...and the interactive preview (B in mT)
plot_EPR_Specs2D_interact(
  data.spectra = spectra.data.tidy.df,
  line.colors = c("blue2","darkorange"),
  var2nd.series = "Radical",
  legend.title = "Radical"
)


}
