% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_MHz_conv_to_mT.R
\name{A_MHz_conv_to_mT}
\alias{A_MHz_conv_to_mT}
\title{Convert Coupling Constants (\eqn{A}, HFCCs in MHz) to Hyperfine Splittings (\eqn{a}, HFSCs in mT).}
\usage{
A_MHz_conv_to_mT(A.MHz, g = 2.0023193)
}
\arguments{
\item{A.MHz}{Numeric value/vector corresponding to HFCCs in \code{MHz}}

\item{g}{Numeric value/vector corresponding to actual \eqn{g}-factor (\code{unitless}),
\strong{default:} \code{g = 2.00231930} (free electron)}
}
\value{
Numeric value/vector corresponding to HFSCs (\eqn{a}) in \code{mT})
}
\description{
Converts hyperfine coupling constants HFCCs (\eqn{A} values in \code{MHz})
to hyperfine splitting constants HFSCs (\eqn{a} values in \code{mT})
}
\examples{
A_MHz_conv_to_mT(A.MHz = 16)
A_MHz_conv_to_mT(20,2.0059)
A_MHz_conv_to_mT(4,g = 2.00036)


}
