% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_MHz_conv_to_mT.R
\name{A_MHz_conv_to_mT}
\alias{A_MHz_conv_to_mT}
\title{Conversion of Hyperfine Coupling Constants (\code{HFCCs}, \eqn{A} in \code{MHz}) to Hyperfine Splitting
  ones (\code{HFSCs}, \eqn{a} in \code{mT})}
\usage{
A_MHz_conv_to_mT(A.MHz, g)
}
\arguments{
\item{A.MHz}{Numeric value/Numeric vector corresponding to HFCCs in \code{MHz}}

\item{g}{Numeric value/Numeric vector corresponding to actual \eqn{g}-factor (\code{unitless})}
}
\value{
Numeric Value/Vector corresponding to HFSCs (\eqn{a}) in \code{mT}
}
\description{
TODO
}
\examples{
TODO
TODO


}
