% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_mT_convert_to_MHz.R
\name{a_mT_convert_to_MHz}
\alias{a_mT_convert_to_MHz}
\title{Conversion of Hyperfine Splitting Constants (\code{HFSCs}, \eqn{a} in \code{mT}) to Hyperfine Coupling
  ones (\code{HFCCs}, \eqn{A} in \code{MHz})}
\usage{
a_mT_convert_to_MHz(a, g)
}
\arguments{
\item{a}{Numeric value/Numeric vector corresponding to HFSCs in \code{mT} ('line distances' from spectrum)}

\item{g}{Numeric value/Numeric vector corresponding to actual \eqn{g}-factor (\code{unitless})}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
