% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_A_MHz_2a.R
\name{convert_A_MHz_2a}
\alias{convert_A_MHz_2a}
\title{Convert Coupling Constants (\eqn{A} in MHz) to Splitting ones (\eqn{a} in mT).}
\usage{
convert_A_MHz_2a(A.MHz, g = 2.0023193)
}
\arguments{
\item{A.MHz}{Numeric value/vector corresponding to HFCCs in \code{MHz}}

\item{g}{Numeric value/vector corresponding to actual \eqn{g}-factor (\code{unitless}).
\strong{Default:} \code{g = 2.00231930} ( corresp. to free electron)}
}
\value{
Numeric value/vector corresponding to HFSCs (\eqn{a}) in \code{mT})
}
\description{
Converts hyperfine coupling constants HFCCs (\eqn{A} values in \code{MHz})
  to hyperfine splitting constants HFSCs (\eqn{a} values in \code{mT}) according to following
  relation:
  \deqn{a = \frac{A\,h}{g\,\mu_{\text{B}}}}
}
\examples{
convert_A_MHz_2a(A.MHz = 16)
convert_A_MHz_2a(20,2.0059)
convert_A_MHz_2a(4,g = 2.00036)


}
