% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_A_MHz_2a.R
\name{convert_A_MHz_2a}
\alias{convert_A_MHz_2a}
\title{Convert Coupling Constants (\eqn{A} in MHz) into Splitting ones (\eqn{a} in mT).}
\usage{
convert_A_MHz_2a(A.MHz, g.val = 2.0023193)
}
\arguments{
\item{A.MHz}{Numeric value/vector corresponding to HFCCs in `MHz`.}

\item{g.val}{Numeric value/vector corresponding to actual \eqn{g}-factor (\code{unitless}).
\strong{Default:} \code{g.val = 2.00231930} (corresponding to free electron).}
}
\value{
Numeric value/vector corresponding to HFSCs (\eqn{a}) in `mT`.
}
\description{
Converting hyperfine coupling constants (HFCCs, \eqn{A} values in `MHz`)
  into hyperfine splitting constants (HFSCs, \eqn{a} values in `mT`).
}
\details{
Conversion done according to following relation:
  \deqn{a = A\,h / (g\,\mu_{\text{B}})}
  where \eqn{h} corresponds to Planck's constant and \eqn{\mu_{\text{B}}} to Bohr's magneton.
  Both latter were obtained by \code{constans::syms$h} and \code{constants::syms$mub}, respectively
  using the \pkg{constants} package (see \code{\link[constants]{syms}}). Conversion is suitable
  for EPR simulations and/or ENDOR.
}
\examples{
convert_A_MHz_2a(A.MHz = 16)
#
convert_A_MHz_2a(20,2.0059)
#
convert_A_MHz_2a(4,g.val = 2.00036)


}
\seealso{
Other Conversions and Corrections: 
\code{\link{convert_B}()},
\code{\link{convert_a_mT_2A}()},
\code{\link{convert_time2var}()},
\code{\link{correct_time_Exp_Specs}()}
}
\concept{Conversions and Corrections}
