% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_a_mT_2A.R
\name{convert_a_mT_2A}
\alias{convert_a_mT_2A}
\title{Convert Splitting Constants (\eqn{a} in mT) to Coupling ones (\eqn{A} in MHz).}
\usage{
convert_a_mT_2A(a.mT, g = 2.0023193)
}
\arguments{
\item{a.mT}{Numeric value/vector of HFSCs in \code{mT} ('line distances' from EPR spectrum)}

\item{g}{Numeric value/vector corresponding to actual \eqn{g}-factor (\code{unitless}).
\strong{Default:} \code{g = 2.00231930} (corresp. to free electron)}
}
\value{
Numeric value/vector corresponding to HFCCs (\eqn{A}) in \code{MHz})
}
\description{
Converts hyperfine splitting constants HFSCs (\eqn{a} values in \code{mT})
  to hyperfine coupling constants HFCCs (\eqn{A} values in \code{MHz}) according to following
  relation:
  \deqn{A = \frac{a\,g\,\mu_{\text{B}}}{h}}
}
\examples{
convert_a_mT_2A(a.mT = 0.5)
convert_a_mT_2A(0.6,2.0059)
convert_a_mT_2A(0.15,g = 2.00036)


}
