% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time2var.R
\name{convert_time2var}
\alias{convert_time2var}
\title{Convert Time (\eqn{t}) into Variable Linearly Depending on \eqn{t}}
\usage{
convert_time2var(time, time.unit = "s", var0, var.switch = NULL, var.rate)
}
\arguments{
\item{time}{Numeric, value or vector corresponding to time (points) where the variable \code{var}
is changed.}

\item{time.unit}{Character string \strong{time unit} defined by \code{"s"},\code{"min"} or \code{"h"}.
\strong{Default}: \code{time.unit = "s"}.}

\item{var0}{Numeric, the initial/starting value (\strong{INCL. ALSO NEGATIVE SIGN, if required}, e.g. like
negative electrochemical potential) of the variable (\code{var}).}

\item{var.switch}{Numeric, the switching point \code{var} value, when a linear \strong{cyclic change}
(or 'triangular ramp') of \code{var} upon time is applied (e.g. like cyclic voltammetry).
\strong{Default}: \code{var.switch = NULL} (in case there is no such cyclic change).}

\item{var.rate}{Numeric, corresponding to rate of linear \code{var} change (\strong{INCL. ALSO NEGATIVE SIGN,
if required}, e.g. like in case of electrochemical reduction or sample cooling).}
}
\value{
Numeric value or vector of variable (e.g. like electrochem. potential or temperature...etc)
  linearly changing upon time.
}
\description{
Conversion of time into variable \code{var} which is linearly changed upon time like
\deqn{var = var0 + rate\times t}
This is especially suitable for time conversion of EPR time series experiments
(see e.g. \code{\link{readEPR_Exp_Specs_kin}}) simultaneously performed either during
electrochemical/voltammetric or variable temperature experiment.
}
\examples{
## Calculate Potential after 30 s, starting from 200 mV
## into cathodic direction (reduction) by 5 mV/s
convert_time2var(30,var0 = 0.2,var.rate = - 0.005)
#
## Heating sample after 5 min starting from 293 K
## by the temperature rate of 0.5 K/s
convert_time2var(5,
                 time.unit = "min",
                 var0 = 293,
                 var.rate = 0.5)
\dontrun{
## Create/Evaluate column in EPR time series data frame
## containing the applied cell potential (in V) from the simultaneously
## performed electrochemical reduction experiment (e.g. cyclic voltammetry).
## Time series column is labeled as "time_s"
data.spectra.series <- data.spectra.series \%>\%
  dplyr::mutate(E_V_potential = convert_time2var(time_s,
                                                 time.unit = "s",
                                                 var0 = -0.1,
                                                 var.switch = -1.5,
                                                 var.rate = -0.004))
}


}
\seealso{
Other Conversions and Corrections: 
\code{\link{convert_A_MHz_2a}()},
\code{\link{convert_a_mT_2A}()},
\code{\link{correct_time_Exp_Specs}()}
}
\concept{Conversions and Corrections}
