% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_time_Exp_Specs.R
\name{correct_time_Exp_Specs}
\alias{correct_time_Exp_Specs}
\title{Time Correction for Experimental CW EPR Spectral Time Series}
\usage{
correct_time_Exp_Specs(time.s, Nscans, sweep.time.s)
}
\arguments{
\item{time.s}{Numeric value/vector/column in \code{data frame} corresponding to \code{time} (in \code{s})
at which the individual EPR spectra were recorded (supplied by the EPR acquisition software).}

\item{Nscans}{Numeric, number of accumulations (number of scans usually denoted as \code{AVGS})
for each spectrum in EPR time series.}

\item{sweep.time.s}{Numeric, time (in \code{s}) for recording individual EPR spectrum \eqn{=} one accumulation.}
}
\value{
Numeric value/vector corresponding to accurate time at which the middle (on \eqn{B} abscissa) EPR spectra
  were recorded during the kinetic measurements (e.g. like radical formation, stability, electrochemical
  and/or photochemical measurements).
}
\description{
Providing the accurate time for EPR spectral line/spectrum appearance. Usually the middle of the EPR spectrum is
 is set up as the CF (`central field`) for the spectrum sweep.
}
\details{
The actual time at the middle/crossing point
 is different from that recorder by the EPR acquisition software, see below. This is especially important
 in determining the kinetics of radical generation or decay.
 The time is recorded according to the following scheme. "^v" in the scheme denotes the derivative form
 of an EPR spectrum:

 \tabular{cccccc}{

           \tab     EPR Spectr.     \tab                   \tab     \tab             \tab   EPR Spectr. \cr

 \code{t[1]} \tab  ------^v------> \tab \code{t[2]}-delay \tab ---->  \tab \code{t[2]} \tab ------^v-------> ... \cr

        \tab  \code{N_scans}*\code{swt}  \tab          \tab \code{delay} \tab        \tab \code{N_scans}*\code{swt}
}
The recorded times are: \code{t[1]},\code{t[2]},\code{t[3]},...
and \code{N_scans} corresponds to \code{number of scans} and \code{swt} to \code{sweep time} for individual scan.
These parameters can be obtained by \code{\link{readEPR_params_slct_kin}} or other functions which can read
the instrumental parameters.
}
\examples{
## 12 s recorded by spectrometer, 6 accumulations by sweep time of 6 s
correct_time_Exp_Specs(12,Nscans = 6,6)


}
\seealso{
Other Conversions and Corrections: 
\code{\link{convert_A_MHz_2a}()},
\code{\link{convert_B}()},
\code{\link{convert_a_mT_2A}()},
\code{\link{convert_time2var}()}
}
\concept{Conversions and Corrections}
