% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_molecules_rcdk.R
\name{draw_molecules_rcdk}
\alias{draw_molecules_rcdk}
\title{Draw Molecule by `rcdk` Defined by SMILES}
\usage{
draw_molecules_rcdk(
  molecules,
  type = "smiles",
  mol.names = NULL,
  sma = NULL,
  annotate = "off",
  style = "cow",
  abbr = "off",
  supressh = TRUE,
  ...
)
}
\arguments{
\item{molecules}{Character vector => either vector of molecular SMILES strings (in case
of \code{type = "smiles"}/\code{type = "SMILES"}) or vector of file path strings
(generated by e.g. \code{\link{list.files}}) pointing to \code{".sdf"} files.}

\item{type}{Character string ...tbc... either \code{type = "smiles"}/\code{type = "SMILES"}
or \code{type = "sdf"}/\code{type = "SDF"}. \strong{Default}: \code{type = "smiles"}.}

\item{mol.names}{Character vector pointing to name of molecules/compounds, e.g. like
\code{mol.names = c("acetone")} or \code{mol.names = c("PBN","DMPO")}.}

\item{sma}{Character string allowing to highlight sub-structures using `SMARTS`
(\strong{SM}ILES \strong{A}rbitrary \strong{T}arget \strong{S}pecification) to highlight
common substructures in a set of molecules, e.g. like \code{sma = "C=O"}.}

\item{annotate}{Character string to interpret the molecular structure elements like =>
\tabular{rcl}{
---------------- \tab | \tab ----------------------------- \cr
\strong{Mol. Structure Element}  \tab | \tab \strong{Annotate Text String} \cr
---------------- \tab | \tab ----------------------------- \cr
`atom numbers` \tab | \tab \code{annotate = "number"} \cr
`atom mapping` \tab | \tab \code{annotate = "mapidx"} \cr
`color map` \tab | \tab \code{annotate = "colmap"} \cr
`atom value` \tab | \tab \code{annotate = "atomvalue"} \cr
---------------- \tab | \tab ----------------------------- \cr
}
\strong{Default}: \code{annotate = "off"}.}

\item{style}{Character string denoting the plotting style like =>
\tabular{rcl}{
---------------- \tab | \tab ----------------------------- \cr
\strong{Plotting Style}  \tab | \tab \strong{Style Text String} \cr
---------------- \tab | \tab ----------------------------- \cr
`color on black` \tab | \tab \code{style = "cob"} \cr
`color on white` \tab | \tab \code{style = "cow"} \cr
`black on white` \tab | \tab \code{style = "bow"} \cr
`white on black` \tab | \tab \code{style = "wob"} \cr
`neon on black` \tab | \tab \code{style = "nob"} \cr
---------------- \tab | \tab ----------------------------- \cr
}
\strong{Default}: \code{style = "cow"}.}

\item{abbr}{Character string which controls how the structure is displayed. Following options
can be set => \code{abbr = "off"} (\strong{default}) pointing to present structure as is;
\code{abbr = "groups"} creating an abbreviation for `groups`; \code{abbr = "reagents"} creating
an abbreviation for `reagents` or \code{abbr = "on"} to abbreviate the both latter.}

\item{supressh}{Logical, denoting whether to suppress displaying the hydrogen atoms.
\strong{Default}: \code{supressh = TRUE}.}

\item{...}{Additional options/arguments for \code{\link[rcdk]{get.depictor}} or displaying
the name of compound(s) like color, font, ...etc (see \code{\link[graphics]{text}}).}
}
\value{
tbc
}
\description{
A short description...
}
\examples{
\dontrun{
tbc
tbc
}

}
\seealso{
Other Visualization and Graphics: 
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{presentEPR_Sim_Spec}()}
}
\concept{Visualization and Graphics}
