% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_molecules_rcdk.R
\name{draw_molecules_rcdk}
\alias{draw_molecules_rcdk}
\title{Draw Molecule by `rcdk` Defined by SMILES}
\usage{
draw_molecules_rcdk(molecules, type = "smiles", name = NULL, sma = NULL, ...)
}
\arguments{
\item{molecules}{Character vector => either vector of molecular SMILES strings (in case
of \code{type = "smiles"}/\code{type = "SMILES"}) or vector of file path strings
(generated by e.g. \code{\link{list.files}}) pointing to \code{".sdf"} files}

\item{type}{Character string ...tbc... either \code{type = "smiles"}/\code{type = "SMILES"}
or \code{type = "sdf"}/\code{type = "SDF"}}

\item{name}{tbc}

\item{sma}{tbc}

\item{...}{tbc}
}
\value{
tbc
}
\description{
A short description...
}
\examples{
\dontrun{
tbc
tbc
}

}
\seealso{
Other Visualization and Graphics: 
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{presentEPR_Sim_Spec}()}
}
\concept{Visualization and Graphics}
