% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_extremeX_Spec.R
\name{eval_extremeX_Spec}
\alias{eval_extremeX_Spec}
\title{Find Intensity Extremes within the EPR/ENDOR Spectrum (incl. Integrated Form).}
\usage{
eval_extremeX_Spec(
  data.spectrum,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  xlim = NULL,
  extreme = "max"
)
}
\arguments{
\item{data.spectrum}{EPR/ENDOR spectrum data frame/table with magnetic flux density \eqn{B} (in \code{mT} or \code{G})
or \eqn{g}-Value or \eqn{RF} (in \code{MHz}) column and that of the derivative \code{dIepr_over_dB}
or integrated \code{Intensity}. \code{Index} column may be included as well.}

\item{x}{Character/String pointing to \code{x}-axis/column quantity like magnetic flux density \eqn{B}, \eqn{g}-Value
or \eqn{RF} (radio frequency), \strong{default}: \code{x = "B_mT"}}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{xlim}{Numeric vector corresponding to border limits of the selected \eqn{x} region,
e.g. like `xlim = c(3495.4,3595.4)` (\eqn{B} in \code{G}) or `xlim = c(12.5,21.2)` (\eqn{RF} in \code{MHz})
or `xlim = c(2.004,2.001)` (\eqn{g} dimensionless). \strong{Default}: \code{xlim = NULL} (corresponding
to entire `x` range)}

\item{extreme}{Character/String with only a two values allowed: \code{"min"} or \code{"max"} (\strong{default})}
}
\value{
Numeric value of \code{x}-axis quantity like \eqn{B},\eqn{g},\eqn{RF} corresponding
 to \code{minimal} or \code{maximal} \code{intensity} within the EPR/ENDOR spectrum (or its integrated form)
}
\description{
Finding \code{x} positions like \eqn{B} ( in \code{mT} or \code{G}) or \eqn{g}
or \eqn{RF} (in \code{MHz}) of intensity minimum and/or maximum within selected region of the EPR/ENDOR spectrum
}
\examples{
\dontrun{
eval_extremeX_Spec(data.spectrum,
                   xlim = c(349.54,359.54),
                   extreme = 'min')
eval_extremeX_Spec(data.frame,
                   "g_Value",
                   Intensity = "Integrated_Intensity",
                   c(2.007,2.000))
}


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()}
}
\concept{Evaluations}
