% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_extremeX_Spec.R
\name{eval_extremeX_Spec}
\alias{eval_extremeX_Spec}
\title{Find Intensity Extremes within the EPR/ENDOR Spectrum (incl. Integrated Form).}
\usage{
eval_extremeX_Spec(
  data.spectr,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  xlim = NULL,
  extreme = "max"
)
}
\arguments{
\item{data.spectr}{EPR/ENDOR spectrum data frame object with magnetic flux density \eqn{B} (in \code{mT} or \code{G})
or \eqn{g}-Value or \emph{RF} (in \code{MHz}) column/variable and that of the derivative \code{dIepr_over_dB}
or integrated \code{Intensity}. \code{Index} column/variable may be included as well.}

\item{x}{Character string pointing to name of the \code{x}-axis/column/variable (in the original \code{data.spectr})
like magnetic flux density \emph{B}, \emph{g}-Value or \emph{RF} (radio frequency), \strong{default}: \code{x = "B_mT"}.}

\item{Intensity}{Character string pointing to name of the \code{intensity column/variable}
(in the original \code{data.spectr}) if other than \code{dIepr_over_dB} name/label
is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}.}

\item{xlim}{Numeric vector corresponding to border limits of the selected \emph{x} region,
e.g. like \code{xlim = c(3495.4,3595.4)} (\emph{B} in \code{G}) or \code{xlim = c(12.5,21.2)} (\emph{RF} in \code{MHz})
or \code{xlim = c(2.004,2.001)} (\emph{g} dimensionless). \strong{Default}: \code{xlim = NULL} (corresponding
to entire \code{x} range).}

\item{extreme}{Character string with only a two values allowed: \code{"min"} or \code{"max"} (\strong{default}).}
}
\value{
Numeric value of \code{x}-axis variable like \emph{B},\emph{g},\emph{RF} corresponding
 to \code{minimal} or \code{maximal} \code{intensity} within the EPR/ENDOR spectrum or its integrated form.
}
\description{
Finding the \code{x} positions like \emph{B} (in \code{mT} or \code{G})) or \emph{g} (unitless)
  or \emph{RF} (in \code{MHz}) of intensity minimum or maximum within the selected region
  of an EPR/ENDOR spectrum.
}
\examples{
\dontrun{
eval_extremeX_Spec(data.spectr,
                   xlim = c(349.54,359.54),
                   extreme = 'min')
#
eval_extremeX_Spec(data.frame,
                   "g_Value",
                   Intensity = "single_Integ",
                   c(2.007,2.000))
}


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()}
}
\concept{Evaluations}
