% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_FWHMx_Spec.R
\name{eval_FWHMx_Spec}
\alias{eval_FWHMx_Spec}
\title{Evaluating Full Width at Half-Maximum (FWHM) from integrated EPR Spectra}
\usage{
eval_FWHMx_Spec(
  data.spectr.integ,
  x = "B_G",
  Intensity = "single_Integ",
  xlim = NULL
)
}
\arguments{
\item{data.spectr.integ}{Data frame object containing \emph{x} column/variable like magnetic flux density, \emph{B}
(in in `mT` or `G`) or \emph{g}-factor/value (unitless) and integrated intensity
(common absorption-like spectrum) column/variable.}

\item{x}{Character string pointing to name of the \code{x}-axis/column/variable (in the original
\code{data.spectr.integ}) like magnetic flux density \emph{B} (in `mT` or `G`) or \emph{g}-Value
(unitless), \strong{default}: \code{x = "B_G"}.}

\item{Intensity}{Character string pointing to name of the \code{intensity column/variable}
(in the original \code{data.spectr.integ}) if other than \code{single_Integ} (\strong{default}) name/label
is used (e.g. "Integral_Intensity" or "integral").}

\item{xlim}{Numeric vector corresponding to border limits of the selected \eqn{x} region,
e.g. like \code{xlim = c(3495.4,3595.4)} (\emph{B} in `G`) or \code{xlim = c(2.004,2.001)} (\emph{g} dimensionless).
\strong{Default}: \code{xlim = NULL} (corresponding to entire \emph{x} range).}
}
\value{
Numeric value of the FWHM directly from EPR spectrum, depending on \code{x} variable =>
  either in `mT`/`G` or unitless in case if \emph{g}-factor is presented on abscissa.
}
\description{
Finding the full width at half-maximum (FWHM) height of the EPR integrated spectrum/intensity. For such purpose
  the EPR spectrum has to be in single integrated form (common absorption-like spectrum) and not in the derivative,
  e.g. like with \code{dIepr_over_dB} y-axis/variable. If this is not the case, the derivative EPR spectrum
  can be integrated by \code{\link{eval_integ_EPR_Spec}}.
}
\examples{
\dontrun{
tbc
}


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()}
}
\concept{Evaluations}
