% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_FWHMx_Spec.R
\name{eval_FWHMx_Spec}
\alias{eval_FWHMx_Spec}
\title{Evaluating Full Width at Half-Maximum (FWHM) from Spectra}
\usage{
eval_FWHMx_Spec(
  data.spec.integ,
  x = "B_G",
  Intensity = "single_Integ",
  xlim = NULL
)
}
\arguments{
\item{data.spec.integ}{tbc}

\item{x}{tbc}

\item{Intensity}{tbc}

\item{xlim}{tbc}
}
\value{
tbc
}
\description{
A short description...
}
\examples{
\dontrun{
tbc
}


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()}
}
\concept{Evaluations}
