% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gFactors.R
\name{eval_gFactor}
\alias{eval_gFactor}
\title{Basic Calculation of \eqn{g}-factor}
\usage{
eval_gFactor(nu.val, nu.unit = "GHz", B.val, B.unit = "mT")
}
\arguments{
\item{nu.val}{Numeric, microwave Frequency value.}

\item{nu.unit}{Character string, frequency unit defined by \code{GHz} or \code{Hz},
\strong{default}: \code{nu.unit = "GHz"}.}

\item{B.val}{Numeric, magnetic flux density value.}

\item{B.unit}{Character string, magnetic flux density unit in \code{G} or \code{mT} or \code{T},
\strong{default}: \code{B.unit = "mT"}.}
}
\value{
\eqn{g}-Value from \eqn{(\nu h)/(\mu_{\text{B}} B)}. Physical constants (\eqn{\mu_{\text{B}}}
  and \eqn{h}) are taken from \href{https://r-quantities.github.io/constants/}{constants}
  package by the \code{\link[constants]{syms}}.
}
\description{
Calculation of \eqn{g}-factor according to fundamental formula (see \code{Value}).
  The magnetic flux density (\code{B.val}) and microwave frequency (\code{nu.val},\eqn{\nu})
  can be defined, having the common units like \code{G} (Gauss) \code{mT}
  (millitesla) or \code{T} (tesla) as well as \code{GHz} or \code{Hz}.
}
\examples{
eval_gFactor(9.8020458,
             nu.unit = "GHz",
             350.214,
             B.unit = "mT")
#
eval_gFactor(nu.val = 9.8020458e+9,
             nu.unit = "Hz",
             B.val = 3502.14,
             B.unit = "G")
#
eval_gFactor(9.5421,"GHz",0.333251,"T")


}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()},
\code{\link{eval_peakPick_Spec}()}
}
\concept{Evaluations}
