% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gFactors.R
\name{eval_gFactor_QCHcomp}
\alias{eval_gFactor_QCHcomp}
\title{Calculation of \eqn{g}-factor from Quantum Chemical Computational Output}
\usage{
eval_gFactor_QCHcomp(path_to_QCHoutput, mean = TRUE, origin = "gaussian")
}
\arguments{
\item{path_to_QCHoutput}{Character string corresponding to path of "Gaussian" or "ORCA" output text files
including all \eqn{g}-factors. Alternatively, the \code{\link[base]{file.path}} can be applied to get
the full/relative path.}

\item{mean}{Logical, whether to calculate the \code{mean value/iso} from the principal components,
\strong{default}: \code{mean = TRUE}, or return the entire vector with the all 3 components.}

\item{origin}{Character string pointing to origin of the EPR calculation parameters <=> which
software package was used. Only two values are available => \code{"Gaussian"} (\strong{default})
or \code{"ORCA"}.}
}
\value{
Numeric mean \eqn{g}-factor value from the principal difference (from \eqn{g_e}) components
  calculated by QCH the method (e.g. by DFT) or numeric vector with the principal \eqn{g}-components
  if \code{mean = FALSE}.
}
\description{
In the Gaussian and ORCA outputs, the \eqn{g}-Values (3 principal ones) are presented in the form
  of differences from the \eqn{g_e} (\eqn{g} of the free electron). Therefore, the function takes these values
  to calculate the entire \eqn{g}-factor components or parses the corresponding \eqn{g}-mean value
  from the outputs.
}
\examples{
## built-in package file example and path:
gauss.file.path <-
  load_data_example(file = "TMPDAradCatEPRa.inp.log.zip")
gauss.file <- unzip(gauss.file.path)
## g_iso-value calculation from Gaussian output file:
eval_gFactor_QCHcomp(gauss.file)


}
\seealso{
Other Evaluations and Quantum Chemistry: 
\code{\link{rearrange_aAiso_QCHcomp}()},
\code{\link{rearrange_aAiso_QCHorgau}()}
}
\concept{Evaluations and Quantum Chemistry}
