% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_interval_cnfd_tVec.R
\name{eval_interval_cnfd_tVec}
\alias{eval_interval_cnfd_tVec}
\title{Confidence Interval of a Vector or Data Frame Column}
\usage{
eval_interval_cnfd_tVec(data.vec.col, level.cnfd = 0.95, separate = TRUE)
}
\arguments{
\item{data.vec.col}{Numeric vector (pointing to column) of interest (within a data frame)
to calculate the confidence interval or uncertainty (margin of error).}

\item{level.cnfd}{Numeric (floating) value, corresponding to confidence level (\strong{default}:
\code{level.cnfd = 0.95}). This value is related to significance level \eqn{alpha} defined
by \eqn{1 - confidence\,\,level}.}

\item{separate}{Logical, whether to separate the mean value and the uncertainty (margin of error),
corresponding to non-negative right/left confidence limit of the mean. If \code{separate = TRUE}
(\strong{default}), the result is shown as a named vector with the (mean) \code{value}
and the \code{uncertaity}. Otherwise, the result is returned in the format of \eqn{value\pm\,\,uncertainty}.}
}
\value{
Named vector of (mean) \code{value} and \code{uncertaity} or \eqn{value\pm\,\,uncertainty}
  format depending on the \code{separate} argument, where the uncertainty actually represents non-negative
  limit for the mean (one side of the confidence interval not including the mean value).
}
\description{
Calculation of the mean value and its confidence limits (according to Student's t-distribution),
 corresponding to data frame column or vector, characterizing dispersion of the individual
 values such as double integrals in quantitative EPR analysis, g-values or linewidths.
}
\details{
The confidence interval evaluation suggests that values/observations obey two-tailed Student's t-distribution,
  which for number of observations \eqn{N > 30} may approach the normal \eqn{z}-distribution.
  Evaluation of the confidence interval and/or its limits can be well documented on \eqn{g}-factor series example:
  \deqn{g = \overline{g}\pm (t_{(1-\alpha/2),df}\,s/\sqrt{N})}
  where \eqn{\overline{g}} is the mean value and the \eqn{t_{(1-\alpha/2),df}} corresponds to the quantile
  of the t-distribution having the significance level of \eqn{\alpha} (\eqn{1-\alpha = condfidence\,\,level},
  see the \code{level.cnfd} argument) and the degrees of freedom \eqn{df = N -1}. Finally, the \eqn{s} represents
  the sample standard deviation, defined by the following relation (regarding the \eqn{g}-value series example):
  \deqn{s = \sqrt{(1/(N-1))\,\sum_{i=1}^N (g_i - \overline{g})^2}}
  which is computed by the \code{\link[stats]{sd}} function. The above-mentioned \eqn{t}-quantile is actually calculated
  by the \code{\link[stats:TDist]{stats::qt}}. Alternatively, one could also evaluate confidence
  interval by the one sample \code{\link[stats]{t.test}} for a certain level of confidence, giving a descriptive output
  with statistical characteristics.
}
\examples{
## double integral/intensity values
## coming from several experiments:
di.vec <- c(0.025,0.020,0.031,0.022,0.035)
#
## evaluation of the confidence interval
## in different formats:
eval_interval_cnfd_tVec(di.vec)
#
eval_interval_cnfd_tVec(di.vec,
                        level.cnfd = 0.95,
                        separate = FALSE)


}
\references{
Miller JN, Miller JC, Miller RD (2018). \emph{Statistics and Chemometrics for Analytical Chemistry}, 7th edition,
  Pearson Education. ISBN 978-1-292-18674-0, \url{https://elibrary.pearson.de/book/99.150005/9781292186726}.

  Hayes A, Moller-Trane R, Jordan D, Northrop P, Lang MN, Zeileis A (2022). \emph{distributions3: Probability Distributions
  as S3 Objects}. \url{https://github.com/alexpghayes/distributions3}, \url{https://alexpghayes.github.io/distributions3/},
  see also \emph{Vignette: t-confidence interval for a mean},
  \url{https://alexpghayes.github.io/distributions3/articles/one-sample-t-confidence-interval.html}.

  National Institute of Standards and Technology (NIST) (2012). “Confidence Limits for the Mean”,
  \url{https://www.itl.nist.gov/div898/handbook/eda/section3/eda352.htm}.

  Kaleta J, Tarábek J, Akdag A, Pohl R, Michl J (2012). “The 16 CB11(CH3)n(CD3)12–N• Radicals with
  5-Fold Substitution Symmetry: Spin Density Distribution in CB11Me12•”,
  \emph{Inorg. Chem.}, \strong{51}(20), 10819–10824, \url{https://doi.org/10.1021/ic301236s}.

  Curley JP, Milewski TM (2020). “PSY317L Guidebook - Confidence Intervals”,
  \url{https://bookdown.org/curleyjp0/psy317l_guides5/confidence-intervals.html}.

  Goodson DZ (2011). \emph{Mathematical Methods for Physical and Analytical Chemistry}, 1st edition,
  John Wiley and Sons, Inc. ISBN 978-0-470-47354-2, \url{https://onlinelibrary.wiley.com/doi/book/10.1002/9781118135204}.
}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_kinR_Eyring_GHS}()},
\code{\link{eval_nu_ENDOR}()},
\code{\link{eval_peakPick_Spec}()}
}
\concept{Evaluations}
