% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_EPR_modelFit.R
\name{eval_kinR_EPR_modelFit}
\alias{eval_kinR_EPR_modelFit}
\title{Radical Kinetic Models Fitted to Experimental Data}
\usage{
eval_kinR_EPR_modelFit(
  data.integs,
  time.unit = "s",
  time = "time_s",
  qvarR = "Area",
  model.react = "(x=1)R --> [k1] B",
  params.guess = c(qvar0R = 0.001, k1 = 0.001),
  fit.kin.method = "diff-levenmarq",
  time.correct = FALSE,
  path_to_dsc_par = NULL,
  origin = NULL
)
}
\arguments{
\item{data.integs}{Data frame object Integrals/Areas \emph{vs.} Time}

\item{time.unit}{Character string ... argument/parameter... tbc}

\item{time}{Character string ... argument/parameter... tbc}

\item{qvarR}{Character string ... argument/parameter... tbc}

\item{model.react}{Character string ... argument/parameter... tbc, the same like in \code{\link{eval_kinR_ODE_model}}}

\item{params.guess}{Named vector ... argument/parameter... tbc}

\item{fit.kin.method}{Character string ... argument/parameter... tbc}

\item{time.correct}{Logical, ... argument/parameter... tbc}

\item{path_to_dsc_par}{Character string ... argument/parameter... tbc}

\item{origin}{Character string ... argument/parameter... tbc}
}
\value{
List ... tbc ...
}
\description{
A short description...(Integrals/Areas \emph{vs.} Time)...
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Abs}()},
\code{\link{quantify_EPR_Norm_const}()}
}
\concept{Evaluations and Quantification}
