% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_EPR_modelFit.R
\name{eval_kinR_EPR_modelFit}
\alias{eval_kinR_EPR_modelFit}
\title{Radical Kinetic Models Fitted to Experimental Data}
\usage{
eval_kinR_EPR_modelFit(
  data.qt.expr,
  time.unit = "s",
  time = "time_s",
  qvarR = "Area",
  model.react = "(r=1)R --> [k1] B",
  elementary.react = TRUE,
  params.guess = c(qvar0R = 0.001, k1 = 0.001),
  params.guess.lower = NULL,
  params.guess.upper = NULL,
  fit.kin.method = "diff-levenmarq",
  time.correct = FALSE,
  path_to_dsc_par = NULL,
  origin = NULL,
  ...
)
}
\arguments{
\item{data.qt.expr}{A data frame object containing the concentrations/integral intensities/areas under
the EPR spectra calculated using the \strong{experimental data} as well as time column. These two essential
columns are described by the character strings like those below \code{time} and \code{qvarR}.}

\item{time.unit}{Character string corresponding to time unit like \code{"s"} (\strong{default}),
\code{"min"} or \code{"h"}.}

\item{time}{Character string pointing to \code{time} \strong{column/variable name} in the original
\code{data.qt.expr} data frame. \strong{Default}: \code{time = "time_s"}.}

\item{qvarR}{Character string pointing to \code{qvarR} \strong{column/variable name} in the original
\code{data.qt.expr} data frame. \strong{Default}: \code{qvarR = "Area"}.}

\item{model.react}{Character string denoting a specific radical (\code{"R"}) reaction related to
changes in integral intensities (or any other \strong{q}uantitative \strong{var}iable) in EPR spectral
time series. Arrow shows direction of the reaction (\code{"-->", irreversible} or \code{"<==>", reversible}).
Rate constants are indicated by square brackets after the arrows. Following examples of the reaction schemes
are predefined and commonly used to describe the integral intensity and/or radical concentration/amount
changes during the EPR time series experiment (the \code{r,a,b,c} stoichiometric coefficients
may be varied, see below).
\tabular{ll}{
\strong{Reaction Scheme} \tab \strong{model.react} \cr
\eqn{(r=1)\text{R} \xrightarrow{k_1} \text{B}} \tab \code{"(r=1)R --> [k1] B"} \cr
 \eqn{(a=2)\text{A} \xrightarrow{k_1} (r=2)\text{R}} \tab \code{"(a=2)A --> [k1] (r=2)R"} \cr
 \eqn{(a=2)\text{A} \xrightarrow{k_1} (r=2)\text{R} \xrightleftharpoons[k_3]{k_2} (c=1)\text{C}} \tab
 \code{"(a=2)A --> [k1] (r=2)R <==> [k2] [k3] (c=1)C"} \cr
 \eqn{(r=1)\text{R} \xrightleftharpoons[k_2]{k_1} (b=1)\text{B}} \tab
 \code{"(r=1)R <==> [k1] [k2] (b=1)B"} \cr
 \eqn{(a=2)\text{A} \xrightleftharpoons[k_2]{k_1} (r=2)\text{R}} \tab
 \code{"(a=2)A <==> [k1] [k2] (r=2)R"} \cr
 \eqn{(a=1)\text{A} + (b=1)\text{B} \xrightarrow{k_1} (r=1)\text{R}} \tab
 \code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \cr
 \eqn{(r=1)\text{R} + (b=1)\text{B} \xrightarrow{k_1} (c=1)\text{C}} \tab
 \code{"(r=1)R + (b=1)B --> [k1] (c=1)C"} \cr
}
Couple of examples are also given in details. The function is relatively flexible and enables
later addition of any other reaction schemes describing the EPR time series experiments
(YOU MAY ASK DEVELOPER(S) via forum/help-channels). The stoichiometric coefficient (e.g. like \code{(r=1)}
or \code{(a=1)}) can be varied within the \code{model.react} character string.
Defined/Allowed values are integers e.g. 1,2,3...etc. The space character within the \code{model.react}
string is not fixed and can be skipped for the sake of flexibility.
If \code{elementary.react = FALSE} (the model reaction is not considered as an elementary one),
a possible non-integer partial stoichiometric coefficients (like e.g. \code{alpha},\code{beta} or \code{gamma})
must be included in \code{kin.params} (see also \code{kin.params} description).}

\item{elementary.react}{Logical, if the model reaction should be considered as elementary one,
i.e. the stoichiometric coefficients equal to the partial reaction orders. Such reaction proceeds without
identifiable intermediate species forming. \strong{Default}: \code{elementary.react = TRUE}.
If \code{elementary.react = FALSE}, i.e. the \code{model.react} cannot be considered like an elementary one,
one must include the parameterized reaction orders \eqn{\alpha}, \eqn{\beta} or \eqn{\gamma} into
the \code{kin.params}, e.g like \code{kin.params = c(k1 = 0.01, qvar0A = 0.05, alpha = 1.5)}.}

\item{params.guess}{Named vector, initial values of \code{kin.params} (see \code{\link{eval_kinR_ODE_model}})
ready for optimization/fitting.}

\item{params.guess.lower}{Numeric vector of lower bounds on each parameter.
If not given, the \strong{default} (\code{params.guess.lower = NULL}) lower bound
corresponds to \code{-Inf} of each \code{params.guess} component.}

\item{params.guess.upper}{Numeric vector of upper bounds on each parameter.
If not given, the \strong{default} (\code{params.guess.upper = NULL}) upper bound
corresponds to \code{+Inf} of each \code{params.guess} component.}

\item{fit.kin.method}{Character string pointing to optimization/fitting method. So far,
the default one (\code{fit.kin.method = "diff-levenmarq"}) is exclusively used.
It corresponds to differential Levenberg-Marquardt (see also \code{\link[minpack.lm]{nls.lm}})
because it is based on the numeric solution of the ordinary differential equations and not on the integration
of rate equations.}

\item{time.correct}{Logical, if the time of the recorded series of EPR spectra needs to be corrected
(see also \code{\link{correct_time_Exp_Specs}}).}

\item{path_to_dsc_par}{Character string ... argument/parameter... tbc}

\item{origin}{Character string ... argument/parameter... tbc}

\item{...}{additional parameters for \code{\link[minpack.lm]{nls.lm}}.}
}
\value{
List with the following components is available:
  \describe{
  \item{df}{Data frame object with the variables/columns such as \code{time},
  experimental quantitative variable like \code{sigmoid_Integ} (sigmoid integral) or \code{Area},
  concentration \code{c_M} or number of radicals of the relevant EPR spectrum and the corresponding
  quantitative variable \code{fitted} vector values.}
  \item{plot}{Plot/Graph object \emph{Quantitative variable} \emph{vs.} \emph{Time} with the experimental
  data and the corresponding fit.}
  \item{df.coeffs}{Data frame object containing the optimized parameter values (\code{Estimates}),
  their corresponding \code{standard errors}, \code{t-} as well as \code{p-values}.}
  \item{N.evals}{Total number of evaluations/iterations before the best fit is found.}
  \item{sum.LSQ.min}{The minimal least-square sum after \code{N.evals}.}
  \item{convergence}{Vector corresponding to residual sum of squares at each iteration/evaluation.
  The length of \code{convergence} is equal to the length of \code{N.evals}.}
  }
}
\description{
A short description...(Integrals/Areas/Concentration \emph{vs.} Time)...
}
\examples{
## loading example data (incl. `Area` and `time` variables)
## from Xenon: decay of a triarylamine radical cation after its generation
## by electrochemical oxidation
triaryl_radCat_path <-
  load_data_example(file = "Triarylamine_radCat_decay_a.txt")
## corresponding data (double integrated EPR spectrum = `Area` vs `time`)
triaryl_radCat_data <-
  readEPR_Exp_Specs(triaryl_radCat_path,
                    header = TRUE,
                    fill = TRUE,
                    select = c(3,7),
                    col.names = c("time_s","Area"),
                    x.unit = "s",
                    x.id = 1,
                    Intensity.id = 2,
                    qValue = 1700) \%>\%
  na.omit()
## data preview
head(triaryl_radCat_data)
#
## loading the `.DSC` file
triaryl_radCat_dsc_path <-
  load_data_example(file = "Triarylamine_radCat_decay_a.DSC")
#
## fit previous data by second order kinetics, where the `model.react`
## is considered as an elementary step (`time.correct` of the CW-sweeps
## is included (`TRUE`))
triaryl_model_kin_fit_01 <-
  eval_kinR_EPR_modelFit(data.qt.expr = triaryl_radCat_data,
                         model.react = "(r=2)R --> [k1] B",
                         elementary.react = TRUE,
                         params.guess = c(qvar0R = 0.019,
                                          k1 = 0.04
                                         ),
                         time.correct = TRUE,
                         path_to_dsc_par = triaryl_radCat_dsc_path,
                         origin = "xenon")
## data frame preview
head(triaryl_model_kin_fit_01$df)
#
## plot preview
triaryl_model_kin_fit_01$plot
#
## coefficients/parameters table preview
triaryl_model_kin_fit_01$df.coeffs
#
## convergence preview
triaryl_model_kin_fit_01$convergence
#
## take the same experimental data and perform fit by first order
## kinetics where the `model.react` is considered as an elementary step
## (`time.correct` of the CW-sweeps is included (`TRUE`))
triaryl_model_kin_fit_02 <-
  eval_kinR_EPR_modelFit(data.qt.expr = triaryl_radCat_data,
                         model.react = "(r=1)R --> [k1] B",
                         elementary.react = TRUE,
                         params.guess = c(qvar0R = 0.019,
                                          k1 = 0.0002
                                         ),
                         time.correct = TRUE,
                         path_to_dsc_par = triaryl_radCat_dsc_path,
                         origin = "xenon")
## plot preview
triaryl_model_kin_fit_02$plot
#
## coefficients/parameters table preview
triaryl_model_kin_fit_02$df.coeffs


}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Abs}()},
\code{\link{quantify_EPR_Norm_const}()}
}
\concept{Evaluations and Quantification}
