% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_EPR_modelFit.R
\name{eval_kinR_EPR_modelFit}
\alias{eval_kinR_EPR_modelFit}
\title{Radical Kinetic Models Fitted to Experimental Data (Integrals/Areas \emph{vs.} Time)}
\usage{
eval_kinR_EPR_modelFit(
  data.spectra.integ,
  time.unit = "s",
  time.series = "time_s",
  qvarR = "Area",
  model.react = "(x=1)R --> [k1] B",
  params.guess = c(qvar0R = 0.001, k1 = 0.001),
  algorithm.fit.kin = "diff-LM",
  time.series.correct = FALSE,
  path_to_DSC_or_par = NULL,
  origin = NULL
)
}
\arguments{
\item{data.spectra.integ}{tbc}

\item{time.unit}{Character string ... argument/parameter... tbc}

\item{time.series}{Character string ... argument/parameter... tbc}

\item{qvarR}{Character string ... argument/parameter... tbc}

\item{model.react}{Character string ... argument/parameter... tbc, the same like in \code{\link{eval_kinR_ODE_model}}}

\item{params.guess}{Named vector ... argument/parameter... tbc}

\item{algorithm.fit.kin}{Character string ... argument/parameter... tbc}

\item{time.series.correct}{Logical, ... argument/parameter... tbc}

\item{path_to_DSC_or_par}{Character string ... argument/parameter... tbc}

\item{origin}{Character string ... argument/parameter... tbc}
}
\value{
List ... tbc ...
}
\description{
A short description...
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Norm_const}()},
\code{\link{quantify_EPR_abs}()},
\code{\link{quantify_EPR_sim}()}
}
\concept{Evaluations and Quantification}
