% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_Eyring_GHS.R
\name{eval_kinR_Eyring_GHS}
\alias{eval_kinR_Eyring_GHS}
\title{Reaction Activation Parameters Obtained by Essential Transition State Theory}
\usage{
eval_kinR_Eyring_GHS(
  data.kvT,
  rate.const,
  rate.const.unit = "s^{-1}",
  Temp,
  Temp.unit = "K",
  transmiss.coeff = 1,
  fit.method = "default"
)
}
\arguments{
\item{data.kvT}{Data frame object, which must include two essential columns: rate constant (\eqn{k} of an elementary
radical reaction) and the corresponding temperatures at which the \eqn{k} was acquired.}

\item{rate.const}{Character string, pointing to rate constant column header in the actual \code{data.kvT} data frame.}

\item{rate.const.unit}{Character string, referring to rate constant unit. This has to be specified using
the \code{\link[grDevices]{plotmath}} notation, like \code{rate.const.unit = "M^{-1}~s^{-1}"}
or \code{rate.const.unit = "s^{-1}"} (\strong{default}), because it is automatically applied as \eqn{y}-axis unit
in the graphical output by the \code{{ggplot2}}.}

\item{Temp}{Character string, pointing to temperature column header within the original \code{data.kvT} data frame.}

\item{Temp.unit}{Character string, corresponding to temperature unit related to \code{Temp}. Temperature can be defined
in the following units: \code{Temp.unit = "K"} (kelvin, \strong{default}), \code{Temp.unit = "oC"} (degree Celsius)
or \code{Temp.unit = "oF"} (degree Fahrenheit). If other than \strong{default} specified, temperature values
(column characterized by the \code{Temp} argument) are automatically converted into \code{"K"} (kelvins).}

\item{transmiss.coeff}{Numeric value, corresponding to probability that the activated complex is transformed into products.
\strong{Default}: \code{transmiss.coeff = 1} (\eqn{100\,\%}).}

\item{fit.method}{Character string, corresponding to method applied to fit the theoretical Eyring relation
(by optimizing the activation parameters, see \code{Details}) to the experimental \eqn{k} \emph{vs} \eqn{T}
dependence. For this purpose, the \code{\link[stats]{nls}} function is used. Therefore, all the methods, defined
under its \code{algorithm} argument, are available: \code{"default"}
(corresponding to \href{https://journal.r-project.org/articles/RJ-2023-091/}{Gauss-Newton algorithm}),
\code{"plinear"}, which is
\href{https://geo-ant.github.io/blog/2020/variable-projection-part-1-fundamentals/}{Golub-Pereyra} algorithm
or \code{"port"}
(\href{https://ms.mcmaster.ca/\%7Ebolker/misc/port.pdf}{Fortran PORT, "portable" library for numerical computation}).}
}
\value{
As a result of the Eyring-relation fit, list with the following components is available:
  \describe{
  \item{df}{Data frame, including the original \code{data.kvT} + the column of \eqn{\Delta^{\ddagger} G^o},
  with the name of \code{DeltaG_active_kJ_per_mol}, as well as \code{fitted}/predicted values of the rate constant
  and finally, the corresponding residuals.}
  \item{df.fit}{Data frame including temperature (in the same region like in the original \code{data.kvT},
  however with the resolution of 1024 points) and the corresponding \code{.fitted} \eqn{k}, according to
  Eyring model.}
  \item{plot}{Static ggplot2-based object/list, showing graphical representation of the non-linear fit,
  together with the Eyring equation.}
  \item{plot.ra}{GGplot2 object (related to simple \strong{r}esidual \strong{a}nalysis), including
  two main plots: Q-Q plot and residuals vs predicted/fitted \eqn{k} \emph{vs} \eqn{T} from the Eyring fit.}
  \item{df.coeffs.HS}{Data frame object, containing the optimized (best fit) parameter values (\code{Estimates}),
  their corresponding \code{standard errors}, \code{t-} as well as \code{p-values} for both \eqn{\Delta^{\ddagger} H^o}
  and \eqn{\Delta^{\ddagger} S^o}.}
  \item{converg}{List, containing fitting/optimization characteristics like number of evaluations/iterations
  (\code{N.evals}); character denoting the (un)successful convergence (\code{message})
  and finally, standard deviation of the residuals (\code{residual.sd}), which is defined as:
  \deqn{\sqrt{\sum_i (y_i - y_{i,\text{fit/model}})^2\,/\,(N - k_{\text{pars}} - 1)}}
  where \eqn{N} is the number of observations and \eqn{k_{\text{pars}}} is the number of optimized parameters.
  Therefore, the smaller the \code{residual.sd}, the better the Eyring-relation fit.}
  }
}
\description{
Finding the temperature-dependence of a rate constant (\eqn{k}) related to the elementary radical reaction, using the essential
  transition state theory (TST). The activation parameters, such as \eqn{\Delta^{\ddagger} S^o} and \eqn{\Delta^{\ddagger} H^o}
  are obtained by the non-linear fit (see the general \code{\link[stats]{nls}} R function) of the Eyring expression
  (its non-linear form, see \code{Details}) on the original \eqn{k} \emph{vs} \eqn{T} relation (please,
  refer to the \code{data.kvT} argument). The latter can be acquired by the \code{\link{eval_kinR_EPR_modelFit}}
  from sigmoid-integrals of the EPR spectra recorded at different temperatures. Finally, the activation Gibbs energy
  (\eqn{\Delta^{\ddagger} G^o}) is calculated, using the optimized \eqn{\Delta^{\ddagger} S^o} and \eqn{\Delta^{\ddagger} H^o},
  for each temperature in the series.
}
\details{
The basic assumption of the Transition State Theory (TST) is the existence of activated state/complex, formed
  by the collision of reactant molecules, which does not actually lead to reaction products directly. The activated state (AS)
  is formed as highly energized, and therefore as an unstable intermediate, decomposing into products of the reaction.
  Accordingly, the reaction rate is given by the rate of its decomposition. Additional important assumption for TST
  is the presence of pre-equilibrium (characterized by the \eqn{K^{\ddagger}} constant) of the reactants with
  the activated complex (AC). Because the latter is not stable, it dissociates with motion along the corresponding
  bond-stretching coordinate. For this reason, the rate constant (\eqn{k}) must be related to the associated vibration
  frequency (\eqn{\nu}). Thus, every time, if the AC is formed, the \eqn{k} of AC-dissociation
  actually equals to \eqn{\nu}. Nevertheless, it is possible that the AC will revert back to reactants and therefore,
  only a fraction of ACs will lead to product(s). Such situation is reflected by the transmission coefficient \eqn{\kappa}
  (see also the argument \code{transmiss.coeff}), where \eqn{k = \kappa\,\nu}.

  According to statistical thermodynamics, the equilibrium constant can be expressed by the partition function (\eqn{q})
  of the reactants and that of the AC. By definition, each \eqn{q} corresponds to ratio of total number of particles
  to the number of particles in the ground state. In essence, it is the measure of degree to which the particles
  are spread out (partitioned among) over the energy levels. Therefore, taking into account the energies of a harmonic quantum
  oscillator vibrating along the reaction coordinate as well as partition functions of the AC and those of the reactants,
  the rate constant can be expressed as follows (see e.g. Ptáček P, Šoukal F, Opravil T (2018) in the \code{References}):
  \deqn{k = \kappa\,(k_{\text{B}}\,T\,/\,h)\,K^{\ddagger}}
  where the \eqn{k_{\text{B}}} and \eqn{h} are the Boltzmann and Planck constants, respectively, \eqn{T} corresponds
  to temperature and finally, the \eqn{K^{\ddagger}} represents the equilibrium constant including the partition functions
  of reactants and that of the AC. In order to evaluate the AC partition function, its structure must be known.
  However, often, due to the lack of structural information, it is difficult (if not impossible) to evaluate
  the corresponding \eqn{q^{\ddagger}(\text{AC})}. Therefore, considering the equilibrium between the reactants and the AC,
  one may express the \eqn{K^{\ddagger}} in terms of \emph{Gibbs} activation energy (\eqn{\Delta^{\ddagger} G^o}),
  because \eqn{\Delta^{\ddagger} G^o = - R\,T\,ln K^{\ddagger}} and thus the Eyring equation reads:
  \deqn{k = \kappa\,(k_{\text{B}}\,T\,/\,h)\,exp[- (\Delta^{\ddagger} G^o)/(R\,T)] =
  \kappa\,(k_{\text{B}}\,T\,/\,h)\,exp[- (\Delta^{\ddagger} H^o)/(R\,T)]\,exp[\Delta^{\ddagger} S^o / R]}
  where \eqn{R\approx 8.31446\,\text{J}\,\text{mol}^{-1}\,\text{K}^{-1}} is the universal gas constant and the upper index \eqn{^o}
  denotes the standard molar state (see IUPAC (2019) in the \code{References}). Previous formula is applied
  as a model to fit the experimental \eqn{k\,\,vs\,\,T} (see the argument \code{data.kvT}) relation, where both
  the \eqn{\Delta^{\ddagger} S^o} and the \eqn{\Delta^{\ddagger} H^o} (in the graphical output, are also denoted as
  \eqn{\Delta^{active} S^o} and \eqn{\Delta^{active} H^o}, respectively) are optimized using the \code{fit.method}
  (by the \code{\link[stats]{nls}} function). In the first approach, both latter are considered as temperature independent
  within the selected temperature range. Often, the Eyring equation is not applied in the original form,
  however in the linearized one. Nevertheless, the latter is not recommended as a model for fitting the experimental \eqn{k(T)}
  (see also Lente G, Fábián I, Poë AJ (2005) in the \code{References}). The reason inherits in the misinterpretation
  of the extrapolation to \eqn{T\rightarrow \infty} (or \eqn{1/T\rightarrow 0}) by which the \eqn{\Delta^{\ddagger} S^o}
  is obtained and thus it is unreliable. Accordingly, \strong{the original exponential Eyring form is recommended}
  as a model to fit the experimental \eqn{k(T)}.
  The \href{https://goldbook.iupac.org/terms/view/E02142}{\eqn{k}-unit depends on the molecularity of the reaction},
  please also refer to the \code{rate.const.unit} argument. Therefore, the left hand site of the Eyring equation above
  must be multiplied by the standard molar concentration \eqn{c^o = 1\,\text{mol}\,\text{dm}^{-3}}:
  \deqn{k\,(c^o)^{- \sum_i \nu_i^{\ddagger}}}
  where the \eqn{\sum_i \nu_i^{\ddagger}} goes through stoichiometric coefficients (including the negative sign for reactants)
  of the AC formation reaction (therefore the index \eqn{^{\ddagger}} is used), i.e. for the bi-molecular reaction,
  the sum results in \code{-1}, however for the mono-molecular one, the sum results in \code{0}.

  While the transition state theory (TST) is a helpful tool to get information about the mechanism
  of an elementary reaction, it has some limitations, particularly for radical reactions. Couple of them are listed below.
  \enumerate{
  \item One should be very careful if applied to elementary steps in a multistep reaction kinetics (like
  consecutive reactions, example shown in \code{\link{eval_kinR_ODE_model}}). If the intermediate (e.g. in the consecutive
  reaction mechanism) possesses a short life-time, the TST probably fails.

  \item For very fast reactions the assumed equilibrium between the reactants and the AC won't be reached.
  Therefore, the spin trapping reactions, which \eqn{k}s may actually fall into the order
  of \eqn{10^9\,\text{dm}^3\,\text{mol}^{-1}\,\text{s}^{-1}} (or oven higher, see Kemp TJ (1999) in the \code{References})
  should be taken with extreme caution in terms of TST.

  \item Formation of AC in TST is based on classical mechanics, that is molecules/atoms will only collide,
  having enough energy (to form the AC), otherwise reaction does not occur. Whereas, taking into account the quantum
  mechanical principle, molecules/atoms with any finite energy may (with a certain probability) tunnel across
  the energy barrier. Such effect will be less probable for high energy barriers, however e.g. for radical-radical
  recombination, where the barriers are typically very low, the tunneling probability is high and TST may fail.
  In addition, such reactions proceed relatively fast and therefore the TST (Eyring fit) can also strongly bias
  the activation parameters. This type of reactions may also exhibit negative \eqn{k} \emph{vs} \eqn{T} dependence
  (see Wardlaw DM and Marcus RA (1986) in the \code{References}).
  }
}
\examples{
## demonstration on raw data, presented
## in https://www.rsc.org/suppdata/nj/b5/b501687h/b501687h.pdf
## considering reaction H+ + (S2O6)2- <==> SO2 + (HSO4)-
kinet.test.data <-
  data.frame(k_per_M_per_s =
               c(9.54e-7,1.91e-6,3.76e-6,
                 7.33e-6,1.38e-5,2.56e-5,
                 4.71e-5,8.43e-5,1.47e-4),
             T_oC = c(50,55,60,65,70,
                      75,80,85,90)
)
activ.kinet.test.data <-
  eval_kinR_Eyring_GHS(
    data.kvT = kinet.test.data,
    rate.const = "k_per_M_per_s",
    rate.const.unit = "M^{-1}~s^{-1}",
    Temp = "T_oC",
    Temp.unit = "oC"
  )
#
## preview of the original data
## + ∆G (activated) + fitted + residuals
activ.kinet.test.data$df
#
## preview of the non-linear fit plot
activ.kinet.test.data$plot
#
## preview of the optimized (activated)
## ∆S and ∆H parameters
activ.kinet.test.data$df.coeffs.HS
#
## compare values with those presented in
## https://www.rsc.org/suppdata/nj/b5/b501687h/b501687h.pdf
## ∆S = (7.2 +- 1.1) kJ/mol*K & ∆H = (118.80 +- 0.41) kJ/mol
#
## preview of the new `.fitted` data frame
activ.kinet.test.data$df.fit
#
## preview of the convergence measures
activ.kinet.test.data$converg


}
\references{
Engel T, Reid P (2013). \emph{Physical Chemistry, 3rd Edition}, Pearson Education, ISBN 978-0-321-81200-1,
  \url{https://elibrary.pearson.de/book/99.150005/9781292035444}.

  Ptáček P, Šoukal F, Opravil T (2018). "Introduction to the Transition State Theory",
  \emph{InTech.}, \url{https://doi.org/10.5772/intechopen.78705}.

  International Union of Pure and Applied Chemistry (IUPAC) (2019). “Transition State Theory”,
  \url{https://goldbook.iupac.org/terms/view/T06470}.

  Anslyn EV, Dougherty DA (2006). \emph{Modern Physical Organic Chemistry}, University Science Books,
  ISBN 978-1-891-38931-3, \url{https://uscibooks.aip.org/books/modern-physical-organic-chemistry/}.

  Lente G, Fábián I, Poë AJ (2005). "A common Misconception about the Eyring Equation",
  \emph{New J. Chem.}, \strong{29}(6), 759–760, \url{https://doi.org/10.1039/B501687H}.

  Kemp TJ (1999), "Kinetic Aspects of Spin Trapping", \emph{Progress in Reaction Kinetics}, \strong{24}(4),
  287-358, \url{https://doi.org/10.3184/007967499103165102}.

  Wardlaw DM and Marcus RA (1986), "Unimolecular reaction rate theory for transition states of any looseness.
  3. Application to methyl radical recombination", \emph{J. Phys. Chem.}, \strong{90}(21), 5383-5393,
  \url{https://doi.org/10.1021/j100412a098}.
}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_nu_ENDOR}()},
\code{\link{eval_peakPick_Spec}()}
}
\concept{Evaluations}
