% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_ODE_model.R
\name{eval_kinR_ODE_model}
\alias{eval_kinR_ODE_model}
\title{Radical Kinetic Models Solved by Ordinary Differential Equations (ODE) for Visualization and Fitting}
\usage{
eval_kinR_ODE_model(
  model.react = "(x=1)R --> [k1] B",
  model.expr.diff = FALSE,
  kin.params = c(k1 = 0.001, qvar0R = 0.02),
  data.expr = NULL,
  time.expr.series = NULL,
  qvar.expr = NULL
)
}
\arguments{
\item{model.react}{parameter/argument tbc}

\item{model.expr.diff}{parmeter/argument tbc}

\item{kin.params}{parameter/argument tbc}

\item{data.expr}{parameter/argument tbc}

\item{time.expr.series}{parameter/argument tbc}

\item{qvar.expr}{parameter/argument tbc}
}
\value{
tbc
}
\description{
A short description...
  \tabular{lcll}{
   \strong{Reaction Scheme} \tab \strong{model.react} \tab \strong{kin.params} \tab \strong{Description} \cr
 \eqn{(x=1)\text{R} \overrightarrow{k_1} \text{B}} \tab "(x=1)R --> [k1] B" \tab \code{k1},\code{qvar0R} \tab Blah \cr
  }
}
\examples{
\dontrun{
tbc
tbc
}


}
