% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_ODE_model.R
\name{eval_kinR_ODE_model}
\alias{eval_kinR_ODE_model}
\title{Kinetic Models Solved by Ordinary Differential Equations (ODE) for Visualization and Fitting of Radical Reactions}
\usage{
eval_kinR_ODE_model(
  model.react = "(x=1)R --> [k1] B",
  model.expr.diff = FALSE,
  kin.params = c(k1 = 0.001, qvar0R = 0.02),
  data.expr = NULL,
  time.expr.series = NULL,
  qvar.expr = NULL
)
}
\arguments{
\item{model.react}{Character string denoting a specific radical (\code{"R"}) reaction related to changes in integral
intensities in EPR spectral time series. Arrow show the direction of reaction (\code{"-->"}, irreversible
or \code{"<==>", reversible}). Rate constants are indicated by square brackets after the arrows.
Notation inherited from \href{https://www.ctan.org/pkg/mhchem?lang=en}{\eqn{\LaTeX} `mhchem` package}.
Following reaction schemes are the most common used to describe the integral intensity and/or radical
concentration/amount changes during the EPR time series experiment.
\tabular{lcl}{
\strong{Reaction Scheme} \tab | \tab \strong{`model.react`} \cr
------------------------ \tab | \tab -------------------- \cr
\eqn{(x=1)\text{R} \xrightarrow{k_1} \text{B}} \tab | \tab \code{"(x=1)R --> [k1] B"} \cr
 \eqn{(x=1)\text{A} \xrightarrow{k_1} \text{R}} \tab | \tab \code{"(x=1)A --> [k1] R"} \cr
 \eqn{(x=1)\text{A} \xrightarrow{k_1} (x=1)\text{R} \xrightarrow{k_2} \text{C}} \tab | \tab
 \code{"(x=1)A --> [k1] (x=1)R --> [k2] C"} \cr
 \eqn{(x=1)\text{R} \xrightleftharpoons[k_2]{k_1} \text{B}} \tab | \tab
 \code{"(x=1)R <==> [k1] [k2] B"} \cr
 \eqn{(x=1)\text{A} \xrightleftharpoons[k_2]{k_1} (x=1)\text{R}} \tab | \tab
 \code{"(x=1)A <==> [k1] [k2] (x=1)R"} \cr
 \eqn{\text{A} \xleftarrow{k_1} (x=1)\text{R} \xrightarrow{k_2} \text{B}} \tab | \tab
 \code{"A [k1] <-- (x=1)R --> [k2] B"} \cr
 \eqn{(x=1)\text{A} + (y=1)\text{B} \xrightarrow{k_1} \text{R}} \tab | \tab
 \code{"(x=1)A + (y=1)B --> [k1] R"} \cr
}
Couple of examples are also given in the description. The function is relatively flexible and enables later addition
of any other reaction schemes describing the EPR time series experiments (you may ask developer(s) via `github issue`).
The stoichiometric coefficient (e.g. like \code{"(x=1)"} or \code{"(y=1)"}) can be varied within
the \code{model.react} character string. Defined/Allowed values are \code{"(x=1)"},  \code{"(x=2)"}
and \code{"(x=n)"}. \strong{The latest NOTATION (\code{"(x=n)"}) is USED ONLY FOR FITTING},
see \code{\link{eval_kinR_EPR_modelFit}}, in order to also consider a possible non-integer stoichiometric
coefficients. If equal coefficients are given for both sides of (partial) reaction
they have to be changed equally (e.g. like \code{"(x=2)A <==> [k1] [k2] (x=2)R"}). Otherwise, the stoichiometric
coefficients may be varied independently (e.g. like in \code{"(x=1)A + (y=2)B --> [k1] R"}).
\strong{Default}: \code{model.react = "(x=1)R --> [k1] B"}.}

\item{model.expr.diff}{Logical, difference between the integral intensities/areas under the curves calculated
using the experimental data and those generated by the model. By \strong{default} the argument is \strong{FALSE}
and it is ONLY ACTIVATED IN CASE KINETIC MODEL FITTING PROCEDURE (see also \code{\link{eval_kinR_EPR_modelFit}}).}

\item{kin.params}{Named numeric vector containing rate constants as well as initial radical or other reactant/product
concentration/integral intensities/areas...etc. Therefore, a general \code{qvar} (\strong{q}uantitative
\strong{var}iable) was defined which may actually reflect all the mentioned quantities.
\strong{Default}: \code{kin.params = c(k1 = 0.001,qvar0R = 0.02)}.
The initial values are denoted as \code{qvar0X} (e.g. qvar0R for radical or qvar0A for reactant).
The components of \code{kin.params} numeric vector depend on \code{model.react} like summarized in the
following table =>
\tabular{ccl}{
\strong{`model.react`} \tab | \tab \strong{Essetial `kin.params` components} \cr
-------------------- \tab | \tab ------------------- \cr
\code{"(x=1)R --> [k1] B"} \tab | \tab \code{k1}, \code{qvar0R}, (\code{n}) \cr
\code{"(x=1)A --> [k1] R"} \tab | \tab \code{k1}, \code{qvar0A}, \code{qvar0R}, (\code{n}) \cr
\code{"(x=1)A --> [k1] (x=1)R --> [k2] C"} \tab | \tab \code{k1}, \code{k2}, \code{qvar0A}, \code{qvar0R}, \code{qvar0C},
(\code{n}) \cr
\code{"(x=1)R <==> [k1] [k2] B"} \tab | \tab \code{k1}, \code{k2}, \code{qvar0R}, \code{qvar0B}, (\code{n}) \cr
\code{"(x=1)A <==> [k1] [k2] (x=1)R"} \tab | \tab \code{k1}, \code{k2}, \code{qvar0A}, \code{qvar0R}, (\code{n}) \cr
\code{"A [k1] <-- (x=1)R --> [k2] B"} \tab | \tab \code{k1}, \code{k2}, \code{qvar0R}, (\code{n}) \cr
\code{"(x=1)A + (y=1)B --> [k1] R"} \tab | \tab \code{k1}, \code{qvar0A}, \code{qvar0B}, \code{qvar0R}, (\code{n}) \cr
}}

\item{data.expr}{R data frame containing the integral intensities/areas under the curves calculated
using the experimental data as well as time corresponding to time series experiment. The essential two columns
are described by character strings like bellow (see the last two arguments). The \code{data.expr} should be used
in case if experimental time ... tbc}

\item{time.expr.series}{parameter/argument tbc}

\item{qvar.expr}{parameter/argument tbc}
}
\value{
tbc
}
\description{
A short description ...

  Additional description ...
  \tabular{ccl}{
  \strong{`model.react`} \tab | \tab \strong{Description} \cr
  -------------------- \tab | \tab -------------------- \cr
  \code{"(x=1)R --> [k1] B"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)A --> [k1] R"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)A --> [k1] (x=1)R --> [k2] C"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)R <==> [k1] [k2] B"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)A <==> [k1] [k2] (x=1)R"} \tab | \tab Blah Blah Blah \cr
  \code{"A [k1] <-- (x=1)R --> [k2] B"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)A + (y=1)B --> [k1] R"} \tab | \tab Blah Blah Blah \cr
  }
}
\examples{
\dontrun{
tbc
tbc
}


}
