% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_ODE_model.R
\name{eval_kinR_ODE_model}
\alias{eval_kinR_ODE_model}
\title{Radical Kinetic Models Solved by Ordinary Differential Equations (ODE) for Visualization and Fitting}
\usage{
eval_kinR_ODE_model(
  model.react = "(x=1)R --> [k1] B",
  model.expr.diff = FALSE,
  kin.params = c(k1 = 0.001, qvar0R = 0.02),
  data.expr = NULL,
  time.expr.series = NULL,
  qvar.expr = NULL
)
}
\arguments{
\item{model.react}{parameter/argument tbc}

\item{model.expr.diff}{parmeter/argument tbc}

\item{kin.params}{parameter/argument tbc}

\item{data.expr}{parameter/argument tbc}

\item{time.expr.series}{parameter/argument tbc}

\item{qvar.expr}{parameter/argument tbc}
}
\value{
tbc
}
\description{
A short description ...
  \tabular{lcl}{
  \strong{Reaction Scheme} \tab | \tab \strong{model.react} \cr
  ------------------------ \tab | \tab -------------------- \cr
  \eqn{(x=1)\text{R} \xrightarrow{k_1} \text{B}} \tab | \tab \code{"(x=1)R --> [k1] B"} \cr
   \eqn{(x=1)\text{A} \xrightarrow{k_1} \text{R}} \tab | \tab \code{"(x=1)A --> [k1] R"} \cr
   \eqn{(x=1)\text{A} \xrightarrow{k_1} (x=1)\text{R} \xrightarrow{k_2} \text{C}} \tab | \tab
   \code{"(x=1)A --> [k1] (x=1)R --> [k2] C"} \cr
   \eqn{(x=1)\text{R} \xrightleftharpoons[k_2]{k_1} \text{B}} \tab | \tab \code{"(x=1)R <==> [k1] [k2] B"} \cr
  }

  Additional description ...
  \tabular{ccl}{
  \strong{model.react} \tab | \tab \strong{Description} \cr
  -------------------- \tab | \tab -------------------- \cr
  \code{"(x=1)R --> [k1] B"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)A --> [k1] R"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)A --> [k1] (x=1)R --> [k2] C"} \tab | \tab Blah Blah Blah \cr
  \code{"(x=1)R <==> [k1] [k2] B"} \tab | \tab Blah Blah Blah \cr
  }

  More additional description ...
  \tabular{ccl}{
  \strong{model.react} \tab | \tab \strong{kin.params} \cr
  -------------------- \tab | \tab ------------------- \cr
  \code{"(x=1)R --> [k1] B"} \tab | \tab \code{k1}, \code{qvar0R}, (\code{n}) \cr
  \code{"(x=1)A --> [k1] R"} \tab | \tab \code{k1}, \code{qvar0A}, \code{qvar0R}, (\code{n}) \cr
  \code{"(x=1)A --> [k1] (x=1)R --> [k2] C"} \tab | \tab \code{k1}, \code{k2}, \code{qvar0A}, \code{qvar0R}, \code{qvar0C},
  (\code{n}) \cr
  \code{"(x=1)R <==> [k1] [k2] B"} \tab | \tab \code{k1}, \code{k2}, \code{qvar0R}, \code{qvar0B}, (\code{n}) \cr
  }
}
\examples{
\dontrun{
tbc
tbc
}


}
