% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kinR_ODE_model.R
\name{eval_kinR_ODE_model}
\alias{eval_kinR_ODE_model}
\title{Quantitative Kinetic Model Profiles by Numeric Solution of the ODE.}
\usage{
eval_kinR_ODE_model(
  model.react = "(r=1)R --> [k1] B",
  model.expr.diff = FALSE,
  elementary.react = TRUE,
  kin.params = c(k1 = 0.001, qvar0R = 0.02),
  time.unit = "s",
  timeLim.model = c(0, 1800),
  data.qt.expr = NULL,
  time.expr = NULL,
  qvar.expr = NULL,
  ...
)
}
\arguments{
\item{model.react}{Character string denoting a specific radical (\code{"R"}) reaction related to
changes in integral intensities (or any other \strong{q}uantitative \strong{var}iable) in EPR spectral
time series. Arrow shows direction of the reaction (\code{"-->", irreversible} or \code{"<==>", reversible}).
Rate constants are indicated by square brackets after the arrows. Following examples of the reaction schemes
are predefined and commonly used to describe the integral intensity and/or radical concentration/amount
changes during the EPR time series experiment (the \code{r,a,b,c} stoichiometric coefficients
may be varied, see below).
\tabular{ll}{
\strong{Reaction Scheme} \tab \strong{model.react} \cr
\eqn{(r=1)\text{R} \xrightarrow{k_1} \text{B}} \tab \code{"(r=1)R --> [k1] B"} \cr
 \eqn{(a=2)\text{A} \xrightarrow{k_1} (r=2)\text{R}} \tab \code{"(a=2)A --> [k1] (r=2)R"} \cr
 \eqn{(a=2)\text{A} \xrightarrow{k_1} (r=2)\text{R} \xrightleftharpoons[k_3]{k_2} (c=1)\text{C}} \tab
 \code{"(a=2)A --> [k1] (r=2)R <==> [k2] [k3] (c=1)C"} \cr
 \eqn{(r=1)\text{R} \xrightleftharpoons[k_2]{k_1} (b=1)\text{B}} \tab
 \code{"(r=1)R <==> [k1] [k2] (b=1)B"} \cr
 \eqn{(a=2)\text{A} \xrightleftharpoons[k_2]{k_1} (r=2)\text{R}} \tab
 \code{"(a=2)A <==> [k1] [k2] (r=2)R"} \cr
 \eqn{(a=1)\text{A} + (b=1)\text{B} \xrightarrow{k_1} (r=1)\text{R}} \tab
 \code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \cr
 \eqn{(r=1)\text{R} + (b=1)\text{B} \xrightarrow{k_1} (c=1)\text{C}} \tab
 \code{"(r=1)R + (b=1)B --> [k1] (c=1)C"} \cr
}
Couple of examples are also given in details. The function is relatively flexible and enables
later addition of any other reaction schemes describing the EPR time series experiments
(YOU MAY ASK DEVELOPER(S) via forum/help-channels). The stoichiometric coefficient (such as \code{(r=1)}
or \code{(a=1)}) can be varied within the \code{model.react} character string.
Defined/Allowed values are integers e.g. 1,2,3...etc. The space character within the \code{model.react}
string is not fixed and can be skipped for the sake of flexibility.
If \code{elementary.react = FALSE} (the model reaction is not considered as an elementary one),
a possible non-integer partial stoichiometric coefficients (e.g. \code{alpha},\code{beta} or \code{gamma})
must be included in \code{kin.params} (see also \code{kin.params} description).}

\item{model.expr.diff}{Logical, difference between the integral intensities/areas under the EPR spectra calculated
using the experimental data and those generated by the model. By \strong{default} the argument
is \strong{FALSE} and it is ONLY ACTIVATED (\code{model.expr.diff = TRUE}) IN THE CASE WHEN THE KINETIC MODEL
FITTING PROCEDURE (see also \code{\link{eval_kinR_EPR_modelFit}} or examples below) IS PERFORMED.}

\item{elementary.react}{Logical, if the model reaction should be considered as elementary one,
i.e. the stoichiometric coefficients equal to the partial reaction orders. Such reaction proceeds without
identifiable intermediate species forming. \strong{Default}: \code{elementary.react = TRUE}.
If \code{elementary.react = FALSE}, i.e. the \code{model.react} cannot be considered like an elementary one,
one must include the parameterized reaction orders \eqn{\alpha}, \eqn{\beta} or \eqn{\gamma} into
the \code{kin.params}, e.g like \code{kin.params = c(k1 = 0.01, qvar0A = 0.05, alpha = 1.5)}.}

\item{kin.params}{Named numeric vector containing rate constants as well as initial radical
or other reactant/product concentration/integral intensities/areas...etc. Therefore, a general
\code{qvar} (\strong{q}uantitative \strong{var}iable) was defined which may actually reflect
all above-mentioned quantities. \strong{Default}: \code{kin.params = c(k1 = 0.001,qvar0R = 0.02)}.
The initial values are denoted as \code{qvar0X} (e.g. qvar0R for radical or qvar0A for the reactant \code{A}).
The components of \code{kin.params} depend on \code{model.react} as well as on the \code{elementary.react}.
If \code{elementary.react = FALSE} additional parameters like partial reaction orders (\code{alpha}
and/or \code{beta} and/or \code{gamma}) must be defined within \code{kin.params}, like summarized
in the following table:
\tabular{ll}{
\strong{model.react} \tab \strong{Essential kin.params components} \cr
\code{"(r=1)R --> [k1] B"} \tab \code{k1}, \code{qvar0R}, (\code{alpha}) \cr
\code{"(a=1)A --> [k1] (r=1)R"} \tab \code{k1}, \code{qvar0A}, \code{qvar0R}, (\code{alpha}) \cr
\code{"(a=1)A --> [k1] (r=1)R <==> [k2] [k3] (c=1)C"} \tab \code{k1}, \code{k2}, \code{k3} \code{qvar0A},
\code{qvar0R}, \code{qvar0C}, (\code{alpha}, \code{beta}, \code{gamma}) \cr
\code{"(r=1)R <==> [k1] [k2] (b=1)B"} \tab \code{k1}, \code{k2}, \code{qvar0R},
\code{qvar0B}, (\code{alpha}, \code{beta}) \cr
\code{"(a=1)A <==> [k1] [k2] (r=1)R"} \tab \code{k1}, \code{k2}, \code{qvar0A},
\code{qvar0R}, (\code{alpha}, \code{beta}) \cr
\code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \tab \code{k1}, \code{qvar0A}, \code{qvar0B},
\code{qvar0R}, (\code{alpha},\code{beta}) \cr
\code{"(r=1)R + (b=1)B --> [k1] (c=1)C"} \tab \code{k1}, \code{qvar0R}, \code{qvar0B},
\code{qvar0C}, (\code{alpha},\code{beta}) \cr
}}

\item{time.unit}{Character string corresponding to time unit like \code{"s"} (\strong{default}),
\code{"min"} or \code{"h"}.}

\item{timeLim.model}{Numeric vector incl. two values corresponding to starting and final time/termination
of the model reaction.}

\item{data.qt.expr}{A data frame object containing the concentrations/integral intensities/areas under
the EPR spectra calculated using the \strong{experimental data} as well as time column. These two essential
columns are described by the character strings like those below \code{time.expr} and \code{qvar.expr}.
The \code{data.qt.expr} MUST BE USED ONLY IN SUCH CASE WHEN THE EXPERIMENTAL TIME HAS TO BE INCLUDED
IN THE KINETIC MODEL (e.g. also for THE FITTING of EXPERIMENTAL DATA BY THE KINETIC MODEL).
\strong{Default}: \code{data.qt.expr = NULL}.}

\item{time.expr}{Character string pointing to \code{time} column/variable name in the original
\code{data.qt.expr} data frame. \strong{Default}: \code{time.expr = NULL} (when the experimental
data aren't taken into account). If the time has to be corrected (e.g. in the case of double integrals),
please use \code{\link{correct_time_Exp_Specs}} function prior to kinetic evaluation.}

\item{qvar.expr}{Character string pointing to \code{qvar} column/variable name in the original
\code{data.qt.expr} data frame. \strong{Default}: \code{qvar.expr = NULL} (when the experimental
data aren't taken into account).}

\item{...}{additional arguments passed to the ODE (see also \code{\link[deSolve]{ode}}).}
}
\value{
If the function \strong{is not used for fitting} of the experimental and processed data,
  the result is \code{list} consisting of:
  \describe{
  \item{df}{Data frame containing \code{time} column and \code{qvar}, quantitative variable,
  columns corresponding to quantities of different relevant species
  denoted as \code{"R"}, \code{"A"}, \code{"B"}, ... etc. + if \code{data.qt.expr} is NOT NULL
  additional experimental quantitative variable is present.}
  \item{plot}{Plot object containing \code{time} as abscissa and \code{qvar}
  (see \code{df} above) as \eqn{y}-axis. + if \code{data.qt.expr} is NOT NULL the experimental
  quantitative variable is presented as well.}
  }
  Applying function \strong{for the fitting} procedure
  requires \code{model.expr.diff = TRUE} and therefore the result is difference between
  the integral intensities/areas under the curves calculated using the experimental data
  and those generated by the model.
}
\description{
Theoretical quantitative kinetic profiles (such as concentration/amount/integral intensity)
  as well as fitting of the experimental data for various predefined model reactions with
  the radical(s) (labeled as "R"). Profiles are provided by the numeric solution of rate equations
  by the \strong{O}rdinary \strong{D}ifferential \strong{E}quations (ODE). This function is inspired by:
  \href{https://www.r-bloggers.com/2013/06/learning-r-parameter-fitting-for-models-involving-differential-equations/}{R-bloggers
  article}.
}
\details{
According to IUPAC \insertCite{IUPACkinetR2019}{eprscope} the rate of a chemical reaction
  with the radicals (\eqn{\text{R}}) involved (example see argument \code{model.react})
  \deqn{r\text{R} + b\text{B} \xrightarrow\,\, c\text{C}}
  is expressed via time change of the reaction extend (\eqn{\text{d}\xi/\text{d}t}):
  \deqn{-(1/r)\,(\text{d}n_{\text{R}}/\text{d}t) = -(1/b)\,(\text{d}n_{\text{B}}/\text{d}t) =
  (1/c)\,(\text{d}n_{\text{C}}/\text{d}t)}
  At constant volume (\eqn{V}) conditions (or if volume changes are negligible) the amount
  (\eqn{n} in mole) of reactant/product can be replaced by its corresponding concentration
  (\eqn{c = n/V}). Such reaction rate (expressed in moles per unit volume) is function of temperature,
  pressure as well as of concentration of reactants and products. For the reaction example shown above
  it applies:
  equation...TBC
  In EPR Spectroscopy the number of radicals
  is directly proportional to (double) integral of the radical EPR spectrum
  (see also \code{\link{quantify_EPR_Abs}}).
  Secondary purpose is to fit the experimental EPR spectral time series outputs, e.g. integral
  intensity (area under the spectral curve) \emph{vs} time in order to gather the rate constants
  (\eqn{k}) of radical ("R") formation and/or decay. Quantitative kinetic profiles
  are not evaluated by integration of the kinetic equations, however by numeric solution of the Ordinary
  Differential Equations (\href{http://desolve.r-forge.r-project.org/index.html}{ODE}, see also
  \pkg{deSolve} package). Therefore, higher number of models is available than for integrated
  differential equations because for complex mechanisms it's quite often highly demanding.
  Several kinetic models for radical reactions in EPR spectroscopy are predefined and summarized below
  (see also \code{model.react} function argument).
  \tabular{ll}{
  \strong{model.react} \tab \strong{Short Description} \cr
  \code{"(r=1)R --> [k1] B"} \tab Basic irreversible forward reaction,
  e.g. irrev. dimerization (if \code{(r=2)}). \cr
  \code{"(a=1)A --> [k1] (r=1)R"} \tab Basic irreversible radical formation,
  e.g. irrev. \eqn{\text{e}^-} transfer. \cr
  \code{"(a=1)A --> [k1] (r=1)R <==> [k2] [k3] (c=1)C"} \tab Consecutive reactions,
  e.g. like comproportionation (for \code{(a=2)} and \code{(r=2)}) + follow-up reversible
  dimerization (\code{(c=1)}). \cr
  \code{"(r=1)R <==> [k1] [k2] (b=1)B"} \tab Basic reversible radical quenching,
  e.g. rev. \eqn{\pi-\pi} dimerization for \code{(r=2)} and \code{(b=1)}. \cr
  \code{"(a=1)A <==> [k1] [k2] (r=1)R"} \tab Basic reversible radical formation,
  e.g. from rev. comproportionation (\eqn{\text{A}^{++} + \text{A}^0 \xrightleftharpoons ~ 2\text{R}^{.+}},
  for \code{(a=2)} and \code{(r=2)}). \cr
  \code{"(a=1)A + (b=1)B --> [k1] (r=1)R"} \tab Radical formation by chemical reaction like oxidation,
  reduction or spin trapping (transient radical is not visible within the common EPR time scale). \cr
  \code{"(r=1)R + (b=1)B --> [k1] (c=1)C"} \tab General radical quenching by chemical reaction. \cr
  }
}
\examples{
## irreversible dimerization quantitative kinetic profile
## table (df) with first 10 observations/rows
kin.test.01 <-
  eval_kinR_ODE_model(model.react = "(r=2)R --> [k1] B",
                      kin.params = c(k1 = 0.012,
                                     qvar0R = 0.08))
## preview
head(kin.test.01$df,n = 10)
#
## consecutive reactions and the corresponding plot
## (`model.react` character string without spaces)
kin.test.02 <-
 eval_kinR_ODE_model(
   model.react = "(a=2)A-->[k1](r=2)R<==>[k2][k3](c=1)C",
   kin.params = c(k1 = 0.1,
                  k2 = 0.1,
                  k3 = 0.0002,
                  qvar0A = 0.02,
                  qvar0R = 0.002,
                  qvar0C = 0)
 )
## plot preview
kin.test.02$plot
#
## data frame/table preview
head(kin.test.02$df)
#
## loading example data (incl. `Area` and `time` variables)
## from Xenon: decay of a triarylamine radical cation
## after its generation by electrochemical oxidation
triaryl_radCat_path <-
  load_data_example(file = "Triarylamine_radCat_decay_a.txt")
## corresponding data (double integrated
## EPR spectrum = `Area` vs `time`)
triaryl_radCat_data <-
  readEPR_Exp_Specs(triaryl_radCat_path,
                    header = TRUE,
                    fill = TRUE,
                    select = c(3,7),
                    col.names = c("time_s","Area"),
                    x.unit = "s",
                    x.id = 1,
                    Intensity.id = 2,
                    qValue = 1700,
                    data.structure = "others") \%>\%
  na.omit()
## data preview
head(triaryl_radCat_data)
#
## comparison of the kinetic model with the experimental
## data `triaryl_radCat_data`, kinetic parameters were estimated
## to be as close as possible to the latter.
compar_model_expr_data_01 <-
  eval_kinR_ODE_model(model.react = "(r=2)R --> [k1] B",
                      kin.params = c(qvar0R = 0.019,
                                     k1 = 0.04),
                      timeLim.model = c(0,1500),
                      data.qt.expr = triaryl_radCat_data,
                      qvar.expr = "Area",
                      time.expr = "time_s")
## plot preview
compar_model_expr_data_01$plot
#
## previous kinetic model with partial reaction
## order ("alpha") corresponding to "R" (radical species).
## In such case REACTION is NOT CONSIDERED
## as an ELEMENTARY one !
compar_model_expr_data_02 <-
  eval_kinR_ODE_model(model.react = "(r=2)R --> [k1] B",
                      elementary.react = FALSE,
                      kin.params = c(qvar0R = 0.019,
                                     k1 = 0.04,
                                     alpha = 1.9
                                    ),
                      timeLim.model = c(0,1500),
                      data.qt.expr = triaryl_radCat_data,
                      qvar.expr = "Area",
                      time.expr = "time_s")
## plot preview
compar_model_expr_data_02$plot


}
\references{
\insertRef{IUPACkinetR2019}{eprscope}

 \insertRef{KinetRJCE2006}{eprscope}

 \insertRef{LevinePCbook2009}{eprscope}

 \insertRef{RbloggKinet2013}{eprscope}
}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{quantify_EPR_Abs}()},
\code{\link{quantify_EPR_Norm_const}()}
}
\concept{Evaluations and Quantification}
