% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_nu_ENDOR.R
\name{eval_nu_ENDOR}
\alias{eval_nu_ENDOR}
\title{ENDOR/Larmor Frequency of Specific Nucleus}
\usage{
eval_nu_ENDOR(nucle_us_i, B.unit = "G", B.val)
}
\arguments{
\item{nucle_us_i}{(Vector) character string in the form like \code{"14N"} or \code{c("1H","13C")},
pointing to specific nucleus/nuclei, for which the frequency should by calculated. The nuclear \emph{g}-factors
for those nuclei are taken from the \code{\link{isotopes_ds}}.}

\item{B.unit}{Character string, denoting the magnetic flux density \emph{B} unit. \strong{Default}:
\code{B.unit = "G"}.}

\item{B.val}{Numeric, magnetic flux density \eqn{B}-\code{value}. This actually corresponds
to \eqn{B} at which the EPR signal saturates to record the ENDOR spectrum/spectra.}
}
\value{
Numeric value or vector of nuclear Larmor/ENDOR frequencies in MHz for selected
  nuclei at \eqn{B} = \code{B.val}.
}
\description{
Larmor/ENDOR frequency calculations for the analysis in EPR/Electron-Nuclear Double Resonance (ENDOR).
  Function is inspired by the similar \href{https://bio.groups.et.byu.net/LarmourFreqCal.phtml}{Larmor Frequency
  Calculator}, available online.
}
\details{
The frequency in MHz is calculated according to relation
  \deqn{\nu_{\text{ENDOR}}^{} = - (1/h)\,\mu_{\text{N}}^{}\,g_{\text{n}}^{}\,B\,10^{-6}}
  where \eqn{h} is the Planck's constant, \eqn{\mu_{\text{N}}^{}} is the nuclear magneton
  available from \href{https://r-quantities.github.io/constants/}{constants} package (\code{constants::syms$mun}),
  \eqn{g_{\text{n}}^{}} is the nuclear \eqn{g}-factor of the specific nucleus (reported in the package
  \code{\link{isotopes_ds}} data frame as \code{g_Nuclear}) and finally, the \eqn{B} denotes the magnetic flux
  density at which the ENDOR spectra are recorded (see also \code{B.val} in arguments). The \eqn{10^{-6}} coefficient
  is referred to the resulting frequency in MHz. The negative sign "\eqn{-}" mirrors the convention to describe the direction
  of magnetic spin moments precession either counter-clockwise (\eqn{+}, if \eqn{\gamma_{\text{n}}^{} < 0})
  or clockwise (\eqn{-}, if \eqn{\gamma_{\text{n}}^{} > 0}, Levitt MH (2013)).
}
\examples{
## Larmor/ENDOR frequency for one selected nucleus
## only, e.g. "14N" at 3486 G
eval_nu_ENDOR(nucle_us_i = "14N",
              B.val = 3486)
#
## Larmor/ENDOR frequency for selected nuclei
## e.g. "1H" and "31P" at saturation
## field of B = 349.9 mT
eval_nu_ENDOR(nucle_us_i = c("1H","31P"),
              B.unit = "mT",
              B.val = 349.9)


}
\references{
Levitt MH (2013). \emph{Spin Dynamics: Basics of Nuclear Magnetic Resonance}.
  Wiley, ISBN 978-1-118-68184-8, \url{https://books.google.cz/books?id=bysFAa4MPQcC}.
}
\seealso{
Other Evaluations: 
\code{\link{eval_DeltaXpp_Spec}()},
\code{\link{eval_FWHMx_Spec}()},
\code{\link{eval_extremeX_Spec}()},
\code{\link{eval_gFactor}()},
\code{\link{eval_gFactor_Spec}()},
\code{\link{eval_interval_cnfd_tVec}()},
\code{\link{eval_peakPick_Spec}()}
}
\concept{Evaluations}
