% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_iso.R
\name{eval_sim_EPR_iso}
\alias{eval_sim_EPR_iso}
\title{Simulation of Isotropic EPR Spectra}
\usage{
eval_sim_EPR_iso(
  g.iso = 2.00232,
  instrum.params = c(cf = 3500, sw = 200, points = 2048, mwGHz = 9.8),
  path_to_DSC_or_par = NULL,
  origin = NULL,
  B.unit = "G",
  nuclear.system = NULL,
  natur.abund = FALSE,
  lineSpecs.form = "derivative",
  lineGL.DeltaB = list(1, NULL),
  lineGL.content = list(1, NULL),
  Intensity.sim = "dIeprSim_over_dB",
  plot.sim.interact = FALSE
)
}
\arguments{
\item{g.iso}{Numeric value, guess of the isotropic \eqn{g}-factor. It may also possess a `NULL`
value if the \eqn{g} corresponding to "central field" is equal to `g.iso`.
\strong{Default}: \code{g.iso = 2.00232} (the approximate \eqn{g} of the free electron).}

\item{instrum.params}{Named numeric vector containing instrumental parameters required
for the simulation =>
\tabular{ll}{
\code{cf} \tab "central field" (magnetic flux density, \eqn{B_{\text{CF}}}) \cr
\code{sw} \tab "sweep width" (magnetic flux density recording region,
\eqn{B_{\text{SW}}}) \cr
\code{points} \tab number of spectral points (corresponding to resolution) within
the "sweep width" \cr
\code{mwGHz} \tab applied microwave frequency in `GHz` to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default} values are chosen to cover the EPR spectra of common organic radical.
If \code{instrum.params = NULL} then parameters are provided by \code{path_to_DSC_or_par}
as well as by \code{origin}.}

\item{path_to_DSC_or_par}{Character string, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} parameter)
\code{text} files including all instrumental parameters and provided by the EPR machine.
\strong{Default}: \code{path_to_DSC_or_par = NULL} in case if the \code{instrum.params}
is already defined. IF ARGUMENT \code{instrum.params = NULL} then BOTH \code{path_to_DSC_or_par}
AS WELL AS \code{origin} MUST BE DEFINED !}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_DSC_or_par = NULL}).}

\item{B.unit}{Character string pointing to unit of magnetic flux density which is to be presented
on \eqn{B} abscissa of the EPR spectrum, like \code{"G"} (`Gauss`) or \code{"mT"} (`millitesla`),
\strong{default}: \code{B.unit = "G"}. THE UNIT MUST BE SHARED ACROSS ALL B ARGUMENTS
like `cf` and `sw` within the \code{instrum.params} AS WELL AS THOSE IN \code{lineGL.DeltaB} !}

\item{nuclear.system}{List containing the information about groups of equivalent nuclei
interacting with the unpaired electron, e.g. like \code{nuclear.system = list("14N",1,45)}.
This corresponds to one group of "14N" interacting nuclei where \strong{the second number}
denotes \strong{the number of nuclei} within the group and \strong{the third number}
is the \strong{guess of the hyperfine coupling constant in MHz}. Therefore, in summary
it corresponds to \eqn{A(1\times ^{14}\text{N}) = 45~\text{MHz}}. If more complex interaction
is considered, e.g. like \eqn{A(3\times ^{1}\text{H}) = 5.06~\text{MHz} +
A(6\times ^{1}\text{H}) = 17.64~\text{MHz}}, such system must be defined by nested lists like
\code{nuclear.system = list(list("1H",3,5.06),list("1H",6,17.64))}...etc. Actually,
the number of groups is limited to 6 which covers the vast majority of organic radicals
and can be extended in the future. \strong{Default}: \code{nuclear.system = NULL} in case
if no interaction with electron surrounding nuclei is considered and only single
line EPR spectrum is expected.}

\item{natur.abund}{...tbc...}

\item{lineSpecs.form}{Character string describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{lineGL.DeltaB}{...tbc...is corresponding either to \eqn{\Delta B_{\text{pp}}}
in case of derivative line form or \eqn{FWHM} in case of integrated line form...tbc..}

\item{lineGL.content}{...tbc...}

\item{Intensity.sim}{...tbc...}

\item{plot.sim.interact}{description ...tbc...}
}
\value{
...list...tbc or interactive simulated `plotly` spectrum
}
\description{
A short description...tbc...of the EPR simulation
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Simulation and Optimization: 
\code{\link{quantify_EPR_sim}()}
}
\concept{Simulation and Optimization}
