% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_iso.R
\name{eval_sim_EPR_iso}
\alias{eval_sim_EPR_iso}
\title{Simulation of Isotropic EPR Spectra}
\usage{
eval_sim_EPR_iso(
  g.iso = 2.00232,
  instrum.params = c(Bcf = 3500, Bsw = 200, Npoints = 2048, mwGHz = 9.8),
  path_to_dsc_par = NULL,
  origin = NULL,
  B.unit = "G",
  nuclear.system = NULL,
  natur.abund = TRUE,
  lineSpecs.form = "derivative",
  lineGL.DeltaB = list(1, NULL),
  lineG.content = 1,
  Intensity.sim = "dIeprSim_over_dB",
  plot.sim.interact = FALSE
)
}
\arguments{
\item{g.iso}{Numeric value, guess of the isotropic \eqn{g}-factor. It may also possess a \code{NULL}
value if the \eqn{g} corresponding to "central field" is equal to \code{g.iso}.
\strong{Default}: \code{g.iso = 2.00232} (the approximate \eqn{g} of the free electron).}

\item{instrum.params}{Named numeric vector, containing instrumental parameters required
for the simulation =>
\tabular{ll}{
\code{Bcf} \tab "central field" (magnetic flux density, \eqn{B_{\text{CF}}}) \cr
\code{Bsw} \tab "sweep width" (magnetic flux density recording region,
\eqn{B_{\text{SW}}}) \cr
\code{Npoints} \tab number of spectral points (corresponding to resolution) within
the "sweep width" \cr
\code{mwGHz} \tab applied microwave frequency in \code{GHz} to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default} values are chosen to cover the EPR spectra of common organic radicals.
If \code{instrum.params = NULL} then parameters must be provided by the \code{path_to_dsc_par}
as well as by \code{origin} arguments.}

\item{path_to_dsc_par}{Character string, path (can be also acquired by the \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on the OS, see \code{origin} argument)
\code{text} files including all instrumental parameters from the EPR machine.
\strong{Default}: \code{path_to_dsc_par = NULL} in case if the \code{instrum.params}
is already defined. IF the \code{instrum.params = NULL} then BOTH the \code{path_to_dsc_par}
AS WELL AS the \code{origin} MUST BE DEFINED !}

\item{origin}{Character string, corresponding to software which was used to obtain the EPR spectra
on spectrometers, because the files are slightly different, whether they
were recorded by the "WinEpr" (\code{origin = "winepr"}) or by the "Xenon".
\strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_dsc_par = NULL}).}

\item{B.unit}{Character string, pointing to unit of magnetic flux density which is to be presented
on \eqn{B}-axis of the EPR spectrum, like \code{"G"} (Gauss), \code{"mT"} (millitesla) or \code{"T"} (tesla),
\strong{default}: \code{B.unit = "G"}. THE UNIT MUST BE SHARED ACROSS ALL RELEVANT B-ARGUMENTS
like \code{cf} and \code{sw} within the \code{instrum.params} AS WELL AS within THOSE IN \code{lineGL.DeltaB} !}

\item{nuclear.system}{List, containing the information about groups of equivalent nuclei
interacting with the unpaired electron like \code{nuclear.system = list("14N",1,45)}.
This corresponds to one group of "14N" interacting nuclei where \strong{the second number}
denotes \strong{the number of nuclei} within the group and \strong{the third number}
is the \strong{guess of the hyperfine coupling constant in MHz}. Therefore, in summary
it refers to \eqn{A(1\times ^{14}\text{N}) = 45~\text{MHz}}. If more complex interaction
is considered, e.g. \eqn{A(3\times ^{1}\text{H}) = 5.06~\text{MHz} +
A(6\times ^{1}\text{H}) = 17.64~\text{MHz}}, such system must be defined by nested lists like
\code{nuclear.system = list(list("1H",3,5.06),list("1H",6,17.64))}. The number of \code{lists}
is not limited and therefore, any number of equivalent nuclei groups can be used to simulate
the EPR spectra. \strong{Default}: \code{nuclear.system = NULL} in case
if no interaction with the unpaired electron surrounding nuclei is considered and only single
line EPR spectrum is expected.}

\item{natur.abund}{Logical, whether the natural abundance of the interacting nuclei
is taken into the calculation of intensity pattern of the simulated EPR spectrum.
\strong{Default}: \code{natur.abund = TRUE}. For a single-line
EPR spectrum without hyperfine splitting(HFS) it is automatically switched to \code{natur.abund = FALSE}.}

\item{lineSpecs.form}{Character string, describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{lineGL.DeltaB}{List of two values referred to as \emph{Gaussian} (G) and \emph{Lorentzian} (L)
spectral line-forms, respectively. For the "pure" \emph{Gaussian} only the first value is numeric
and the second one is \code{NULL} => e.g. \code{lineGL.DeltaB = list(1,NULL)} (\strong{default}).
For the "pure" \emph{Lorentzian} the opposite expression must be used => e.g.
\code{lineGL.DeltaB = list(NULL,0.5)}. If the linear combination of both
line forms is taken into account (see \code{lineG.content}), that is so called \emph{pseudo-Voigt},
then both values are numeric (e.g. \code{lineGL.DeltaB = list(0.5,0.5)}) and are related
to \emph{Gaussian} and \emph{Lorentzian} forms, respectively. The \code{DeltaB} corresponds either to
\eqn{\Delta B_{\text{pp}}} (if \code{lineSpecs.form = "derivative"}) or to \eqn{FWHM}
(if \code{lineSpecs.form = "integrated"} or if \code{lineSpecs.form = "absorption"}).
The unit of values must coincide with those used in \code{instrum.params} as well as with \code{B.unit}.}

\item{lineG.content}{Numeric value between \code{0} and \code{1} referring to content of \emph{Gaussian} line form.
If \code{lineG.content = 1} (\strong{default}) it corresponds to "pure" \emph{Gaussian} line form
and if \code{lineG.content = 0} it corresponds to \emph{Lorentzian} one. The value from (0,1)
(e.g. \code{lineG.content = 0.5}) represents the linear combination (for the example above
with the coefficients 0.5 and 0.5) of both line forms ➨ so called \emph{pseudo-Voigt}.}

\item{Intensity.sim}{Character string, pointing to column of simulated EPR intensity within the related output
data frame. \strong{Default}: \code{Intensity.sim = "dIeprSim_over_dB"}.}

\item{plot.sim.interact}{Logical, whether to display the simulated spectrum by interactive \code{plotly} graph
(see also \code{\link{plot_EPR_Specs2D_interact}}). If \code{plot.sim.interact = FALSE} (\strong{dafault}),
then the output contains the data frame as well as \code{ggplot2} based plot of the simulated EPR spectrum
within a list.}
}
\value{
If \code{plot.sim.interact = TRUE}, function returns an interactive plot object with the simulated EPR spectrum.
  Otherwise (if \code{plot.sim.interact = FALSE}), the output is represented by the \code{list} with the following
  elements:
  \describe{
  \item{plot}{\code{ggplot2} static object showing the simulated EPR spectrum.}

  \item{df}{Data frame/table object related to the simulated spectrum.}
  }
}
\description{
EPR spectra often display hyperfine structure, where the spectral lines split into several other ones,
  as a consequence of the electron-nuclear interaction with nuclei close to the paramagnetic center.
  Sometimes, such splitting can be complicated and requires \strong{simulations}, i.e. computational "synthesis"
  of spectral structure based on quantum chemistry as well as on mathematical description of spectral lineshapes
  (see also \code{References}). The actual function helps to analyze the hyperfine structure (HFS)
  of the isotropic EPR spectra by simulations. For such purpose, properties like nuclear \eqn{g}-value,
  spin quantum number as well as natural abundance of the isotopes, related to interacting nuclei,
  must be known and are collected from the \code{\link{isotopes_ds}}. EPR spectra can be simulated
  for the derivative as well as for the integrated line forms (see the argument \code{lineSpec.form}).
}
\details{
Theoretical predictions or computations of a spectrum (and its comparison with the experiment) represent
  an important step in the analysis of EPR spectra. However, such step requires an iterative process with modelling
  the above-described \strong{electron-nuclear system} followed by a numerical simulations of EPR spectra
  to match the experimental ones (see also \code{\link{eval_sim_EPR_isoFit}}). Commonly, quantum chemical calculations
  (usually DFT, see also \code{vignette("functionality")} are involved in this process.
  EPR simulations in the isotropic regime assume that the molecules tumble/move extremely fast causing a total averaging
  of any anisotropic properties out. An EPR spectrum corresponding to the latter consists of a series of symmetric
  lines with equal widths. In such case the spectrum is characterized by the isotropic HF coupling/splitting
  constants \eqn{A_{\text{iso}}/a_{\text{iso}}} (see also the \code{nuclear.system} argument) as well as
  by the above-mentioned linewidth. Many organic radicals exhibit such EPR spectra at room temperature
  and in solutions of low viscosities (Gerson F, Huber W (2003), see \code{References}).

  In the first step the \eqn{B}-region (magnetic flux density) and the resolution must be defined
  by the \code{instrum.params} argument or can be directly acquired from the parameter file, using
  the \code{path_to_dsc_par} argument. Position of the spectrum (within the desired \eqn{B}-region)
  as well as those of HFS-lines are evaluated by the resonance condition (see also the \code{\link{eval_gFactor}})
  and by the Breit-Rabi analytical expression for energy levels of the interacting nuclei.
  The related \eqn{B}s are computed by the fixed-point iterations,
  because the corresponding \eqn{g}-value for each of the HFS-lines is not known (Weil JA (1971)
  and Stoll A, Schweiger A (2006)). The shape of spectral lines are calculated by the analytical formula
  of linear combination of the Gaussian and the Lorentzian line-shapes (also referred to as pseudo-Voigt,
  Weil JA, Bolton JR (2007) and Stoll S (2024)).  The linear coefficients
  are defined by the \code{lineG.content} argument, actually, corresponding to Gaussian line content (the Lorentzian one
  is computed as 1-\code{lineG.content}, accordingly). The linewidth, from that linear combination,
  is defined individually for the Gaussian and the Lorentzian (please, refer to the \code{lineGL.DeltaB} argument).
  The multiplicities (relative intensity ratios) are computed by the binomial/multinomial coefficients
  taking into account the spin quantum numbers of the interacting nuclei, in each of the equivalent groups (please, also
  refer to the \code{\link{plot_eval_EPRtheo_mltiplet}}), as well as their natural abundance (if \code{natur.abund = TRUE}).
}
\examples{
## simulation of simple EPR spectrum (without hyperfine
## structure) with g(iso) = 1.9804 and the linewidth
## ∆Bpp = 3.2 G, only Gaussian lineform is considered:
sim.simple.a <-
  eval_sim_EPR_iso(g.iso = 1.9804,
    instrum.params = c(Bcf = 3490,
                       Bsw = 200,
                       Npoints = 1600,
                       mwGHz = 9.8943),
    lineGL.DeltaB = list(3.2,NULL)
  )
## simulation preview:
sim.simple.a$plot
#
## check the g_{iso}-value
## of the previous spectrum
eval_gFactor_Spec(
  sim.simple.a$df,
  nu.GHz = 9.8943,
  Intensity = "dIeprSim_over_dB",
  B = "Bsim_G"
 )
#
## simulation of luteolin radical anion with
## the following four hyperfine coupling constants
## A(1 x 1H) = 3.1 MHz, A(1 x 1H) = 2.8 MHz,
## A(1 x 1H) = 8.0 MHz and A(1 x 1H) = 4.1 MHz,
## one may check out the simulation
## at https://doi.org/10.1016/j.electacta.2013.06.136
## (see Fig. 6 in that article):
sim.luteol <-
  eval_sim_EPR_iso(g.iso = 2.00495,
    instrum.params = c(Bcf = 339.367,
                       Bsw = 5.9,
                       Npoints = 2048,
                       mwGHz = 9.5294),
    nuclear.system = list(list("1H",1,3.1),
                          list("1H",1,2.8),
                          list("1H",1,8.0),
                          list("1H",1,4.1)),
    lineGL.DeltaB = list(0.034,0.034),
    lineG.content = 0.6,
    B.unit = "mT"
 )
#
## simulated spectrum preview within
## the B = (338-341) mT region:
sim.luteol$plot +
  ggplot2::coord_cartesian(xlim = c(338,341))
#
## ...and the corresponding data frame:
head(sim.luteol$df)
#
## simulation of phenalenyl/perinaphthenyl
## (PNT) radical in the integrated form:
eval_sim_EPR_iso(g = 2.0027,
  instrum.params = c(Bcf = 3500, # central field
                     Bsw = 100, # sweep width
                     Npoints = 4096,
                     mwGHz = 9.8), # MW Freq. in GHz
  B.unit = "G",
  nuclear.system = list(
      list("1H",3,5.09), # 3 x A(1H) = 5.09 MHz
      list("1H",6,17.67) # 6 x A(1H) = 17.67 MHz
   ),
  lineSpecs.form = "integrated",
  lineGL.DeltaB = list(0.54,NULL), # Gauss. FWHM in G
  Intensity.sim = "single_Integ",
  plot.sim.interact = TRUE
 )
#
## simulation of methyl-viologen radical cation (MV*+)
## as reported in following papers:
## https://doi.org/10.1039/C5CP04259C (PCCP 2015)
## https://doi.org/10.1021/ja00297a006 (JACS 1985)
## https://doi.org/10.1021/j150668a021 (JPC 1984)
## https://doi.org/10.1016/0022-2364(80)90258-9
## (J MAGN RESON 1969)
## additionally, showing time (in seconds) spent
## for the computation/simulation (by the `system.time()`)
system.time(
  methylViol.rad.cat.sim <-
    eval_sim_EPR_iso(
      g.iso = 2.0023,
      instrum.params = c(
        Bcf = 3400,
        Bsw = 60,
        Npoints = 2048,
        mwGHz = 9.53
      ),
      nuclear.system = list(
        list("14N",2,11.85), # 4.23 G
        list("1H",6,11.18), # 3.99 G
        list("1H",4,4.46), # 1.59 G
        list("1H",4,3.78) # 1.35 G
      ),
      lineG.content = 0.5,
      lineGL.DeltaB = list(0.2,0.2)
    )
 )
#
## simulated spectrum of MV*+
methylViol.rad.cat.sim$plot +
  ggplot2::coord_cartesian(
    xlim = c(3373,3427)
  )


}
\references{
Weil JA, Bolton JR (2007). \emph{Electron paramagnetic resonance: elementary theory and practical applications},
  2nd edition. John Wiley and Sons. ISBN 978-0-471-75496-1,
  \url{https://onlinelibrary.wiley.com/doi/book/10.1002/0470084987}.

  Weil JA (1971). “The Analysis of Large Hyperfine Splitting in Paramagnetic Resonance Spectroscopy.”
  \emph{J. Magn. Reson.} (1969), \strong{4}(3), 394–399, \url{https://doi.org/10.1016/0022-2364(71)90049-7}.

  Stoll S, Schweiger A (2006). “EasySpin, A Comprehensive Software Package for Spectral Simulation
  and Analysis in EPR.” \emph{J. Magn. Reson.}, \strong{178}(1), 42–55,
  \url{https://doi.org/10.1016/j.jmr.2005.08.013}.

  Stoll S (2024). “EasySpin Documentation - Line Shapes.”,
  \url{https://easyspin.org/easyspin/documentation/lineshapes.html}.

  Gerson F, Huber W (2003). \emph{Electron Spin Resonance Spectroscopy of Organic Radicals},
  Biotechnology Series, Wiley-VCH, ISBN 978-3-527-30275-8, \url{https://books.google.cz/books?id=SEPeNjG3IvYC}.
}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_ABIC_forFit}()},
\code{\link{eval_sim_EPR_isoFit}()},
\code{\link{eval_sim_EPR_isoFit_space}()},
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{optim_for_EPR_fitness}()},
\code{\link{plot_eval_EPRtheo_mltiplet}()},
\code{\link{plot_eval_RA_forFit}()},
\code{\link{quantify_EPR_Sim_series}()},
\code{\link{smooth_EPR_Spec_by_npreg}()}
}
\concept{Simulations and Optimization}
