% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_iso.R
\name{eval_sim_EPR_iso}
\alias{eval_sim_EPR_iso}
\title{Simulation of Isotropic EPR Spectra}
\usage{
eval_sim_EPR_iso(
  g.iso = 2.00232,
  B.CF,
  B.SW,
  B.unit = "G",
  Npoints,
  nu.GHz = 9.8,
  nuclear.system = NULL,
  natur.abund = FALSE,
  lineGL.DeltaBpp = list(1, NULL),
  lineGL.content = list(1, NULL),
  Intensity.sim = "dIeprSim_over_dB",
  plot.sim.interact = FALSE
)
}
\arguments{
\item{g.iso}{Numeric value, guess of the isotropic \eqn{g}-factor. It may also possess a `NULL`
(\code{g.iso = NULL}) in case of the `B.CF` corresponding \eqn{g} is equal to `g.iso`. \strong{Default}:
\code{g.iso = 2.00232} (the approximate \eqn{g} of the free electron).}

\item{B.CF}{Numeric value of the "\strong{c}entral \strong{f}ield" i.e. the central position
of magnetic flux density \eqn{B} within the spectral sweep => see also argument \code{B.SW}.
\code{B.CF} UNIT (the corresp. B.CF value) MUST BE EQUAL TO \code{B.unit} !}

\item{B.SW}{Numeric value equal to the \eqn{B} region (\eqn{B.CF - B.SW / 2}, eqn{B.CF + B.SW / 2})
of the spectral "\strong{s}weep \strong{w}idth". \code{B.SW} UNIT (the corresp. B.CF value) MUST
BE EQUAL TO \code{B.unit} !}

\item{B.unit}{Character string pointing to unit of magnetic flux density which is to be presented
on \eqn{B} abscissa of the EPR spectrum, like \code{"G"} (`Gauss`) or \code{"mT"} (`millitesla`),
\strong{default}: \code{B.unit = "G"}. THE UNIT MUST BE SHARED ACROSS ALL B ARGUMENTS like \code{B.CF},
\code{B.SW} AND \code{lineGL.DeltaBpp} !}

\item{Npoints}{Numeric value corresponding to number of points of the entire \code{B.SW}
(\eqn{\equiv } resolution).}

\item{nu.GHz}{Numeric value equal to either hypothetical or applied microwave frequency in \strong{GHz}
to record/simulate the EPR spectrum. \strong{Default}: \code{nu.GHz = 9.8} (usually applied X-band frequency).}

\item{nuclear.system}{List containing the information about groups of equivalent nuclei
interacting with the unpaired electron, like \code{nuclear.system} ...tbc...if only one,
it could be either nested or simple.}

\item{natur.abund}{...tbc...}

\item{lineGL.DeltaBpp}{...tbc...}

\item{lineGL.content}{...tbc...}

\item{Intensity.sim}{...tbc...}

\item{plot.sim.interact}{description ...tbc...}
}
\value{
...list...tbc or interactive simulated `plotly` spectrum
}
\description{
A short description...tbc...of the EPR simulation
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Evaluations and Quantum Chemistry: 
\code{\link{eval_gFactor_QCHcomp}()},
\code{\link{rearrange_aAiso_QCHcomp}()},
\code{\link{rearrange_aAiso_QCHorgau}()}
}
\concept{Evaluations and Quantum Chemistry}
