% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_iso_combo.R
\name{eval_sim_EPR_iso_combo}
\alias{eval_sim_EPR_iso_combo}
\title{Simulation of Isotropic EPR Spectra Consisting of Several Components}
\usage{
eval_sim_EPR_iso_combo(
  g.iso.vec,
  instrum.params = c(cf = 3500, sw = 200, points = 2048, mwGHz = 9.8),
  B.unit = "G",
  path_to_DSC_or_par = NULL,
  origin = NULL,
  nuclear.system,
  natur.abund.vec,
  lineSpecs.form = "derivative",
  lineGL.DeltaB,
  lineG.content.vec,
  Intensity.sim = "dIeprSim_over_dB",
  Intensity.sim.coeffs.vec,
  plot.sim.interact = FALSE
)
}
\arguments{
\item{g.iso.vec}{Nemeric vector...tbc...}

\item{instrum.params}{Named numeric vector containing instrumental parameters required
for the simulation =>
\tabular{ll}{
\code{cf} \tab "central field" (magnetic flux density, \eqn{B_{\text{CF}}}) \cr
\code{sw} \tab "sweep width" (magnetic flux density recording region,
\eqn{B_{\text{SW}}}) \cr
\code{points} \tab number of spectral points (corresponding to resolution) within
the "sweep width" \cr
\code{mwGHz} \tab applied microwave frequency in `GHz` to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default} values are chosen to cover the EPR spectra of common organic radicals.
If \code{instrum.params = NULL} then parameters are provided by \code{path_to_DSC_or_par}
as well as by \code{origin}.}

\item{B.unit}{Character string pointing to unit of magnetic flux density which is to be presented
on \eqn{B} abscissa of the EPR spectrum, like \code{"G"} (`Gauss`) or \code{"mT"} (`millitesla`),
\strong{default}: \code{B.unit = "G"}. THE UNIT MUST BE SHARED ACROSS ALL B ARGUMENTS
like `cf` and `sw` within the \code{instrum.params} AS WELL AS THOSE IN \code{lineGL.DeltaB} !}

\item{path_to_DSC_or_par}{Character string, path (can be also gathered by \code{\link[base]{file.path}})
to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} argument)
\code{text} files including all instrumental parameters from the EPR machine.
\strong{Default}: \code{path_to_DSC_or_par = NULL} in case if the \code{instrum.params}
is already defined. IF ARGUMENT \code{instrum.params = NULL} then BOTH \code{path_to_DSC_or_par}
AS WELL AS \code{origin} MUST BE DEFINED !}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_DSC_or_par = NULL}).}

\item{nuclear.system}{Nested list ...tbc...}

\item{natur.abund.vec}{Logical vector ...tbc ...}

\item{lineSpecs.form}{Character string describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{lineGL.DeltaB}{Nested list ...tbc ...}

\item{lineG.content.vec}{Numeric vector ...tbc...}

\item{Intensity.sim}{Character string pointing to column of simulated EPR intensity within the related
data frame, which is also available in the output list. \strong{Default}: \code{Intensity.sim = "dIeprSim_over_dB"}.}

\item{Intensity.sim.coeffs.vec}{Numeriv vector ...tbc...}

\item{plot.sim.interact}{Logical ...tbc ...}
}
\value{
List or interactive plotly ...tbc...
}
\description{
A short description...tbc...
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Simulation and Optimization: 
\code{\link{eval_sim_EPR_iso}()},
\code{\link{quantify_EPR_sim}()}
}
\concept{Simulation and Optimization}
