% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_iso_combo.R
\name{eval_sim_EPR_iso_combo}
\alias{eval_sim_EPR_iso_combo}
\title{Simulation of Isotropic EPR Spectra Consisting of Several Components}
\usage{
eval_sim_EPR_iso_combo(
  g.iso.vec,
  instrum.params = c(Bcf = 3500, Bsw = 200, Npoints = 2048, mwGHz = 9.8),
  B.unit = "G",
  path_to_dsc_par = NULL,
  origin = NULL,
  nuclear.system,
  natur.abund.vec,
  lineSpecs.form = "derivative",
  lineGL.DeltaB,
  lineG.content.vec,
  Intensity.sim = "dIeprSim_over_dB",
  Intensity.sim.coeffs.vec,
  plot.sim.interact = NULL
)
}
\arguments{
\item{g.iso.vec}{Numeric vector...tbc...}

\item{instrum.params}{Named numeric vector containing instrumental parameters required
for the simulation =>
\tabular{ll}{
\code{Bcf} \tab "central field" (magnetic flux density, \eqn{B_{\text{CF}}}) \cr
\code{Bsw} \tab "sweep width" (magnetic flux density recording region,
\eqn{B_{\text{SW}}}) \cr
\code{Npoints} \tab number of spectral points (corresponding to resolution) within
the "sweep width" \cr
\code{mwGHz} \tab applied microwave frequency in `GHz` to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default} values are chosen to cover the EPR spectra of common organic radicals.
If \code{instrum.params = NULL} then parameters are provided by \code{path_to_dsc_par}
as well as by \code{origin}.}

\item{B.unit}{Character string pointing to unit of magnetic flux density which is to be presented
on \eqn{B} abscissa of the EPR spectrum, like \code{"G"} (`Gauss`) or \code{"mT"} (`millitesla`),
\strong{default}: \code{B.unit = "G"}. THE UNIT MUST BE SHARED ACROSS ALL B ARGUMENTS
like `cf` and `sw` within the \code{instrum.params} AS WELL AS THOSE IN \code{lineGL.DeltaB} !}

\item{path_to_dsc_par}{Character string, path (can be also gathered by \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on the OS, see \code{origin} argument)
\code{text} files including all instrumental parameters from the EPR machine.
\strong{Default}: \code{path_to_dsc_par = NULL} in case if the \code{instrum.params}
is already defined. IF ARGUMENT \code{instrum.params = NULL} then BOTH \code{path_to_dsc_par}
AS WELL AS \code{origin} MUST BE DEFINED !}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_dsc_par = NULL}).}

\item{nuclear.system}{Nested list ...tbc...}

\item{natur.abund.vec}{Logical vector ...tbc ...}

\item{lineSpecs.form}{Character string describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{lineGL.DeltaB}{Nested list ...tbc ...}

\item{lineG.content.vec}{Numeric vector ...tbc...}

\item{Intensity.sim}{Character string pointing to column of simulated EPR intensity within the related
data frame, which is also available in the output list. \strong{Default}: \code{Intensity.sim = "dIeprSim_over_dB"}.}

\item{Intensity.sim.coeffs.vec}{Numeric vector ...tbc...}

\item{plot.sim.interact}{Character string "components" or "sum" or NULL (\code{default})}
}
\value{
List of the following data frames and plots in case of \code{plot.sim.interact = NULL} =>

  * long-format data frame with simulation components A, B, C, ...(e.g. representing the individual
    radicals) as categorical variable + magnetic flux density and intensity denoted as \code{df}

  * data frame with simulation components A, B, C ...and their corresponding double/single
    integrals (`areas`) and with their related ratios to overall integral sum (`weighted_areas`)
    denoted as \code{df.areas}

  * data frame with the overall intensity (+ magnetic flux density) as well as integral sum
    from all simulation components denoted as \code{df.sum}

  * overlay plot with all simulated components with their corresponding intensities defined
    by \code{Intensity.sim.coeffs.vec} denoted as \code{plot.comps}

  * plot displaying the sum of all simulation components denoted as \code{plot.sum}.

  If \code{plot.sim.interact} is activated (i.e. possesses either \code{components}
  or \code{sum} values) interactive plots (based on `plotly`) are presented either with
  all individual components or with the overall simulated sum EPR spectrum, respectively.
}
\description{
A short description...tbc...
}
\examples{
## Simulation of EPR spectrum of TEMPO (nitroxyl) radical with 13C satellites
## and hyperfine coupling constants A(1 x 14N) = 45 MHz and A(1 x 13C) = 18 MHz.
sim.tempo.13c <- eval_sim_EPR_iso_combo(g.iso.vec = c(2.0059,2.0059),
                                        nuclear.system = list(list("14N",1,45),
                                                              list(list("14N",1,45),
                                                                   list("13C",1,18))),
                                        natur.abund.vec = c(TRUE,TRUE),
                                        lineGL.DeltaB = list(list(1.2,NULL),
                                                             list(1.2,NULL)),
                                        lineG.content.vec = c(1,1),
                                        Intensity.sim.coeffs.vec = c(1,1))
sim.tempo.13c$plot.sum + ggplot2::coord_cartesian(xlim = c(3425,3550))
#
## ...and the corresponding data frame =>
sim.tempo.13c$df.sum[1000:1005,]
#
## The data frame with all components =>
sim.tempo.13c$df[1000:1005,]
#
## The data frame with all integrals (areas)
sim.tempo.13c$df.areas


}
\seealso{
Other Simulation and Optimization: 
\code{\link{eval_sim_EPR_iso}()},
\code{\link{optim_fit_EPR_by_nloptr}()},
\code{\link{quantify_EPR_sim}()}
}
\concept{Simulation and Optimization}
