% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_iso_combo.R
\name{eval_sim_EPR_iso_combo}
\alias{eval_sim_EPR_iso_combo}
\title{Simulation of Isotropic EPR Spectra Consisting of Several Components}
\usage{
eval_sim_EPR_iso_combo(
  g.iso.vec,
  instrum.params = c(Bcf = 3500, Bsw = 200, Npoints = 2048, mwGHz = 9.8),
  B.unit = "G",
  path_to_dsc_par = NULL,
  origin = NULL,
  nuclear.system,
  natur.abund.vec,
  lineSpecs.form = "derivative",
  lineGL.DeltaB,
  lineG.content.vec,
  Intensity.sim = "dIeprSim_over_dB",
  Intensity.sim.coeffs.vec,
  plot.sim.interact = NULL
)
}
\arguments{
\item{g.iso.vec}{Numeric vector with all \eqn{g_{\text{iso}}} values fro each component.}

\item{instrum.params}{Named numeric vector, containing instrumental parameters required
for the simulation =>
\tabular{ll}{
\code{Bcf} \tab "central field" (magnetic flux density, \eqn{B_{\text{CF}}}) \cr
\code{Bsw} \tab "sweep width" (magnetic flux density recording region,
\eqn{B_{\text{SW}}}) \cr
\code{Npoints} \tab number of spectral points (corresponding to resolution) within
the "sweep width" \cr
\code{mwGHz} \tab applied microwave frequency in \code{GHz} to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default} values are chosen to cover the EPR spectra of common organic radicals.
If \code{instrum.params = NULL} then parameters must be provided by the \code{path_to_dsc_par}
as well as by \code{origin} arguments.}

\item{B.unit}{Character string, pointing to unit of magnetic flux density which is to be presented
on \eqn{B}-axis of the EPR spectrum, like \code{"G"} (Gauss) or \code{"mT"} (millitesla),
\strong{default}: \code{B.unit = "G"}. THE UNIT MUST BE SHARED ACROSS ALL RELEVANT B-ARGUMENTS
like \code{cf} and \code{sw} within the \code{instrum.params} AS WELL AS within THOSE IN \code{lineGL.DeltaB} !}

\item{path_to_dsc_par}{Character string, path (can be also acquired by the \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on the OS, see \code{origin} argument)
\code{text} files including all instrumental parameters from the EPR machine.
\strong{Default}: \code{path_to_dsc_par = NULL} in case if the \code{instrum.params}
is already defined. IF the \code{instrum.params = NULL} then BOTH the \code{path_to_dsc_par}
AS WELL AS the \code{origin} MUST BE DEFINED !}

\item{origin}{Character string, corresponding to software which was used to obtain the EPR spectra
on spectrometers, because the files are slightly different, whether they
were recorded by the "WinEpr" (\code{origin = "winepr"}) or by the "Xenon".
\strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_dsc_par = NULL}).}

\item{nuclear.system}{Nested list with the elements corresponding to considered interacting nuclei for each EPR
spectral component. For example, the \code{list(list("1H",2,24),NULL,list(list("14N",1,45),list("1H",4,15)))}
refers to the following components: 1. \eqn{2\times A(\text{1H}) = 24\,\text{MHz}}, 2. single line spectrum
without HF structure, 3. \eqn{1\times A(\text{14N}) = 45\,\text{MHz} + 4\times A(\text{1H}) = 15\,\text{MHz}}.}

\item{natur.abund.vec}{Logical vector, whether to consider natural abundance of the interacting nuclei within
the components (see also \code{Examples}) like \code{c(TRUE,FALSE,TRUE)}.}

\item{lineSpecs.form}{Character string, describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{lineGL.DeltaB}{Nested list of the Gaussian and Lorentzian linewidths for all individual components
like \code{list(list(1,NULL),list(3,NULL),list(1,NULL))}.}

\item{lineG.content.vec}{Numeric vector, corresponding to Gaussian line content for all individual components
of the EPR spectrum like \code{1,1,1} (all spectral components are described by the pure Gaussian line).}

\item{Intensity.sim}{Character string, pointing to column of simulated EPR intensity within the related output
data frame. \strong{Default}: \code{Intensity.sim = "dIeprSim_over_dB"}.}

\item{Intensity.sim.coeffs.vec}{Numeric vector of multiplication coefficients related to EPR component intensities
like \code{c(2,10,0.2)}.}

\item{plot.sim.interact}{Character string, indicating the interactive plot outputs, to visualize either individual
\code{"components"} or \code{"sum"} of all components. As \strong{default}, the interactive plot is switched
off (\code{plot.sim.interact = NULL}).}
}
\value{
List of the following data frames and plots in case of
  \code{plot.sim.interact = NULL} =>
   \describe{
   \item{df}{Long-format data frame with the simulated EPR spectral components A, B, C, ...
    (e.g. representing the individual radicals) as a categorical variable
    + magnetic flux density, intensity as well as their sigmoid integral
    column/variable.}
    \item{df.areas}{Data frame with simulation components A, B, C ...and their
    corresponding double/single integrals (or \code{areas}) and their relative ratios
    to overall integral sum (\code{weighted_areas}).}
    \item{df.sum}{Data frame with the overall intensity (+ magnetic flux density)
    as well as integral sum from all simulation components.}
    \item{plot.comps}{Overlay plot object with all simulated components with their
    corresponding intensities defined by \code{Intensity.sim.coeffs.vec}.}
    \item{plot.sum}{Plot object displaying the sum of all simulation components.}
   }
  If \code{plot.sim.interact} is activated (i.e. possesses either \code{"components"}
  or \code{"sum"} values) interactive plots (based on \code{plotly}) are presented either with
  all individual components or with the overall simulated EPR spectrum sum, respectively.
}
\description{
This is an extension of the "basic" EPR simulation provided by the \code{\link{eval_sim_EPR_iso}} function,
  where one can combine several simulated EPR spectra (components), even with (partial) overlay,
  into one spectrum corresponding to sum of all components. Such processing might be useful for the simulation
  of EPR spectra with satellites, especially of those consisted of naturally occurring
  isotopes, like the one presented in \code{Examples}.
}
\examples{
## Simulation of EPR spectrum of TEMPO (aminoxyl)
## radical with 13C satellites and hyperfine coupling
## constants A(1 x 14N) = 48 MHz, A(1 x 13C) = 18.2 MHz,
## the latter 13C may appear on 4 different
## positions (methyl carbons) => therefore the overall
## probability to find them is approx. 4,4\% (1.1\% per one 13C),
## the additional two quaternary α-Carbons are not considered
## see e.g. https://doi.org/10.1016/j.mencom.2014.09.018
sim.tempo.13c <-
eval_sim_EPR_iso_combo(g.iso.vec = c(2.0059,2.0059),
                       nuclear.system = list(list("14N",1,48),
                                             list(list("14N",1,48),
                                                  list("13C",1,18.5)
                                               )
                                            ),
                       natur.abund.vec = c(FALSE,FALSE),
                       lineGL.DeltaB = list(list(1.3,NULL),
                                            list(1.3,NULL)
                                            ),
                       lineG.content.vec = c(1,1),
                       Intensity.sim.coeffs.vec = c(0.956/3,0.044/6))
#
## simulated spectrum/plot:
sim.tempo.13c$plot.sum +
ggplot2::coord_cartesian(xlim = c(3425,3550))
#
## ...and the corresponding data frame:
options(pillar.sigfig = 5) ## prevent rounding
#
tempo.df.sum <- sim.tempo.13c$df.sum
tempo.df.sum[1000:1005, ]
#
## data frame with all components:
sim.tempo.13c$df[1000:1005,]
#
## areas/integrals
sim.tempo.13c$df.areas


}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_sim_EPR_iso}()},
\code{\link{eval_sim_EPR_isoFit}()},
\code{\link{optim_for_EPR_fitness}()},
\code{\link{quantify_EPR_Sim_series}()},
\code{\link{smooth_EPR_Spec_by_npreg}()}
}
\concept{Simulations and Optimization}
