% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_isoFit.R
\name{eval_sim_EPR_isoFit}
\alias{eval_sim_EPR_isoFit}
\title{Least-Squares Fitting of Isotropic EPR Spectra by Simulations}
\usage{
eval_sim_EPR_isoFit(
  data.spectr.expr,
  Intensity.expr = "dIepr_over_dB",
  Intensity.sim = "dIeprSim_over_dB",
  nu.GHz,
  B.unit = "G",
  Blim = NULL,
  nuclear.system.noA = NULL,
  baseline.correct = "constant",
  lineG.content = 0.5,
  lineSpecs.form = "derivative",
  optim.method = "neldermead",
  optim.params.init,
  optim.params.fix.id = NULL,
  optim.params.lower = NULL,
  optim.params.upper = NULL,
  Nmax.evals = 512,
  check.fit.plot = TRUE,
  msg.optim.progress = TRUE,
  output.list.forFitSp = FALSE,
  ...
)
}
\arguments{
\item{data.spectr.expr}{Data frame object/table, containing the experimental spectral data with the magnetic flux density
(\code{"B_mT"} or \code{"B_G"}) and the intensity (see the \code{Intensity.expr} argument) columns.}

\item{Intensity.expr}{Character string, pointing to column name of the experimental EPR intensity within
the original \code{data.spectr.expr}. \strong{Default}: \code{dIepr_over_dB}.}

\item{Intensity.sim}{Character string, pointing to column name of the simulated EPR intensity within the related output
data frame. \strong{Default}: \code{Intensity.sim = "dIeprSim_over_dB"}.}

\item{nu.GHz}{Numeric value, microwave frequency in \code{GHz}.}

\item{B.unit}{Character string, denoting the magnetic flux density unit e.g. \code{B.unit = "G"}
(gauss, \strong{default}) or \code{B.unit = "mT"}/\code{"T"} (millitesla/tesla).}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G}
corresponding to lower and upper \strong{visual limit} of the selected \eqn{B}-region,
such as \code{Blim = c(3495.4,3595.4)}. \strong{Default}: \code{Blim = NULL} (corresponding to the entire
\eqn{B}-range of EPR spectrum). \strong{This does not correspond to simulation fit data region !}.
If narrower \eqn{B}-region (in comparison to the original one) is required to fit the EPR spectrum,
the filtering has to be done prior to own fitting procedure. For example, if the original data frame
(\code{df.spectr.orgin} within 200 G), of the experimental EPR spectrum, should be fitted within
the region of \code{B = c(3450,3550)} (100 G), following operation must be performed:
\code{df.spectr.actuall <- df.spectr.origin |> dplyr::filter(dplyr::between(B_G,3450,3550))},
where the \code{df.spectr.actuall} serves as an input (represented by the \code{data.spectr.expr}
argument) for the function.}

\item{nuclear.system.noA}{List or nested list \strong{without estimated hyperfine coupling constant values},
such as \code{list("14N",1)} or \code{list(list("14N", 2),list("1H", 4),list("1H", 12))}. The \eqn{A}-values
are already defined as elements of the \code{optim.params.init} argument/vector. If the EPR spectrum
does not display any hyperfine splitting, the argument definition reads \code{nuclear.system.noA = NULL} (\strong{default}).}

\item{baseline.correct}{Character string, referring to baseline correction of the simulated/fitted spectrum.
Corrections like \code{"constant"} (\strong{default}), \code{"linear"} or \code{"quadratic"} can be applied.}

\item{lineG.content}{Numeric value between \code{0} and \code{1}, referring to content of the \emph{Gaussian} line form.
If \code{lineG.content = 1} (\strong{default}) it corresponds to "pure" \emph{Gaussian} line form
and if \code{lineG.content = 0} it corresponds to \emph{Lorentzian} one. The value from (0,1)
(e.g. \code{lineG.content = 0.5}) represents the linear combination (for the example above,
with the coefficients 0.5 and 0.5) of both line forms => so called \emph{pseudo-Voigt}.}

\item{lineSpecs.form}{Character string, describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{optim.method}{Character string (vector), setting the optimization method(s) gathered within
the \code{\link{optim_for_EPR_fitness}}. \strong{Default}: \code{optim.method = "neldermead"}. Additionally,
several consecutive methods can be defined like \code{optim.method = c("levenmarq","neldermead")}, where
the best fit parameters from the previous method are used as input for the next one. In such case, the output
is \code{list} with the elements/vectors from each method, in order to see the progress of the optimization.}

\item{optim.params.init}{Numeric vector with the initial parameter guess (elements) where the \strong{first five
  elements are immutable}
  \enumerate{
  \item g-value (g-factor)

  \item \strong{G}aussian linewidth

  \item \strong{L}orentzian linewidth

  \item baseline constant (intercept or offset)

  \item intensity multiplication constant

  \item baseline slope (only if \code{baseline.correct = "linear"} or \code{baseline.correct = "quadratic"}),
  if \code{baseline.correct = "constant"} it corresponds to the \strong{first HFCC} (\eqn{A_1})

  \item baseline quadratic coefficient (only if \code{baseline.correct = "quadratic"}),
  if \code{baseline.correct = "constant"} it corresponds to the \strong{second HFCC} (\eqn{A_2}),
  if \code{baseline.correct = "linear"} it corresponds to the \strong{first HFCC} (\eqn{A_1})

  \item additional HFCC (\eqn{A_3}) if \code{baseline.correct = "constant"} or if
  \code{baseline.correct = "linear"} (\eqn{A_2}), if \code{baseline.correct = "quadratic"} it corresponds
  to the \strong{first HFCC} (\eqn{A_1})

  \item ...additional HFCCs (\eqn{A_k...}, each vector element is reserved only for one \eqn{A})
  }
  DO NOT PUT ANY OF THESE PARAMETERS to \code{NULL}. If the lineshape is expected to be pure
  \strong{L}orentzian or pure \strong{G}aussian then put the corresponding vector element to \code{0}.}

\item{optim.params.fix.id}{Numeric value/vector of index/indices of the \code{optim.params.init}, corresponding
to optimization/simulation parameter(s) to be fixed. For example, if the g-Value and the (intensity) multiplication
constant should not be optimized (their values should be fixed during the procedure), the argument must be defined
as follows: \code{optim.params.fix.id = c(1,5)} (see also the \code{optim.params.init} description
for the simulation/optimized parameters order). \strong{Default}: \code{optim.params.fix.id = NULL}, indicating
that none of the \code{optim.params.init} is fixed, i.e. all parameters are optimized within their default/defined
boundaries (see the \code{optim.params.lower} and the \code{optim.params.upper}). Alternatively, the parameter value(s)
can be also adjusted by assigning the \code{optim.params.init} + \code{optim.params.lower} + \code{optim.params.upper}
elements to the same value, as already demonstrated in the \code{Examples}.}

\item{optim.params.lower}{Numeric vector (with the same element order like \code{optim.params.init})
with the lower bound constraints. \strong{Default}: \code{optim.params.lower = NULL} which actually
equals to \eqn{g_{\text{init}} - 0.001}, \eqn{0.8\,\Delta B_{\text{G,init}}},
\eqn{0.8\,\Delta B_{\text{L,init}}}, baseline intercept initial constant \eqn{- 0.001},
intensity multiplication initial constant \eqn{= 0.75\,\text{init}}, baseline initial slope \eqn{- 5} (in case
the \code{baseline.correct} is set either to \code{"linear"} or \code{"quadratic"}) and finally,
the baseline initial quadratic coefficient \eqn{- 5} (in case the \code{baseline.correct} is set to
\code{"quadratic"}). Default lower limits of all hyperfine coupling constant (HFCCs) are set to \eqn{0.875\,A_{\text{init}}}.}

\item{optim.params.upper}{Numeric vector (with the same element order like \code{optim.params.init})
with the upper bound constraints. \strong{Default}: \code{optim.params.upper = NULL} which actually
equals to \eqn{g_{\text{init}} + 0.001}, \eqn{1.2\,\Delta B_{\text{G,init}}},
\eqn{1.2\,\Delta B_{\text{L,init}}}, baseline intercept initial constant \eqn{+ 0.001},
intensity multiplication initial constant \eqn{= 1.25\,\text{init}}, baseline initial slope \eqn{+ 5} (in case
the \code{baseline.correct} is set either to \code{"linear"} or \code{"quadratic"}) and finally,
the baseline initial quadratic coefficient \eqn{+ 5} (in case the \code{baseline.correct} is set to
\code{"quadratic"}). Default upper limits of all HFCCs are set to \eqn{1.125\,A_{\text{init}}}.}

\item{Nmax.evals}{Numeric value, maximum number of function evaluations and/or iterations.
The only one method, limited by this argument, is \code{\link[minpack.lm]{nls.lm}}, where
\code{Nmax.evals = 1024}. Higher \code{Nmax.evals} may extremely extend the optimization
time, therefore the \strong{default} value reads \code{Nmax.evals = 512}. However, the \code{"pswarm"}
method requires at least the default or even higher values.}

\item{check.fit.plot}{Logical, whether to return overlay plot with the initial simulation + the best simulation
fit + experimental spectrum (including residuals in the lower part of the plot,
\code{check.fit.plot = TRUE}, \strong{default}) or with the following three spectra
(\code{check.fit.plot = FALSE}): 1. experimental, 2. the best simulated one with the baseline fit
and 3. the best simulated spectrum with the baseline fit subtracted. The latter two are offset for clarity,
within the plot.}

\item{msg.optim.progress}{Logical, whether to display message (in the R console) about progress of the \code{optim.method}
during the optimization/fitting procedure, e.g.
\code{"EPR simulation parameters are currently being optimized by NELDERMEAD, method 1 of 2..."} and at the end
it shows the elapsed time. \strong{Default}: \code{msg.optim.progress = TRUE}. If \code{FALSE}, no message
is displayed. Latter option is especially suitable for the case when \code{output.list.forFitSp = TRUE} (see below).
This argument can be combined with the optional \code{eval.optim.progress} (see below or described in the
\code{\link{optim_for_EPR_fitness}}).}

\item{output.list.forFitSp}{Logical. If \code{TRUE}, \code{list} with the following components will be exclusively returned:
1. optimized parameters from the best fit (together with the minimum sum of residual squares)
and 2. Plot of the experimental as well as simulated EPR spectrum depending on \code{check.fit.plot} (see \code{Value}).
Such output is applied for the \code{\link{eval_sim_EPR_isoFit_space}},
therefore, the \strong{default} value reads \code{output.list.final = FALSE}.}

\item{...}{additional arguments specified, see also \code{\link{optim_for_EPR_fitness}},
like \code{eval.optim.progress = TRUE} (which is \code{FALSE} by \strong{default}), \code{pswarm.size},
\code{pswarm.diameter}, \code{pswarm.type}, \code{tol.step} + \code{optim.method} depended additional arguments.}
}
\value{
Optimization/Fitting procedure results in vector or data frame or list depending on the \code{check.fit.plot}
  and \code{output...} arguments.
  \enumerate{
  \item If \code{check.fit.plot = TRUE} or \code{check.fit.plot = FALSE}, the result corresponds
  to list with the following components:
  \describe{
  \item{plot}{Visualization of the experimental as well as the best fitted EPR simulated spectra.
  If \code{check.fit.plot = TRUE}, the overlay plot consists of the initial simulation + the best simulation
  fit + experimental spectrum, including residuals in the plot lower part. Whereas, if \code{check.fit.plot = FALSE},
  following three spectra are available: 1. experimental, 2. the best simulated one with the baseline fit
  and 3. the best simulated spectrum with the baseline fit subtracted. The latter two are offset for clarity.}
  \item{ra}{Simple residual analysis - a list consisting of 4 elements: diagnostic plots
  \code{plot.rqq()} function, \code{plot.histDens}; original data frame (\code{df}) with residuals and their corresponding
  standard deviation (\code{sd}). For details, please refer to the \code{\link{plot_eval_RA_forFit}}.}
  \item{best.fit.params}{Vector of the best (final) fitting (optimized) parameters, for each corresponding
  \code{optim.method}, to simulate the experimental EPR spectrum, see also description of the \code{optim.params.init}.}
  \item{df}{Tidy data frame (table) with the magnetic flux density and intensities of the experimental,
  the best simulated/fitted, as well as the initially simulated EPR spectrum and residuals
  (if \code{check.fit.plot = TRUE}), or wide data frame with the following variables / columns
  (for \code{check.fit.plot = FALSE}): magnetic flux density, intensity of the experimental
  spectrum, intensity of the best simulated one (including the baseline fit), residual intensity and finally,
  the best simulated spectrum intensity without the baseline fit.}
  \item{min.rss}{Minimum sum of residual squares (vector) after the least-square procedure.}
  \item{cov.df}{Covariance \code{matrix} of a data frame, consisting of EPR experimental, simulated (best fit) and the residual
  intensities as columns/variables. Covariance between the experiment and simulation should be positive and strong
  for a decent fit. Contrary, the \code{cov} between the simulation and residuals should be ideally close to \code{0},
  indicating no systematic relationship. However, the covariance is scale-depended and must be "normalized". Therefore,
  for such a purpose, the correlation is defined as shown below.}
  \item{cor.df}{Correlation \code{matrix} of a data frame, consisting of EPR experimental,
  simulated (best fit) and residual intensities as columns/variables. Such matrix can be additionally nicely visualized
  by a correlation \code{plot} created by the \code{\link[corrplot]{corrplot}} function.
  A higher positive correlation (between the experiment
  and the best fit), with the value close to \code{1}, indicates that simulation best fit nicely follows
  the experimental spectrum. Contrary, no clear correlation between the residuals and the experimental/fitted
  EPR intensities must be visible. Therefore, such correlation should be ideally close to \code{0}.}
  \item{abic}{A list consisting of Akaike and Bayesian information criteria (AIC & BIC) vector (\code{abic.vec})
  and \code{message}, denoting the residuals/errors distribution, applied to evaluate
  those criteria. To be used when comparing different simulation fits. The lower the (negative) values,
  the better the fit. Please, refer to the \code{\link{eval_ABIC_forFit}}.}
  \item{N.evals}{Number of iterations/function evaluations completed before termination.
  If the \code{pswarm} optimization algorithm is included in \code{optim.method}, the \code{N.evals}
  equals to vector with the following elements: number of function evaluations, number of iterations (per one particle)
  and the number of restarts.}
  \item{N.converg}{Vector or simple integer code indicating the successful completion
  of the optimization/fit. In the case of \code{"levenmarq"} method, the vector elements coincide with
  the sum of residual squares at each iteration. If the \code{optim.method = "pswarm"} is applied, one of the following
  codes can be returned: \code{0}: algorithm terminated by reaching the absolute tolerance,
  \code{1}: maximum number of function evaluations reached, \code{2}: maximum number of iterations reached,
  \code{3}: maximum number of restarts reached, \code{4}: maximum number of iterations without improvement reached.
  For all the other remaining methods (coming from \code{{nloptr}} package), the integers have to be positive
  to indicate successful convergence.}
  }

  \item If \code{output.list.forFitSp = TRUE}, the function exclusively returns list with the two components,
  which is to be applied for the \code{\link{eval_sim_EPR_isoFit_space}} or similar functions.
  \describe{
  \item{params}{A vector, containing the following elements:
  \enumerate{
  \item The best fitting (optimized) parameters (related to the \code{optim.params.init} argument).

  \item Minimum sum of residual squares (corresponding to previous item).

  \item Standard deviation of residuals, after the (final) \code{optim.method} procedure.

  \item Akaike Information Criterion/AIC metric (refer to \code{\link{eval_ABIC_forFit}}),
  after the (final) \code{optim.method}.

  \item Bayesian Information Criterion/BIC metric (refer to \code{\link{eval_ABIC_forFit}}),
  after the (final) \code{optim.method}.
  }}
  \item{plot}{Visualization of the experimental as well as the best fitted EPR simulated spectra depending
  on the \code{check.fit.plot}. It corresponds either to EPR spectra with residuals or to those with baseline correction,
  (please, refer to the \code{check.fit.plot} argument description).}
  }
  }
}
\description{
Ordinary least-squares fitting of the simulated spectrum onto the experimental one represents an important step in the analysis
  of EPR spectra. Parameters of the simulated spectrum like \eqn{g_{\text{iso}}}; coupling constants
  (in \code{MHz}) \eqn{A_{\text{iso}}} for each group of equivalent nuclei; linewidth
  (either \eqn{\Delta B_{\text{pp}}} or \eqn{FWHM} depending on the \code{lineSpecs.form} argument);
  spectral baseline (see the \code{baseline.correct} argument) and finally the intensity (multiplication coefficient)
  are optimized by methods listed in the \code{\link{optim_for_EPR_fitness}}.
  The \code{lineG.content} corresponding parameter is the only one,
  which needs to be varied "manually". For an extended version of this function
  (including automatic \code{lineG.content} variations), please refer to the \code{\link{eval_sim_EPR_isoFit_space}}.
}
\note{
In order to guess suitable intensity multiplication constant (please, refer to the \code{optim.params.init}
  argument), one might compare the intensities of the experimental (\code{expr}) and simulated (\code{sim})
  EPR spectrum by one of the interactive or static plot functions (e.g. \code{\link{plot_EPR_Specs}}
  or \code{\link{plot_EPR_Specs2D_interact}}) as well as by the \code{\link{eval_sim_EPR_iso}}. Accordingly,
  \strong{the initial intensity multiplication constant} can be estimated as the ratio
  \strong{max(\code{expr} intensity)/max(\code{sim} intensity) * 0.25}. The coefficient \code{0.25(1/4)} is
  selected in order to be sure that both of the experimental and initially simulated EPR spectra,
  in the graphical output (see the \code{Value}), are clearly visible if \code{check.fit.plot = TRUE}.
  Otherwise, the spectra may overlay and it will be difficult to differentiate between them. Of course,
  any intensity multiplication coefficient, close to the experimental EPR intensity, can be applied as a starting
  point to evaluate the fit.

  To fix one or more simulation parameter(s) (i.e. parameter(s) which is/are not optimized) during
  the fitting procedure, the corresponding \code{optim.parms.lower} as well as the \code{optim.params.upper} vector
  element(s) must equal to that of the \code{optim.params.init}. Please, refer to the \code{Examples} below
  for the simulation fit of aminoxyl radical EPR spectrum with bound constraints and fixed A(1 x 14N).
  An alternative selection of a non-optimized/fixed simulation parameter can be also done by using
  the \code{optim.params.fix.id} argument, which can be also applied to fix the simulation parameters when searching
  for the best fit by the \code{\link{eval_sim_EPR_isoFit_space}}. Therefore, the \code{optim.params.fix.id} argument
  just selects an element of the \code{optim.params.init} (by its corresponding index) and accordingly,
  the selected simulation parameter value won't be optimized.

  A simple EPR spectrum analysis by interactive simulation using the \code{\link{plot_eval_ExpSim_app}}
  may return an \code{.R} script/code snippet for the \strong{initial} simulation \strong{fit} and therefore
  the user does not have to write (or remember) the entire code required to run the \code{eval_sim_EPR_isoFit}.
  Rather, she/he can instantly and seamlessly analyze an isotropic EPR spectrum by optimizing the simulation parameters,
  right after saving and then subsequently running of the above-mentioned script.
}
\examples{
## loading built-in example dataset which is simple
## EPR spectrum of the aminoxyl radical:
aminoxyl.data.path <-
  load_data_example(file = "Aminoxyl_radical_a.DTA")
aminoxyl.data <-
  readEPR_Exp_Specs(
    path_to_file = aminoxyl.data.path,
    qValue = 2100
 )
#
## EPR spectrum simulation fit with "Nelder-Mead"
## optimization method with `check.fit.plot = FALSE`:
tempo.test.sim.fit.a <-
  eval_sim_EPR_isoFit(data.spectr.expr = aminoxyl.data,
    nu.GHz = 9.806769,
    lineG.content = 0.5,
    optim.method = "neldermead",
    nuclear.system.noA = list("14N",1),
    baseline.correct = "linear",
    optim.params.init =
      c(2.006, # g-value
        4.8, # G Delta Bpp
        4.8, # L Delta Bpp
        0, # intercept (constant) lin. baseline
        0.018, # Sim. intensity multiply
        1e-6, # slope lin. baseline
        49), # A in MHz
    check.fit.plot = FALSE,
    msg.optim.progress = FALSE
  )
## OUTPUTS:
## best fit parameters:
tempo.test.sim.fit.a$best.fit.params
#
## spectrum plot with experimental spectrum,
## simulated one with the linear baseline fit
## and simulated one with the linear baseline
## fit subtracted:
tempo.test.sim.fit.a$plot
#
## minimum sum of residual squares:
tempo.test.sim.fit.a$min.rss
#
## number of evaluations / iterations:
tempo.test.sim.fit.a$N.evals
#
## convergence, in this case it is represented
## by the integer code indicating the successful
## completion (it must be > 0):
tempo.test.sim.fit.a$N.converg
#
## preview of data frame including all EPR spectra:
head(tempo.test.sim.fit.a$df)
#
## similar EPR spectrum simulation fit with "particle swarm"
## optimization algorithm and `check.fit.plot = TRUE` option
## as well as user defined bound constraints, including
## fixed A(1 x 14N) = 52.68 MHz:
tempo.test.sim.fit.b <-
  eval_sim_EPR_isoFit(data.spectr.expr = aminoxyl.data,
    nu.GHz = 9.806769,
    lineG.content = 0.75,
    optim.method = "pswarm",
    nuclear.system.noA = list("14N",1),
    baseline.correct = "constant",
    optim.params.init = c(2.0062,4.8,4.8,0,8e-3,52.68),
    optim.params.lower = c(2.0048,4.6,4.6,-1e-4,5e-3,52.68),
    optim.params.upper = c(2.0068,5.0,5.0,1e-4,1.5e-2,52.68),
    check.fit.plot = TRUE,
    eval.optim.progress = TRUE ## iterations, progress
  )
## OUTPUTS:
## minimum sum of residual squares:
tempo.test.sim.fit.b$min.rss
#
## check and compare the previous value
## by residual analysis (`ra`)
sum((tempo.test.sim.fit.b$ra$df$Residuals)^2)
#
## number of evaluations / iterations:
tempo.test.sim.fit.b$N.evals
#
## best fit parameters:
tempo.test.sim.fit.b$best.fit.params
#
## correlation matrix of the EPR simulation fit:
tempo.test.sim.fit.b$cor.df
#
## visualization of the previous matrix:
tempo.test.sim.fit.b$cor.df \%>\%
  corrplot::corrplot(addCoef.col = "#c2c2c2")
#
## quick simulation check by plotting the both
## simulated and the experimental EPR spectra
## together with the initial simulation
## and the residuals (differences between the
## experiment and the best fit)
tempo.test.sim.fit.b$plot
#
## simple residual density plot
## together with standard deviation
tempo.test.sim.fit.b$ra$plot.histDens
tempo.test.sim.fit.b$ra$sd
#
## Akaike and Bayesian Criteria (AIC & BIC)
## + information about the residuals distribution
tempo.test.sim.fit.b$abic
#
## residual & Q-Q plots for the proposed normal
## distribution of residuals and confidence level 99\%
## (default: 95\%/0.95)
tempo.test.sim.fit.b$ra$plot.rqq(confidence = 0.99)
#
## residual & Q-Q plots for the proposed Student's
## distribution of residuals with degrees of freedom
## df = 2.2 (see the "$message" right above)
tempo.test.sim.fit.b$ra$plot.rqq(
  residuals.distro = "t",
  df = 2.2
)
#
## fitting of the aminoxyl EPR spectrum
## by the combination of the 1. "Levenberg-Marquardt"
## and 2. "Nelder-Mead" algorithms
tempo.test.sim.fit.c <-
  eval_sim_EPR_isoFit(aminoxyl.data,
                      nu.GHz = 9.86769,
                      lineG.content = 0.5,
                      optim.method = c("levenmarq",
                                       "neldermead"),
                      nuclear.system.noA = list("14N",1),
                      baseline.correct = "constant",
                      optim.params.init = c(2.0060,
                                            4.8,
                                            4.8,
                                            0,
                                            7e-3,
                                            49),
                      check.fit.plot = FALSE
                    )
## OUTPUTS:
## best fit parameters for both procedures within a list:
tempo.test.sim.fit.c$best.fit.params
#
## compare the results with the example in the `readMAT_params_file`,
## corresponding to the best fit from `Easyspin`
#
## `N.converg` also consists of two components
## each corresponding to result of the individual
## optimization method where the "levenmarq" returns
## the sum of squares at each iteration, therefore the 1st
## component is vector and the 2nd one is integer code
## as already stated above:
tempo.test.sim.fit.c$N.converg
#
## Akaike and Bayesian Criteria (AIC & BIC)
## + information about the residuals distribution
tempo.test.sim.fit.c$abic



}
\seealso{
Other Simulations and Optimization:
\code{\link[=eval_ABIC_forFit]{eval_ABIC_forFit()}},
\code{\link[=eval_sim_EPR_iso]{eval_sim_EPR_iso()}},
\code{\link[=eval_sim_EPR_isoFit_space]{eval_sim_EPR_isoFit_space()}},
\code{\link[=eval_sim_EPR_iso_combo]{eval_sim_EPR_iso_combo()}},
\code{\link[=optim_for_EPR_fitness]{optim_for_EPR_fitness()}},
\code{\link[=plot_eval_EPRtheo_mltiplet]{plot_eval_EPRtheo_mltiplet()}},
\code{\link[=plot_eval_RA_forFit]{plot_eval_RA_forFit()}},
\code{\link[=quantify_EPR_Sim_series]{quantify_EPR_Sim_series()}},
\code{\link[=smooth_EPR_Spec_by_npreg]{smooth_EPR_Spec_by_npreg()}}
}
\concept{Simulations and Optimization}
