% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sim_EPR_isoFitb.R
\name{eval_sim_EPR_isoFitb}
\alias{eval_sim_EPR_isoFitb}
\title{Interactive Least-Squares Fitting of Isotropic EPR Spectra by Simulation Iteratioins/Evaulations}
\usage{
eval_sim_EPR_isoFitb(
  data.spectr.expr,
  Intensity.expr = "dIepr_over_dB",
  Intensity.sim = "dIeprSim_over_dB",
  nu.GHz,
  B.unit = "G",
  Blim = NULL,
  nuclear.system.noA = NULL,
  baseline.correct = "constant",
  lineG.content = 0.5,
  lineSpecs.form = "derivative",
  optim.method = c("neldermead", "cobyla", "lbfgs"),
  optim.params.init,
  optim.params.lower = NULL,
  optim.params.upper = NULL,
  optim.params.fix.id = NULL,
  Niters.per.eval = 128,
  Nevals = 16
)
}
\arguments{
\item{data.spectr.expr}{Data frame object/table, containing the experimental spectral data with the magnetic flux density
(\code{"B_mT"} or \code{"B_G"}) and the intensity (see the \code{Intensity.expr} argument) columns.}

\item{Intensity.expr}{Character string, pointing to column name of the experimental EPR intensity within
the original \code{data.spectr.expr}. \strong{Default}: \code{dIepr_over_dB}.}

\item{Intensity.sim}{Character string, pointing to column name of the simulated EPR intensity within the related output
data frame. \strong{Default}: \code{Intensity.sim = "dIeprSim_over_dB"}.}

\item{nu.GHz}{Numeric value, microwave frequency in \code{GHz}.}

\item{B.unit}{Character string, denoting the magnetic flux density unit e.g. \code{B.unit = "G"}
(gauss, \strong{default}) or \code{B.unit = "mT"}/\code{"T"} (millitesla/tesla).}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G}
corresponding to lower and upper \strong{visual limit} of the selected \eqn{B}-region,
such as \code{Blim = c(3495.4,3595.4)}. \strong{Default}: \code{Blim = NULL} (corresponding to the entire
\eqn{B}-range of EPR spectrum). \strong{This does not correspond to simulation fit data region !}.
If narrower \eqn{B}-region (in comparison to the original one) is required to fit the EPR spectrum,
the filtering has to be done prior to own fitting procedure. For example, if the original data frame
(\code{df.spectr.orgin} within 200 G), of the experimental EPR spectrum, should be fitted within
the region of \code{B = c(3450,3550)} (100 G), following operation must be performed:
\code{df.spectr.actuall <- df.spectr.origin |> dplyr::filter(dplyr::between(B_G,3450,3550))},
where the \code{df.spectr.actuall} serves as an input (represented by the \code{data.spectr.expr}
argument) for the function.}

\item{nuclear.system.noA}{List or nested list \strong{without estimated hyperfine coupling constant values},
such as \code{list("14N",1)} or \code{list(list("14N", 2),list("1H", 4),list("1H", 12))}. The \eqn{A}-values
are already defined as elements of the \code{optim.params.init} argument/vector. If the EPR spectrum
does not display any hyperfine splitting, the argument definition reads \code{nuclear.system.noA = NULL} (\strong{default}).}

\item{baseline.correct}{Character string, referring to baseline correction of the simulated/fitted spectrum.
Corrections like \code{"constant"} (\strong{default}), \code{"linear"} or \code{"quadratic"} can be applied.}

\item{lineG.content}{Numeric value between \code{0} and \code{1}, referring to content of the \emph{Gaussian} line form.
If \code{lineG.content = 1} (\strong{default}) it corresponds to "pure" \emph{Gaussian} line form
and if \code{lineG.content = 0} it corresponds to \emph{Lorentzian} one. The value from (0,1)
(e.g. \code{lineG.content = 0.5}) represents the linear combination (for the example above,
with the coefficients 0.5 and 0.5) of both line forms => so called \emph{pseudo-Voigt}.}

\item{lineSpecs.form}{Character string, describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} which can be used as well) line form
of the analyzed EPR spectrum/data.}

\item{optim.method}{Character string, setting the optimization method/algorithm. Even though, by default, the argument
is defined as a vector, \code{optim.method = c("neldermead","cobyla","lbfgs")}, only one method from those three
can be selected. For example \code{optim.method = "neldermead"} (\strong{default}). For additional information
to all three available methods, please refer to the \code{\link{optim_for_EPR_fitness}}.}

\item{optim.params.init}{Numeric vector with the initial parameter guess (elements) where the \strong{first five
  elements are immutable}
  \enumerate{
  \item g-value (g-factor)

  \item \strong{G}aussian linewidth

  \item \strong{L}orentzian linewidth

  \item baseline constant (intercept or offset)

  \item intensity multiplication constant

  \item baseline slope (only if \code{baseline.correct = "linear"} or \code{baseline.correct = "quadratic"}),
  if \code{baseline.correct = "constant"} it corresponds to the \strong{first HFCC} (\eqn{A_1})

  \item baseline quadratic coefficient (only if \code{baseline.correct = "quadratic"}),
  if \code{baseline.correct = "constant"} it corresponds to the \strong{second HFCC} (\eqn{A_2}),
  if \code{baseline.correct = "linear"} it corresponds to the \strong{first HFCC} (\eqn{A_1})

  \item additional HFCC (\eqn{A_3}) if \code{baseline.correct = "constant"} or if
  \code{baseline.correct = "linear"} (\eqn{A_2}), if \code{baseline.correct = "quadratic"} it corresponds
  to the \strong{first HFCC} (\eqn{A_1})

  \item ...additional HFCCs (\eqn{A_k...}, each vector element is reserved only for one \eqn{A})
  }
  DO NOT PUT ANY OF THESE PARAMETERS to \code{NULL}. If the lineshape is expected to be pure
  \strong{L}orentzian or pure \strong{G}aussian then put the corresponding vector element to \code{0}.}

\item{optim.params.lower}{Numeric vector (with the same element order like \code{optim.params.init})
with the lower bound constraints. \strong{Default}: \code{optim.params.lower = NULL} which actually
corresponds to relative default limits (refer to the \code{\link{eval_sim_EPR_isoFit}} arguments)
of all fitted/optimized parameters for the actual evaluation. If specified
(e.g. \code{optim.params.lower = c(2.004,0.3,0.3,-1e-3,0.001,46)}),
it represents the general lower boundaries for the entire optimization/fitting procedure.}

\item{optim.params.upper}{Numeric vector (with the same element order like \code{optim.params.init})
with the upper bound constraints. \strong{Default}: \code{optim.params.upper = NULL} which actually
corresponds to relative default limits (refer to the \code{\link{eval_sim_EPR_isoFit}} arguments)
of all fitted/optimized parameters for the actual evaluation. If specified
(e.g. \code{optim.params.upper = c(2.007,0.5,0.5,1e-3,0.004,48)}),
it represents the general upper boundaries for the entire optimization/fitting procedure.}

\item{optim.params.fix.id}{Numeric value/vector of index/indices of the \code{optim.params.init}, corresponding
to optimization/simulation parameter(s) to be fixed. For example, if the g-Value and the (intensity) multiplication
constant should not be optimized (their values should be fixed during the procedure), the argument must be defined
as follows: \code{optim.params.fix.id = c(1,5)} (see also the \code{optim.params.init} description
for the simulation/optimized parameters order). \strong{Default}: \code{optim.params.fix.id = NULL}, indicating
that none of the \code{optim.params.init} is fixed, i.e. all parameters are optimized within their default/defined
boundaries (see the \code{optim.params.lower} and the \code{optim.params.upper}). Alternatively, the parameter value(s)
can be also adjusted by assigning the \code{optim.params.init} + \code{optim.params.lower} + \code{optim.params.upper}
elements to the same value.}

\item{Niters.per.eval}{Numeric value, equal to the number of iterations per one the \code{Nevals} (see the related \code{Nevals}
argument description). \strong{Default}: \code{Niters.per.eval = 128}. This argument, among other things,
depends on the complexity of the \code{nuclear.system} (\code{nuclear.system.noA}). For example, if an aminoxyl
radical with 1 x 14N nucleus interaction is considered, the \code{Niters.per.eval} might be lower than the default
one (e.g. 64). However, the more complex system like N,N,N′,N′-Tetramethyl-p-phenylenediamine radical cation,
with 2 x 14N, 4 x 1H and 12 x 1H nuclei interaction may require higher \code{Niters.per.eval} (e.g. 128). The higher
the \code{Niters.per.eval}, the longer the computational fitting/optimization time.}

\item{Nevals}{Numeric value, corresponding to total number of evaluations/cycles (i.e. how many loops/runs
will be considered for the entire fitting/optimization procedure). This argument is related to the \code{Niters.per.eval},
where the total number of iterations reads \code{Nevals} x \code{Niters.per.eval}. For example, for \code{Nevals = 16}
(\strong{default}) and \code{Niters.per.eval = 128}, the number of iterations = 16 x 128 = 2048.
Higher \code{Nevals} requires longer computational time. For interactive visualization of EPR spectra (during
the optimization/fitting procedure) no parallel computing is supported
(contrary to the \code{\link{eval_sim_EPR_isoFit_space}}).}
}
\value{
List with the following elements:
  \describe{
  \item{plot.fit}{A graph panel (2 rows x 2 columns) with 4 components. Simulated + experimental EPR spectra,
  residuals vs \eqn{B} magnetic flux density, sum of the residual squares (RSS) vs iteration as well as AIC+BIC information
  criteria vs iteration. Even though it represents the final status of those four dependencies, the actual version,
  appears at each of the \code{Nevals} (see the \code{Arguments}) in order to follow the progress of the fitting procedure
  interactively.}
  \item{best.fit.params}{Named vector of the optimized (best fitting) simulation parameters, corresponding
  to minimum \code{RSS}. The actual values also appear at each of the \code{Nevals} (see the \code{Arguments}),
  interactively during the fitting procedure.}
  \item{df.params.optim}{A data frame object of all simulation parameters + \code{RSS} (residual sum of squares)
  + \code{AIC} + \code{BIC} (Akaike and Bayessian information criteria, respectively, see also the \code{\link{eval_ABIC_forFit}}),
  evaluated during the fitting process at corresponding actual iteration. Visual progress of all those variables
  may be nicely followed by the \code{\link[esquisse]{esquisser}} or other \emph{R} plotting/visualization function.}
  \item{nuclear.system}{List consisting of all considered nuclei, and their optimized (best fitted) coupling
  constants \eqn{A} in MHz, which may be used in any other additional EPR simulation (see the \code{\link{eval_sim_EPR_iso}}).}
  \item{ra}{Final (related to the minimum of \code{RSS}) \strong{r}esidual \strong{a}nalysis list
  (refer to the \code{\link{plot_eval_RA_forFit}} and/or to \code{\link{eval_sim_EPR_isoFit}}) extended by the \code{message}
  which distribution (Normal/Gaussian, Student's t-distribution or Cauchy) fits the residuals at best and was actually
  applied to evaluate AIC and BIC (consult the \code{\link{eval_ABIC_forFit}} for a detailed description).}
  \item{cor.df}{Function to evaluate correlation \code{matrix} of a data frame, consisting of EPR experimental,
  simulated (best fit) and residual intensities as columns/variables. Such matrix can be additionally nicely visualized
  by a correlation \code{plot} created by the \code{\link[corrplot]{corrplot}} function. Three \code{methods}
  are available: \code{"pearson"} (\strong{default}), \code{"spearman"} (captures monotonic relationships)
  and \code{"kendall"} (for small data ensembles), see also \code{\link[stats]{cor}}.}
  \item{plot.best.sim.expr}{Visualization function to plot either static \code{ggplot2} (\code{interactive = FALSE},
  \strong{default}) or interactive \code{plotly} (\code{interactive = TRUE}) comparison between the experimental
  and the simulated EPR spectrum. In addition, both spectra, within the static \code{ggplot}, can be presented
  either in the \code{overlay = TRUE} (\strong{default}) or in the offset (\code{overlay = FALSE}) mode.}
  \item{df.best.sim.expr}{A data frame object, containing experimental and simulated spectra in long/tidy form,
  actually corresponding to \code{plot.best.sim.expr} while \code{interactive = TRUE}.}
  }
}
\description{
Ordinary least-squares fitting of the isotropic EPR spectra by simulation iterations/evaluations.
  In principle, this function is based on the \code{\link{eval_sim_EPR_isoFit}} however,
  it represents a more interactive version of the \code{\link{eval_sim_EPR_isoFit}}.
  Namely, it provides \code{{ggplot2}} objects (graphs, see the \code{Value} and the \code{plot.fit} description)
  in order to simultaneously check/explore the optimization/fitting process at each of the evaluations
  (refer to the \code{Nevals} argument). In addition, it also simultaneously shows current values
  of all the fitting parameters in the \emph{R} console. The function was built, because during the parallel processing
  (see the \code{\link{eval_sim_EPR_isoFit_space}}) it is not possible to display the actual
  EPR spectra during the optimization/fitting procedure. In the upcoming package updates, it will be also implemented
  into the \code{\link{plot_eval_ExpSim_app}}.
}
\examples{
\dontrun{
test.list <- eval_sim_EPR_isoFitb(
  data.spectr.expr = data.tmpd.spec,
  nu.GHz = data.tmpd.params.values[1,2],
  nuclear.system.noA = list(
    list("14N", 2),
    list("1H", 4),
    list("1H", 12)
  ),
  optim.method = "cobyla",
  optim.params.init = c(
    2.00305, ## g_iso
    0.521, ## Gaussian linewidth
    0.52, ## Lorentz linewidth
    0, ## offset/baseline constant
    3.2e5, ## intensity multiplication coeff.
    19.5, 5.5, 19.5 ## required As in MHz
  ),
  ## number of iterations per evaluation
  Niters.per.eval = 128,
  Nevals = 17 ## total number of evaluations
  ## total number of iterations =
  ## = Niters.per.eval * Nevals
)
#
## simulation fit with the lower and upper
## bound constraints
epr.spectrum.sim.fit <-
  eval_sim_EPR_isoFitb(
    data.spectr.expr = epr.spectrum.data,
    nu.GHz = 9.793116,
    B.unit = 'G',
    Blim = c(3435,3537),
    lineG.content = 0.93,
    optim.method = 'neldermead',
    optim.params.init = c(
      2.00581,0.57,0.77,0,0.0006,41.1,7.94,2.84
    ),
    optim.params.lower = c(
      2.0055,0.3,0.45,-1e-4,5e-4,40,6,1
    ),
    optim.params.upper = c(
      2.0059,0.7,1.2,1e-4,1e-3,42,9,4
    ),
    nuclear.system.noA =
      list(list('14N',1),list('1H',1),list('1H',1)),
    baseline.correct = 'constant',
    Nevals = 32,
    Niters.per.eval = 116
 )
}


}
\seealso{
Other Simulations and Optimization:
\code{\link[=eval_ABIC_forFit]{eval_ABIC_forFit()}},
\code{\link[=eval_sim_EPR_iso]{eval_sim_EPR_iso()}},
\code{\link[=eval_sim_EPR_isoFit]{eval_sim_EPR_isoFit()}},
\code{\link[=eval_sim_EPR_isoFit_space]{eval_sim_EPR_isoFit_space()}},
\code{\link[=eval_sim_EPR_iso_combo]{eval_sim_EPR_iso_combo()}},
\code{\link[=optim_for_EPR_fitness]{optim_for_EPR_fitness()}},
\code{\link[=plot_eval_EPRtheo_mltiplet]{plot_eval_EPRtheo_mltiplet()}},
\code{\link[=plot_eval_RA_forFit]{plot_eval_RA_forFit()}},
\code{\link[=quantify_EPR_Sim_series]{quantify_EPR_Sim_series()}},
\code{\link[=smooth_EPR_Spec_by_npreg]{smooth_EPR_Spec_by_npreg()}}
}
\concept{Simulations and Optimization}
