% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue.R
\name{gValue}
\alias{gValue}
\title{Basic Calculation of \eqn{g}-factor}
\usage{
gValue(nu, nu.unit = "GHz", B, B.unit = "mT")
}
\arguments{
\item{nu}{Numeric, microwave Frequency}

\item{nu.unit}{String, frequency unit defined by \code{"GHz"} or \code{"Hz"}, \strong{default}: \code{nu.unit = "GHz"}}

\item{B}{Numeric, magnetic flux density}

\item{B.unit}{String, magnetic flux density unit in \code{"G"} or \code{"mT"} or \code{"T"}, \strong{default}:
\code{B.unit = "mT"}}
}
\value{
g-value from \eqn{(\nu h)/(\mu_{B} B)}. For variables and constants =>
  see description above
}
\description{
Calculation of {g}-factor according to fundamental formula.
  The magnetic flux density (\code{B}) and microwave frequency (\code{nu},\eqn{\nu})
  can be entered with common units like \code{G} (Gauss) \code{mT}
  (millitesla) or \code{T} (tesla) as well as \code{GHz} or \code{Hz}, respectively.
  The Planck constant (\eqn{h}) and Bohr magneton (\eqn{\mu_{B}}) are included
  in \code{\link[constants]{syms}} function and their values are taken by \code{syms$h}
  and \code{syms$mub} commands, respectively.
}
\examples{
gValue(9.8020458,
       nu.unit = "GHz",
       350.214,
       B.unit = "mT")
gValue(nu = 9.8020458e+9,
       nu.unit = "Hz",
       B = 3502.14,
       B.unit = "G")
gValue(9.5421,"GHz",0.333251,"T")

}
