% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue.R
\name{gValue}
\alias{gValue}
\title{Basic Calculation of g-Factor}
\usage{
gValue(nu, unit_nu = "GHz", B, unit_B = "mT")
}
\arguments{
\item{nu}{Frequency (number)}

\item{unit_nu}{Frequency unit defined by "GHz" or "Hz"}

\item{B}{Magnetic flux density (number)}

\item{unit_B}{Magnetic flux density unit in "G" or "mT" or "T"}
}
\value{
g-value from \eqn{frac{h nu}{B mu_{B}}}
}
\description{
Calculation of g-factor from the basic formula (see below).
  The magnetic flux density (\code{B}) and frequency (\code{nu}) can be entered
  with common units like \code{GHz} or \code{Hz} as well as \code{G} (Gauss)
  or \code{mT} (millitesla), respectively. SI units like \code{Hz} and \code{T}
  can be used as well. The Planck constant and Bohr magneton are included in
  \code{\link{constants}} package and their values are taken by
  \code{syms$h} and \code{syms$muB} commands, respectively.
}
\examples{
\dontrun{
gValue(9.8020458,unit_nu = "GHz",350.214,unit_B = "mT")
gvalue(9.8020458e+9,unit_nu = "Hz",3502.14,unit_B = "G")
}


}
