% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue_fromQCHcomp.R
\name{gValue_fromQCHcomp}
\alias{gValue_fromQCHcomp}
\title{Calculation of \eqn{g}-factor from Quantum Chemical Computational Output}
\usage{
gValue_fromQCHcomp(delta_g_vec, mean = TRUE)
}
\arguments{
\item{delta_g_vec}{Numeric vector of principal g-values from e.g. Gaussian output}

\item{mean}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
gValue_fromQCHcomp(c(-806.1,1099.9,1417.1))
gValue_fromQCHcomp(c(-543.1,832.8,1089.5),mean = FALSE)


}
