% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gValue_Spec.R
\name{gValue_Spec}
\alias{gValue_Spec}
\title{Calculation of \eqn{g}-factor ("Position") from the EPR Spectrum/Data}
\usage{
gValue_Spec(
  data.spectrum,
  nu.GHz,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  Blim,
  iso = TRUE
)
}
\arguments{
\item{data.spectrum}{Spectrum data frame/table where the magnetic flux density (in \code{mT} or \code{G}) column
must be labeled as \code{B_mT} (or \code{B_G}) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column can be included as well}

\item{nu.GHz}{Numeric, microwave frequency in \code{GHz}}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{data.spectrum} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (\strong{default})
or \code{B = "B_G"} or \code{B = "B_G_Sim"} to include simulated EPR spectra as well}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(3495.4,3595.4)`}

\item{iso}{Boolean, whether to calculate the \eqn{g}-factor from the \eqn{B} value corresponding to
that between the \code{min.} and \code{max.} derivative intensities (\code{dIepr_over_dB}, that is \eqn{g_{iso}}
(this is the \strong{default}: \code{iso = TRUE}), or by finding the the \eqn{B} value corresponding
to \code{dIepr_over_dB = 0} (close/near zero, which is \code{iso = FALSE})}
}
\value{
Numeric \eqn{g_{iso}}-value ('iso' = 'isotropic') according to \eqn{(\nu h)/(\mu_{B} B)}
}
\description{
Calculation of g-value according to fundamental formula (\code{\link{gValue}}).
  \eqn{g}-related magnetic flux density (like \eqn{B_{iso}} or \eqn{B_{center}}) is directly taken
  from the EPR spectrum. If positive and negative derivative intensities of the spectral line are similar
  and their distance from the middle of the spectrum equals, the \eqn{B_{iso}} should be be considered,
  otherwise the \eqn{B_{center}} must be taken into account.
}
\examples{
\dontrun{
gValue_Spec(data.spectrum,
            9.82451,
            "B_mT",
            Intensity = "dIepr_over_dB_Sim",
            c(349.8841,351.112))
gValue_Spec(data.spectrum,
            nu.GHz = 9.82451,
            B = "B_G",
            Blim = c(3498.841,3511.12),
            iso = FALSE)
gValue_Spec(data.spectrum,
            9.91024,
            B = "B_G_Sim",
            c(3499,3501))
}


}
