% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integ_correct_EPRspecs.R
\name{integ_correct_EPRspecs}
\alias{integ_correct_EPRspecs}
\title{Polynomial Baseline Correction for single-Integrated EPR Spectra + Double Integral Calculation/Presentation}
\usage{
integ_correct_EPRspecs(
  data.spec.integ,
  B = "B_G",
  Blim,
  BpeaKlim,
  poly.degree,
  double.integ = FALSE
)
}
\arguments{
\item{data.spec.integ}{Spectrum data frame/table with magnetic flux density (in \code{mT} or \code{G}) column
must be labeled as \code{B_mT} (or \code{B_G}) and that of the derivative intensity as \code{dIepr_over_dB},
\code{index} column may be included as well, \code{sinlge integrated EPR} \strong{spectrum must be included}
in column named by \code{"single|sinteg|s_integ|single_|singleinteg|sintegral|sInteg_"} ("|" == "or" operator),
this can be obtained by \code{\link[pracma:trapz]{pracma::cumtrapz}} function from \emph{B} and \emph{dIepr_over_dB}}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"}
or \code{B = "B_G"} (\strong{default})}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `Blim = c(3495.4,3595.4)`}

\item{BpeaKlim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to border limits
of the selected \eqn{B} region, e.g. like `BpeaKlim = c(3535.4,3555.4)`}

\item{poly.degree}{Numeric, degree of polynomial function used to fit the baseline under the single integrated
curve of \emph{dIepr_over_dB}}

\item{double.integ}{Boolean, whether to present (column in data frame) the double integral of \emph{dIepr_over_dB},
which is required for quantitative analysis, \strong{default}: \code{double.integ = FALSE}}
}
\value{
Data frame/table including the EPR spectral data (\emph{dIepr_over_dB} vs \eqn{B}) as well as its
  corresponding \code{single} (column \code{sIntegCorr}) and/or \code{double} (column \code{dIntegCorr})
  integral corrected against the baseline fit
}
\description{
tbc
}
\examples{
tbc


}
