% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_cnfdnc_t_vecCol.R
\name{interval_cnfdnc_t_vecCol}
\alias{interval_cnfdnc_t_vecCol}
\title{Calculate Confidence Interval of a Vector/Data Frame-Column According to Sudent's t-Distribution}
\usage{
interval_cnfdnc_t_vecCol(
  data.vec.col,
  level.cnfdnc = 0.95,
  lw.tail = TRUE,
  separate = FALSE
)
}
\arguments{
\item{data.vec.col}{tbc}

\item{level.cnfdnc}{tbc}

\item{lw.tail}{tbc}

\item{separate}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}

}
