% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_example.R
\name{load_data_example}
\alias{load_data_example}
\title{Loading Built-In Data Files for Package Examples}
\usage{
load_data_example(file = NULL)
}
\arguments{
\item{file}{Character string corresponding to file name + extension.}
}
\description{
Loads built-in package data files which are required either to demonstrate the package functionality
  by examples within the documentation, or to run functions involving essential characteristics
  of nuclei important for EPR spectroscopy (e.g. \code{\link{eval_nu_ENDOR}} or \code{\link{eval_sim_EPR_iso}},
  see also \code{\link{isotopes_ds}}). Details of all data are summarized in \code{vignette("datasets")}.
}
\examples{
## list of all files within the "extdata" directory =>
load_data_example()
#
## additionally, you may refer to several function examples


}
