% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_for_EPR_fitness.R
\name{optim_for_EPR_fitness}
\alias{optim_for_EPR_fitness}
\title{General Function for Non-Linear Optimization/Fitting of EPR Parameters/Data}
\usage{
optim_for_EPR_fitness(
  method = "neldermead",
  x.0,
  fn,
  lower,
  upper,
  data,
  Nmax.evals = 512,
  tol.step = 5e-07,
  pswarm.size = NULL,
  pswarm.diameter = NULL,
  pswarm.type = NULL,
  eval.optim.progress = FALSE,
  fix.optim.x.0.id = NULL,
  ...
)
}
\arguments{
\item{method}{Character string, pointing to applied optimization method/algorithm. One may choose one from
those listed in \code{Details}, \strong{default}: \code{method = "neldermead"}, setting up
the \href{https://brandewinder.com/2022/03/31/breaking-down-Nelder-Mead/}{"Nelder-Mead" simplex method}.}

\item{x.0}{Numeric vector with the initial values to be optimized in order to fit onto the experimental data.}

\item{fn}{Objective function that is to be minimized. Usually it is the function calculating the sum of residual squares,
where a more general parameterized one can be implemented in (see \code{Details} and \code{Examples}).}

\item{lower, upper}{lower and upper bound constraints.}

\item{data}{Data frame object, containing columns/variables (e.g. intensity of an EPR spectrum),
required to undergo a fitting/optimization process.}

\item{Nmax.evals}{Numeric value, maximum number of function evaluations and/or iterations.
The only one method, limited by this argument, is \code{\link[minpack.lm]{nls.lm}}, where
\code{Nmax.evals = 1024}. Higher \code{Nmax.evals} may extremely extend the optimization
time, therefore the \strong{default} value reads \code{Nmax.evals = 512}. However, the \code{"pswarm"}
method requires at least the default or even higher values.}

\item{tol.step}{Numeric, the smallest optimization step (relative change) between
2 iterations to stop the optimization procedure. For the \code{method = "pswarm"}
(particle swarm optimization procedure) it actually corresponds to tolerance for restarting.
Once the maximum distance between the "best" particle and all the others is less
than \code{tol.step} * \code{pswarm.diameter}) the algorithm restarts.
See also \code{\link[pso]{psoptim}}. \strong{Default}: \code{tol.step = 5e-7}.}

\item{pswarm.size}{Numeric value, which equals to particle swarm size (i.e. number of particles),
if \code{method = "pswarm"}. The \strong{default} value (\code{pswarm.size = NULL}) actually
corresponds to \code{floor(10+2*sqrt(length(Np)))} (for \code{SPSO2007}, see the \code{pswarm.type}
argument), e.g. to optimize 8 parameters, number of particles = 15. The \code{length(Np)} corresponds
either to \code{length(x.0)} or to difference between the \code{length(x.0)} and number of parameters,
for which \code{(upper - lower) == 0} (i.e. those \code{x.0} parameters are actually fixed/won't be optimized).
For the \code{SPSO2011} the default number of particles equals to \code{40}.}

\item{pswarm.diameter}{Numeric value, corresponding to diameter of the particle swarm search space
(in case \code{method = "pswarm"}). The \strong{default} value (\code{pswarm.diameter = NULL})
refers to the Euclidean distance, defined as:
\deqn{\sqrt{\sum_k\,(\text{upper}[k] - \text{lower}[k])^2}}}

\item{pswarm.type}{Character string, setting the type/version of particle swarm algorithm
if \code{method = "pswarm"}. There are two types available: \code{pswarm.type = "SPSO2007"}
and \code{pswarm.type = "SPSO2011"}. The latter introduced an adaptive random topology,
which allows the swarm to dynamically adjust its communication structure.
This helps in maintaining diversity in the swarm and improves the algorithm's ability
to escape local optima. This type generally offers better performance on larger multidimensional spaces
than the \code{pswarm.type = "SPSO2007"}, which uses a more static topology. Details may be found
in the \code{References}. \strong{Default}: \code{pswarm.type = NULL} (actually corresponding
to \code{"SPSO2007"}, that performs slightly better on smaller scales such as common simulations of EPR spectra
with lower number of parameters like hyperfine coupling constants).}

\item{eval.optim.progress}{Logical. If \code{TRUE} a progress of the optimization/fitting, defined by the \code{method}
argument, is monitored/tracked in the R console. The \strong{default} value is set to \code{FALSE}
because higher number of evaluations/iterations might result in several tens or hundreds of rows
with the information, depending on the applied \code{method}. In the case of \code{{nloptr}} methods/functions
as well as \code{method = "pswarm"} it displays the iteration number (each 10-th iteration per particle
shown for \code{pswarm}) and the value of the objective/fitness function (e.g. least-square minimization or RSS).
Additionally, \code{pswarm} method shows possible shrinking of the particle swarm diameter by the convergence.
For the \code{method = "levenmarq"} it shows the iteration/evaluation number, sum of residual squares (RSS)
and the relevant parameter (Par.) value. In the \code{\link{eval_sim_EPR_isoFit}}
and \code{\link{quantify_EPR_Sim_series}} this argument can be combined with the \code{msg.optim.progress}.}

\item{fix.optim.x.0.id}{Numeric value/vector of the \code{x.0} indices, corresponding to \code{x.0} elements that
will be fixed during the optimization/fitting, i.e. they won't be optimized because their \code{lower}
and \code{upper} limits equal to related elements of the initial \code{x.0}. For example, if the 1st and the 3rd
parameter/element of the \code{x.0} are supposed to be fixed, put \code{fix.optim.x.0.id = c(1,3)}.
\strong{Default}: \code{fix.optim.x.0.id = NULL}, i.e. none of the \code{x.0} elements is fixed and all
parameters will be optimized as required (see also \code{\link{eval_sim_EPR_isoFit}}). However, even
in the case if \code{fix.optim.x.0.id = NULL}, the parameters can be fixed by adjustment of the corresponding
\code{lower} and \code{upper} arguments.}

\item{...}{additional arguments passed to the function.}
}
\value{
For all listed algorithms the function returns \code{list} with the elements like
  (please, refer to e.g. \code{Value} in \code{\link{eval_sim_EPR_isoFit}})
  \enumerate{
  \item The best parameters found (\code{par} vector, depending on the initial \code{x.0} set of parameters).

  \item The value of \code{fn} (minimum value) corresponding to the best \code{par}.

  \item Number of evaluations and/or iterations (in case of the \code{method = "pswarm"},
  see also \code{Value} in the \code{\link[pso]{psoptim}}) before the termination.

  \item (Un)successful termination information (\code{convergence} or \code{rsstrace}), usually
  corresponding either to integer value showing the (un)successful termination like \code{2: Maximum number
  of iterations reached} (or integer code > 0 indicating successful completion) or in the case
  of \code{\link[minpack.lm]{nls.lm}}, it returns a vector with the values equal
  to sum of the residual squares at each iteration.

  \item A descriptive message/character string, giving the additional information about the optimization
  procedure/termination. \strong{By default} (e.g. within the \code{\link{eval_sim_EPR_isoFit}}),
  this is however \strong{"turned off"}, for the sake of simplicity,
  because most of the information can be found in the previous convergence list element
  or can be activated by the \code{eval.optim.progress} argument.

  \item Additional objects (lists/vectors/values), providing information about the fitting/optimization
  process and depending on the \code{method} argument. It may or may not be returned by the specific
  function(s) (e.g. by the \code{\link{eval_sim_EPR_isoFit}}).
  }
}
\description{
General-purpose optimization of the objective \code{fn} function (also called "fitness")
  which is to be minimized in order to fit theoretical models (EPR simulations)
  onto the experimental data. Several methods/algorithms are implemented (see also \code{Details}):
  from the \href{https://astamm.github.io/nloptr/}{nloptr} package: \code{\link[nloptr]{slsqp}},
  \code{\link[nloptr]{neldermead}}, \code{\link[nloptr]{crs2lm}}, \code{\link[nloptr]{sbplx}},
  \code{\link[nloptr]{cobyla}}, \code{\link[nloptr]{lbfgs}}; from the
  \href{https://cran.r-universe.dev/minpack.lm/doc/manual.html}{minpack.lm} package:
  \code{\link[minpack.lm]{nls.lm}} and finally
  from the \href{https://cran.r-project.org/web/packages/pso/pso.pdf}{pso} package:
  \code{\link[pso]{psoptim}}.
}
\details{
All algorithms are based on the least-square minimization however,
  the \code{fn} definition in case of \code{nls.lm} must be provided as a difference/residual
  vector (see also \code{\link{eval_kinR_EPR_modelFit}}) and not as sum of differences/residual squares.
  The applied optimization/fitting methods are summarized in the following table (please, consult the details
  in \code{References} or in the individual function documentation - links in the \code{Description}) =>
  \tabular{lcl}{
  \strong{Method/Algorithm} \tab \strong{Package} \tab \strong{Short Description} \cr
  \code{slsqp} \tab \code{{nloptr}} \tab Sequential quadratic programming method for non-linearly
  constrained, gradient-based optimization. \cr
  \code{cobyla} \tab \code{{nloptr}} \tab Constrained optimization by linear approximations,
  algorithm for derivative-free optimization with nonlinear inequality and equality constraints. \cr
  \code{lbfgs} \tab \code{{nloptr}} \tab Low-storage version of the Broyden-Fletcher-Goldfarb-Shanno (BFGS) method.
  This is a quasi-Newton method well suited for the optimization problems with a large number of variables. \cr
  \code{neldermead} \tab \code{{nloptr}} \tab Nelder-Mead ("N-M") simplex algorithm. \cr
  \code{crs2lm} \tab \code{{nloptr}} \tab Controlled Random Search (CRS) algorithm (and in particular,
  the CRS2 variant) with the `local mutation' modification. \cr
  \code{sbplx} \tab \code{{nloptr}} \tab Subplex algorithm, which is a variant of the "N-M" method
  on a sequence of sub-spaces. \cr
  \code{nls.lm} (\code{levenmarq}) \tab \code{{minpack.lm}} \tab Modified Levenberg-Marquardt algorithm.
  It is a combination of gradient descent and Gauss-Newton method. \cr
  \code{psoptim} (\code{pswarm}) \tab \code{{pso}} \tab Particle swarm optimization, which is a population-based
  stochastic optimization algorithm motivated by the intelligent collective behavior of some animals such
  as flocks of birds or schools of fish. \cr
  }
  Not all \code{{nloptr}}-methods are implemented into the \code{optim_for_EPR_fitness}. Those summarized above
  were tested by the EPR simulation fit (see \code{\link{eval_sim_EPR_isoFit}}) on the experimental spectra
  of TEMPOL and Wuster's blue radicals. They provide the best results (without extensive "playing"
  with \code{\link[nloptr]{nl.opts}}, i.e. with options to control the optimization procedure) and proceed
  relatively fast.
}
\examples{
## generate Gaussian peak UV-Vis spectrum
set.seed(42)
wl <- seq(400, 600, by = 2) ## Wavelength
y  <- 2 * exp(- (wl - 500)^2/(2 * 15^2)) +
  rnorm(length(wl), sd = 0.05) ## Intensity
#
## data frame
df.model.expr <- data.frame(
  Wavelength_nm = wl,
  Intensity = y
)
#
## parametrize the fitting function
## `x` and `Intensity.fit` are column headers
Gaussian_Fit_x0 <-
  function(data,x,Intensity.fit,x0) {
  data[[Intensity.fit]] <-
    x0[1] * exp(- (data[[x]] - x0[2])^2 / (2 * x0[3]^2))
  return(data[[Intensity.fit]])
}
#
## fitness function
min_residuals_x0 <-
  function(data,x,Intensity.fit,x0) {
  sum(
    ## `"Intensity"`, see `df.model.expr`:
    (data[["Intensity"]] -
       Gaussian_Fit_x0(data,x,Intensity.fit,x0))^2
  )
}
#
## own optimization
optim.fit.list <-
  optim_for_EPR_fitness(
    method = "neldermead",
    x.0 = c(1.4,480,17), ## initial params. guess
    fn = min_residuals_x0,
    lower = c(1.3,470,13), ## lower bound for `x.0`
    upper = c(2.2,520,18), ## upper bound for `x.0`
    data = df.model.expr,
    x = "Wavelength_nm",
    Intensity.fit = "Fit",
    Nmax.evals = 256
  )
#
## best optimized parameters
optim.fit.list$par
#
## compare these parameters with those
## returned by the `stats::nls()` function,
## which by default, applies `Gauss-Newton`
## optim. algorithm
optim.fit.nls <-
  stats::nls(
    Intensity ~ A * exp(- (Wavelength_nm - mu)^2/(2 * sigma^2)),
    data = df.model.expr,
    start = list(A = 1.4, mu = 480, sigma = 17)
  )
unname(coef(optim.fit.nls))
#
## additional returned objects by `optim.fit.list`
## min RSS (corresponding to `fn`)
optim.fit.list$value
#
## number of iterations/evaluations
optim.fit.list$iter
#
## message + convergence (must be > 0)
optim.fit.list$message
optim.fit.list$convergence
#
## best fit ("Fit" column) into data frame
df.model.expr[["Fit"]] <-
  Gaussian_Fit_x0(
    data = df.model.expr,
    x = "Wavelength_nm",
    Intensity.fit = "Fit",
    x0 = optim.fit.list$par
  )
#
## plot experimental + best fit
df.model.expr \%>\% {
  graphics::plot(
    .$Wavelength_nm,
    .$Intensity,
    xlab = bquote(italic(Wavelength)~~"("~nm~")"),
    ylab = bquote(italic(Intensity)~~~"("~p.d.u.~")"),
    col = "darkcyan",
    pch = 16,
    cex = 1.2
  )
  graphics::lines(
    .$Wavelength_nm,
    .$Fit,
    col = "magenta",
    lwd = 2.4
  )
}


}
\references{
Johnson SG (2023). “The NLopt nonlinear-optimization package.” \url{https://github.com/stevengj/nlopt}.

 Stamm A (2023). “nloptr.” \url{https://github.com/astamm/nloptr/}.

 Mullen KM, Elzhov TV, Spiess A, Bolker B (2023). “minpack.lm.” \url{https://github.com/cran/minpack.lm}.

 Gavin HP (2019). “The Levenberg-Marquardt algorithm for nonlinear least squares curve-fitting problems.”
 \emph{Department of civil and environmental engineering, Duke University},
 \url{https://people.duke.edu/~hpgavin/lm.pdf}.

 Adyatama A (2019). “Particle Swarm Optimization.” \url{https://rpubs.com/argaadya/intro-pso}.

 Tam A (2021). “A Gentle Introduction to Particle Swarm Optimization.”
 \url{https://machinelearningmastery.com/a-gentle-introduction-to-particle-swarm-optimization/}.

 Ugolotti R, Cagnoni S (2016). "A Fair Comparison Between Standard PSO Versions."
 In: Rossi F, Mavelli F, Stano P, Caivano D (eds), \emph{Advances in Artificial Life, Evolutionary
 Computation and Systems Chemistry}, WIVACE 2015, \emph{Communications in Computer and Information Science},
 Springer, \url{https://doi.org/10.1007/978-3-319-32695-5_1}.

 Zambrano-Bigiarini M, Clerc M, Rojas-Mujica R (2013). "Standard Particle Swarm Optimisation 2011
 at CEC-2013: A baseline for future PSO improvements." \emph{2013 IEEE Congress on Evolutionary Computation},
 \url{https://ieeexplore.ieee.org/document/6557848}.
}
\seealso{
Other Simulations and Optimization:
\code{\link[=eval_ABIC_forFit]{eval_ABIC_forFit()}},
\code{\link[=eval_sim_EPR_iso]{eval_sim_EPR_iso()}},
\code{\link[=eval_sim_EPR_isoFit]{eval_sim_EPR_isoFit()}},
\code{\link[=eval_sim_EPR_isoFit_space]{eval_sim_EPR_isoFit_space()}},
\code{\link[=eval_sim_EPR_isoFitb]{eval_sim_EPR_isoFitb()}},
\code{\link[=eval_sim_EPR_iso_combo]{eval_sim_EPR_iso_combo()}},
\code{\link[=plot_eval_EPRtheo_mltiplet]{plot_eval_EPRtheo_mltiplet()}},
\code{\link[=plot_eval_RA_forFit]{plot_eval_RA_forFit()}},
\code{\link[=quantify_EPR_Sim_series]{quantify_EPR_Sim_series()}},
\code{\link[=smooth_EPR_Spec_by_npreg]{smooth_EPR_Spec_by_npreg()}}
}
\concept{Simulations and Optimization}
