% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_for_EPR_fitness.R
\name{optim_for_EPR_fitness}
\alias{optim_for_EPR_fitness}
\title{General Function to Non-Linear Optimization of the EPR Spectral Data}
\usage{
optim_for_EPR_fitness(
  method = "neldermead",
  x.0,
  fn,
  lower,
  upper,
  data,
  Nmax.evals = 1000,
  tol.step = 1e-06,
  pswarm.size = NULL,
  pswarm.diameter = NULL,
  ...
)
}
\arguments{
\item{method}{Character string, method...}

\item{x.0}{Numeric vector ...}

\item{fn}{Objective function that is to be minimized ...+ examples}

\item{lower, upper}{lower and upper bound constraints.}

\item{data}{Data object ...}

\item{Nmax.evals}{Numeric, maximum number of function evaluations and/or iterations.}

\item{tol.step}{Numeric, the smallest optimization step (relative change) between
2 iterations to stop the optimization or fitting procedure. For the \code{method == "pswarm"}
(particle swarm optimization procedure) it actually corresponds to tolerance for restarting
Once the maximal distance between the best particle and all other particles is less
than \code{tol.step} * \code{pswarm.diameter}) the algorithm restarts.
See also \code{\link[pso]{psoptim}}.}

\item{pswarm.size}{Numeric value equal to particle swarm size (i. e. number of particles).}

\item{pswarm.diameter}{Numeric value corresponding to diameter of search space.}

\item{...}{additional arguments passed to the function.}
}
\value{
description
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_sim_EPR_isoFit}()},
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{eval_sim_EPR_iso}()},
\code{\link{quantify_EPR_Sim_series}()}
}
\concept{Simulations and Optimization}
