% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_for_EPR_fitness.R
\name{optim_for_EPR_fitness}
\alias{optim_for_EPR_fitness}
\title{General Function to Non-Linear Optimization of the EPR Data}
\usage{
optim_for_EPR_fitness(
  method = "neldermead",
  x.0,
  fn,
  lower,
  upper,
  data,
  Nmax.evals = 1024,
  tol.step = 5e-07,
  pswarm.size = NULL,
  pswarm.diameter = NULL,
  ...
)
}
\arguments{
\item{method}{Character string, method...}

\item{x.0}{Numeric vector ...}

\item{fn}{Objective function that is to be minimized ...+ examples}

\item{lower, upper}{lower and upper bound constraints.}

\item{data}{Data object ...}

\item{Nmax.evals}{Numeric, maximum number of function evaluations and/or iterations.}

\item{tol.step}{Numeric, the smallest optimization step (relative change) between
2 iterations to stop the optimization or fitting procedure. For the \code{method == "pswarm"}
(particle swarm optimization procedure) it actually corresponds to tolerance for restarting
Once the maximal distance between the best particle and all other particles is less
than \code{tol.step} * \code{pswarm.diameter}) the algorithm restarts.
See also \code{\link[pso]{psoptim}}.}

\item{pswarm.size}{Numeric value equal to particle swarm size (i. e. number of particles).}

\item{pswarm.diameter}{Numeric value corresponding to diameter of search space.}

\item{...}{additional arguments passed to the function.}
}
\value{
description
}
\description{
General-purpose optimization of the objective \code{fn} function (also called "fitness")
  which is to be minimized in order to fit theoretical models (radical kinetic ones
  as well as EPR simulations) onto the experimental data. Several methods/algorithms implemented
  in \pkg{nloptr} (\code{\link[nloptr]{slsqp}}, \code{\link[nloptr]{neldermead}}, \code{\link[nloptr]{crs2lm}},
  \code{\link[nloptr]{sbplx}}, \code{\link[nloptr]{cobyla}}, \code{\link[nloptr]{lbfgs}}),
  \pkg{minpack.lm} (\code{\link[minpack.lm]{nls.lm}}) and \pkg{pso} (\code{\link[pso]{psoptim}})
  are used. All algorithms are based on the least-square minimization however,
  the \code{fn} definition in case of \code{nls.lm} must be provided as a difference/residual
  vector (see also \code{\link{eval_kinR_EPR_modelFit}}) and not as sum of difference/residual squares.
}
\details{
The algorithms applied for the optimization/fitting are summarized in the following table =>
  \tabular{lcl}{
  ----------------------- \tab ------------ \tab ---------------------------------------------------- \cr
  \strong{Method/Algorithm} \tab \strong{Package} \tab \strong{Short Description} \cr
  ----------------------- \tab ------------ \tab ---------------------------------------------------- \cr
  \code{slsqp} \tab \pkg{nloptr} \tab Sequential quadratic programming method for non-linearly
  constrained, gradient-based optimization. \cr
  \code{cobyla} \tab \pkg{nloptr} \tab Constrained optimization by linear approximations,
  algorithm for derivative-free optimization with nonlinear inequality and equality constraints. \cr
  \code{lbfgs} \tab \pkg{nloptr} \tab Low-storage version of the Broyden-Fletcher-Goldfarb-Shanno (BFGS) method.
  This is a quasi-Newton method well suited for the optimization problems with a large number of variables. \cr
  \code{neldermead} \tab \pkg{nloptr} \tab Nelder-Mead ("N-M") simplex algorithm. \cr
  \code{crs2lm} \tab \pkg{nloptr} \tab Controlled Random Search (CRS) algorithm (and in particular,
  the CRS2 variant) with the `local mutation' modification. \cr
  \code{sbplx} \tab \pkg{nloptr} \tab Subplex algorithm, which is a variant of the "N-M" method
  on a sequence of subspaces. \cr
  \code{nls.lm} (\code{levenmarq}) \tab \pkg{minpack.lm} \tab Modified Levenberg-Marquardt algorithm.
  It is a combination of gradient descent and Guass-Newton method. \cr
  \code{psoptim} (\code{pswarm}) \tab \pkg{pso} \tab Particle swarm optimization, which is a population-based
  stochastic optimization algorithm motivated by the intelligent collective behavior of some animals such
  as flocks of birds or schools of fish. \cr
  ----------------------- \tab ------------ \tab ---------------------------------------------------- \cr
  }
  Not all \pkg{nloptr}-methods are implemented into the \code{optim_for_EPR_fitness}. Those summarized above
  were tested by the EPR simulation fit (see \code{\link{eval_sim_EPR_isoFit}}) on the experimental spectra
  of TEMPOL and Wuster's blue radicals. They provide the best results (without extensive "playing"
  with \code{\link[nloptr]{nl.opts}}, i.e. with options to control the optimization procedure) and proceeds
  relatively fast.
}
\examples{
\dontrun{
TODO
TODO
}


}
\references{
\insertRef{nlopt2023}{eprscope}

 \insertRef{nloptr2023}{eprscope}

 \insertRef{levenmarq2023}{eprscope}

 \insertRef{gavin2019levenberg}{eprscope}

 \insertRef{AdyatamaPSO2019}{eprscope}

 \insertRef{TamPSO2021}{eprscope}
}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_sim_EPR_isoFit}()},
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{eval_sim_EPR_iso}()},
\code{\link{quantify_EPR_Sim_series}()}
}
\concept{Simulations and Optimization}
