% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_present_interact.R
\name{plot_EPR_present_interact}
\alias{plot_EPR_present_interact}
\title{Presents/Saves Interactive Plotly-Plot/-Graph Depending on Output Document Format}
\usage{
plot_EPR_present_interact(
  plot.name,
  size.width = 700,
  size.height = 500,
  size.resolv.f = 2
)
}
\arguments{
\item{plot.name}{tbc}

\item{size.width}{tbc}

\item{size.height}{tbc}

\item{size.resolv.f}{tbc}
}
\value{
Plot format depending on the output document one
}
\description{
Presents/Saves Interactive Plotly-Plot/-Graph Depending on Output Document Format
}
\examples{
\dontrun{
TODO
TODO
}


}
