% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs.R
\name{plot_EPR_Specs}
\alias{plot_EPR_Specs}
\title{Essential Plotting of EPR/ENDOR Spectrum/Spectra}
\usage{
plot_EPR_Specs(
  data.spectra,
  x = "B_mT",
  x.unit = "mT",
  xlim = NULL,
  var2nd.series = NULL,
  var2nd.series.slct.by = NULL,
  Intensity = "dIepr_over_dB",
  lineSpecs.form = "derivative",
  line.colors = "steelblue",
  line.width = 0.75,
  border.line.width = 0.5,
  border.line.color = "black",
  theme.basic = "theme_gray",
  axis.title.size = 15,
  axis.text.size = 14,
  legend.title = NULL,
  legend.title.size = NULL,
  legend.text.size = NULL,
  grid = TRUE,
  yTicks = TRUE
)
}
\arguments{
\item{data.spectra}{Spectrum data frame/table containing magnetic flux density, \eqn{g}-value
or radio-frequency columns as \code{x} variable. They can be labeled as \code{Field}, \code{B_mT}
in mT (or \code{B_G} in gauss), see also \code{x} parameter/argument. The \code{y/Intensity} variable
can be labeled as \code{dIepr_over_dB}, in case of derivative intensity, or if
integrated or simulated spectra intensities are present, they can be labeled accordingly.
See also \code{Intensity} parameter/argument. For spectral series the second independent variable
\code{var2nd.series} column (like e.g. \code{var2nd.series = "time_s"}) must be available. In such case
the entire \code{data.spectra} has to be in form of `tidy` table format (see also parameter/argument
\code{var2nd.series}).}

\item{x}{Character/String pointing to \code{x}-axis/column quantity in the original \code{data.spectra}
like magnetic flux density \eqn{B}, \eqn{g}-Value or \eqn{RF} (radio frequency),
\strong{default}: \code{x = "B_mT"}.}

\item{x.unit}{Character/String pointing to unit of quantity (coming from original ASCII data, see also
\code{column.names} parameter) which is to be presented on \eqn{x} abscissa of the EPR spectrum,
like \code{"G"} (`Gauss`), \code{"mT"} (`millitesla`), \code{"MHz"} (`megahertz` in case of ENDOR spectra)
or \code{"Unitless"} in case of \eqn{g}-values, \strong{default}: \code{x.unit = "mT"}.}

\item{xlim}{Numeric vector corresponding to border limits of the selected \eqn{x} region,
e.g. like `xlim = c(3495.4,3595.4)` (\eqn{B} in \code{G}) or `xlim = c(12.5,21.2)` (\eqn{RF} in \code{MHz})
or `xlim = c(2.004,2.001)` (\eqn{g} dimensionless). \strong{Default}: \code{xlim = NULL} (corresponding
to entire `x` range)}

\item{var2nd.series}{String/Character referred to name of the second independent variable/quantity
column in the original \code{data.spectra} (e.g. like `time`,`Temperature`, `Electrochemical Potential`,
`Microwave Power`...etc) altered upon individual experiments as a second variable
(\code{var2nd.series}) and related to spectra/data. Data must be available in \strong{long table}
(or \strong{tidy}) \strong{format} (see also \code{\link{readEPR_Exp_Specs_multif}}).
\strong{Default}: \code{var2nd.series = NULL}. Otherwise \strong{usually} \code{var2nd.series = "time_s"}.}

\item{var2nd.series.slct.by}{Numeric, number corresponding to each \eqn{n-th} presented spectrum in the plot,
like e.g. display each second (\code{var2nd.series.slct.by = 2}), third (\code{var2nd.series.slct.by = 3}),
fourth (\code{var2nd.series.slct.by = 4})...etc. spectrum. The argument is used in case
\code{var2nd.series} is \strong{NOT NULL} (e.g. \code{var2nd.series = "time_s"}) and one wants to present
separated labels/levels for spectra (not the continuous one). Recommended max. number of spectra/lines
is 12.}

\item{Intensity}{Character/String pointing to \code{intensity column} in the original \code{data.spectra}
if other than \code{dIepr_over_dB} name/label is used (e.g. for simulated or integrated spectra),
\strong{default}: \code{Intesity = "dIepr_over_dB"}.}

\item{lineSpecs.form}{Character string describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} or sigmoid-integrated which can be used as well)
line form of the analyzed EPR spectrum/data.}

\item{line.colors}{Character string, line color to plot EPR spectrum/spectra. All \pkg{ggplot2} compatible
  colors are allowed for plotting individual spectrum, \strong{default}: \code{line.colors = "steelblue"}.
  For series of EPR spectra two colorscales are used
  \enumerate{
  \item \strong{Continuous.} This is the case when \code{var2nd.series} \strong{IS NOT} \code{NULL}
  and \code{var2nd.series.slct.by = NULL}. The \code{line.colors} argument is identical with the continuous
  \code{colorscales} one from \code{\link[ggplot2]{scale_colour_gradientn}}. Following color definitions
  are allowed =>
  \itemize{
    \item an arbitrary vector color like \code{c("blue","green","red")} with the length of \eqn{\geq 2}
    \item any color definition from \pkg{grDevices} like \code{hcl.colors(n,pallete)}, \code{rainbow(n)},
    \code{heat.colors(n)}, \code{terrain.colors(n)}, \code{topo.colors(n)}, \code{cm.colors(n)} where the number
    of colors \eqn{n \geq 2} should be specified.
    See also \href{https://www.rdocumentation.org/packages/grDevices/versions/3.6.2/topics/Palettes}{grDevices Palettes}
    and \href{https://developer.r-project.org/Blog/public/2019/04/01/hcl-based-color-palettes-in-grdevices/}{HCL Color
    Palettes}
  }

  \item \strong{Discrete.} This is the case when both \code{var2nd.series}
  as well as \code{var2nd.series.slct.by} are \strong{DISTINCT} from \code{NULL}. Following color definitions
  are allowed =>
  \itemize{
  \item an arbitrary vector color like \code{c("blue","green","red")} with the length of \eqn{\geq 2}
  \item any color definition from \code{\link[ggplot2]{scale_color_viridis_d}} \code{"option"}.
  These involve \code{"magma"} (or \code{"A"}), \code{"inferno"} (or \code{"B"}), \code{"plasma"} (or \code{"C"}),
  \code{"viridis"} (or \code{"D"}), \code{"cividis"} (or \code{"E"}), \code{"rocket"} (or \code{"F"}),
  \code{"mako"} (or \code{"G"}) and \code{"turbo"} (or \code{"H"})
  }
  }}

\item{line.width}{Numeric, linewidth of the plot line in \code{pt}, \strong{default}: \code{line.width = 0.75}}

\item{border.line.width}{tbc}

\item{border.line.color}{tbc}

\item{theme.basic}{Character/String, which calls a ggplot theme base. The following ones are defined:
\describe{
  \item{\code{"theme_gray"}}{(\strong{default} one) => the gray background with white grid lines}
  \item{\code{"theme_bw"}}{ => the white background with thin gray grid lines, theme is proposed \strong{for publications}}
  \item{\code{"theme_light"}}{ => similar to \code{theme_bw()} but without the pronounced axis black lines}
  \item{\code{"theme_classic"}}{ => without grid, pronounced axis lines, however no opposite ones}
  \item{\code{"theme_linedraw"}}{ => pronounced axis lines (both for origin and opposite)
  as well as the grid-lines, theme is proposed \strong{for publications}
  (if the \code{grid} is set to \code{FALSE})}
}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{legend.title}{Character string tbc}

\item{legend.title.size}{tbc}

\item{legend.text.size}{description}

\item{grid}{Logical, whether to dislay the \code{grid} within the plot/graph, \strong{default}: \code{grid = TRUE}}

\item{yTicks}{Logical, whether to display the \code{y} (\code{dIepr_over_dB}) ticks and the corresponding text
(not the axis title!), which is usually skipped in the EPR community, \strong{default}: \code{yTicks = TRUE}}
}
\value{
EPR spectrum/spectra ('overlay' plot) by \pkg{ggplot2} with key parameter
  (e.g. line-color and theme,grid...etc.) variation
}
\description{
Graph/Plot of an EPR/ENDOR spectrum/spectra ('overlay' plot) based on \pkg{ggplot2}-functionality.
  Spectral data are in the form of data frame (details see below).
  Theme of the graphic spectrum representation as well its line colors can be varied like
  in \pkg{ggplot2}. Within a theme \code{y} ticks can be displayed
  or skipped \code{y} (e.g. \code{dIepr_over_dB} in 'procedure defined unit',
  see \href{http://www.iupac.org/divisions/VII/VII.C.1/C-NPU_Uppsala_081023_25_minutes_confirmed.pdf}{p.d.u.}),
  this is common for presenting the EPR spectra. Function can be additionally combined by \code{+} sign
  with other functions (e.g. with \code{\link{plot_theme_In_ticks}}) like in \pkg{ggplot2},
  e.g. present or skip \code{grid} within the code.
}
\examples{
\dontrun{
## simple plot of an EPR spectrum with `B` in `mT`
## and `dIepr_over_dB_Sim` in `p.d.u.` (derivative intensity)
plot_EPR_Specs(data.spectrum)
#
## simple plot of an EPR spectrum with `B` in `G`
## and `dIepr_over_dB_Sim` in `p.d.u.` (derivative intensity)
plot_EPR_Specs(data.spectrum,
              x = "B_G",
              x.unit = "G",
              Intensity = "dIepr_over_dB")
#
## single integrated EPR spectrum with `B` in `G`
## and `single_Integ` in `p.d.u.` (single integrated intensity)
plot_EPR_Specs(data.spectrum.integ,
              x = "B_G",
              x.unit = "G"
              Intensity = "single_Integ",
              lineSpecs.form = "absorption")
#
## Continuous display of time series EPR spectra selecting
## continuous colorscale (see argument `line.colors`
## and color definitions `1.`)
plot_EPR_Specs(data.spectra,
               var2nd.series = "time_s",
               xlim = c(340,357.5),
               line.colors = grDevices::hcl.colors(8,palette = "Roma"),
               legend.title = "Time (s)",
               yTicks = F)
#
## Separated display of time series EPR spectra selecting
## discrete colorscale (see argument `line.colors`
## and color definitions `2.`). Each 8-th spectrum from
## original plot is presented
plot_EPR_Specs(data.spectra,
               var2nd.series = "time_s",
               var2nd.series.slct.by = 8,
               xlim = c(340,357.5),
               line.colors = c("darkblue","darkorange","darkviolet"),
               legend.title = "Time (s)")
#
## Plotting simple CW ENDOR spectrum displayed
## as a "darkred" `line.colors`
plot_EPR_Specs(data.spectrum,
               x = "RF_MHz",
               x.unit = "MHz",
               line.colors = "darkred",
               line.width = 1.2,
               lineSpecs.form = "derivative")
}


}
\seealso{
Other Visualization and Graphics: 
\code{\link{draw_molecules_by_rcdk}()},
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{presentEPR_Sim_Spec}()}
}
\concept{Visualization and Graphics}
