% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs2D_interact.R
\name{plot_EPR_Specs2D_interact}
\alias{plot_EPR_Specs2D_interact}
\title{Interactive (incl. Zooming, Data Reading...etc) EPR Spectrum Plot}
\usage{
plot_EPR_Specs2D_interact(
  data.spectra,
  x = "B_mT",
  x.unit = "mT",
  Intensity = "dIepr_over_dB",
  var2nd.series = NULL,
  lineSpecs.form = "derivative",
  line.colors = "darkviolet",
  line.width = 0.75,
  line.type = 1,
  bg.color = "#e5ecf6",
  grid.color = "#ffff",
  border.line.width = 1.2,
  border.line.color = "black",
  legend.title = NULL,
  legend.title.size = NULL,
  axis.title.size = 15,
  axis.text.size = 14
)
}
\arguments{
\item{data.spectra}{Spectrum data frame/table containing magnetic flux density, \eqn{g}-value
or radio-frequency columns as \code{x} variable. They can be labeled as \code{Field}, \code{B_mT}
in mT (or \code{B_G} in gauss), see also \code{x} parameter/argument. The \code{y/Intensity} variable
can be labeled as \code{dIepr_over_dB}, in case of derivative intensity, or if
integrated or simulated spectra intensities are present, they can be labeled accordingly.
See also \code{Intensity} parameter/argument. For spectral series the second independent variable
\code{var2nd.series} column (like e.g. \code{var2nd.series = "time_s"}) must be available. In such case
the entire \code{data.spectra} has to be in form of `tidy` table format (see also parameter/argument
\code{var2nd.series}).}

\item{x}{Character/String pointing to \code{x}-axis/column quantity in original \code{data.spectrum} like
magnetic flux density \eqn{B}, \eqn{Field}, \eqn{g}-Value or \eqn{RF} (radio frequency),
\strong{default}: \code{x = "B_mT"}.}

\item{x.unit}{Character/String pointing to unit of quantity (coming from original ASCII data, see also
\code{column.names} parameter) which is to be presented on \eqn{x} abscissa of the EPR spectrum,
like \code{"G"} (`Gauss`), \code{"mT"} (`millitesla`), \code{"MHz"} (`megahertz` in case of ENDOR spectra)
or \code{"Unitless"} in case of \eqn{g}-values, \strong{default}: \code{x.unit = "mT"}.}

\item{Intensity}{Character/String pointing to \code{intensity column} in the original \code{data.spectrum}
if other than \code{dIepr_over_dB} name/label is used (e.g. for integrated or simulated spectra),
\strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{var2nd.series}{String/Character referred to name of the second independent variable/quantity
column in the original \code{data.spectra} (e.g. like `time`,`Temperature`, `Electrochemical Potential`,
`Microwave Power`...etc) altered upon individual experiments as a second variable
(\code{var2nd.series}) and related to spectra/data. Data must be available in \strong{long table}
(or \strong{tidy}) \strong{format} (see also \code{\link{readEPR_Exp_Specs_multif}}).
\strong{Default}: \code{var2nd.series = NULL}. Otherwise \strong{usually} \code{var2nd.series = "time_s"}.}

\item{lineSpecs.form}{Character string describing either \code{"derivative"} (\strong{default})
or \code{"integrated"} (i.e. \code{"absorption"} or sigmoid-integrated which can be used as well)
line form of the analyzed EPR spectrum/data.}

\item{line.colors}{Character string or its vector. In case of \strong{of SIMPLE SPECTRUM}
(NOT FOR \code{var2nd.series}) ONLY ONE COLOR CHARCTER STRING IS REQUIRED => therefore,
\strong{default:} \code{line.color = "darkviolet"}. For the SERIES OF SPECTRA CHARACTER COLOR VECTOR
WITH LENGTH \eqn{\geq 2} must be DEFINED (e.g. like \code{line.colors = c("darkorange","darkblue")}).}

\item{line.width}{Numeric, linewidth of the plot line in \code{mm}, \strong{default}: \code{line.width = 0.75}.}

\item{line.type}{Character string or integer corresponding to width of the (spectral) line(s). Following types
can be specified: \code{0 = "blank"}, \code{1 = "solid"} (\strong{default}), \code{2 = "dashed"}, \code{3 = "dotted"},
\code{4 = "dotdash"}, \code{5 = "longdash"} and \code{6 = "twodash"}.}

\item{bg.color}{Character string corresponding to \strong{background color}}

\item{grid.color}{Character string corresponding to \strong{grid color}}

\item{border.line.width}{tbc}

\item{border.line.color}{tbc}

\item{legend.title}{tbc}

\item{legend.title.size}{tbc}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}
}
\value{
Interactive EPR spectrum/spectra plot/graph based on \pkg{plotly}
}
\description{
tbc
}
\examples{
## load built-in EPR spectral data
data.file.path <-
  load_data_example(file = "TMPD_specelchem_accu_b.asc")
## spectrum recorded as accumulation of 20 scans
data.epr <-
  readEPR_Exp_Specs(path_to_ASC = data.file.path,
                    col.names = c("B_G", "dIepr_over_dB"),
                    qValue = 3500,
                    norm.vec.add = 20,
                    origin = "winepr")
## interactive plot or screenshot
plot_EPR_Specs2D_interact(data.spectra = data.epr)


}
\seealso{
Other Visualizations and Graphics: 
\code{\link{draw_molecule_by_rcdk}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{present_EPR_Sim_Spec}()}
}
\concept{Visualizations and Graphics}
