% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs3D_interact.R
\name{plot_EPR_Specs3D_interact}
\alias{plot_EPR_Specs3D_interact}
\title{Interactive (incl. Zooming, Data Reading...etc) 3D Plot for Time Series of EPR Spectra}
\usage{
plot_EPR_Specs3D_interact(
  data.time.spectra,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  time = "time_s",
  plot.type = "surface",
  scheme.color = "Viridis",
  contour.labels = FALSE,
  Blim,
  xlab = "<i>B</i> (mT)",
  ylab = "<i>Time</i> (s)",
  zlab = "d <i>I</i><sub>EPR</sub> / d <i>B</i> (p.d.u.)",
  bg.color.x = "rgb(220, 220,220)",
  grid.color.x = "rgb(255, 255, 255)",
  bg.color.y = "rgb(220, 220,220)",
  grid.color.y = "rgb(255, 255, 255)",
  bg.color.z = "rgb(220, 220,220)",
  grid.color.z = "rgb(255, 255, 255)",
  output.matrix.table = FALSE
)
}
\arguments{
\item{data.time.spectra}{tbc}

\item{x}{tbc}

\item{Intensity}{tbc}

\item{time}{tbc}

\item{plot.type}{tbc}

\item{scheme.color}{tbc}

\item{contour.labels}{tbc}

\item{Blim}{tbc}

\item{xlab}{tbc}

\item{ylab}{tbc}

\item{zlab}{tbc}

\item{bg.color.x}{tbc}

\item{grid.color.x}{tbc}

\item{bg.color.y}{tbc}

\item{grid.color.y}{tbc}

\item{bg.color.z}{tbc}

\item{grid.color.z}{tbc}

\item{output.matrix.table}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
