% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EPR_Specs_integ.R
\name{plot_EPR_Specs_integ}
\alias{plot_EPR_Specs_integ}
\title{Plotting Integrated Forms of EPR Spectra Acquired by \code{\link{eval_integ_EPR_Spec}}}
\usage{
plot_EPR_Specs_integ(
  data.spectra.integ,
  B = "B_G",
  B.unit = "G",
  Blim = NULL,
  ylim = NULL,
  slct.integs = c("single_Integ", "baseline_Integ_fit", "single_Integ_correct"),
  line.width = 0.75,
  line.type = 1,
  axis.title.size = 15,
  axis.text.size = 14,
  legend.title.size = 13,
  legend.text.size = 11,
  separate.integs = FALSE,
  separate.integ.scales = NULL,
  output.df = FALSE
)
}
\arguments{
\item{data.spectra.integ}{Data frame object, inherited output from the \code{\link{eval_integ_EPR_Spec}},
if \code{output.vecs = FALSE} (argument from the latter), corresponding to data frame including the original
EPR spectral data and the integral(s).}

\item{B}{Character string, pointing to magnetic flux density \code{column} of the original (\code{data.spectra.integ})
data frame, either in "millitesla" or in "gauss", that is \code{B = "B_G"} (\strong{default})
or \code{B = "B_mT"} or \code{B = "Bsim_G"} to include simulated EPR spectra as well.}

\item{B.unit}{Character string, denoting the magnetic flux density unit e.g. \code{B.unit = "G"}
(gauss, \strong{default}), \code{B.unit = "mT"} (millitesla) or \code{B.unit = T} (tesla).}

\item{Blim}{Numeric vector, magnetic flux density in \code{mT}/\code{G} corresponding to lower and upper limit
of the selected \eqn{B}-region, such as \code{Blim = c(3495.4,3595.4)}. \strong{Default}: \code{Blim = NULL}
(corresponding to the entire \eqn{B}-range of the integrated EPR spectrum).}

\item{ylim}{Numeric vector, corresponding to lower and upper limit of the \eqn{y}-axis scale
(e.g. \code{ylim = c(-1e-4,1e-3)}). This doesn't apply for separated integrals (if \code{separate.integs = TRUE})
and works only in overlay mode (in one graph/panel). \strong{Default}: \code{ylim = NULL} corresponding to the entire
\eqn{y}-range of presented integrals.}

\item{slct.integs}{Character string vector, pointing to selected integrals/columns/variables (of the original
\code{data.spectra.integ} data frame) to be presented in the actual plot. \strong{Default}:
\code{slct.integs = c("single_Integ","baseline_Integ_fit","single_Integ_correct")}.}

\item{line.width}{Numeric, linewidth of the plot line in \code{mm}, \strong{default}: \code{line.width = 0.75}.}

\item{line.type}{Character string or integer, corresponding to width of the (spectral) line(s). Following types
can be specified: \code{0 = "blank"}, \code{1 = "solid"} (\strong{default}), \code{2 = "dashed"}, \code{3 = "dotted"},
\code{4 = "dotdash"}, \code{5 = "longdash"} and \code{6 = "twodash"}.}

\item{axis.title.size}{Numeric, axis text title size in \code{pt}. \strong{Default}: \code{axis.title.size = 15}.}

\item{axis.text.size}{Numeric, text size in \code{pt} for the axes unit values/descriptions,
\strong{default}: \code{axis.text.size = 14}.}

\item{legend.title.size}{Numeric, legend text title size in \code{pt},\strong{default}: \code{legend.title.size = NULL},
actually corresponding to 13/13pt.}

\item{legend.text.size}{Numeric, legend text size in \code{pt}, \strong{default}: \code{legend.text.size = NULL},
actually corresponding to 11/11pt.}

\item{separate.integs}{Logical, should be the integrals presented in overlay mode (in one graph/panel)
or on separated panels (by \code{\link[ggplot2]{facet_wrap}}, see also the next argument) ?
\strong{Default}: \code{separate.integs = FALSE} (integrals are presented in overlay mode).}

\item{separate.integ.scales}{Character string related to \eqn{y}-axes scales, unless the \code{separate.integs = FALSE},
inherited from \code{\link[ggplot2]{facet_wrap}}. Following expressions are available => \code{"fixed"}, \code{"free"}
or in one dimension \code{"free_x"} or \code{"free_y"}. \strong{Default}: \code{separate.integ.scales = NULL}
in case of \code{separate.integs = FALSE}.}

\item{output.df}{Logical, whether a transformed \code{data.spectra.integ} data frame into
\href{https://r4ds.had.co.nz/tidy-data.html}{tidy/long table format} is required for additional processing
or plotting. \strong{Default}: \code{output.df = TRUE}.}
}
\value{
Depending on \code{output.df} argument, function returns plot object including all selected
  integrated EPR spectra (\code{output.df = FALSE},\strong{default}) or list (\code{output.df = TRUE})
  consisting of
  \describe{
  \item{df}{Data frame object with intensities of all selected integrals and magnetic flux density \eqn{B}
  variables/columns in tidy/long table format.}
  \item{plot}{Plot object showing all integrated EPR spectra corresponding to \code{df}.}
  }
}
\description{
Visualization (static plots based on \href{https://ggplot2.tidyverse.org/}{ggplot2}) of integrated EPR spectra
  coming as data frame outputs from \code{\link{eval_integ_EPR_Spec}}.
}
\examples{
## loading the package built-in example
TMPD.data.path <-
  load_data_example(file = "TMPD_specelchem_accu_b.asc")
TMPD.data <-
  readEPR_Exp_Specs(TMPD.data.path,
                    col.names = c("B_G","dIepr_over_dB"),
                    qValue = 3500,
                    norm.vec.add = c(20,0.001),
                    origin = "winepr")
## integration of the `TMPD` EPR spectrum
TMPD.data.integs <-
  eval_integ_EPR_Spec(TMPD.data,sigmoid.integ = TRUE)
#
## plotting integrals in overlay mode
plot_EPR_Specs_integ(TMPD.data.integs,
  slct.integs = c("single_Integ",
                  "sigmoid_Integ"),
  B = "B_mT",
  B.unit = "mT"
  )
#
## separate integrals within the plot
plot_EPR_Specs_integ(TMPD.data.integs,
  slct.integs = c("single_Integ",
                  "sigmoid_Integ"),
  B = "B_mT",
  B.unit = "mT",
  separate.integs = TRUE,
  separate.integ.scales = "free_y"
  )



}
\seealso{
Other Visualizations and Graphics: 
\code{\link{draw_molecule_by_rcdk}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{present_EPR_Sim_Spec}()}
}
\concept{Visualizations and Graphics}
