% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout2D_interact.R
\name{plot_layout2D_interact}
\alias{plot_layout2D_interact}
\title{Additional Layout for Interactive Plots Similar to Layout from Plotly Package}
\usage{
plot_layout2D_interact(
  p,
  data = NULL,
  xlab = "<i>B</i> (mT)",
  ylab = "d <i>I</i> <sub>EPR</sub> / d <i>B</i>  (p.d.u.)",
  axis.title.x.family = "Arial",
  axis.title.x.color = "black",
  axis.title.x.size = 15,
  axis.text.x.family = "Arial",
  axis.text.x.color = "black",
  axis.text.x.size = 14,
  axis.title.y.family = "Arial",
  axis.title.y.color = "black",
  axis.title.y.size = 15,
  axis.text.y.family = "Arial",
  axis.text.y.color = "black",
  axis.text.y.size = 14,
  legend.title = NULL,
  legend.title.family = "Arial",
  legend.title.color = "black",
  legend.title.size = 13,
  bg.color = "#e5ecf6",
  grid.color = "#ffff",
  grid.line.width = 1,
  grid.line.type = "solid",
  border.line.width = 1.4,
  border.line.color = "black"
)
}
\arguments{
\item{p}{A \code{plotly} object corresponding to previous/actual graph/plot.}

\item{data}{A data frame object to associate with the layout (optional). If not provided, arguments are evaluated
using the data frame in \code{link[plotly]{plot_ly}} function of the previous/actual plot. Therefore,
\strong{default}: \code{data = NULL}.}

\item{xlab}{tbc...}

\item{ylab}{tbc}

\item{axis.title.x.family}{...tbc...These include "Arial", "Balto", "Courier New", "Droid Sans", "Droid Serif",
"Droid Sans Mono", "Gravitas One", "Old Standard TT", "Open Sans", "Overpass", "PT Sans Narrow", "Raleway",
"Times New Roman".}

\item{axis.title.x.color}{tbc}

\item{axis.title.x.size}{tbc}

\item{axis.text.x.family}{tbc}

\item{axis.text.x.color}{tbc}

\item{axis.text.x.size}{tbc}

\item{axis.title.y.family}{tbc}

\item{axis.title.y.color}{tbc}

\item{axis.title.y.size}{tbc}

\item{axis.text.y.family}{tbc}

\item{axis.text.y.color}{tbc}

\item{axis.text.y.size}{tbc}

\item{legend.title}{tbc}

\item{legend.title.family}{tbc}

\item{legend.title.color}{tbc}

\item{legend.title.size}{tbc}

\item{bg.color}{tbc}

\item{grid.color}{tbc}

\item{grid.line.width}{description...width in pixels...}

\item{grid.line.type}{description...could be also "dot", "dash", "longdash", "dashdot", "longdashdot",
or a dash length list in px (eg "5px,10px,2px,2px").}

\item{border.line.width}{tbc}

\item{border.line.color}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Visualizations and Graphics: 
\code{\link{draw_molecule_by_rcdk}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{plot_theme_Out_ticks}()},
\code{\link{present_EPR_Sim_Spec}()}
}
\concept{Visualizations and Graphics}
