% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_NoY_ticks.R
\name{plot_theme_NoY_ticks}
\alias{plot_theme_NoY_ticks}
\title{TODO}
\usage{
plot_theme_NoY_ticks(
  base.theme = theme_gray(),
  axis.text.size,
  axis.title.size,
  bg.transparent = FALSE
)
}
\arguments{
\item{base.theme}{TODO}

\item{axis.text.size}{TODO}

\item{axis.title.size}{TODO}

\item{bg.transparent}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
