% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_NoY_ticks.R
\name{plot_theme_NoY_ticks}
\alias{plot_theme_NoY_ticks}
\title{Custom \code{ggplot2} Theme with NO \code{Y} Axis Ticks (as Common for EPR Spectra Presentation)}
\usage{
plot_theme_NoY_ticks(
  axis.text.size = 14,
  axis.title.size = 15,
  grid = TRUE,
  plot.bg.transparent = FALSE
)
}
\arguments{
\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 15}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 17}}

\item{grid}{Boolean, whether to dislay the \code{grid} within the plot/graph, \strong{default}: \code{grid = TRUE}}

\item{plot.bg.transparent}{Boolean, whether the \code{entire plot background} (NOT the \code{panel}=\code{own graph})
should be transparent, \strong{default}: \code{plot.bg.transparent = FALSE}, i.e. no transparent background}
}
\value{
Custom \pkg{ggplot2} \code{theme} \code{without Y axis ticks}, to show opposite X axis ticks
  use: \code{scale_x_continuous(sec.axis = dup_axis(name = "",labels = NULL))}
}
\description{
tbc
}
\examples{
tbc
tbc


}
