% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{plot_theme_Out_ticks}
\alias{plot_theme_Out_ticks}
\title{Custom \code{ggplot2} Theme with Axis Ticks Oriented Outside the Panel}
\usage{
plot_theme_Out_ticks(
  axis.text.size = 14,
  axis.title.size = 15,
  grid = TRUE,
  border.line.color = "black",
  border.line.type = 1,
  border.line.width = 0.5,
  bg.transparent = FALSE,
  ...
)
}
\arguments{
\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}.}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}.}

\item{grid}{Logical, whether to display the \code{grid} within the plot/graph panel, \strong{default}: \code{grid = TRUE}.}

\item{border.line.color}{Character string, setting up the color of the plot panel border line, \strong{default}:
\code{border.line.color = "black"}.}

\item{border.line.type}{Character string or integer, corresponding to width of the graph/plot panel border line. Following types
can be specified: \code{0 = "blank"}, \code{1 = "solid"} (\strong{default}), \code{2 = "dashed"}, \code{3 = "dotted"},
\code{4 = "dotdash"}, \code{5 = "longdash"} and \code{6 = "twodash"}..}

\item{border.line.width}{Numeric, width (in \code{mm}) of the plot panel border line, \strong{default}:
\code{border.line.width = 0.5}.}

\item{bg.transparent}{Logical, whether the \strong{entire plot background} (excluding the \strong{panel})
should be transparent, \strong{default}: \code{bg.transparent = FALSE}, i.e. no transparent background.}

\item{...}{additional arguments specified by the \code{\link[ggplot2]{theme}} (such as \code{panel.backgroud},
\code{axis.line},...etc), which are not specified otherwise.}
}
\value{
Custom \code{ggplot2} \code{theme} (list) with \code{x,y-axis} ticks pointing outside of the graph/plot panel.
}
\description{
Change the \href{https://ggplot2.tidyverse.org/}{ggplot2}-based theme in order to meet the needs
  of graph (such as EPR spectrum, kinetic profiles...etc)
  visuals/non-data components of the actual plot. The theme can be mainly applied for the basic
  \code{ggplot2} components like \code{ggplot() + geom_...() + ...} and consists of highlighted panel borders,
  grid and axis ticks pointing \strong{outside of the plot panel}. For details of \code{ggplot2} theme elements please,
  refer to \href{https://ggplot2.tidyverse.org/reference/theme.html}{Modify Components of a Theme}
  (see also \code{\link[ggplot2]{theme}}) or to
  \href{https://henrywang.nl/ggplot2-theme-elements-demonstration/}{ggplot2 Elements Demonstration by Henry Wang}.
}
\examples{
## loading TMPD built-in example file:
tmpd.data.file.path <-
  load_data_example(file = "TMPD_specelchem_accu_b.asc")
## reading data:
tmpd.data.file <-
  readEPR_Exp_Specs(path_to_ASC = tmpd.data.file.path,
                    col.names = c("B_G","dIepr_over_dB"),
                    qValue = 3500,
                    norm.vec.add = 20,
                    origin = "winepr")
#
ggplot2::ggplot(data = tmpd.data.file,
       ggplot2::aes(x = B_G,y = dIepr_over_dB)
       ) +
  ggplot2::geom_line(linewidth = 0.75,color = "darkgreen") +
  ggplot2::labs(
    x = "B (G)",
    y = "dIepr / dB  (p.d.u.)",
    title = "EPR Spectrum"
  ) +
  plot_theme_Out_ticks()


}
\seealso{
Other Visualizations and Graphics: 
\code{\link{draw_molecule_by_rcdk}()},
\code{\link{plot_EPR_Specs}()},
\code{\link{plot_EPR_Specs2D_interact}()},
\code{\link{plot_EPR_Specs3D_interact}()},
\code{\link{plot_EPR_Specs_integ}()},
\code{\link{plot_EPR_present_interact}()},
\code{\link{plot_labels_xyz}()},
\code{\link{plot_layout2D_interact}()},
\code{\link{plot_theme_In_ticks}()},
\code{\link{plot_theme_NoY_ticks}()},
\code{\link{present_EPR_Sim_Spec}()}
}
\concept{Visualizations and Graphics}
