% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRspecs_interact.R
\name{plotEPRspecs_interact}
\alias{plotEPRspecs_interact}
\title{Interactive (incl. Zooming, Data Reading...etc) EPR Spectrum Plot}
\usage{
plotEPRspecs_interact(
  spectrum.data,
  B = "B_mT",
  Intensity = "dIepr_over_dB",
  line.color = "steelblue",
  bg.color = "#e5ecf6",
  grid.color = "#ffff",
  line.width = 0.75,
  axis.title.size = 15,
  axis.text.size = 14,
  time.series = FALSE
)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} in mT (or \code{B_G} in gauss) and that of the derivative
intensity as \code{dIepr_over_dB}, \code{index} column can be included as well, integrated/simulated spectra
(incl. other \code{Intensity} and \code{B} columns) can be read as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (\strong{default})
or \code{B = "B_G"} or \code{B = "B_G_Sim"} to include simulated EPR spectra as well}

\item{Intensity}{Character/String pointing to \code{intensity column} if other than \code{dIepr_over_dB}
name/label is used (e.g. for integrated or simulated spectra), \strong{default}: \code{Intesity = "dIepr_over_dB"}}

\item{line.color}{Character/String corresponding to \strong{line color} in case \strong{of simple spectrum}
(not for \code{time.series}), therefore \strong{default:} \code{line.color = "steelblue"}}

\item{bg.color}{Character/String corresponding to \strong{background color}}

\item{grid.color}{Character/String corresponding to \strong{grid color}}

\item{line.width}{Numeric, linewidth of the plot line in \code{pt}, \strong{default}: \code{line.width = 0.75}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
\strong{default}: \code{axis.title.size = 15}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
\strong{default}: \code{axis.text.size = 14}}

\item{time.series}{Boolean, whether the input ASCII spectrum data comes from the time series experiment
with the additional \code{time} (labeled as \code{time_s}) column (ONLY IN CASE of "Xenon" software)}
}
\value{
Interactive EPR spectrum plot/graph based on \pkg{plotly}
}
\description{
tbc
}
\examples{
TODO
TODO


}
