% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPRspectr.R
\name{plotEPRspectr}
\alias{plotEPRspectr}
\title{EPR Spectrum Simple Plot}
\usage{
plotEPRspectr(
  spectrum.data,
  B = "B_mT",
  line.color = "steelblue",
  line.size = 0.75,
  basic.theme = "theme_gray",
  axis.text.size = 15,
  axis.title.size = 17,
  grid = TRUE,
  yTicks = TRUE
)
}
\arguments{
\item{spectrum.data}{Spectrum data frame/table where the magnetic flux density (in \code{mT}) column
must be labeled as \code{B_mT} in mT (or \code{B_G} in gauss) and that of the derivative
intensity as \code{dIepr_over_dB}, \code{index} column can be included as well}

\item{B}{Character/String pointing to magnetic flux density \code{column} of EPR spectrum data frame
\code{spectrum.data} either in \code{millitesla} or in \code{Gauss}, that is \code{B = "B_mT"} (default)
or \code{B = "B_G"}}

\item{line.color}{String, line color to plot simple EPR spectrum. All \pkg{ggplot2} compatible
colors are allowed, default: \code{line.color = "steelblue"}}

\item{line.size}{Numeric, linewidth of the plot line in \code{pt}, default: \code{line.size = 0.75}}

\item{basic.theme}{Character/String, which calls a ggplot theme base. The following ones are defined:
\itemize{
  \item \code{"theme_gray"} (default one) => the gray background with white grid lines, default theme
  \item \code{"theme_bw"} => the white background with thin gray grid lines
  \item \code{"theme_light"} => similar to \code{theme_bw()} but without the pronounced axis black lines
  \item \code{"theme_classic"} => without grid, pronounced axis lines, however no opposite ones
  \item \code{"theme_linedraw"} => pronounced axis lines (both for origin and opposite) as well as the grid-lines,
  theme is proposed \strong{for publications} (if the \code{grid} is set to \code{FALSE})
}}

\item{axis.text.size}{Numeric, text size (in \code{pt}) for the axes units/descriptions,
default: \code{axis.text.size = 15}}

\item{axis.title.size}{Numeric, text size (in \code{pt}) for the axes title,
default: \code{axis.title.size = 17}}

\item{grid}{Boolean, whether to dislay the \code{grid} within the plot/graph, default: \code{grid = TRUE}}

\item{yTicks}{Boolean, whether to display the \code{y} (\code{dIepr_over_dB}) ticks and the corresponding text
(not the axis title!), which is usually skipped in the EPR community, default: \code{yTicks = TRUE}}
}
\value{
EPR simple spectrum graph/plot with key parameter (e.g. line-color and theme,grid...etc.) variation
}
\description{
Graph/Plot of an EPR spectrum based on \code{\link{ggplot2}}-functionality. Spectral data
  are in the form of data frame, which must contain the \code{dIepr_over_dB} and \code{B_mT}/\code{B_G}
  (depending on units, data frame may include both of them) columns,
  i.e. derivative EPR intensity vs. magnetic flux density, respectively. Theme of the graphic
  spectrum representation as well its line color can be varied like in \pkg{ggplot2} (see below).
  Within a theme \code{y} ticks can be displayed or skipped \code{y} (\code{dIepr_over_dB} in 'procedure defined unit',
  see \href{http://www.iupac.org/divisions/VII/VII.C.1/C-NPU_Uppsala_081023_25_minutes_confirmed.pdf}{p.d.u.}),
  this is common for presenting the EPR spectra. Function can be additionally combined by \code{+} sign
  with other functions like in \pkg{ggplot2}, e.g. present or skip \code{grid} within the code.
}
\examples{
\dontrun{
plotEPRspectr(spectrum.data)
plotEPRspectr(spectrum.data,"B_G")
plotEPRspectr(spectrum.data,"blue",basic.theme = theme_linedraw(),yTicks = FALSE)
plotEPRspectr(spectrum.data,line.color = "steelblue",B = "B_G",theme_bw(),grid = TRUE)
plotEPRspectr(spectrum.data,"B_mT","darkred",line.size = 1.2)
}


}
