% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentSimEPRspectrum.R
\name{presentSimEPRspectrum}
\alias{presentSimEPRspectrum}
\title{TODO}
\usage{
presentSimEPRspectrum(
  exp.spectrum.data,
  sim.spectrum.data,
  B = "B_mT",
  Intensity.shift.ratio = 1.2,
  B.shift = 0,
  line.color.exp = "red",
  line.color.sim = "blue",
  line.size = 0.75,
  output.table = FALSE
)
}
\arguments{
\item{exp.spectrum.data}{TODO}

\item{sim.spectrum.data}{TODO}

\item{B}{TODO}

\item{Intensity.shift.ratio}{TODO}

\item{B.shift}{TODO}

\item{line.color.exp}{TODO}

\item{line.color.sim}{TODO}

\item{line.size}{TODO}

\item{output.table}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO
TODO


}
