% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_abs.R
\name{quantify_EPR_abs}
\alias{quantify_EPR_abs}
\title{TBC (Absolute EPR Quantity)}
\usage{
quantify_EPR_abs(
  integ.double,
  nu.GHz,
  power.mW,
  modul.amp.mT,
  qValue = 1,
  tube.sample.id.mm,
  fill.sample.h.mm,
  Norm.constant = 1,
  Temp.K = 298,
  S = 0.5,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{integ.double}{tbc}

\item{nu.GHz}{tbc}

\item{power.mW}{tbc}

\item{modul.amp.mT}{tbc}

\item{qValue}{tbc}

\item{tube.sample.id.mm}{tbc}

\item{fill.sample.h.mm}{tbc}

\item{Norm.constant}{tbc}

\item{Temp.K}{tbc}

\item{S}{tbc}

\item{microW.cavity}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
