% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_abs.R
\name{quantify_EPR_abs}
\alias{quantify_EPR_abs}
\title{Absolute Quantification of Radicals/Spins in the Most Common EPR Cavities}
\usage{
quantify_EPR_abs(
  integ.sigmoid.max,
  instrum.params = NULL,
  path_to_DSC_or_par,
  origin = "xenon",
  qValue = NULL,
  tube.sample.id.mm,
  fill.sample.h.mm,
  Norm.constant = NULL,
  Temp.K = NULL,
  S = 0.5,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{integ.sigmoid.max}{Numeric...tbc...}

\item{instrum.params}{Named numeric vector containing instrumental parameters required
for the quantification =>
\tabular{ll}{
\code{PmW} \tab power of the MW source in mW \cr
\code{BmmT} \tab modulation amplitude (magnetic flux density modulation,
\eqn{B_{\text{m}}}) in mT \cr
\code{TK} \tab temperature in K \cr
\code{mwGHz} \tab applied microwave frequency in `GHz` to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default}: \code{instrum.params = NULL} because they are primarily extracted
from the \code{path_to_DSC_or_par} based on the \code{origin}.}

\item{path_to_DSC_or_par}{Character string, path (can be also gathered by \code{\link[base]{file.path}})
to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} argument)
\code{text} files including all instrumental parameters from the EPR machine.
\strong{Default}: \code{path_to_DSC_or_par = NULL} in case if the \code{instrum.params}
is already defined. IF ARGUMENT \code{instrum.params = NULL} then BOTH \code{path_to_DSC_or_par}
AS WELL AS \code{origin} MUST BE DEFINED !}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_DSC_or_par = NULL}).}

\item{qValue}{Numeric, ...tbc... default NULL}

\item{tube.sample.id.mm}{Numeric, ...tbc...}

\item{fill.sample.h.mm}{Numeric, ...tbc...}

\item{Norm.constant}{Numeric, ...tbc...}

\item{Temp.K}{Numeric ...description... tbc default NULL}

\item{S}{Numeric, ...tbc...}

\item{microW.cavity}{Character string, ...tbc... \strong{Default}:
\code{microW.cavity = "rectangular"}.}
}
\value{
List with \code{"Ncm"} number of spins per effective centimeter (which is the cm
  around, \eqn{\pm 5\,\text{mm}}, the maximum of the intensity distribution curve within the cavity
  \eqn{f(B_1,B_{\text{m}})} from the equation above); with \code{"Ncm3"} corresponding
  to number of spins per \eqn{\text{cm}^3} and finally the \code{"cM"} denotes the concentration
  in \eqn{\text{mol}\,\text{dm}^{-3}}.
}
\description{
A short description...TBC...
}
\details{
There are two approaches how to quantify the number of paramagnetic species/radicals/spins.
  The \strong{relative} one needs a standard sample with a known number and can be evaluated
  by the sigmoid integral ratio of the sample under study and that of the standard.
  Whereas the \strong{absolute} method do not need the reference sample however, it requires
  a precise cavity signal calibration as well as standardized cell geometry. Both are provided
  by the EPR instrument and lab-glass manufacturers (see e.g.
  \href{https://hirschmann-laborgeraete.de/en/artikelgruppe/96001}{Hirschmann Capillaries}).
  In case of absolute quantitative EPR analysis the sigmoid integral (its maximum value),
  \eqn{I_{\text{sigmoid}}},can be expressed as follows =>
  \deqn{I_{\text{sigmoid}} = (c/f(B_1,B_{\text{m}}))\,(G_{\text{R}}\,t_{\text{C}}\,N_{\text{Scans}})\,
  [\sqrt{P_{\text{MW}}}\,B_{\text{m}}\,Q\,n_{\text{B}}\,S(S+1)]\,N_{\text{Spins}}}
  where  \eqn{c} is the point sample calibration factor (supplied by the spectrometer manufacturer);
  \eqn{f(B_1,B_{\text{m}})} is the function depicting the spatial distribution of \eqn{B_1}
  and \eqn{B_{\text{m}}} and actually it corresponds to integrated intensity distribution within
  the cavity/probehead for different sample length and positions. Such intensity distribution
  is expressed by polynomial and is supplied by the manufacturer as well. Additional quantities
  are the following => \eqn{G_{\text{R}}} is receiver gain; \eqn{t_{\text{C}}} corresponds to
  conversion time; \eqn{N_{\text{Scans}}} is the number of scans/accumulations. Because the receiver
  gain is expressed in \eqn{\text{dB}}...tbc.
}
\examples{
\dontrun{
tbc
tbc
}


}
\references{
{
  \insertRef{eatonQepr2010}{eprscope}

  \insertRef{hirschCapill2023}{eprscope}
}
}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Norm_const}()}
}
\concept{Evaluations and Quantification}
