% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_abs.R
\name{quantify_EPR_abs}
\alias{quantify_EPR_abs}
\title{Absolute Quantification of Radicals/Spins in the Most Common EPR Cavities}
\usage{
quantify_EPR_abs(
  integ.sigmoid.max,
  nu.GHz,
  P.mW,
  Bm.mT,
  qValue = NULL,
  tube.sample.id.mm,
  fill.sample.h.mm,
  Norm.constant = NULL,
  Temp.K = 298,
  S = 0.5,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{integ.sigmoid.max}{tbc}

\item{nu.GHz}{tbc}

\item{P.mW}{tbc}

\item{Bm.mT}{tbc}

\item{qValue}{tbc}

\item{tube.sample.id.mm}{tbc}

\item{fill.sample.h.mm}{tbc}

\item{Norm.constant}{tbc}

\item{Temp.K}{tbc}

\item{S}{tbc}

\item{microW.cavity}{tbc}
}
\value{
tbc
}
\description{
There are two approaches how to quantify the number of paramagnetic species/radicals/spins.
  The \strong{relative} one needs a standard sample with a known number and can be evaluated
  by the sigmoid integral ratio of the sample under study and that of the standard.
  Whereas the \strong{absolute} method do not need the reference sample however, it requires
  a precise cavity signal calibration as well as standardized cell geometry. Both are provided
  by the EPR instrument and lab-glass manufacturers. In case of absolute quantitative EPR analysis
  the sigmoid integral (its maximum value), \eqn{I_{\text{sigmoid}}},can be expressed as follows =>
  \deqn{I_{\text{sigmoid}} = (c/f(B_1,B_{\text{m}}))\,(G_{\text{R}}\,t_{\text{C}}\,N_{\text{Scans}})\,
  [\sqrt{P_{\text{MW}}}\,B_{\text{m}}\,Q\,n_{\text{B}}\,S(S+1)]\,N_{\text{Spins}}}
  where  \eqn{c} is the point sample calibration factor (supplied by the spectrometer manufacturer);
  \eqn{f(B_1,B_{\text{m}})} is the function depicting the spatial distribution of \eqn{B_1}
  and \eqn{B_{\text{m}}} and actually it corresponds to integrated intensity distribution within
  the cavity/probehead for different sample length and positions. Such intensity distribution
  is expressed by polynomial and is supplied by the manufacturer as well. Additional quantities
  ...tbc.
}
\examples{
\dontrun{
tbc
tbc
}


}
\references{
{
  \insertRef{eatonQepr2010}{eprscope}
}
}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Norm_const}()},
\code{\link{quantify_EPR_sim}()}
}
\concept{Evaluations and Quantification}
