% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_Abs.R
\name{quantify_EPR_Abs}
\alias{quantify_EPR_Abs}
\title{Absolute Quantification of Radicals/Spins}
\usage{
quantify_EPR_Abs(
  integ.sigmoid.max,
  instrum.params = NULL,
  path_to_dsc_par,
  origin = "xenon",
  qValue = NULL,
  tube.sample.id.mm,
  point.sample.factor = 8.51e-09,
  fill.sample.h.mm,
  eff.cavity.h.mm = 23,
  fn.B1.Bm.fit = c(1.00179, -0.00307086, -0.0265409, 0.000297603, 0.000223277,
    -4.53833e-06, -4.1451e-07, 1.89417e-08, -1.48241e-09),
  Norm.const = NULL,
  Temp.K = NULL,
  S = 0.5
)
}
\arguments{
\item{integ.sigmoid.max}{Numeric value or vectoe of entire EPR spectrum sigmoid integral.}

\item{instrum.params}{Named numeric vector containing instrumental parameters required
for the quantification =>
\tabular{ll}{
\code{PmW} \tab power of the MW source in mW \cr
\code{BmmT} \tab modulation amplitude (magnetic flux density modulation,
\eqn{B_{\text{m}}}) in mT \cr
\code{TK} \tab temperature in K \cr
\code{mwGHz} \tab applied microwave frequency in `GHz` to record the continuous wave (CW)
EPR spectrum \cr
}
\strong{Default}: \code{instrum.params = NULL} because they are primarily extracted
from the \code{path_to_dsc_par} based on the \code{origin}.}

\item{path_to_dsc_par}{Character string, path (can be also gathered by \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on the OS, see \code{origin} argument)
\code{text} files including all instrumental parameters from the EPR machine.
\strong{Default}: \code{path_to_dsc_par = NULL} in case if the \code{instrum.params}
is already defined. IF ARGUMENT \code{instrum.params = NULL} then BOTH \code{path_to_dsc_par}
AS WELL AS \code{origin} MUST BE DEFINED !}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = NULL} in case no file is used to extract
the parameters (i.e. exactly if \code{path_to_dsc_par = NULL}).}

\item{qValue}{Numeric value of the sensitivity `Q` factor. For the processed EPR spectra by
the `{eprscope}` package the \code{integ.sigmoid.max} is usually normalized by the `Q` value.
Therefore, \strong{default}: \code{qValue = NULL}.}

\item{tube.sample.id.mm}{Numeric value equal to internal diameter (in `mm`) of the tube/cell used
for the quantitative EPR experiment.}

\item{point.sample.factor}{Numeric value ... tbc ...}

\item{fill.sample.h.mm}{Numeric value equal to sample height (in `mm`) within the tube/cell.}

\item{eff.cavity.h.mm}{Numeric value equal to effective cavity/probehead height/length,
usually provided by the probehead manufacturer.}

\item{fn.B1.Bm.fit}{Numeric vector (coefficients) of the polynomial degree from 6 to 11
or character string ("theoretical").}

\item{Norm.const}{Numeric value corresponding to normalization constant (see
\code{\link{quantify_EPR_Norm_const}}). \strong{Default}: \code{Norm.const = NULL} in case
if the EPR spectrum was normalized by such constant either upon measurement or during processing.
Otherwise it must be provided by \code{\link{quantify_EPR_Norm_const}}.}

\item{Temp.K}{Numeric temperature value in `K`. Because the \code{instrum.params} also contains temperature
input one may choose which definition (\code{Temp.K} or \code{TK}) is taken for calculation.
Either \code{Temp.K} or \code{TK} CAN BE ALSO `NULL` but NOT BOTH !! In the latter case default value `298 K`
is considered.}

\item{S}{Numeric value, total spin sample quantum number. For radicals \code{S = 0.5}
(\strong{default}).}
}
\value{
List of the following quantities:
  \describe{
  \item{N.cm}{Number of spins per effective centimeter. It is defined
    as the cm around the maximum, \eqn{\pm 5\,\text{mm}}, of the intensity
    distribution curve within the cavity \eqn{f(B_1,B_{\text{m}})} from
    the equation above shown in details.}
   \item{N.cm3}{Number of spins per \eqn{\text{cm}^3}.}
   \item{c.M}{Concentration of spins/radicals in \eqn{\text{mol}\,\text{dm}^{-3}}.}
  }
}
\description{
Estimate the number or concentration of "spins"/radicals/paramagnetic species using the absolute
  quantitative method by sigmoid integral as well as by instrumental parameters without the need
  for a standard sample with known concentration/amount of radicals/"spins".
}
\details{
There are two approaches how to quantify the number of paramagnetic species/radicals/spins.
  The \strong{relative} one needs a standard sample with a known number and can be evaluated
  by the sigmoid integral ratio of the sample under study and that of the standard.
  Whereas the \strong{absolute} method do not need the reference sample however, it requires
  a precise cavity signal calibration as well as standardized cell geometry. Both are provided
  by the EPR instrument and lab-glass manufacturers (see e.g.
  \href{https://hirschmann-laborgeraete.de/en/artikelgruppe/96001}{Hirschmann Capillaries}).
  In case of absolute quantitative EPR analysis the sigmoid integral (its maximum value),
  \eqn{I_{\text{sigmoid}}},can be expressed as follows =>
  \deqn{I_{\text{sigmoid}} = (c/f(B_1,B_{\text{m}}))\,(G_{\text{R}}\,t_{\text{C}}\,N_{\text{Scans}})\,
  [\sqrt{P_{\text{MW}}}\,B_{\text{m}}\,Q\,n_{\text{B}}\,S(S+1)]\,N_{\text{Spins}}}
  where  \eqn{c} is the point sample calibration factor (supplied by the spectrometer manufacturer);
  \eqn{f(B_1,B_{\text{m}})} is the function depicting the spatial distribution of \eqn{B_1}
  and \eqn{B_{\text{m}}} and actually it corresponds to integrated intensity distribution within
  the cavity/probehead for different sample length and positions. Such intensity distribution
  is expressed by polynomial and is supplied by the manufacturer as well. Additional quantities
  are the following => \eqn{G_{\text{R}}} is receiver gain; \eqn{t_{\text{C}}} corresponds to
  conversion time; \eqn{N_{\text{Scans}}} is the number of scans/accumulations. Because the receiver
  gain is expressed in \eqn{\text{dB}}...tbc.
}
\examples{
\dontrun{
tbc
tbc
}


}
\references{
\insertRef{eatonQepr2010}{eprscope}

  \insertRef{RWeberXenon2011}{eprscope}

  \insertRef{hirschCapill2023}{eprscope}

  \insertRef{mazurAnalysis2000}{eprscope}

  \insertRef{portisElectro1953}{eprscope}

  \insertRef{mailerQuant1977}{eprscope}
}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Norm_const}()}
}
\concept{Evaluations and Quantification}
