% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_Norm_const.R
\name{quantify_EPR_Norm_const}
\alias{quantify_EPR_Norm_const}
\title{Normalization Constant Calculation for Quantitative EPR Analysis}
\usage{
quantify_EPR_Norm_const(conv.time.ms, Nscans, Bsw = NULL, rg, rg.unit = "dB")
}
\arguments{
\item{conv.time.ms}{Numeric, conversion time in milliseconds.}

\item{Nscans}{Numeric, number of scans.}

\item{Bsw}{Numeric, experimental sweep width (magnetic flux density recording region,
\eqn{B_{\text{SW}}}) in "G". \strong{Default}: \code{Bsw = NULL}.}

\item{rg}{Numeric, receiver gain value.}

\item{rg.unit}{Character string corresponding to unit of the receiver gain.
Either \code{rg.unit = "db"} (\code{rg.unit = "dB"}, \strong{default})
or \code{rg.unit = "unitless"} (\code{rg.unit = "Unitless"}).}
}
\value{
Numeric value of the normalization constant for quantitative EPR and intensity normalization.
}
\description{
Normalization constant used by \code{\link{quantify_EPR_Abs}} or to normalize
  the EPR spectrum intensity. The calculation depends on expression of the receiver gain.
}
\details{
For the receiver gain expressed in \eqn{\text{dB}} units the normalization constant is defined
  by the following relation  =>
  \deqn{N_{\text{norm}} = t_{\text{C}}(\text{ms})\,N_{\text{Scans}}\,(20)\,
  10^{(G_{\text{R}}(\text{dB})/20)}}
  where \eqn{t_{\text{C}}(\text{ms})} depicts the conversion time in \eqn{\text{ms}};
  \eqn{N_{\text{Scans}}} corresponds to number of scans and \eqn{G_{\text{R}}(\text{dB})}
  is the receiver gain in \eqn{\text{dB}}. In case the receiver gain is unitless, the normalization
  constant is defined by =>
  \deqn{N_{\text{norm}} = t_{\text{C}}(\text{ms})\,G_{\text{R}}\,
  (N_{\text{Scans}} - 1)\,/\,B_{\text{SW}}(\text{G})}
  where \eqn{G_{\text{R}}} and \eqn{B_{\text{SW}}(\text{G})} correspond to unitless receiver gain
  and sweep width in Gauss, respectively. One can gather all parameters
  by \code{\link{readEPR_param_slct}} or by \code{\link{readEPR_params_tabs}} functions from
  the corresponding `.DSC`/`.dsc` or `.par` file. \strong{If during the recording of EPR spectra
  the option} `Normalize Acquisition` (in Spectrometer Configuration/Acquisition Options)
  \strong{is activated, THE INTENSITY is ALREADY NORMALIZED and DOESN'T REQUIRED ANY ADDITIONAL
  NORMALIZATION !}. See also \code{\link{quantify_EPR_Abs}}.
}
\examples{
quantify_EPR_Norm_const(conv.time.ms = 8.2,
                        Nscans = 10,
                        rg = 32)
#
quantify_EPR_Norm_const(conv.time.ms = 13.1,
                        Bsw = 180,
                        Nscans = 10,
                        rg = 3.2e+4,
                        rg.unit = "Unitless")


}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_Abs}()}
}
\concept{Evaluations and Quantification}
