% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_Norm_const.R
\name{quantify_EPR_Norm_const}
\alias{quantify_EPR_Norm_const}
\title{Normalization Constant Calculation for Quantitative EPR Analysis}
\usage{
quantify_EPR_Norm_const(conv.time.ms, Nscans, rg.dB)
}
\arguments{
\item{conv.time.ms}{Numeric, conversion time in milliseconds.}

\item{Nscans}{Numeric, number of scans.}

\item{rg.dB}{Numeric, receiver gain in dB.}
}
\value{
Numeric value of normalization constant for quantitative EPR.
}
\description{
Normalization constant used by \code{\link{quantify_EPR_abs}}. This constant
  can be evaluated from the following expression =>
  \deqn{N_{\text{norm}} = t_{\text{C}}\,N_{\text{Scans}}\,20\,10^{(G_{\text{R}}/20)}}
  where \eqn{t_{\text{C}}} depicts the conversion time in \eqn{\text{ms}}; \eqn{N_{\text{Scans}}}
  corresponds to number of scans and \eqn{G_{\text{R}}} is the receiver gain in \eqn{\text{dB}}.
  One can gather all three parameters by \code{\link{readEPR_param_slct}} function from the
  corresponding `.DSC` or `.par` file. \strong{If during the recording of EPR spectra the option}
  `Normalize Acquisition` (in Spectrometer Configuration/Acquisition Options) \strong{is activated
  the intensity is already normalized and it doesn't require any additional normalization.}.
  See also \code{\link{quantify_EPR_abs}}.
}
\examples{
quantify_EPR_Norm_const(conv.time.ms = 6.4,
                        Nscans = 6,
                        rg.dB = 30)
#
quantify_EPR_Norm_const(conv.time.ms = 13.52,
                        Nscans = 100,
                        rg.dB = 24)

}
\seealso{
Other Evaluations and Quantification: 
\code{\link{eval_integ_EPR_Spec}()},
\code{\link{eval_kinR_EPR_modelFit}()},
\code{\link{eval_kinR_ODE_model}()},
\code{\link{quantify_EPR_abs}()},
\code{\link{quantify_EPR_sim}()}
}
\concept{Evaluations and Quantification}
