% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_sim.R
\name{quantify_EPR_sim}
\alias{quantify_EPR_sim}
\title{Quantify Area of Simulated EPR Spectral Series Instead of Experimental One}
\usage{
quantify_EPR_sim(
  data.spectra.series,
  path_to_ASC_sim,
  var2nd = "time_s",
  B.unit = "G",
  Intensity.exp = "dIepr_over_dB",
  Intensity.sim = "dIeprSim_over_dB",
  single.integ = "sIntegral_Sim",
  double.integ = "dIntegral_Sim",
  output.area = TRUE
)
}
\arguments{
\item{data.spectra.series}{tbc}

\item{path_to_ASC_sim}{tbc}

\item{var2nd}{tbc}

\item{B.unit}{tbc}

\item{Intensity.exp}{tbc}

\item{Intensity.sim}{tbc}

\item{single.integ}{tbc}

\item{double.integ}{tbc can be also \code{NULL} if case of single integral spectral series input}

\item{output.area}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
