% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_sim.R
\name{quantify_EPR_sim}
\alias{quantify_EPR_sim}
\title{Quantify (Components) Area of Simulated EPR Spectral Series Instead of Experimental One}
\usage{
quantify_EPR_sim(
  data.spectra.series,
  dir_ASC_sim,
  name_pattern_sim,
  origin_sim = "easyspin",
  var2nd.series = "time_s",
  B.unit = "G",
  Intensity.expr = "dIepr_over_dB",
  Intensity.sim = "dIeprSim_over_dB",
  optim.method = "slsqp",
  optim.params.init,
  optim.params.lower = NULL,
  optim.params.upper = NULL,
  single.integ = "single_IntegSim",
  double.integ = "double_IntegSim",
  output.area.stat = TRUE
)
}
\arguments{
\item{data.spectra.series}{tbc}

\item{dir_ASC_sim}{tbc}

\item{name_pattern_sim}{description}

\item{origin_sim}{description from}

\item{var2nd.series}{String/Character referred to name of the second independent variable/quantity
column in the original \code{data.spectra} (e.g. like `time`,`Temperature`, `Electrochemical Potential`,
`Microwave Power`...etc) altered upon individual experiments as a second variable
(\code{var2nd.series}) and related to spectra/data. Data must be available in \strong{long table}
(or \strong{tidy}) \strong{format} (see also \code{\link{readEPR_Exp_Specs_multif}}).
\strong{Default}: \code{var2nd.series = NULL}. Otherwise \strong{usually} \code{var2nd.series = "time_s"}.}

\item{B.unit}{tbc}

\item{Intensity.expr}{tbc}

\item{Intensity.sim}{tbc}

\item{optim.method}{Character string description tbc...following methods from \pkg{nloptr} are available:
\code{optim.method = "slsqp"} (\strong{default}), \code{optim.method = "neldermead"},
\code{optim.method = "mma"} and \code{optim.method = "ccsaq"}...tbc}

\item{optim.params.init}{Numeric vector...description tbc...inherited from \code{x0} parameter/argument
of a \pkg{nloptr} function (see e.g. \code{\link[nloptr]{mma}})...tbc}

\item{optim.params.lower}{Numeric vector...description tbc...inherited from \code{lower} parameter/argument
of a \pkg{nloptr} function (see e.g. \code{\link[nloptr]{mma}})...tbc.}

\item{optim.params.upper}{Numeric vector...description tbc...inherited from \code{upper} parameter/argument
of a \pkg{nloptr} function (see e.g. \code{\link[nloptr]{mma}})...tbc.}

\item{single.integ}{tbc}

\item{double.integ}{tbc can be also \code{NULL} in case of single integral spectral series input}

\item{output.area.stat}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Simulation and Optimization: 
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{eval_sim_EPR_iso}()},
\code{\link{optim_fit_EPR_by_nloptr}()}
}
\concept{Simulation and Optimization}
