% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_Sim_series.R
\name{quantify_EPR_Sim_series}
\alias{quantify_EPR_Sim_series}
\title{Quantify EPR Simulated Spectral Components in the Experimental Series}
\usage{
quantify_EPR_Sim_series(
  data.spectra.series,
  dir_ASC_sim,
  name.pattern.sim,
  origin.sim = "easyspin",
  var2nd.series = "time_s",
  B.unit = "G",
  col.names.sim = c("Bsim_G", "dIeprSim_over_dB"),
  x.sim.id = 1,
  Intensity.sim.id = 2,
  Intensity.expr = "dIepr_over_dB",
  optim.method = "sbplx",
  optim.params.init,
  optim.params.lower = NULL,
  optim.params.upper = NULL,
  Nmax.evals = 512,
  single.integ = "single_IntegSim",
  double.integ = "double_IntegSim",
  msg.optim.progress = TRUE,
  output.area.stat = TRUE,
  ...
)
}
\arguments{
\item{data.spectra.series}{Spectrum data frame/table object containing magnetic flux density
as \code{x} variable. They can be labeled as \code{Field}, \code{B_mT}
in mT (or \code{B_G} in gauss). The \code{y/Intensity} variable
can be labeled as \code{dIepr_over_dB}, in case of derivative intensity, or if
integrated spectral intensities are present, they can be labeled accordingly.
See also \code{Intensity.expr} parameter/argument. A second independent variable
\code{var2nd.series} column (e.g. \code{var2nd.series = "time_s"}) must be available. In such case,
the entire \code{data.spectra} must be present in the form of
\href{https://r4ds.had.co.nz/tidy-data.html}{tidy/long table format}
(see also the parameter/argument \code{var2nd.series}). Such data frame can be created, e.g.
by the \code{\link{readEPR_Exp_Specs_kin}} or the \code{\link{readEPR_Exp_Specs_multif}} function.}

\item{dir_ASC_sim}{Character string, pointing to folder where the simulated EPR spectra (data files) of all
components are stored. Path can be alternatively specified by the \code{\link[base]{file.path}} function.}

\item{name.pattern.sim}{Character string pattern from file names related to simulated EPR spectral data
like \code{name.pattern.sim = "DHMB0_1st_04_SimA"}
or \code{name.pattern.sim = "DHMB0_1st_04_Sim[[:upper:]]"} (for the file names \code{..._SimA},
\code{..._SimB},...etc). It assumes, those files must have similar names and this pattern appears
at the beginning of the file name. One may also consult
how to \href{https://r4ds.hadley.nz/regexps}{use regular expressions in R}.}

\item{origin.sim}{Character string, referring to "origin" of the simulated ASCII data.
There are four possibilities \eqn{\Rightarrow} \code{sim.orimgin = "easyspin"} (\strong{default}),
\code{"xenon"}, \code{"simfonia"} as well as universal \code{"csv"}.}

\item{var2nd.series}{Character string referred to name of the second independent variable/quantity
column in the original \code{data.spectra.series} (such as time, temperature, electrochemical potential,
Microwave Power) altered during individual experiments as a second variable.
Data must be available in tidy/long table format.
\strong{Default}: \code{var2nd.series = "time_s"}.}

\item{B.unit}{Character string, pointing to unit of magnetic flux density
like \code{"G"} (Gauss) or \code{"mT"} (millitesla), \strong{default}: \code{B.unit = "G"}.
THE UNIT MUST BE SHARED ACROSS ALL RELEVANT B-ARGUMENTS/DATAFRAMES.}

\item{col.names.sim}{Character string vector, pointing to column names/headers of the original ASCII data
series (refer also to the \code{\link{readEPR_Sim_Spec}} function).
\strong{Default}: \code{col.names.sim = c("Bsim_G","dIeprSim_over_dB")},
corresponding to the \code{origin.sim = "easyspin"}.}

\item{x.sim.id}{Numeric index related to the \code{col.names.sim} vector pointing to independent variable (like "B" or "Bsim"
- magnetic flux density), which corresponds to \eqn{x}-axis in the simulated spectra (refer also to the original ASCII data
of the simulated spectrum). \strong{Default}: \code{x.sim.id = 1} (see also default of the \code{col.names.sim}).}

\item{Intensity.sim.id}{Numeric index related to the \code{col.names.sim} vector pointing to simulated EPR intensity
column name/header in the original ASCII data.
\strong{Default}: \code{x.sim.id = 2} (see also default of the \code{col.names.sim}).}

\item{Intensity.expr}{Character string, pointing to column name of the experimental EPR intensity within
the original \code{data.spectra.series}. \strong{Default}: \code{dIepr_over_dB}.}

\item{optim.method}{Character string, pointing to applied optimization method/algorithm.
One may choose \strong{one} from those listed in \code{\link{optim_for_EPR_fitness}}, \strong{default}:
\code{method = "sbplx"}, setting up
the \href{https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/#sbplx-based-on-subplex}{"Subplex" method}.}

\item{optim.params.init}{Numeric vector with the elements: baseline constant/intercept
followed by simulated intensity multiplication coefficient for each EPR spectral component.
Therefore, the length of this vector is equal to number of components + 1.}

\item{optim.params.lower}{Numeric vector with the length of \code{optim.params.init} and the lower
bound constraints. \strong{Default}: \code{optim.params.init = NULL}, actually corresponding
to vector with all \code{0} value elements.}

\item{optim.params.upper}{Numeric vector with the length of \code{optim.params.init}) and the upper
bound constraints. \strong{Default}: \code{optim.params.init = NULL}, actually corresponding
to vector with all \code{0.9} value elements.}

\item{Nmax.evals}{Numeric value, maximum number of function evaluations and/or iterations.
The only one method, limited by this argument, is \code{\link[minpack.lm]{nls.lm}}, where
\code{Nmax.evals = 1024}. Higher \code{Nmax.evals} may extremely extend the optimization
time, therefore the \strong{default} value reads \code{Nmax.evals = 512}. However, the \code{"pswarm"}
method requires at least the default or even higher values.}

\item{single.integ}{Character string, setting up the column/variable name related to single-integrated spectrum
within the output data frame, \strong{default}: \code{single.integ = "single_IntegSim"}.}

\item{double.integ}{Character string, setting up the column/variable name related to double-integrated spectrum
within the output data frame, \strong{default}: \code{double.integ = "double_IntegSim"}.
If \code{double.integ = NULL}, only single integrals are calculated/returned (e.g. in the case of
single integrated spectral data).}

\item{msg.optim.progress}{Logical, whether to display message (in the R console) about progress of the \code{optim.method}
during the optimization/fitting procedure, e.g.
\code{Intensities of simulated EPR spectral component(s) are currently being evaluated/optimized}
\code{by  NELDERMEAD  method  (4 Component(s))} and at the end it shows the elapsed time.
\strong{Default}: \code{msg.optim.progress = TRUE}. If \code{FALSE}, no message
is displayed. This argument can be combined with the optional \code{eval.optim.progress}
(see below or described in the \code{\link{optim_for_EPR_fitness}}).}

\item{output.area.stat}{Logical, whether to summarize all fitted EPR spectral components, in columns,
for each time/temperature/...etc. point in row. Additional optimization measures are presented as well
(see \code{Value}).\strong{Default}: \code{output.area.stat = TRUE}. Otherwise, tidy/long format
table is returned, including all EPR spectra and their corresponding integrals.}

\item{...}{additional arguments specified, see also \code{\link{optim_for_EPR_fitness}},
like \code{eval.optim.progress = TRUE} (which is \code{FALSE} by \strong{default}), \code{pswarm.size},
\code{pswarm.diameter}, \code{pswarm.type}, \code{tol.step} + \code{optim.method} depended additional arguments
as well as \code{fix.optim.x.0.id} to fix one or several parameters defined by the \code{optim.params.init}
argument.}
}
\value{
Function provides data frame object, depending on the \code{output.area.stat} argument,
  as listed below:
  \enumerate{
  \item If \code{output.area.stat = TRUE} (\strong{default}), the resulting data frame consists
  of columns/variables like integrals/areas for each simulated and fitted EPR spectrum, where
  the components are denoted by the uppercase letters (\code{Area_Sim_A}, \code{Area_Sim_B},...etc.);
  best fitted/optimized coefficients to multiply the intensities (\code{Optim_CoeffInt_Sim_A},
  \code{Optim_CoeffInt_Sim_B},...etc); best fitted/optimized intercept (or baseline constant,
  \code{Optim_intercept}); minimum sum of residual squares (\code{min_RSS}); number
  of evaluations/iterations (\code{N_evals}) and finally convergence information/number (\code{N_converg},
  like already described in the \code{\link{optim_for_EPR_fitness}}). These variables are presented for each
  \code{var2nd.series} (e.g. time) point like example for one EPR spectral component:
  \tabular{lcccccc}{
  \strong{time_s} \tab \strong{Area_Sim_A} \tab \strong{Optim_CoeffInt_Sim_A} \tab \strong{Optim_intercept} \tab
  \strong{min_RSS} \tab \strong{N_evals} \tab \strong{N_converg} \cr
  6 \tab 0.020624473 \tab 0.052843937 \tab 5.508809e-10 \tab 2.289953e-07 \tab 198 \tab 4 \cr
  21 \tab 0.020217930\tab 0.051802287\tab 5.401823e-10 \tab 2.438172e-07 \tab 177 \tab 4 \cr
  36 \tab 0.018836579 \tab 0.048263010 \tab 5.029705e-10 \tab 2.662651e-07 \tab 201 \tab 4 \cr
  }

  \item If \code{output.area.stat = FALSE} Tidy/long table format of the original \code{data.spectra.series}
  with additional columns/variables (best fitted simulated intensities and their corresponding integrals)
  for all spectral components: A, B, C, ...etc.
  }
}
\description{
Evaluating the linear combination of spectral intensities of components (loaded as ASCII text files corresponding
  to simulated EPR spectra). The related intensity multiplication coefficients
  (please, refer to the \code{optim.params.init} argument) are optimized by methods gathered in the
  \code{\link{optim_for_EPR_fitness}}. The goal is to fit the sum of the simulated components
  onto each experimental spectrum within series. The maximum number of components is set to 10.
}
\details{
Analyzed EPR spectra may consists of several components (see also the \code{\link{eval_sim_EPR_iso_combo}}
  function), like the overlapped EPR spectra of several radicals. In order to follow the concentration/amount
  variations of each individual radical during the kinetic/temperature/...series, one must figure out,
  how those individual spectral components are actually changed. In the first approximation, it means
  to follow the corresponding EPR intensities/integrals, whereas the component positions (\eqn{g}-values)
  are assumed to be fixed (or those changes can be neglected). Therefore, the actual function takes the linear
  combination of the spectral intensities of components (simulated spectra) and optimizes
  the related multiplication coefficients. Additional analysis, where the positions of spectral
  components (simulated spectra) are not fixed and can be optimized as well, is under development.
}
\examples{
\dontrun{
## example with default arguments corresponding
## to one simulated spectral component,
## `optim.params.init` has the length
## of => number of components + 1,
## because of included intercept/baseline constant
quant.data.sim.test.a <-
  quantify_EPR_Sim_series(data.spectra.series,
     dir_ASC_sim = "./",
     optim.method = "pswarm",
     name.pattern.sim = "DHMB0_1st_04_SimA",
     optim.params.init = c(0,0.8),
     output.area.stat = TRUE)
## progress bar - activated
#
## similar example with two components
## (simulated spectra) and tidy data frame
## output (not the summarized one) with all spectra
## and their corresponding integrals, additionally,
## display of the iteration progress for each EPR
## spectrum within the series is activated
quant.data.sim.test.b <-
  quantify_EPR_Sim_series(data.spectra.series,
     dir_ASC_sim = "./",
     optim.method = "sbplx",
     name.pattern.sim = "DHMB0_1st_04_Sim[[:upper:]]",
     optim.params.init = c(0,0.8,0.2),
     output.area.stat = FALSE,
     eval.optim.progress = TRUE
     )
## progress bar - deactivated
#
## to implement the `.csv` outputs
## from `plot_eval_ExpSim_app()`, the following
## arguments must be used:
##  origin.sim = "csv",
##  col.names.sim = c(
##    "Bsim_G", # or mT
##    "B_G", # or mT
##    "dIeprSim_over_dB",
##    "dIepr_over_dB",
##    "Norm_dIeprSim_over_dB"
##  ),
##  x.sim.id = 1,
##  Intensity.sim.id = 3
#
}


}
\seealso{
Other Simulations and Optimization:
\code{\link[=eval_ABIC_forFit]{eval_ABIC_forFit()}},
\code{\link[=eval_sim_EPR_iso]{eval_sim_EPR_iso()}},
\code{\link[=eval_sim_EPR_isoFit]{eval_sim_EPR_isoFit()}},
\code{\link[=eval_sim_EPR_isoFit_space]{eval_sim_EPR_isoFit_space()}},
\code{\link[=eval_sim_EPR_iso_combo]{eval_sim_EPR_iso_combo()}},
\code{\link[=optim_for_EPR_fitness]{optim_for_EPR_fitness()}},
\code{\link[=plot_eval_EPRtheo_mltiplet]{plot_eval_EPRtheo_mltiplet()}},
\code{\link[=plot_eval_RA_forFit]{plot_eval_RA_forFit()}},
\code{\link[=smooth_EPR_Spec_by_npreg]{smooth_EPR_Spec_by_npreg()}}
}
\concept{Simulations and Optimization}
