% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_EPR_Sim_series.R
\name{quantify_EPR_Sim_series}
\alias{quantify_EPR_Sim_series}
\title{Quantify (Components) Area of Simulated EPR Spectral Series Instead of Experimental One}
\usage{
quantify_EPR_Sim_series(
  data.spectra.series,
  dir_ASC_sim,
  name_pattern_sim,
  sim.origin = "easyspin",
  var2nd.series = "time_s",
  B.unit = "G",
  Intensity.expr = "dIepr_over_dB",
  Intensity.sim = "dIeprSim_over_dB",
  optim.method = "slsqp",
  optim.params.init,
  optim.params.lower = NULL,
  optim.params.upper = NULL,
  Nmax.evals = 1024,
  tol.step = 5e-07,
  pswarm.size = NULL,
  pswarm.diameter = NULL,
  single.integ = "single_IntegSim",
  double.integ = "double_IntegSim",
  output.area.stat = TRUE
)
}
\arguments{
\item{data.spectra.series}{Data frame object with...corresponding to...TBC...}

\item{dir_ASC_sim}{tbc}

\item{name_pattern_sim}{description}

\item{sim.origin}{description from}

\item{var2nd.series}{Character string referred to name of the second independent variable/quantity
column in the original \code{data.spectra} (e.g. like `time`,`Temperature`, `Electrochemical Potential`,
`Microwave Power`...etc) altered upon individual experiments as a second variable
(\code{var2nd.series}) and related to spectra/data. Data must be available in \strong{long table}
(or \strong{tidy}) \strong{format} (see also \code{\link{readEPR_Exp_Specs_multif}}).
\strong{Default}: \code{var2nd.series = NULL}. Otherwise \strong{usually} \code{var2nd.series = "time_s"}.}

\item{B.unit}{Character string ...tbc...}

\item{Intensity.expr}{Character string ...tbc ...}

\item{Intensity.sim}{Character string ...tbc ...}

\item{optim.method}{Character string description tbc...following methods from \code{\link{optim_for_EPR_fitness}}
are available ...TBC...}

\item{optim.params.init}{Numeric vector...description tbc...1. element = basiline constant, 2.,3...etc (following)
elements multiplication intensity constants of all spectral components}

\item{optim.params.lower}{Numeric vector (with the length of \code{optim.params.init}) with the lower bound constraints.}

\item{optim.params.upper}{Numeric vector (with the length of \code{optim.params.init}) with the upper bound constraints.}

\item{Nmax.evals}{Numeric, maximum number of function evaluations or iterations.}

\item{tol.step}{Numeric, the smallest optimization step (relative change) to stop
the optimization or fitting procedure.}

\item{pswarm.size}{Numeric value equal to particle swarm size (i. e. number of particles).}

\item{pswarm.diameter}{Numeric value corresponding to diameter of search space.}

\item{single.integ}{tbc}

\item{double.integ}{tbc can be also \code{NULL} in case of single integral spectral series input}

\item{output.area.stat}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_sim_EPR_isoFit}()},
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{eval_sim_EPR_iso}()},
\code{\link{optim_for_EPR_fitness}()},
\code{\link{smooth_EPR_Spec_by_npreg}()}
}
\concept{Simulations and Optimization}
