% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativ_EPR_abs.R
\name{quantitativ_EPR_abs}
\alias{quantitativ_EPR_abs}
\title{TODO}
\usage{
quantitativ_EPR_abs(
  double.integ,
  nu.GHz,
  power.mW,
  modul.amp.mT,
  qValue,
  tube.sample.id.mm,
  fill.sample.h.mm,
  Norm.constantant = 1,
  Temp.K = 298,
  S = 0.5,
  centr.sample.h.mm = 61,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{double.integ}{TODO}

\item{nu.GHz}{TODO}

\item{power.mW}{TODO}

\item{modul.amp.mT}{TODO}

\item{qValue}{TODO}

\item{tube.sample.id.mm}{TODO}

\item{fill.sample.h.mm}{TODO}

\item{Norm.constantant}{TODO}

\item{Temp.K}{TODO}

\item{S}{TODO}

\item{centr.sample.h.mm}{TODO}

\item{microW.cavity}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
