% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativ_EPR_abs.R
\name{quantitativ_EPR_abs}
\alias{quantitativ_EPR_abs}
\title{TBC (Absolute EPR Quantity)}
\usage{
quantitativ_EPR_abs(
  double.integ,
  nu.GHz,
  power.mW,
  modul.amp.mT,
  qValue,
  tube.sample.id.mm,
  fill.sample.h.mm,
  Norm.constant = 1,
  Temp.K = 298,
  S = 0.5,
  centr.sample.h.mm = 61,
  microW.cavity = "rectangular"
)
}
\arguments{
\item{double.integ}{tbc}

\item{nu.GHz}{tbc}

\item{power.mW}{tbc}

\item{modul.amp.mT}{tbc}

\item{qValue}{tbc}

\item{tube.sample.id.mm}{tbc}

\item{fill.sample.h.mm}{tbc}

\item{Norm.constant}{tbc}

\item{Temp.K}{tbc}

\item{S}{tbc}

\item{centr.sample.h.mm}{tbc}

\item{microW.cavity}{tbc}
}
\value{
tbc
}
\description{
tbc
}
\examples{
tbc
tbc


}
