% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Exp_Specs_kin.R
\name{readEPR_Exp_Specs_kin}
\alias{readEPR_Exp_Specs_kin}
\title{Read and Process Spectral Data of Time Dependent CW EPR Experiments}
\usage{
readEPR_Exp_Specs_kin(
  name.root,
  dir_ASC,
  dir_dsc_par,
  time.unit = "s",
  time.delta.slice.s = NULL,
  col.names = c("index", "B_G", "time_s", "dIepr_over_dB"),
  x.unit = "G",
  convertB.unit = TRUE,
  qValue = NULL,
  norm.vec.add = NULL,
  origin = "xenon",
  ...
)
}
\arguments{
\item{name.root}{Character string, corresponding to entire \strong{file name without extension}.}

\item{dir_ASC}{Character string, path (can be also defined by the \code{\link[base]{file.path}})
to directory where the \code{ASCII} spectral data is stored.}

\item{dir_dsc_par}{Character string, path (can be also defined by the \code{\link[base]{file.path}})
to directory where the \code{.DSC}/\code{.dsc} or \code{.par} parameter file is stored
(in order to calculate \eqn{g}-value and/or normalize intensities).}

\item{time.unit}{Character string, specifying the \code{"s"},\code{"min"}, \code{"h"}
or \code{time.unit = "unitless"} (if \code{time.delta.slice.s} is different from \code{NULL}).
\strong{Default}: \code{time.unit = "s"}}

\item{time.delta.slice.s}{Numeric, time interval in seconds between \code{slices},
in the case if \code{origin = "winepr"}. \strong{Default}: \code{time.delta.slice = NULL} (actually,
corresponding to \code{1 s}).}

\item{col.names}{Character/String vector inherited from \code{\link[data.table]{fread}}, corresponding to
column/variable names. A safe rule of thumb is to use column names incl. physical quantity notation
with its units, \code{Quantity_Unit} like \code{"B_G"}, \code{"RF_MHz"}, \code{"Bsim_mT"} (e.g. pointing
to simulated EPR spectrum \eqn{x}-axis)...etc,
\strong{default}: \code{col.names = c("index","B_G","time_s","dIepr_over_dB")} (for \code{origin = "xenon"}).}

\item{x.unit}{Character string, corresponding to original \code{x} variable/column unit, such as \code{"G"},
\code{"mT"} or \code{"MHz"}.}

\item{convertB.unit}{Logical (\strong{default}: \code{convertB.unit = TRUE}), whether upon reading, an automatic
conversion between \code{G} and \code{mT} should be performed. If default is chosen, a new column/variable
\eqn{B} in \code{mT}/\code{G} is created.}

\item{qValue}{Numeric, Q value (quality factor, number) displayed at specific \code{dB} by spectrometer.
In case of "Xenon" software the parameter is included in \code{.DSC} file, therefore \strong{default}:
\code{qValue = NULL} (actually corresponding to value \code{1}). If EPR spectra were acquired
by the "Winepr" software, the Q value must be defined like \code{qValue = 3400}.}

\item{norm.vec.add}{Numeric vector, additional (division) normalization constant(s) in the form of vector, involving
all (in addition to \code{qValue}) normalization(s) such as concentration, powder sample
weight, number of scans, ...etc (e.g. \code{norm.vec.add = c(2000,0.5,2)}). \strong{Default}:
\code{norm.vec.add = NULL} (actually corresponding to value(s) \code{1}).}

\item{origin}{Character string, corresponding to \strong{origin} of the ASCII data, like from the
most common spectrometers (from which the data are loaded automatically using the default parameters).
Options are summarized in the following table (any other specific \code{origin} may be added later) =>
\tabular{rl}{
\strong{String} \tab \strong{Description} \cr
"xenon" \tab \strong{default} automatically loads data from the "Xenon" software with the default parameters. \cr
"winepr" \tab automatically loads data from the "WinEpr" software. \cr
"magnettech" \tab automatically loads data from the new "Magnettech" software (ESR5000 [11-0422]). \cr
"other" (arbitrary string, e.g. "csv") \tab general, loads any other original
data like \code{csv}, \code{txt}, \code{asc} incl. also data from other instrumental/spectrometer software.
\strong{In such case, all the arguments for} \code{readEPR_Exp_Specs} \strong{have to be set up accordingly}. \cr
}}

\item{...}{additional arguments specified, see also the \code{\link{readEPR_Exp_Specs}}
and \code{\link[data.table]{fread}}.}
}
\value{
List of EPR spectrum data (including time) in tidy long table format (\code{df}) + corrected
   time vector (\code{time}). For the \code{origon = "winepr"} "time" slices/indices must be already converted
   into time domain by \code{time.delta.slice.s} (see arguments and examples).
}
\description{
Reading the continuous wave (CW) EPR time series spectral data (recorded by e.g. \code{2D_Field_Delay}
 experiment in "Xenon" acquisition/processing software). Function (based on \code{\link{readEPR_Exp_Specs}}) includes
 automatic time correction for CW EPR \code{time.series} experiments (see also \code{\link{correct_time_Exp_Specs}}).
}
\examples{
## loading the built-in package example to demonstrate
## the reading of time series EPR spectra/kinetics:
triarylam.decay.series.dsc.path <-
  load_data_example(file = "Triarylamine_radCat_decay_series.DSC")
triarylam.decay.series.asc.path <-
  load_data_example(file = "Triarylamine_radCat_decay_series.zip")
unzip(triarylam.decay.series.asc.path,exdir = tempdir())
#
## loading the kinetics:
triarylam.decay.series.data <-
  readEPR_Exp_Specs_kin(name.root = "Triarylamine_radCat_decay_series",
                        dir_ASC = tempdir(),
                        dir_dsc_par =
                          system.file("extdata",
                                      package = "eprscope")
                       )
#
## data preview
head(triarylam.decay.series.data$df)
#
## preview of corrected time vector
## (the uncorrected one actually starts from `0`)
triarylam.decay.series.data$time
#
\dontrun{
## reading by the "WinEPR" software
readEPR_Exp_Specs_kin("Sample_spectra_irradiation",
                      file.path(".","ASCII_data_dir"),
                      file.path(".","dsc_data_dir"),
                      time.unit = "s",
                      time.delta.slice.s = 24.1,
                      col.names = c("B_G",
                                    "Slice",
                                    "Intensity"),
                      x.unit = "G",
                      qValue = 2900,
                      origin = "winepr")

}


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
