% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Exp_Specs_kin.R
\name{readEPR_Exp_Specs_kin}
\alias{readEPR_Exp_Specs_kin}
\title{Read Spectral Data of Time Dependent Experiments (e.g. like Kinetics)}
\usage{
readEPR_Exp_Specs_kin(
  file.rootname,
  dir_ASC,
  dir_DSC_or_par,
  origin = "xenon",
  qValue = NULL
)
}
\arguments{
\item{file.rootname}{String/Character corresponding to `file name` without `extension`}

\item{dir_ASC}{path (defined by \code{\link[base]{file.path}}, String/Character) to directory where
the `ascii` file is stored}

\item{dir_DSC_or_par}{path (defined by \code{\link[base]{file.path}} String/Character) to directory
where the file (`.DSC` or `.par`) with instrumental parameters (to calculate \eqn{g}-value
or normalize intensities) is stored}

\item{origin}{String/Character corresponding to \strong{software} used to acquire the EPR spectra
on BRUKER spectrometers, i.e. whether they were recorded by the windows based softw. ("WinEpr",
\code{origin = "winepr"}) or by the Linux one ("Xenon"), \strong{default}: \code{origin = "xenon"}}

\item{qValue}{Numeric, Q value (quality factor, number) displayed at specific \code{dB} by spectrometer.
In case of "Xenon" software the parameter is included in \code{.DSC} file, therefore \strong{default}:
\code{qValue = NULL}. If EPR spectra were acquired by the "Winepr" software Q value must be defined
like \code{qValue = 3400}}
}
\value{
List of spectral data (incl. time) in tidy long table format (`data`) + corrected time vector (`time`)
}
\description{
The function reads the EPR time series spectral data (recorded by e.g. `2D_Field_Delay Experiment`
 in "Xenon" acquisition/processing software)
}
\examples{
\dontrun{
## Reading by the "Xenon" software
readEPR_Exp_Specs_kin("Sample_spectra_irradiation",
                      file.path(".","ASCII_data_dir"),
                      file.path(".","DSC_data_dir"),
                      origin = "xenon")

## Reading by the "WinEPR" software
readEPR_Exp_Specs_kin("Sample_spectra_irradiation",
                      file.path(".","ASCII_data_dir"),
                      file.path(".","DSC_data_dir"),
                      origin = "winepr",
                      qValue = 2900)

}


}
