% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Exp_Specs_multif.R
\name{readEPR_Exp_Specs_multif}
\alias{readEPR_Exp_Specs_multif}
\title{Load Several/Multiple EPR Spectral Data Files into one List/Database.}
\usage{
readEPR_Exp_Specs_multif(
  pattern,
  path_to_ASC,
  path_to_DSC_or_par,
  origin = "xenon",
  qValues = NULL,
  names
)
}
\arguments{
\item{pattern}{tbc}

\item{path_to_ASC}{tbc}

\item{path_to_DSC_or_par}{tbc}

\item{origin}{tbc}

\item{qValues}{tbc}

\item{names}{tbc}
}
\value{
List of Data Frames Corresponding to Multiple Spectral Data Files
}
\description{
Function the load EPR spectra from several/multiple files. tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
