% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_param_slct.R
\name{readEPR_param_slct}
\alias{readEPR_param_slct}
\title{Read the \strong{Selected} EPR Instrumental Parameters and Information}
\usage{
readEPR_param_slct(path_to_DSC_or_par, string, origin = "xenon")
}
\arguments{
\item{path_to_DSC_or_par}{String, path to \code{.DSC} or \code{.par} file including all instrumental
parameters provided by the EPR machine, path can be provided by \code{\link[base]{file.path}}}

\item{string}{String, within the \code{.DSC} or \code{.par} (at the line beginning) file
 corresponding to instrumental parameter,
following \strong{strings are defined} (\strong{in parenthesis for "winepr" software}):
\tabular{rcl}{
 ---------------- \tab | \tab ----------------------------- \cr
 \strong{String}  \tab | \tab \strong{Instrumental Parameter} \cr
 ---------------- \tab | \tab ----------------------------- \cr
  "OPER" ("JON") \tab | \tab  operator (of the EPR instrument) \cr
  "CMNT" ("JCO") \tab | \tab  comment (in order to describe the measurement) \cr
  "DATE" ("JDA") \tab | \tab  date (when the EPR spectrum was recorded) \cr
  "TIME" ("JTM")  \tab | \tab  time (when the EPR spectrum was recorded) \cr
  "SAMP" \tab | \tab   name/decsript. of the sample \cr
  "B0MF" \tab | \tab  modulation frequency in \code{Hz} \cr
  "MWFQ"  ("MF") \tab | \tab microwave frequency in \code{Hz} (\code{GHz}) \cr
  "QValue" \tab | \tab  recorded quality-Factor (required for intensity norm.) \code{unitless} \cr
  "A1CT" ("HCF") \tab | \tab central field (B) in \code{T} (\code{G}) \cr
  "A1SW" ("HSW") \tab | \tab  sweep width in \code{T} (\code{G}) \cr
  "STMP" ("TE")  \tab | \tab  temperature in \code{K} \cr
  "B0MA" ("RMA") \tab | \tab  modulation amplitude in \code{T} (\code{G}) \cr
  "AVGS" ("JSD") \tab | \tab  number of accumulations for each spectrum \cr
  "A1RS" ("RES") \tab | \tab  number of points/resolution \cr
  "MWPW" ("MP") \tab | \tab microwave power in \code{W} (\code{mW}) \cr
  "SPTP" ("RCT") \tab | \tab  conversion time in \code{s} (\code{ms}) \cr
  "RCTC" ("RTC") \tab | \tab  time constant in \code{s} (ms) \cr
  "RCAG" ("RRG") \tab | \tab signal receiver gain in \code{dB} (unitless) \cr
  "ConvFact" \tab | \tab conversion factor/instr. calibration constant for quantitative analysis \code{unitless} \cr
  ---------------- \tab | \tab ----------------------------- \cr
}}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}
}
\value{
Numeric or String/Character (e.g. date or comment) corresponding to selected (\code{slct}) instrumental
  parameter applied to record the EPR spectra
}
\description{
Function takes the \strong{selected} instrumental parameters or information
 from \code{.DSC} or \code{.par} file of an EPR Spectrum (written by the \code{Xenon}
 or \code{WinEpr} Software, respectively)
}
\examples{
\dontrun{
## Reading modulation amplitude from 'Xenon' spectrometer file
readEPR_param_slct(path_to_DSC_or_par,
                   string = "B0MA")

## Reading Q Value from 'Xenon' spectrometer file
read_param_slct(file.path(".",
                          "dir_DSC",
                          "EPR_spectrum.DSC"),
                string = "QValue")

## Reading `date` from 'WinEPR' spectrometer file
read_param_slct(file.path(".",
                          "dir_par",
                          "EPR_spectrum.par"),
                string = "JDA",
                origin = "winepr")
}


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
