% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_slct.R
\name{readEPR_param_slct}
\alias{readEPR_param_slct}
\title{Read Selected EPR Instrumental Parameters and Information}
\usage{
readEPR_param_slct(path_to_dsc_par, string, origin = "xenon")
}
\arguments{
\item{path_to_dsc_par}{Character string, path to \code{.DSC/.dsc} or \code{.par} file including the instrumental
parameters provided by the EPR machine. File path can be also defined by the \code{\link[base]{file.path}} function.}

\item{string}{Character (vector) string (appeared at the line beginning) within the \code{.DSC/.dsc} or \code{.par} file
 corresponding to instrumental parameter or information.
Following \strong{strings are defined for all three main acquisition softwares described-above}
 (\strong{in parenthesis for the "winepr" origin}):
\tabular{ll}{
 \strong{String} \tab \strong{Instrumental Parameter} \cr
  "OPER" ("JON") \tab operator (of the EPR instrument) \cr
  "CMNT" ("JCO") \tab comment (in order to describe the measurement) \cr
  "DATE" ("JDA") \tab date (when the EPR spectrum was recorded) \cr
  "TIME" ("JTM") \tab time (when the EPR spectrum was recorded) \cr
  "SAMP" \tab name/decsript. of the sample, not available in "magnettech" \code{.dsc} \cr
  "B0MF" \tab modulation frequency in \code{Hz} \cr
  "MWFQ" ("MF") \tab microwave frequency in \code{Hz} (\code{GHz}) \cr
  "QValue" \tab recorded quality-Factor (required for intensity normalization) \code{unitless} \cr
  "A1CT" ("HCF") \tab central field (B) in \code{T} (\code{G}) \cr
  "A1SW" ("HSW") \tab sweep width in \code{T} (\code{G}) \cr
  "STMP" ("TE")  \tab temperature in \code{K} \cr
  "B0MA" ("RMA") \tab modulation amplitude in \code{T} (\code{G}) \cr
  "AVGS" ("JSD") \tab number of accumulations for each spectrum \cr
  "A1RS" ("RES") \tab number of points/resolution \cr
  "MWPW" ("MP") \tab microwave power in \code{W} (\code{mW}) \cr
  "SPTP" ("RCT") \tab conversion time in \code{s} (\code{ms}) \cr
  "RCTC" ("RTC") \tab time constant in \code{s} (ms),
  not available in "magnettech" \code{.dsc} \cr
  "RCAG" ("RRG") \tab signal receiver gain in \code{dB} (unitless),
  not available in "magnettech" \code{.dsc} \cr
  "ConvFact" \tab conversion factor/instr. calibration constant for quantitative
  analysis \code{unitless}, not available in "magnettech" \code{.dsc} \cr
}}

\item{origin}{Character string, corresponding to software used to acquire EPR spectra.
The files are slightly different depending on whether
they were recorded by the "WinEpr",\code{origin = "winepr"}, "Xenon"
(\strong{default}: \code{origin = "xenon"}) or by the "Magnettech" (ESR5000 [11-0422], \code{origin = "magnettech"}).}
}
\value{
Numeric or character string (e.g. date or comment), corresponding to selected instrumental parameter(s)
  applied to record the EPR spectra. In case of \code{string} character vector, named list, containing
  either character and/or numeric values, is returned with the names corresponding to \code{string}.
}
\description{
Taking selected instrumental parameters or information
 from the \code{.DSC/.dsc} or \code{.par} file of an EPR spectrum (written by the \code{Xenon}/\code{Magnettech}
 or \code{WinEpr} software, respectively).
}
\examples{
## loading `.DSC` (`Xenon`) parameter file example
triaryl_radCat_dsc_path <-
  load_data_example(file = "Triarylamine_radCat_decay_a.DSC")
#
## reading modulation amplitude (in T) from the `Xenon` spectrometer file
readEPR_param_slct(triaryl_radCat_dsc_path,string = "B0MA")
#
## reading Q-Value from the `Xenon` spectrometer file
readEPR_param_slct(triaryl_radCat_dsc_path,string = "QValue")
#
## reading `CMNT` (comment) and `MWFQ` (microwave frequency in Hz)
## from the `Xenon` spectrometer file
readEPR_param_slct(triaryl_radCat_dsc_path,
                   string = c("CMNT","MWFQ"))
#
## loading `.par` (`WinEPR`) parameter file example
TMPD_radCat_par_path <-
  load_data_example(file = "TMPD_specelchem_accu_b.par")
#
## reading `JDA` (date) from `WinEPR` spectrometer file
readEPR_param_slct(TMPD_radCat_par_path,
                   string = "JDA",
                   origin = "winepr")
#
## reading `RMA` (modulation amplitude in G) and `TE`
## (temperature in K) as well as `JCO` (comment)
## from `WinEPR` spectrometer file
readEPR_param_slct(TMPD_radCat_par_path,
                   string = c("RMA","TE","JCO"),
                   origin = "WinEPR")
#
## loading and reading the `.DSC` file from `Xenon`
## corresponding to phenalenyl (PNT) CW ENDOR spectrum,
## read expr. date (`TIME`), microwave frequency (`MWFQ`)
## in Hz and the corresponding field for saturation (`B0VL`)
## in Tesla:
pnt_endor_dsc_path <-
  load_data_example(file = "PNT_ENDOR_a.DSC")
readEPR_param_slct(pnt_endor_dsc_path,
                string = c("TIME","MWFQ","B0VL")
               )



}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
