% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_slct_kin.R
\name{readEPR_params_slct_kin}
\alias{readEPR_params_slct_kin}
\title{Read the \strong{Selected} Instrumental Parameters Relevant to \strong{Time Series} Experiment}
\usage{
readEPR_params_slct_kin(path_to_DSC_or_par, origin = "xenon")
}
\arguments{
\item{path_to_DSC_or_par}{String, path to \code{.DSC} or \code{.par} file including all instrumental
parameters provided by the EPR machine}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}
}
\value{
List containing number of scans (\code{Nscans}) and sweep time (\code{sweepTime}) required
  for time correction during the \code{2D_Field_Delay} experiment ("kinetic" time series)
}
\description{
Function takes the \strong{selected} instrumental parameters relevant to \strong{time series ("kinetic")}
 experiment from \code{.DSC} or \code{.par} file of an EPR Spectrum (written by the `Xenon`
 or `WinEpr` software, respectively). These parameters are required for the time correction of EPR
 spectra, see \code{\link{time_correct_EPR_Specs}}
}
\examples{
\dontrun{
readEPR_params_slct_kin(path_to_DSC_or_par)
readEPR_params_slct_kin(file.path(".",
                                  "dir_par",
                                  "EPR_spectrum.par"),
                        origin = "winepr")
}

}
