% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_slct.R
\name{readEPR_params_slct_quant}
\alias{readEPR_params_slct_quant}
\title{Read the \strong{Selected} Instrumental Parameters Relevant to EPR Quantitative Analysis}
\usage{
readEPR_params_slct_quant(path_to_dsc_par, origin = "xenon")
}
\arguments{
\item{path_to_dsc_par}{Character string, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on OS, see \code{origin} parameter)
\code{text} files including all instrumental parameters and provided by the EPR machine.}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = "xenon"}.}
}
\value{
List consisting of:
  \describe{
  \item{BmmT}{Modulation amplitude value in `mT`.}
  \item{PmW}{Microwave source power in `mW`.}
  \item{TK}{Experimental temperature in `K`.}
  \item{mwGHz}{Microwave frequency value in `GHz`.}
  }
}
\description{
Reading the \code{.DSC/.dsc} or \code{.par} file to extract the important parameters like
  "modulation amplitude", "temperature", "microwave power" as well as "microwave frequency"
  which are are required for absolute quantitative analysis of the EPR spectra (\eqn{\equiv}
  radical or paramagnetic species number determination).
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
