% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_slct.R
\name{readEPR_params_slct_sim}
\alias{readEPR_params_slct_sim}
\title{Read the \strong{Selected} Instrumental Parameters Relevant to \strong{Simulation} of EPR Spectra}
\usage{
readEPR_params_slct_sim(path_to_DSC_or_par, origin = "xenon", B.unit = "G")
}
\arguments{
\item{path_to_DSC_or_par}{Character string, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} parameter)
\code{text} files including all instrumental parameters and provided by the EPR machine.}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they
were recorded by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux
one ("Xenon"). \strong{Default}: \code{origin = "xenon"}.}

\item{B.unit}{Character string pointing to unit of magnetic flux density which is the output
`unit`, \code{"G"} (`Gauss`) or \code{"mT"} (`millitesla`), for \code{"sweep width"}
and \code{"central field"} (see \code{\link{eval_sim_EPR_iso}}).
\strong{Default}: \code{B.unit = "G"}.}
}
\value{
List consisting of:

  * \code{Bcf}, "central field" value in `B.unit`

  * \code{Bsw}, "sweep width" value in `B.unit`)

  * \code{Npoints}, "number of points" value

  * \code{mwGHz}, "microwave frequency" value in `GHz`
}
\description{
Reading the \code{.DSC} or \code{.par} file to extract the important parameters like
  "sweep width", "central field", "number of points" as well as "microwave frequency"
  which are are required for the simulations of EPR spectra (see \code{\link{eval_sim_EPR_iso}}).
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
