% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_tabs.R
\name{readEPR_params_tabs}
\alias{readEPR_params_tabs}
\title{Read the EPR Instrumental Parameters and Experiment Information}
\usage{
readEPR_params_tabs(path_to_dsc_par, origin = "xenon", interact = NULL)
}
\arguments{
\item{path_to_dsc_par}{Character string, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC/.dsc} or \code{.par} (depending on \code{origin} parameter)
\code{text} files including instrumental parameters and provided by the EPR machine.}

\item{origin}{Character string, corresponding to software which was used to acquire the EPR spectra
on BRUKER/MAGNETTECH spectrometers, because the files are slightly different depending on whether they
were recorded by the "WinEpr",\code{origin = "winepr"} softw. or by the "Xenon"
(\strong{default}: \code{origin = "xenon"}) one or by the "Magnettech" softw. (\code{origin = "magnetech"}).}

\item{interact}{Character string, whether or not to display interactive tables by \code{\link[DT]{datatable}}.
\strong{Default}: \code{interact = NULL}. Interactive table with parameters can be displayed by
\code{interact = "params"} or to display the additional information table:
\code{interact = "info"}.}
}
\value{
List of data frames/tables containing:
 \describe{
 \item{params}{Instrumental parameters with their numeric values and units.}
 \item{info}{Information character string, such as date, operator, comment...etc.}
 }
 Both data frames may be depicted in the form of interactive tables
 by \code{interact} function argument.
}
\description{
Taking the instrumental parameters from the \code{.DSC/.dsc} or \code{.par} files, applied
  to record the EPR Spectra and transferring them into list of \code{Tables/Data Frames} incl.
  parameter values as well as character/string information about the measurement, see also
  \code{\link{readEPR_param_slct}}.
}
\examples{
## loading built-in example file =>
## "AcridineDeriv_Irrad_365nm.dsc" by `Magnettech`
## spectrometer software
AcridinRad.data.file <-
  load_data_example("AcridineDeriv_Irrad_365nm.dsc")
## reading nd displaying parameters as data frame
AcridinRad.params.data <-
  readEPR_params_tabs(AcridinRad.data.file,
                      origin = "magnettech")
#
## parameters preview
AcridinRad.params.data$params
#
## info preview
AcridinRad.params.data$info
#
## built-in example file => "TMPD_specelchem_accu_b.par"
## by the `WinEPR` spectrometer software
tmpd.params.file <-
  load_data_example(file = "TMPD_specelchem_accu_b.par")
## reading and displaying parameters as data frame
tmpd.params.tab <-
  readEPR_params_tabs(tmpd.params.file,
                      origin = "winepr")
#
## preview
tmpd.params.tab$params
##
## the same data frame, now in interactive table form
readEPR_params_tabs(tmpd.params.file,
                    origin = "winepr",
                    interact = "params")


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
