% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_params_tabs.R
\name{readEPR_params_tabs}
\alias{readEPR_params_tabs}
\title{Read the EPR Instrumental Parameters and Experiment Information}
\usage{
readEPR_params_tabs(path_to_DSC_or_par, origin = "xenon", interact = NULL)
}
\arguments{
\item{path_to_DSC_or_par}{String, path (also provided by \code{\link[base]{file.path}})
to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} parameter)
\code{text} files including all instrumental parameters and provided by the EPR machine}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}

\item{interact}{Charater, whether or not display tables by \code{\link[DT]{datatable}}. \strong{Default}:
\code{interact = NULL}. To display interactive table with parameters: \code{interact = "params"} aa well as
to display that of the additional information: \code{interact = ""info}.}
}
\value{
List of data frames/tables containing instrumental parameters (\code{params}) and information (\code{info}),
 i.e. numeric values & character strings related to an EPR experiment. Both data frames tables may be depicted
 in the form of interactive tables by \code{interact} function argument.
}
\description{
Function takes the instrumental parameters (from \code{.DSC} or \code{.par} file) applied
to record the EPR Spectra and transfers them into list of \code{Tables/Data Frames} incl.
parameter values as well as character/string information about the measurement see also
\code{\link{readEPR_param_slct}}
}
\examples{
\dontrun{
readEPR_params_tabs(path_to_DSC_or_par)
readEPR_params_tabs(file.path(".",
                              "dir_par",
                              "EPR_spectrum.par"),
                    origin = "winepr",
                    interact = params)
}


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_for_sim}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
