% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Sim_Spec.R
\name{readEPR_Sim_Spec}
\alias{readEPR_Sim_Spec}
\title{Read the Simulated ASCII (\code{.txt}) EPR Spectrum from \emph{MATLAB} and Transfer it into Data Frame}
\usage{
readEPR_Sim_Spec(path_to_ASC)
}
\arguments{
\item{path_to_ASC}{String, path to ASCII file/table (\code{.txt})
with spectral data (\eqn{Intensity vs B}(Field) obtained from \emph{MATLAB})}
}
\value{
tbc
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}

}
