% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Sim_Spec.R
\name{readEPR_Sim_Spec}
\alias{readEPR_Sim_Spec}
\title{Read the Simulated ASCII EPR Spectrum}
\usage{
readEPR_Sim_Spec(
  path_to_ASC,
  B.unit = "mT",
  Intensity.sim = "dIeprSim_over_dB",
  sim.origin = "easyspin"
)
}
\arguments{
\item{path_to_ASC}{Character string, path to ASCII file/table
with simulated spectral data (\eqn{Intensity vs B}(Field) obtained from various sources.
The path can be also defined by \code{\link[base]{file.path}}.}

\item{B.unit}{Character string pointing to unit of magnetic flux density \strong{coming from the original data}
which is to be presented on the \eqn{B} abscissa of an EPR spectrum,
like \code{"G"} (`Gauss`), \code{"mT"} (`millitesla`). \strong{Default}: \code{B.unit = "mT"}.}

\item{Intensity.sim}{Character string pointing to \strong{intensity column} of the data frame
corresponding to EPR simulated spectrum. If used together with quantification of radicals,
this argument must be equal to that of the \code{\link{quantify_EPR_sim}}.}

\item{sim.origin}{Character string referring to "origin" of the simulated ASCII data.
There are four possibilities \eqn{\Rightarrow} \code{sim.orimgin = "easyspin"} (\strong{default}),
\code{"xenon"}, \code{"simfonia"} as well as universal \code{"csv"}.}
}
\value{
Data frame consisting of magnetic flux density and intensity variable/column corresponding
  to simulated EPR spectrum.
}
\description{
Loading the raw `.txt`,`.asc` or `.csv` data of a simulated EPR spectrum from different sources like
  `EasySpin` (Matlab), `Xenon` (EPR spectrometer), `SimFonia` (WinEPR system) or `csv` (universal format
  or MS Excel). The `ASCII` data is automatically converted into data frame.
}
\examples{
\dontrun{
tbc
tbc
}


}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_for_sim}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
