% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_Sim_Spec.R
\name{readEPR_Sim_Spec}
\alias{readEPR_Sim_Spec}
\title{Read Simulated ASCII EPR Spectrum}
\usage{
readEPR_Sim_Spec(
  path_to_ASC,
  B.unit = "G",
  Intensity.sim = "dIeprSim_over_dB",
  sim.origin = "easyspin"
)
}
\arguments{
\item{path_to_ASC}{Character string, path to ASCII file/table
with simulated spectral data (\eqn{Intensity\,\,vs\,\,B}(Field) obtained from various sources.
Path can be alternatively defined by the \code{\link[base]{file.path}} function.}

\item{B.unit}{Character string pointing to unit of magnetic flux density coming from the original data
which is to be presented on the \eqn{B}-axis of an EPR spectrum,
like \code{"G"} ("Gauss"), \code{"mT"} ("millitesla"). \strong{Default}: \code{B.unit = "G"}.}

\item{Intensity.sim}{Character string pointing to intensity data frame column corresponding
to simulated EPR spectrum. If used together with quantification of radicals,
this argument must be equal to that of the \code{\link{quantify_EPR_Sim_series}}.}

\item{sim.origin}{Character string referring to "origin" of the simulated ASCII data.
There are four possibilities \eqn{\Rightarrow} \code{sim.orimgin = "easyspin"} (\strong{default}),
\code{"xenon"}, \code{"simfonia"} as well as universal \code{"csv"}.}
}
\value{
Data frame consisting of magnetic flux density (\code{Bsim_mT} or \code{Bsim_G}) and intensity
  variable/column (depending on \code{Intensity.sim} argument) corresponding to simulated EPR spectrum.
}
\description{
Loading the ASCII data like \code{.txt},\code{.asc} or \code{.csv}, related to simulated EPR spectrum from different
  sources like "EasySpin" (\emph{Matlab}), "Xenon" (EPR spectrometer), "SimFonia" (WinEPR system)
  or "csv" (comma separated values, universal format or MS Excel). Finally, they are automatically converted
  into data frame by the \code{\link[data.table]{fread}}.
}
\examples{
\dontrun{
readEPR_Sim_Spec(path_to_ASC =
                  "./Simulations/TEMPO_simulation.txt",
                 sim.origin = "xenon")
#
readEPR_Sim_Spec("Cu_complex_simulation.txt",
                 B.unit = "mT",
                 sim.origin = "easyspin")
}



}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readEPR_solvent_props}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
