% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPR_solvent_props.R
\name{readEPR_solvent_props}
\alias{readEPR_solvent_props}
\title{Reading Solvent Properties from \code{solvents_ds} Dataset}
\usage{
readEPR_solvent_props(solvent, prop = NULL)
}
\arguments{
\item{solvent}{Character string pointing to solvent name (or any string from the solvent name/abbreviation),
such as \code{solvent = "DMSO"},\code{solvent = "acetone"}, \code{solvent = "xylene"}.
If more than one rows/observations are being returned (e.g. in case of \code{solvent = "xylene"}) => additional
solvent specification must be provided e.g. \code{solvent = "o-xylene"}.}

\item{prop}{Character string related to a data frame variable/column/property e.g. \code{"boiling"},
\code{"formula"}, \code{"dens"}, \code{"dielectric"}...etc (see also \code{\link{solvents_ds}}).}
}
\value{
Data frame with row/rows of selected solvent(s) and the corresponding properties. If a specific property
  is called like \code{code = "melting"} (\eqn{\equiv} melting point in °C), the function returns
  \code{value/character}.
}
\description{
Gathering the solvent properties from \code{\link{solvents_ds}} in order to pick-up
  specific \code{solvent} and its corresponding properties (by the \code{prop} argument).
  See also \code{vignette("datasets")}.
}
\examples{
## Properties of `DMSO`:
solvent_01 <- readEPR_solvent_props("DMSO")
head(solvent_01)
#
## All `xylene` solvent specifications:
solvent_02 <- readEPR_solvent_props(solvent = "xylene")
head(solvent_02)
#
## Boiling point of `o-xylene`:
readEPR_solvent_props(solvent = "o-xylene",
                      prop = "boiling")



}
\seealso{
Other Data Reading: 
\code{\link{readEPR_Exp_Specs}()},
\code{\link{readEPR_Exp_Specs_kin}()},
\code{\link{readEPR_Exp_Specs_multif}()},
\code{\link{readEPR_Sim_Spec}()},
\code{\link{readEPR_param_slct}()},
\code{\link{readEPR_params_slct_kin}()},
\code{\link{readEPR_params_slct_quant}()},
\code{\link{readEPR_params_slct_sim}()},
\code{\link{readEPR_params_tabs}()},
\code{\link{readMAT_params_file}()}
}
\concept{Data Reading}
