% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPRparam_slct.R
\name{readEPRparam_slct}
\alias{readEPRparam_slct}
\title{Read the \strong{selected} Instrumental Parameters from \code{.DSC} to Record
  the EPR Spectra (within the \code{Xenon} Software)}
\usage{
readEPRparam_slct(path_to_DSC, string)
}
\arguments{
\item{path_to_DSC}{String, path to \code{.DSC} file including all instrumental parameters provided by the EPR machine}

\item{string}{String, within the \code{.DSC} file corresponding to instrumental parameter,
following \strong{strings are defined}:
\tabular{rl}{
 \strong{String}  \tab \strong{Instrumental Parameter} \cr
  "OPER" \tab  operator (of the EPR instrument) \cr
  "CMNT" \tab  comment (in order to describe the measurement) \cr
  "DATE" \tab  date (when the EPR spectrum was recorded) \cr
  "TIME"  \tab  time (when the EPR spectrum was recorded) \cr
  "SAMPLE" \tab   name/decsript. of the sample \cr
  "MWFQ"  \tab microwave frequency in \code{Hz} \cr
  "QValue" \tab   recorded quality-Factor (required for intensity norm.) \cr
  "A1CT" \tab   central field (B) in \code{T} \cr
  "A1SW" \tab   sweep width in \code{T} \cr
  "B0MA" \tab   modulation amplitude in \code{T} \cr
  "AVGS" \tab   number of accumulations for each spectrum \cr
  "A1RS" \tab   number of points/resolution \cr
  "MWPW" \tab   microwave power in \code{W} \cr
  "SPTP" \tab   conversion time in \code{s} \cr
  "RCTC" \tab   time constant in \code{s} \cr
}}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
TODO


}
