% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPRparams_vals_tab.R
\name{readEPRparams_vals_tab}
\alias{readEPRparams_vals_tab}
\title{Read the Instrumental Parameters (from \code{.DSC} or \code{.par} file) to Record the EPR Spectra
  and Transfer it into \code{Table/Data Frame}}
\usage{
readEPRparams_vals_tab(path_to_DSC_or_par, origin = "xenon")
}
\arguments{
\item{path_to_DSC_or_par}{String, path to \code{.DSC} or \code{.par} (depending on OS, see \code{origin} parameter)
\code{text} files including all instrumental parameters and provided by the EPR machine}

\item{origin}{String, corresponding to software which was used to acquire the EPR spectra
on BRUKER spectrometers, because the files are slightly different depending on whether they were recorded
by the windows based softw. ("WinEpr",\code{origin = "winepr"}) or by the Linux one ("Xenon"),
\strong{default}: \code{origin = "xenon"}}
}
\value{
Data frame/table containing instrumental parameters (numeric values not strings) applied
  to acquire the EPR spectra
}
\description{
TODO
}
\examples{
TODO
TODO


}
