% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAT_params_file.R
\name{readMAT_params_file}
\alias{readMAT_params_file}
\title{Reads all Parameters (incl. the Simulation ones) from \emph{MATLAB} \code{.mat} File}
\usage{
readMAT_params_file(path_to_MAT, str.var = NULL, field.var = NULL)
}
\arguments{
\item{path_to_MAT}{Character string, path to \code{.mat} MATLAB file with all variables saved in workspace.
The file path can be also defined by \code{\link[base]{file.path}}.}

\item{str.var}{Character string, \code{structure/variable}, which may contain \code{fields}, e.g. like \code{Sys}
and \code{g} => \strong{Sys}.g, respectively. \strong{Default}: \code{str.var = NULL}.}

\item{field.var}{Character string, \code{field variable after 'dot'}, which is available only for certain
structures/variables, see e.g. example above (Sys.\strong{g}), therefore
the \strong{default} 'value' is \code{NULL} and the \code{string} \strong{is applied only for structures with fields}.}
}
\value{
Unless the \code{str.var} and/or \code{field.var} not specified, the output is \code{list} with all original
  parameters/structures from MATLAB file. Otherwise, the function returns either numeric or character \code{vector/value},
  depending on `class` of the original parameter/field variable.
}
\description{
tbc
}
\examples{
\dontrun{
tbc
tbc
}


}
