% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMATfile_params.R
\name{readMATfile_params}
\alias{readMATfile_params}
\title{Read Parameters/Variables/Fields (including the Simulation ones) from \emph{MATLAB} \code{.mat} File}
\usage{
readMATfile_params(path_to_MAT, str.var, field.var = NULL)
}
\arguments{
\item{path_to_MAT}{String, path to \code{.mat} MATLAB file with all variables saved in workspace}

\item{str.var}{String, \code{structure/variable}, which may contain \code{fields}, e.g. like \code{\strong{Sys}}
and \code{g} => \code{\strong{Sys}.g}, respectively}

\item{field.var}{String, \code{field variable after 'dot'}, which is available only for certain
structures/variables, see e.g. example above (\code{Sys.\strong{g}}), therefore
the \strong{default} 'value' is \code{NULL} and \code{string} \strong{is applied only for structures with fields}.}
}
\value{
Numeric or String/Character depending on MATLAB variable used for simulation of EPR spectra
}
\description{
TODO
}
\examples{
TODO
TODO


}
