% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_aAiso.R
\name{rearrange_aAiso_QCHcomp}
\alias{rearrange_aAiso_QCHcomp}
\title{Rearrangement of \eqn{A_{iso}}/\eqn{a_{iso}} from Quantum Chemical Computations}
\usage{
rearrange_aAiso_QCHcomp(path_to_ASC, col.names, nuclei.list.slct)
}
\arguments{
\item{path_to_ASC}{Character string pointing to path of ASCII file (\code{txt},\code{csv}...etc,
it may be also provided by \code{\link[base]{file.path}}) incl. characteristic
  \eqn{A_{iso}} or \eqn{a_{iso}} values.}

\item{col.names}{Character string vector containing names of all columns from QCH Computational output,
for the names see the example in \code{path_to_ASC}, they must contain atomic/structure number, isotopic value
with element label (nucleus characterization) and \eqn{A} in MHz as well as \eqn{a} in Gauss.}

\item{nuclei.list.slct}{List of numeric values for the rearrangement of selected atoms/nuclei according to symmetry,
e.g. \code{nuclei.list.slct <- list(3,c(21,22),c(20,23),c(24,25),c(27,26))} where the numbers
correspond to indices of atoms/nuclei in the ASCII text file.}
}
\value{
Data frame/Table of \eqn{A_{iso}}/\eqn{a_{iso}} mean values corresponding to nuclei group
  structure/symmetry
}
\description{
Provides table from Gaussian/ORCA output text file to summarize \eqn{A_{iso}}/\eqn{a_{iso}} values
  according to proposed molecular structure/symmetry.
}
\details{
The \eqn{A_{iso}}/\eqn{a_{iso}} values are computed for each atom/nucleus
  (with its corresponding \code{atomic number within the structure} as well as with the characteristic
  \code{isotopic number/value}), such an entire table can be copied from \strong{Gaussian} output
  (after \code{'Isotropic Fermi Contact Couplings'} line) or can be constructed from \strong{ORCA} output,
  example for such a file structure (from \strong{Gaussian}):
  \tabular{rcccc}{
   \strong{No_atom} \tab \strong{Atom_Nucleus} \tab \strong{MegaHertz} \tab \strong{Gauss} \tab \strong{1e-4_cm-1} \cr
   1 \tab N(14) \tab 0.00643 \tab 0.00229 \tab 0.00214 \cr
   17 \tab N(14) \tab 13.99707 \tab 4.9945 \tab 4.66892 \cr
   28 \tab H(1) \tab 16.34971 \tab 5.83398 \tab 5.45368 \cr
  }
}
\examples{
\dontrun{
rearrange_aAiso_QCHcomp(
"./iso_values_Gaussian.txt",
c("No","Nucleus","au","Megahertz","Gauss","10^n4_cm^n1"),
list(3,c(21,22),c(20,23),c(24,25),c(27,26))
)
}


}
\seealso{
Other Evaluations and Quantum Chemistry: 
\code{\link{eval_gFactor_QCHcomp}()},
\code{\link{rearrange_aAiso_QCHorgau}()}
}
\concept{Evaluations and Quantum Chemistry}
