% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_aAiso.R
\name{rearrange_aAiso_QCHorgau}
\alias{rearrange_aAiso_QCHorgau}
\title{Rearrangement of \eqn{A_{iso}}/\eqn{a_{iso}} from Gaussian & ORCA Computations}
\usage{
rearrange_aAiso_QCHorgau(
  path_to_QCHoutput,
  Nnuclei,
  nuclei.list.slct,
  origin = "gaussian",
  output.text.origin = FALSE,
  output.text.path = NULL
)
}
\arguments{
\item{path_to_QCHoutput}{Character string corresponding to path of \code{Gaussian} or \code{ORCA} output text files.}

\item{Nnuclei}{Numeric value equal to number of atoms/nuclei within the calculated structure.}

\item{nuclei.list.slct}{List of numeric values for the rearrangement of selected atoms/nuclei according to symmetry,
e.g. \code{nuclei.list.slct <- list(3,c(21,22),c(20,23),c(24,25),c(27,26))} where the numbers
correspond to indices of atoms/nuclei in the ASCII text file.}

\item{origin}{Character string pointing to origin of DFT EPR calculation parameters <=> which
software package was used. Only two values are available => \code{"gaussian"} (\strong{default})
or \code{"orca"}.}

\item{output.text.origin}{Logical, whether to write a text file containing the extracted
\eqn{A_{iso}}/\eqn{a_{iso}} values from the the original output file defined by the \code{path_to_QCHoutput}.
\strong{Default}: \code{output.text.origin = FALSE}.}

\item{output.text.path}{Character string, setting the path to file containing
the extracted \eqn{A_{iso}}/\eqn{a_{iso}} values from the original uotput file defined
by the \code{path_to_QCHoutput}. See the previous argument.}
}
\value{
Data frame/Table of \eqn{A_{iso}}/\eqn{a_{iso}} mean values corresponding to nuclear group
  structure/symmetry constructed directly form \emph{Gaussian} or \emph{ORCA} output files.
}
\description{
Providing table, specifically from \code{Gaussian} or \code{ORCA} output text file to summarize
  the \eqn{A_{iso}}/\eqn{a_{iso}} mean values of a nuclear group, according to proposed molecular structure/symmetry
  (see also \code{\link{rearrange_aAiso_QCHcomp}}).
}
\examples{
## built-in file and path
gauss.file.path <-
  load_data_example(file = "TMPDAradCatEPRa.inp.log.zip")
gauss.file <- unzip(gauss.file.path)
symmetry.As.df <-
  rearrange_aAiso_QCHorgau(gauss.file,
    Nnuclei = 28,
    nuclei.list.slct =
    list(c(7, 8), ## 2 x 14N
         c(13, 14, 15, 16), ## 4 x 1H (aromatic)
         c(17, 18, 19, 20,
           21, 22, 23, 24,
           25, 26, 27, 28) ## 12 x 1H (methyls)
         )
     )
#
## preview
symmetry.As.df


}
\seealso{
Other Evaluations and Quantum Chemistry: 
\code{\link{eval_gFactor_QCHcomp}()},
\code{\link{rearrange_aAiso_QCHcomp}()}
}
\concept{Evaluations and Quantum Chemistry}
