% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_aAiso_QCHorgau.R
\name{rearrange_aAiso_QCHorgau}
\alias{rearrange_aAiso_QCHorgau}
\title{Rearrangement of \eqn{A_{iso}}/\eqn{a_{iso}} from Gaussian & ORCA Computations}
\usage{
rearrange_aAiso_QCHorgau(
  path_to_QCHoutput,
  No_nuclei,
  nuclei.list.slct,
  origin = "gaussian",
  output.text.origin = FALSE,
  output.text.path = NULL
)
}
\arguments{
\item{path_to_QCHoutput}{tbc}

\item{No_nuclei}{tbc}

\item{nuclei.list.slct}{tbc}

\item{origin}{tbc}

\item{output.text.origin}{tbc}

\item{output.text.path}{tbc}
}
\value{
Data frame/Table of \eqn{A_{iso}}/\eqn{a_{iso}} mean values corresponding to nuclei group
  structure/symmetry constructed directly form \emph{Gaussian} or \emph{ORCA} computational output
  files.
}
\description{
A short description...the same like \code{\link{rearrange_aAiso_QCHcomp}} something...
}
\examples{
\dontrun{
tbc
tbc
}

}
