% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_EPR_Spec_by_npreg.R
\name{smooth_EPR_Spec_by_npreg}
\alias{smooth_EPR_Spec_by_npreg}
\title{Smoothing and Fitting of an EPR Spectrum by Splines}
\usage{
smooth_EPR_Spec_by_npreg(
  data.spectr,
  B = "B_mT",
  B.unit = "mT",
  lineSpecs.form = "derivative",
  Intensity = "dIepr_over_dB",
  method = "BIC",
  output.vec = FALSE,
  ...
)
}
\arguments{
\item{data.spectr}{Data fame object ...TBC...}

\item{B}{Character string ...TBC...}

\item{B.unit}{Character string ...TBC...}

\item{lineSpecs.form}{Character string ...TBC...}

\item{Intensity}{Character string ...TBC...}

\item{method}{Character string ...TBC...}

\item{output.vec}{Logical, description ...TBC...}

\item{...}{additional arguments passed to the function (see also \code{\link[npreg]{ss}}).}
}
\value{
List with the following values: ...TBC...
}
\description{
A short description...Smoothing of EPR spectra...TBC...
}
\examples{
\dontrun{
TODO
TODO
}


}
\seealso{
Other Simulations and Optimization: 
\code{\link{eval_sim_EPR_isoFit}()},
\code{\link{eval_sim_EPR_iso_combo}()},
\code{\link{eval_sim_EPR_iso}()},
\code{\link{optim_for_EPR_fitness}()},
\code{\link{quantify_EPR_Sim_series}()}
}
\concept{Simulations and Optimization}
